% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_parameter_to_xml.R
\name{alpha_parameter_to_xml}
\alias{alpha_parameter_to_xml}
\title{Internal function}
\usage{
alpha_parameter_to_xml(alpha_parameter, beauti_options)
}
\arguments{
\item{alpha_parameter}{an alpha parameter,
as created by \link{create_alpha_param}}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the parameter as XML text
}
\description{
Converts an alpha parameter to XML
}
\examples{
remove_beautier_folder()
check_empty_beautier_folder()

# The alpha parameter must be initialized, i.e. have an ID
alpha_parameter_to_xml(
  alpha_parameter = create_alpha_param(id = "1"),
  beauti_options = create_beauti_options()
)

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
