% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_beast2_input_lines.R
\name{are_beast2_input_lines}
\alias{are_beast2_input_lines}
\title{Would these lines of text, when written to a file,
  result in a valid BEAST2 input file?}
\usage{
are_beast2_input_lines(text, verbose = FALSE)
}
\arguments{
\item{text}{lines of text}

\item{verbose}{if TRUE, BEAST2 output is shown,
no output otherwise}
}
\value{
TRUE if the text is valid, FALSE if not
}
\description{
Would these lines of text, when written to a file,
  result in a valid BEAST2 input file?
}
\seealso{
Use \code{\link{is_beast2_input_file}} to check a file
}
\author{
Richel J.C. Bilderbeek
}
