% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_clock_models.R
\name{are_clock_models}
\alias{are_clock_models}
\title{Determine if x consists out of clock_models objects}
\usage{
are_clock_models(x)
}
\arguments{
\item{x}{the object to check if it consists out of clock_models objects}
}
\value{
TRUE if x, or all elements of x, are clock_model objects
}
\description{
Determine if x consists out of clock_models objects
}
\examples{
library(testthat)

rln_clock_model <- create_rln_clock_model()
strict_clock_model <- create_strict_clock_model()
both_clock_models <- list(rln_clock_model, strict_clock_model)
expect_true(are_clock_models(rln_clock_model))
expect_true(are_clock_models(strict_clock_model))
expect_true(are_clock_models(both_clock_models))

expect_false(are_clock_models(NA))
expect_false(are_clock_models(NULL))
expect_false(are_clock_models("nonsense"))
expect_false(are_clock_models(create_jc69_site_model()))
}
\author{
Richèl J.C. Bilderbeek
}
