% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_equal_tracelogs.R
\name{are_equal_tracelogs}
\alias{are_equal_tracelogs}
\title{Determine if two tracelogs are equal.}
\usage{
are_equal_tracelogs(tracelog_1, tracelog_2)
}
\arguments{
\item{tracelog_1}{an tracelog, as created by \link{create_tracelog}}

\item{tracelog_2}{an tracelog, as created by \link{create_tracelog}}
}
\value{
TRUE if the two tracelogs are equal
}
\description{
Will \link{stop} if the arguments are not tracelogs.
}
\examples{
library(testthat)

tracelog_1 <- create_tracelog(log_every = 1000)
tracelog_2 <- create_tracelog(log_every = 314)
expect_true(are_equal_tracelogs(tracelog_1, tracelog_1))
expect_false(are_equal_tracelogs(tracelog_1, tracelog_2))
}
\seealso{
Use \link{create_tracelog} to create an tracelog
}
\author{
Richèl J.C. Bilderbeek
}
