% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_equal_xml_files.R
\name{are_equal_xml_files}
\alias{are_equal_xml_files}
\title{Determine if XML files result in equal trees}
\usage{
are_equal_xml_files(filename_1, filename_2, section)
}
\arguments{
\item{filename_1}{name of a first XML file}

\item{filename_2}{name of a second XML file}

\item{section}{name of an XML section.
Assumes that there is one line that starts with \code{<section}
(excluding whitespace)
and one line that is \code{</section>} (also excluding whitespace)}
}
\value{
TRUE if the two sections of the XML files are equal,
  FALSE otherwise
}
\description{
Determine if XML files result in equal trees
}
\seealso{
to check for equivalence, use \code{\link{are_equivalent_xml_files}}
}
\author{
Richèl J.C. Bilderbeek
}
