% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_equivalent_xmls.R
\name{are_equivalent_xml_files}
\alias{are_equivalent_xml_files}
\title{Determine if XML files result in equivalent trees}
\usage{
are_equivalent_xml_files(filename1, filename2)
}
\arguments{
\item{filename1}{name of a first XML file}

\item{filename2}{name of a second XML file}
}
\value{
TRUE if the two XML files result in equivalent trees,
  FALSE otherwise
}
\description{
Determine if XML files result in equivalent trees
}
\author{
Richel J.C. Bilderbeek
}
