% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_equivalent_xmls.R
\name{are_equivalent_xml_lines}
\alias{are_equivalent_xml_lines}
\title{Determine if XML lines result in equivalent trees}
\usage{
are_equivalent_xml_lines(lines1, lines2, verbose = FALSE)
}
\arguments{
\item{lines1}{lines of a first XML file}

\item{lines2}{lines of a second XML file}

\item{verbose}{print the reason why the XML lines differ}
}
\value{
TRUE if the two XML lines result in equivalent trees,
  FALSE otherwise
}
\description{
Determine if XML lines result in equivalent trees
}
\author{
Richel J.C. Bilderbeek
}
