% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_equivalent_xmls.R
\name{are_equivalent_xml_lines_all}
\alias{are_equivalent_xml_lines_all}
\title{Determine if XML lines result in equivalent trees}
\usage{
are_equivalent_xml_lines_all(lines_1, lines_2, verbose = FALSE)
}
\arguments{
\item{lines_1}{lines of a first XML file}

\item{lines_2}{lines of a second XML file}

\item{verbose}{if TRUE, additional information is displayed, that
is potentially useful in debugging}
}
\value{
TRUE if the two XML lines result in equivalent trees,
  FALSE otherwise
}
\description{
Determine if XML lines result in equivalent trees
}
\author{
Richel J.C. Bilderbeek
}
