% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_ids.R
\name{are_ids}
\alias{are_ids}
\title{Determine if x consists out of IDs}
\usage{
are_ids(x)
}
\arguments{
\item{x}{the object to check if it consists out of IDs}
}
\value{
TRUE if x, or all elements of x, are IDs
}
\description{
Determine if x consists out of IDs
}
\seealso{
to check one ID, use \code{\link{is_id}}
}
\author{
Richel J.C. Bilderbeek
}
