% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_init_clock_models.R
\name{are_init_clock_models}
\alias{are_init_clock_models}
\title{Determine if x consists out of initialized clock_models objects}
\usage{
are_init_clock_models(x)
}
\arguments{
\item{x}{the object to check if it consists out of
initialized clock_models objects}
}
\value{
TRUE if x, or all elements of x, are initialized clock_model objects
}
\description{
Determine if x consists out of initialized clock_models objects
}
\author{
Richel J.C. Bilderbeek
}
