% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_init_mrca_priors.R
\name{are_init_mrca_priors}
\alias{are_init_mrca_priors}
\title{Determine if x consists out of initialized MRCA priors}
\usage{
are_init_mrca_priors(x)
}
\arguments{
\item{x}{the object to check if it consists out of
initialized MRCA priors}
}
\value{
TRUE if x, or all elements of x, are initialized MRCA priors
}
\description{
Determine if x consists out of initialized MRCA priors
}
\author{
Richèl J.C. Bilderbeek
}
