% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_init_site_models.R
\name{are_init_site_models}
\alias{are_init_site_models}
\title{Determine if x consists out of initialized site_models objects}
\usage{
are_init_site_models(x)
}
\arguments{
\item{x}{the object to check if it consists out of
initialized site_models objects}
}
\value{
TRUE if x, or all elements of x, are initialized site_model objects
}
\description{
Determine if x consists out of initialized site_models objects
}
\author{
Richèl J.C. Bilderbeek
}
