% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_init_tree_priors.R
\name{are_init_tree_priors}
\alias{are_init_tree_priors}
\title{Determine if x consists out of initialized tree_priors objects}
\usage{
are_init_tree_priors(x)
}
\arguments{
\item{x}{the object to check if it consists out of
initialized tree_priors objects}
}
\value{
TRUE if x, or all elements of x, are initialized tree_prior objects
}
\description{
Determine if x consists out of initialized tree_priors objects
}
\author{
Richèl J.C. Bilderbeek
}
