% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_initial_phylogenies.R
\name{are_initial_phylogenies}
\alias{are_initial_phylogenies}
\title{Determine if x is exactly one phylogeny}
\usage{
are_initial_phylogenies(phylos)
}
\arguments{
\item{phylos}{the object to check if it is exactly one phylogeny}
}
\value{
TRUE if x is exactly one phylogeny
}
\description{
Determine if x is exactly one phylogeny
}
\examples{
  phylogeny <- ape::read.tree(text = "(A:1,B:1);")
  testthat::expect_false(are_initial_phylogenies(phylogeny))
  testthat::expect_true(are_initial_phylogenies(NA))
  testthat::expect_true(are_initial_phylogenies(c(phylogeny)))
  testthat::expect_true(are_initial_phylogenies(c(phylogeny, phylogeny)))
}
\author{
Richèl J.C. Bilderbeek
}
