% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_mrca_alignment_ids_in_fastas.R
\name{are_mrca_align_ids_in_fastas}
\alias{are_mrca_align_ids_in_fastas}
\title{Determine if the MRCA priors' alignment IDs are present in the FASTA files}
\usage{
are_mrca_align_ids_in_fastas(mrca_priors, fasta_filenames)
}
\arguments{
\item{mrca_priors}{a list of one or more Most Recent Common Ancestor priors,
as returned by \code{\link{create_mrca_prior}}}

\item{fasta_filenames}{One or more FASTA filenames.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}
}
\description{
Determine if the MRCA priors' alignment IDs are present in the FASTA files
}
\author{
Richel J.C. Bilderbeek
}
