% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_mrca_taxa_names_in_fastas.R
\name{are_mrca_taxa_names_in_fastas}
\alias{are_mrca_taxa_names_in_fastas}
\title{Determine if the MRCA priors' taxa names are present in the FASTA files}
\usage{
are_mrca_taxa_names_in_fastas(mrca_taxa_names, fasta_filenames)
}
\arguments{
\item{fasta_filenames}{One or more FASTA filenames.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}
}
\description{
Determine if the MRCA priors' taxa names are present in the FASTA files
}
\author{
Richel J.C. Bilderbeek
}
