% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_mrca_taxa_non_intersecting.R
\name{are_mrca_taxa_non_intersecting}
\alias{are_mrca_taxa_non_intersecting}
\title{Determine if MRCA priors that are monophyletic (i.e., assume
no other taxons share a same MRCA) are compatible, that
is, there are no taxa that intersect. For example, one
MRCA prior with taxon 1 and 2 is incompatible with an MRCA prior
with taxon 3 and 4}
\usage{
are_mrca_taxa_non_intersecting(mrca_priors)
}
\arguments{
\item{mrca_priors}{a list of one or more Most Recent Common Ancestor priors,
as returned by \code{\link{create_mrca_prior}}}
}
\description{
Determine if MRCA priors that are monophyletic (i.e., assume
no other taxons share a same MRCA) are compatible, that
is, there are no taxa that intersect. For example, one
MRCA prior with taxon 1 and 2 is incompatible with an MRCA prior
with taxon 3 and 4
}
\author{
Richel J.C. Bilderbeek
}
