% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_mrca_taxon_names_in_fasta.R
\name{are_mrca_taxon_names_in_fasta}
\alias{are_mrca_taxon_names_in_fasta}
\title{Determine if the MRCA priors' taxa names are present in the FASTA files}
\usage{
are_mrca_taxon_names_in_fasta(mrca_prior, fasta_filename)
}
\arguments{
\item{mrca_prior}{a Most Recent Common Ancestor prior,
as returned by \code{\link{create_mrca_prior}}}

\item{fasta_filename}{a FASTA filename.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}
}
\value{
TRUE if the MRCA priors' taxa names are
  present in the FASTA files. FALSE otherwise.
}
\description{
Determine if the MRCA priors' taxa names are present in the FASTA files
}
\author{
Richèl J.C. Bilderbeek
}
