% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_rln_clock_models.R
\name{are_rln_clock_models}
\alias{are_rln_clock_models}
\title{Are the clock models Relaxed Log-Normal clock models?}
\usage{
are_rln_clock_models(clock_models)
}
\arguments{
\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}
}
\description{
Are the clock models Relaxed Log-Normal clock models?
}
\author{
Richel J.C. Bilderbeek
}
