% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_site_models.R
\name{are_site_models}
\alias{are_site_models}
\title{Determine if x consists out of site_models objects}
\usage{
are_site_models(x)
}
\arguments{
\item{x}{the object to check if it consists out of site_models objects}
}
\value{
TRUE if x, or all elements of x, are site_model objects
}
\description{
Determine if x consists out of site_models objects
}
\author{
Richel J.C. Bilderbeek
}
