% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_tree_priors.R
\name{are_tree_priors}
\alias{are_tree_priors}
\title{Determine if x consists out of tree_priors objects}
\usage{
are_tree_priors(x)
}
\arguments{
\item{x}{the object to check if it consists out of tree_priors objects}
}
\value{
TRUE if x, or all elements of x, are tree_prior objects
}
\description{
Determine if x consists out of tree_priors objects
}
\author{
Richel J.C. Bilderbeek
}
