% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_list.R
\name{as_list}
\alias{as_list}
\title{Create a list of the element, in which index 1 contains
object a, index 2 contains object b, etcetera}
\usage{
as_list(a, b)
}
\arguments{
\item{a}{desired first element of the list}

\item{b}{desired second element of the list}
}
\value{
a list with object a as its first element, object b
  as its second element, etcetera
}
\description{
Create a list of the element, in which index 1 contains
object a, index 2 contains object b, etcetera
}
\examples{
  v <- as_list(create_hky_site_model(), create_jc69_site_model())
  testit::assert(is_hky_site_model(v[[1]]))
  testit::assert(is_jc69_site_model(v[[2]]))
}
\author{
Richel J.C. Bilderbeek
}
