% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_pop_sizes_parameter_to_xml.R
\name{b_pop_sizes_parameter_to_xml}
\alias{b_pop_sizes_parameter_to_xml}
\title{Internal function}
\usage{
b_pop_sizes_parameter_to_xml(
  b_pop_sizes_parameter,
  beauti_options = create_beauti_options()
)
}
\arguments{
\item{b_pop_sizes_parameter}{a Bayesian population size parameter,
as created by \link{create_b_pop_sizes_param}}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the parameter as XML text
}
\description{
Converts a Bayesian population sizes parameter to XML
}
\examples{
b_pop_sizes_parameter_to_xml(
  b_pop_sizes_parameter = create_b_pop_sizes_param(id = 42),
  beauti_options = create_beauti_options()
)
b_pop_sizes_parameter_to_xml(
  b_pop_sizes_parameter = create_b_pop_sizes_param(id = 42, upper = Inf),
  beauti_options = create_beauti_options()
)
}
\author{
Richèl J.C. Bilderbeek
}
