% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distr.R
\name{bd_tree_prior_to_xml_prior}
\alias{bd_tree_prior_to_xml_prior}
\title{Creates the tree prior section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file for a Birth-Death tree prior}
\usage{
bd_tree_prior_to_xml_prior(bd_tree_prior)
}
\arguments{
\item{bd_tree_prior}{a Birth-Death tree_prior,
as created by \code{\link{create_bd_tree_prior}}}
}
\description{
Creates the tree prior section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file for a Birth-Death tree prior
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
