% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastscriptr.R
\name{beast_scriptr}
\alias{beast_scriptr}
\title{Create a BEAST2 XML parameter file}
\usage{
beast_scriptr(input_fasta_filename, mcmc_chainlength, tree_prior,
  output_xml_filename, fixed_crown_age = FALSE, initial_phylogeny = NA)
}
\arguments{
\item{input_fasta_filename}{Filename of a fasta file}

\item{mcmc_chainlength}{Length of MCMC chain}

\item{tree_prior}{The tree prior, can be 'birth_death' or
'coalescent_constant_population'}

\item{output_xml_filename}{Name of the XML parameter file created by this
function. BEAST2 uses this file as input.}

\item{fixed_crown_age}{determines if the phylogeny its crown age is
fixed. If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}

\item{initial_phylogeny}{the MCMC chain its initial phylogeny. If
this is set to NA, BEAST2 will use a random phylogeny. Else
a phylogeny must be supplied of class ape::phylo.}
}
\description{
Create a BEAST2 XML parameter file
}
