% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastscriptr.R
\name{beast_scriptr_data}
\alias{beast_scriptr_data}
\title{Creates the data section of a BEAST2 XML parameter file}
\usage{
beast_scriptr_data(filename_base, input_fasta_filename)
}
\arguments{
\item{filename_base}{base of the filenames}

\item{input_fasta_filename}{name of the FASTA file}
}
\description{
Creates the data section of a BEAST2 XML parameter file
}
