% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastscriptr.R
\name{beast_scriptr_init}
\alias{beast_scriptr_init}
\title{Creates the map section of a BEAST2 XML parameter file}
\usage{
beast_scriptr_init(filename_base, initial_phylogeny)
}
\arguments{
\item{filename_base}{base of the filenames}

\item{initial_phylogeny}{initial phylogeny}
}
\description{
Creates the map section of a BEAST2 XML parameter file
}
