% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastscriptr.R
\name{beast_scriptr_map}
\alias{beast_scriptr_map}
\title{Creates the map section of a BEAST2 XML parameter file}
\usage{
beast_scriptr_map()
}
\arguments{
\item{filename_base}{base of the filenames}

\item{input_fasta_filename}{name of the FASTA file}
}
\description{
Creates the map section of a BEAST2 XML parameter file
}
