% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastscriptr.R
\name{beast_scriptr_operators}
\alias{beast_scriptr_operators}
\title{Creates the operators section of a BEAST2 XML parameter file}
\usage{
beast_scriptr_operators(filename_base, tree_prior, fixed_crown_age)
}
\arguments{
\item{filename_base}{base of the filenames}

\item{tree_prior}{The tree prior, can be 'birth_death' or
'coalescent_constant_population'}

\item{fixed_crown_age}{determines if the phylogeny its crown age is
fixed. If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}
}
\description{
Creates the operators section of a BEAST2 XML parameter file
}
