% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastscriptr.R
\name{beast_scriptr_state}
\alias{beast_scriptr_state}
\title{Creates the state section of a BEAST2 XML parameter file}
\usage{
beast_scriptr_state(filename_base, tree_prior, initial_phylogeny)
}
\arguments{
\item{filename_base}{filename its base}

\item{tree_prior}{tree prior}

\item{initial_phylogeny}{initial phylogeny or NA}
}
\description{
Creates the state section of a BEAST2 XML parameter file
}
