% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastscriptr.R
\name{beast_scriptr_xml}
\alias{beast_scriptr_xml}
\title{Creates the xml section of a BEAST2 XML parameter file}
\usage{
beast_scriptr_xml()
}
\description{
Creates the xml section of a BEAST2 XML parameter file
}
