% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beautier-package.R
\docType{package}
\name{beautier-package}
\alias{beautier}
\alias{beautier-package}
\title{beautier: 'BEAUti' from R}
\description{
'BEAST2' (\url{https://www.beast2.org}) is a widely used Bayesian phylogenetic tool, that uses DNA/RNA/protein data and many model priors to create a posterior of jointly estimated phylogenies and parameters. 'BEAUti 2' (which is part of 'BEAST2') is a GUI tool that allows users to specify the many possible setups and generates the XML file 'BEAST2' needs to run. This package provides a way to create 'BEAST2' input files without active user input, but using R function calls instead.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/beautier/ (website) https://github.com/ropensci/beautier/}
  \item Report bugs at \url{https://github.com/ropensci/beautier}
}

}
\author{
\strong{Maintainer}: Richèl J.C. Bilderbeek \email{richel@richelbilderbeek.nl} (\href{https://orcid.org/0000-0003-1107-7049}{ORCID})

Other contributors:
\itemize{
  \item Joëlle Barido-Sottani (Joëlle reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/209) [reviewer]
  \item David Winter (David reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/209) [reviewer]
  \item Paul Van Els (\href{https://orcid.org/0000-0002-9499-8873}{ORCID}) [contributor]
  \item Raphael Scherrer (\href{https://orcid.org/0000-0002-1447-7630}{ORCID}) [contributor]
  \item Yacine B. Chehida (\href{https://orcid.org/0000-0001-7269-9082}{ORCID}) [contributor]
  \item Katharine S. Walter (\href{https://orcid.org/0000-0003-0065-2204}{ORCID}) [contributor]
  \item Gary Napier (\href{https://orcid.org/0000-0002-1077-0055}{ORCID}) [contributor]
  \item Jason Griffiths (\href{https://orcid.org/0000-0002-1667-8233}{ORCID}) [contributor]
  \item Thijs Janzen (\href{https://orcid.org/0000-0002-4162-1140}{ORCID}) [contributor]
  \item Pedro Taucce (\href{https://orcid.org/0000-0002-3088-4543}{ORCID}) [contributor]
  \item Jana Riederer (\href{https://orcid.org/0000-0001-6951-9984}{ORCID}) [contributor]
}

}
\keyword{internal}
