% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beautier.R
\docType{package}
\name{beautier}
\alias{beautier}
\alias{beautier-package}
\title{beautier: A package to create a BEAST2 input file,}
\description{
'beautier' allows to create a BEAST2 input file, using
an R interface. 'beautier' closely follows the interface
of BEAUti 2, a GUI tool bundled with BEAST2, including
its default settings.
}
