% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_parameter_to_xml.R
\name{beta_parameter_to_xml}
\alias{beta_parameter_to_xml}
\title{Internal function}
\usage{
beta_parameter_to_xml(beta_parameter, beauti_options = create_beauti_options())
}
\arguments{
\item{beta_parameter}{a beta parameter,
as created by \link{create_beta_param}}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the parameter as XML text
}
\description{
Converts a beta parameter to XML
}
\author{
Richèl J.C. Bilderbeek
}
