% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distr.R
\name{cbs_tree_prior_to_xml_prior_distr}
\alias{cbs_tree_prior_to_xml_prior_distr}
\title{Creates the tree prior section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file for a Birth-Death tree prior}
\usage{
cbs_tree_prior_to_xml_prior_distr(cbs_tree_prior)
}
\arguments{
\item{cbs_tree_prior}{a Coalenscent Bayesian Skyline tree_prior,
as created by \code{\link{create_cbs_tree_prior}}}
}
\description{
Creates the tree prior section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file for a Birth-Death tree prior
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
