% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp_tree_prior_to_xml_state.R
\name{ccp_tree_prior_to_xml_state}
\alias{ccp_tree_prior_to_xml_state}
\title{Convert a CCP tree prior to the XML as part of the \code{state} section}
\usage{
ccp_tree_prior_to_xml_state(tree_prior)
}
\description{
Convert a CCP tree prior to the XML as part of the \code{state} section
}
