% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distr.R
\name{cep_tree_prior_to_xml_prior_distr}
\alias{cep_tree_prior_to_xml_prior_distr}
\title{Creates the tree prior section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file for a
Coalescent Exponential Population tree prior}
\usage{
cep_tree_prior_to_xml_prior_distr(cep_tree_prior)
}
\arguments{
\item{cep_tree_prior}{a Coalescent Exponential Population tree prior,
as created by \code{\link{create_cep_tree_prior}}}
}
\description{
Creates the tree prior section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file for a
Coalescent Exponential Population tree prior
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
