% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bool.R
\name{check_bool}
\alias{check_bool}
\title{Determine if `x` is one boolean}
\usage{
check_bool(x)
}
\arguments{
\item{x}{the object to be determined to be one boolean}
}
\value{
Nothing. Will raise an exception if the value is not one boolean
}
\description{
Determine if `x` is one boolean
}
\note{
Created by modifying check_bool from import-standalone-type-check.
}
\examples{
check_bool(TRUE)
check_bool(FALSE)
}
\author{
Richèl J.C. Bilderbeek
}
