% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_empty_beautier_folder.R
\name{check_empty_beautier_folder}
\alias{check_empty_beautier_folder}
\title{Check there are no files in the default \link{beautier} folder}
\usage{
check_empty_beautier_folder(beautier_folder = get_beautier_folder())
}
\arguments{
\item{beautier_folder}{the path to
the \link{beautier} temporary files folder}
}
\value{
Nothing.
}
\description{
Check there are no files in the default \link{beautier} folder.
The goal is to make sure no temporary files are left undeleted.
Will \link{stop} if there are files in the \link{beautier} folder
}
\examples{
check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
