% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_true_false.R
\name{check_false}
\alias{check_false}
\title{Determine if `x` is one FALSE}
\usage{
check_false(
  x,
  ...,
  allow_na = FALSE,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{x}{the object to be determined to be one FALSE}

\item{...}{other arguments, no idea why this is needed}

\item{allow_na}{set to TRUE to allow NA to be valid}

\item{allow_null}{set to TRUE to allow NULL to be valid}

\item{arg}{no idea why this is needed}

\item{call}{no idea why this is needed}
}
\value{
Nothing. Will raise an exception if the value is not one FALSE
}
\description{
Determine if `x` is one FALSE
}
\note{
From
[`https://github.com/r-lib/rlang`](https://github.com/r-lib/rlang),
file `R/import-standalone-type-check.R`
}
\examples{
check_false(FALSE)
}
\author{
[`olivroy`](https://github.com/olivroy) and Richèl J.C. Bilderbeek
}
