% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_and_model_agree.R
\name{check_file_and_model_agree}
\alias{check_file_and_model_agree}
\title{Checks if the input FASTA file and the inference model agree.}
\usage{
check_file_and_model_agree(input_filename, inference_model)
}
\arguments{
\item{input_filename}{A FASTA filename.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}

\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\description{
Will \link{stop} if not
}
