% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filename.R
\name{check_filename}
\alias{check_filename}
\title{Check if the `filename` is valid}
\usage{
check_filename(filename, allow_empty_str = FALSE, allow_na = FALSE)
}
\arguments{
\item{allow_empty_str}{allow a string to be empty}

\item{allow_na}{allow \link{NA}}
}
\value{
nothing
}
\description{
Calls \code{stop} if the filename is invalid
}
\examples{
check_filename("trace.log")
check_filename("my.trees")
}
\author{
Richèl J.C. Bilderbeek
}
