% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gamma_site_model.R
\name{check_gamma_site_model}
\alias{check_gamma_site_model}
\title{Checks if the parameter is a valid gamma site model}
\usage{
check_gamma_site_model(gamma_site_model)
}
\arguments{
\item{x}{the object to be determined if it is a valid gamma site object}
}
\value{
nothing. Will call \code{stop} if the argument is not a valid
  gamma site model
}
\description{
Checks if the parameter is a valid gamma site model
}
\examples{
  testthat::expect_silent(
    beautier:::check_gamma_site_model(
      create_gamma_site_model()
    )
  )
  testthat::expect_error(
    check_gamma_site_model(
      "not a gamma site model"
    )
  )
}
\author{
Richèl J.C. Bilderbeek
}
