% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gamma_site_model.R
\name{check_gamma_site_model_names}
\alias{check_gamma_site_model_names}
\title{Checks if the gamma site model has the right list elements' names}
\usage{
check_gamma_site_model_names(gamma_site_model)
}
\arguments{
\item{gamma_site_model}{a site model's gamma site model,
as returned by \code{\link{create_gamma_site_model}}}
}
\value{
nothing. Will call \code{stop} if the argument is not a valid
  gamma site model
}
\description{
Checks if the gamma site model has the right list elements' names
}
\author{
Richèl J.C. Bilderbeek
}
