% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gtr_site_model.R
\name{check_gtr_site_model}
\alias{check_gtr_site_model}
\title{Check if the \code{gtr_site_model} is a valid
GTR nucleotide substitution model.}
\usage{
check_gtr_site_model(gtr_site_model)
}
\arguments{
\item{gtr_site_model}{a GTR site model,
as returned by \code{\link{create_gtr_site_model}}}
}
\description{
Use \link{create_gtr_site_model} to create a valid
GTR nucleotide substitution model.
}
\examples{
check_gtr_site_model(create_gtr_site_model())
}
