% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gtr_site_model.R
\name{check_gtr_site_model_names}
\alias{check_gtr_site_model_names}
\title{Check if the \code{gtr_site_model} has the list elements
of a valid \code{gtr_site_model} object.}
\usage{
check_gtr_site_model_names(gtr_site_model)
}
\arguments{
\item{gtr_site_model}{a GTR site model,
as returned by \code{\link{create_gtr_site_model}}}
}
\value{
nothing
}
\description{
Calls \code{stop} if an element is missing
}
\seealso{
Use \link{create_gtr_site_model}
to create a valid \code{gtr_site_model}
}
\author{
Richèl J.C. Bilderbeek
}
