% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_is_monophyletic.R
\name{check_is_monophyletic}
\alias{check_is_monophyletic}
\title{Check if \code{is_monophyletic} has a valid value.}
\usage{
check_is_monophyletic(is_monophyletic)
}
\arguments{
\item{is_monophyletic}{boolean to indicate monophyly is assumed in
a Most Recent Common Ancestor prior,
as returned by \code{\link{create_mrca_prior}}}
}
\description{
Will \link{stop} if not.
}
