% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_log_mode.R
\name{check_log_mode}
\alias{check_log_mode}
\title{Check if the supplied \code{mode} is a valid logging mode.}
\usage{
check_log_mode(mode)
}
\arguments{
\item{mode}{mode how to log.
Valid are \code{tree}, \code{autodetect} and \code{compound}}
}
\value{
No return value, called for side effects
}
\description{
Check if the supplied \code{mode} is a valid logging mode.
}
