% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_log_sort.R
\name{check_log_sort}
\alias{check_log_sort}
\title{Check if the supplied \code{sort} is a valid logging sorting option.}
\usage{
check_log_sort(sort)
}
\arguments{
\item{sort}{how to sort the entries in a log.
Valid are \code{smart}, \code{none} and \code{alphabetic}}
}
\description{
Check if the supplied \code{sort} is a valid logging sorting option.
}
