% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-types-check.R
\name{check_logical}
\alias{check_logical}
\title{Determine if `x` is one logical value}
\usage{
check_logical(
  x,
  ...,
  allow_null = FALSE,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{the object to be determined to be one logical value}

\item{...}{other arguments, no idea why this is needed}

\item{allow_null}{set to TRUE to allow NULL to be valid}

\item{arg}{no idea why this is needed}

\item{call}{no idea why this is needed}
}
\value{
Nothing. Will raise an exception if the value is not
one logical value
}
\description{
Determine if `x` is one logical value
}
\note{
From
[r-lib](https://github.com/r-lib/rlang/blob/main/R/standalone-types-check.R)
}
\examples{
check_logical(TRUE)
check_logical(FALSE)
check_logical(NA)
}
\author{
[`olivroy`](https://github.com/olivroy) and Richèl J.C. Bilderbeek
}
