% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mcmc.R
\name{check_mcmc_list_element_names}
\alias{check_mcmc_list_element_names}
\title{Check if the MCMC has the list elements of a valid MCMC object.}
\usage{
check_mcmc_list_element_names(mcmc)
}
\arguments{
\item{mcmc}{one MCMC
as returned by \code{\link{create_mcmc}}}
}
\value{
nothing
}
\description{
Calls \code{stop} if an element is missing
}
\seealso{
Use \link{create_mcmc} to create a valid MCMC
}
\author{
Richèl J.C. Bilderbeek
}
