% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mcmc_nested_sampling.R
\name{check_mcmc_nested_sampling}
\alias{check_mcmc_nested_sampling}
\title{Check if this an MCMC that uses Nested Sampling
to estimate a marginal likelihood.}
\usage{
check_mcmc_nested_sampling(mcmc)
}
\description{
Will \link{stop} if not, else will do nothing
}
\seealso{
use \link{create_nested_sampling_mcmc}
to create an MCMC that uses Nested Sampling
to estimate a marginal likelihood
}
\author{
Richel J.C. Bilderbeek
}
