% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mcmc.R
\name{check_mcmc_values}
\alias{check_mcmc_values}
\title{Check if the MCMC has the list elements with valid values
for being a valid MCMC object.}
\usage{
check_mcmc_values(mcmc)
}
\arguments{
\item{mcmc}{one MCMC
as returned by \code{\link{create_mcmc}}}
}
\value{
nothing
}
\description{
Calls \code{stop} if a value is invalid
}
\seealso{
Use \link{create_mcmc} to create a valid MCMC
}
\author{
Richèl J.C. Bilderbeek
}
