% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_misc_options.R
\name{check_misc_options}
\alias{check_misc_options}
\title{Check if the misc options is a valid misc options object.}
\usage{
check_misc_options(misc_options)
}
\arguments{
\item{misc_options}{one misc_options object,
as returned by \code{\link{create_misc_options}}}
}
\value{
nothing
}
\description{
Calls \code{stop} if the misc options object is invalid
}
\examples{
 testthat::expect_silent(check_misc_options(create_misc_options()))

 # Must stop on non-MCMCs
 testthat::expect_error(check_misc_options(misc_options = "nonsense"))
 testthat::expect_error(check_misc_options(misc_options = NULL))
 testthat::expect_error(check_misc_options(misc_options = NA))
}
\seealso{
Use \link{create_misc_options} to create a valid MCMC
}
\author{
Richel J.C. Bilderbeek
}
