% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mrca_prior.R
\name{check_mrca_prior}
\alias{check_mrca_prior}
\title{Check if the MRCA prior is a valid MRCA prior.
Calls \code{stop} if the MRCA prior is invalid.}
\usage{
check_mrca_prior(mrca_prior)
}
\arguments{
\item{mrca_prior}{a Most Recent Common Ancestor prior,
as returned by \code{\link{create_mrca_prior}}}
}
\value{
nothing
}
\description{
Check if the MRCA prior is a valid MRCA prior.
Calls \code{stop} if the MRCA prior is invalid.
}
\examples{
 testthat::expect_silent(check_mrca_prior(create_mrca_prior()))

 # Must stop on non-MRCA priors
 testthat::expect_error(check_mrca_prior(mrca_prior = "nonsense"))
 testthat::expect_error(check_mrca_prior(mrca_prior = NULL))
 testthat::expect_error(check_mrca_prior(mrca_prior = NA))
}
\seealso{
Use \link{create_mrca_prior} to create a valid MRCA prior
}
\author{
Richel J.C. Bilderbeek
}
