% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mrca_prior_name.R
\name{check_mrca_prior_name}
\alias{check_mrca_prior_name}
\title{Check if \code{mrca_prior_name} is a valid MRCA prior name.}
\usage{
check_mrca_prior_name(mrca_prior_name)
}
\arguments{
\item{mrca_prior_name}{the unique name of the MRCA prior,
for example a genus, family,
order or even class name.
Leave at \link{NA} to have it named automatically.}
}
\value{
No return value, called for side effects
}
\description{
A valid MRCA prior name is either \link{NA} or one character string.
Will \link{stop} if not.
}
