% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mrca_prior_taxa_names.R
\name{check_mrca_prior_taxa_names}
\alias{check_mrca_prior_taxa_names}
\title{Check the MRCA prior's taxon names are valid.}
\usage{
check_mrca_prior_taxa_names(taxa_names)
}
\description{
Will \link{stop} if not.
}
