% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-types-check.R
\name{check_number_whole}
\alias{check_number_whole}
\title{Determine if `x` is one whole number}
\usage{
check_number_whole(
  x,
  ...,
  min = NULL,
  max = NULL,
  allow_infinite = FALSE,
  allow_na = FALSE,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{x}{the object to be determined to be one whole number}

\item{...}{other arguments, no idea why this is needed}

\item{min}{lowest number allowed}

\item{max}{lowest number allowed}

\item{allow_infinite}{set to TRUE to allow `Inf` to be valid}

\item{allow_na}{set to TRUE to allow NA to be valid}

\item{allow_null}{set to TRUE to allow NULL to be valid}

\item{arg}{no idea why this is needed}

\item{call}{no idea why this is needed}
}
\value{
Nothing. Will raise an exception if the value is
not one whole number
}
\description{
Determine if `x` is one whole number
}
\note{
From
[r-lib](https://github.com/r-lib/rlang/blob/main/R/standalone-types-check.R)
}
\examples{
check_number_whole(1)
check_number_whole(1.0)
}
\author{
[`olivroy`](https://github.com/olivroy) and Richèl J.C. Bilderbeek
}
