% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_param.R
\name{check_param}
\alias{check_param}
\title{Check if the parameter is a valid parameter}
\usage{
check_param(param)
}
\arguments{
\item{param}{a parameter, as can be created by \code{\link{create_param}}.}
}
\value{
nothing
}
\description{
Calls \code{stop} if the parameter is invalid
}
\examples{
check_empty_beautier_folder()

check_param(create_alpha_param())
check_param(create_beta_param())

check_empty_beautier_folder()
}
\seealso{
Use \link{create_param} to create a valid parameter
}
\author{
Richèl J.C. Bilderbeek
}
