% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_param.R
\name{check_param_names}
\alias{check_param_names}
\title{Check if the \code{param} has the list elements
of a valid \code{param} object.}
\usage{
check_param_names(param)
}
\arguments{
\item{param}{a parameter, as can be created by \code{\link{create_param}}.}
}
\value{
nothing
}
\description{
Calls \code{stop} if an element is missing
}
\seealso{
Use \link{create_param} to create a valid \code{param}
}
\author{
Richèl J.C. Bilderbeek
}
