% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_param.R
\name{check_param_types}
\alias{check_param_types}
\title{Check if the \code{param} has the list elements
of the right type for a valid \code{param} object.}
\usage{
check_param_types(param)
}
\arguments{
\item{param}{a parameter, as can be created by \code{\link{create_param}}.}
}
\value{
nothing
}
\description{
Calls \code{stop} if an element has the incorrect type
}
\seealso{
Use \link{create_param} to create a valid \code{param}
}
\author{
Richèl J.C. Bilderbeek
}
