% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phylo.R
\name{check_phylo}
\alias{check_phylo}
\title{Check if the phylogeny is a valid phylogeny object.}
\usage{
check_phylo(phylo)
}
\value{
nothing
}
\description{
Calls \code{stop} if the phylogeny is invalid
}
\examples{
 phylogeny <- ape::read.tree(text = "((A:1, B:1):1;")
 testthat::expect_silent(check_phylo(phylogeny))

 # Must stop on non-phylogenies
 testthat::expect_error(check_phylo(phylo = "nonsense"))
 testthat::expect_error(check_phylo(phylo = NULL))
 testthat::expect_error(check_phylo(phylo = NA))
}
\author{
Richel J.C. Bilderbeek
}
