% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_screenlog.R
\name{check_screenlog}
\alias{check_screenlog}
\title{Check if a \code{screenlog} is valid.}
\usage{
check_screenlog(screenlog)
}
\arguments{
\item{screenlog}{a \code{screenlog},
as created by \link{create_screenlog}}
}
\description{
Will call \link{stop} if not.
}
