% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_screenlog.R
\name{check_screenlog_names}
\alias{check_screenlog_names}
\title{Check if the \code{screenlog} has the list elements
of a valid \code{screenlog} object.}
\usage{
check_screenlog_names(screenlog)
}
\arguments{
\item{screenlog}{a \code{screenlog},
as created by \link{create_screenlog}}
}
\value{
nothing
}
\description{
Calls \code{stop} if an element is missing
}
\seealso{
Use \link{create_screenlog} to create a valid \code{screenlog}
}
\author{
Richèl J.C. Bilderbeek
}
