% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_screenlog.R
\name{check_screenlog_values}
\alias{check_screenlog_values}
\title{Check if the screenlog has the list elements with valid values
for being a valid screenlog object.}
\usage{
check_screenlog_values(screenlog)
}
\arguments{
\item{screenlog}{a \code{screenlog},
as created by \link{create_screenlog}}
}
\value{
nothing
}
\description{
Calls \code{stop} if a value is invalid
}
\seealso{
Use \link{create_screenlog} to create a valid screenlog
}
\author{
Richèl J.C. Bilderbeek
}
