% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_site_model.R
\name{check_site_model_types}
\alias{check_site_model_types}
\title{Check if the \code{site_model} has the list elements
of the right type for a valid \code{site_model} object.}
\usage{
check_site_model_types(site_model)
}
\arguments{
\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}
}
\value{
nothing
}
\description{
Calls \code{stop} if an element has the incorrect type
}
\seealso{
Use \link{create_site_model} to create a valid \code{site_model}
}
\author{
Richèl J.C. Bilderbeek
}
