% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_site_models.R
\name{check_site_models}
\alias{check_site_models}
\title{Check if the object is a list of one or more site models.}
\usage{
check_site_models(site_models)
}
\arguments{
\item{site_models}{the object to be checked if it is a list of one
or more valid site models}
}
\value{
nothing.
  Will \link{stop} if the object is not a list of one or more site models.
}
\description{
Will \link{stop} if the object is not a list of one or more site models.
}
\examples{
  testthat::expect_silent(check_site_models(create_jc69_site_model()))
  testthat::expect_silent(check_site_models(list(create_jc69_site_model())))
  testthat::expect_silent(
    check_site_models(
      list(create_jc69_site_model(), create_gtr_site_model())
    )
  )

  testthat::expect_error(check_site_models("nonsense"))
  testthat::expect_error(check_site_models(3.14))
  testthat::expect_error(check_site_models(42))
  testthat::expect_error(check_site_models(NA))
  testthat::expect_error(check_site_models(NULL))
}
\seealso{
Use \link{create_site_model} to create a valid site model
}
\author{
Richèl J.C. Bilderbeek
}
