% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_store_every.R
\name{check_store_every}
\alias{check_store_every}
\title{Check if \code{store_every} holds a valid value}
\usage{
check_store_every(store_every)
}
\arguments{
\item{store_every}{number of states the MCMC will process
before the posterior's state will be saved to file.
Use -1 or \code{NA} to use the default frequency.}
}
\value{
No return value, called for side effects
}
\description{
Will \link{stop} if not
}
