% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-types-check.R
\name{check_string}
\alias{check_string}
\title{Determine if `x` is one string}
\usage{
check_string(
  x,
  ...,
  allow_empty = TRUE,
  allow_na = FALSE,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{x}{the object to be determined to be one string}

\item{...}{other arguments, no idea why this is needed}

\item{allow_empty}{set to TRUE to allow an empty string to be valid}

\item{allow_na}{set to TRUE to allow NA to be valid}

\item{allow_null}{set to TRUE to allow NULL to be valid}

\item{arg}{no idea why this is needed}

\item{call}{no idea why this is needed}
}
\value{
Nothing. Will raise an exception if the value is not one string
}
\description{
Determine if `x` is one string
}
\note{
From
[r-lib](https://github.com/r-lib/rlang/blob/main/R/standalone-types-check.R)
}
\examples{
check_string("This is a string")
}
\author{
[`olivroy`](https://github.com/olivroy) and Richèl J.C. Bilderbeek
}
