% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tipdates_file.R
\name{check_tipdates_file}
\alias{check_tipdates_file}
\title{Check if the tip dates file is valid}
\usage{
check_tipdates_file(tipdates_filename)
}
\arguments{
\item{tipdates_filename}{name of the file containing the tip dates.
This file is assumed to have two columns, separated by a tab.
The first column contains the taxa names, the second column contains
the date.}
}
\value{
nothing
}
\description{
Check if the tip dates file is valid
}
\examples{
check_empty_beautier_folder()

check_tipdates_file(
  get_beautier_path("babette_issue_109.tsv")
) # OK

try(
  check_tipdates_file(
    get_beautier_path("babette_issue_109_no_tabs.tsv")
  ) # ERROR
)

try(
  check_tipdates_file(
    get_beautier_path("babette_issue_109_with_header.tsv")
  ) # ERROR
)

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
