% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tn93_site_model.R
\name{check_tn93_site_model}
\alias{check_tn93_site_model}
\title{Check if the \code{tn93_site_model} is a valid
TN93 nucleotide substitution model.}
\usage{
check_tn93_site_model(tn93_site_model)
}
\arguments{
\item{tn93_site_model}{a TN93 site model,
as returned by \code{\link{create_tn93_site_model}}}
}
\value{
No return value, called for side effects
}
\description{
Use \link{create_tn93_site_model} to create a valid
TN93 nucleotide substitution model.
}
\examples{
check_empty_beautier_folder()

check_tn93_site_model(create_tn93_site_model())

check_empty_beautier_folder()
}
