% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tracelog.R
\name{check_tracelog}
\alias{check_tracelog}
\title{Check if a \code{tracelog} is valid.}
\usage{
check_tracelog(tracelog)
}
\arguments{
\item{tracelog}{a \code{tracelog},
as created by \link{create_tracelog}}
}
\value{
No return value, called for side effects
}
\description{
Will call \link{stop} if not.
}
