% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tracelog.R
\name{check_tracelog_names}
\alias{check_tracelog_names}
\title{Check if the \code{tracelog} has the list elements
of a valid \code{tracelog} object.}
\usage{
check_tracelog_names(tracelog)
}
\arguments{
\item{tracelog}{a \code{tracelog},
as created by \link{create_tracelog}}
}
\value{
nothing
}
\description{
Calls \code{stop} if an element is missing
}
\seealso{
Use \link{create_tracelog} to create a valid \code{tracelog}
}
\author{
Richèl J.C. Bilderbeek
}
