% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tracelog.R
\name{check_tracelog_values}
\alias{check_tracelog_values}
\title{Check if the tracelog has the list elements with valid values
for being a valid tracelog object.}
\usage{
check_tracelog_values(tracelog)
}
\arguments{
\item{tracelog}{a \code{tracelog},
as created by \link{create_tracelog}}
}
\value{
nothing
}
\description{
Calls \code{stop} if a value is invalid
}
\seealso{
Use \link{create_tracelog} to create a valid tracelog
}
\author{
Richèl J.C. Bilderbeek
}
