% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tree_prior.R
\name{check_tree_prior}
\alias{check_tree_prior}
\title{Check if the tree prior is a valid tree prior}
\usage{
check_tree_prior(tree_prior)
}
\arguments{
\item{tree_prior}{a tree priors,
as returned by \code{\link{create_tree_prior}}}
}
\value{
nothing
}
\description{
Calls \code{stop} if the tree priors are invalid
}
\examples{
check_tree_prior(create_yule_tree_prior())
check_tree_prior(create_bd_tree_prior())
check_tree_prior(create_cbs_tree_prior())
check_tree_prior(create_ccp_tree_prior())
check_tree_prior(create_cep_tree_prior())

# Can use list of one tree prior
check_tree_prior(list(create_yule_tree_prior()))
}
\seealso{
Use \link{create_tree_prior} to create a valid tree prior
}
\author{
Richèl J.C. Bilderbeek
}
