% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tree_priors.R
\name{check_tree_priors}
\alias{check_tree_priors}
\title{Check if the object is a list of one or more tree priors.}
\usage{
check_tree_priors(tree_priors)
}
\arguments{
\item{tree_priors}{the object to be checked if it is a list of one
or more valid tree priors}
}
\value{
nothing.
  Will \link{stop} if the object is not a list of one or more tree priors.
}
\description{
Will \link{stop} if the object is not a list of one or more tree priors.
}
\examples{
check_empty_beautier_folder()

check_tree_priors(create_yule_tree_prior())
check_tree_priors(list(create_yule_tree_prior()))
check_tree_priors(list(create_yule_tree_prior(), create_bd_tree_prior()))

check_empty_beautier_folder()
}
\seealso{
Use \link{create_tree_prior} to create a valid tree prior
}
\author{
Richèl J.C. Bilderbeek
}
