% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_treelog.R
\name{check_treelog}
\alias{check_treelog}
\title{Check if a \code{treelog} is valid.}
\usage{
check_treelog(treelog)
}
\arguments{
\item{treelog}{a \code{treelog},
as created by \link{create_treelog}}
}
\description{
Will call \link{stop} if not.
}
