% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_treelog.R
\name{check_treelog_names}
\alias{check_treelog_names}
\title{Check if the \code{treelog} has the list elements
of a valid \code{treelog} object.}
\usage{
check_treelog_names(treelog)
}
\arguments{
\item{treelog}{a \code{treelog},
as created by \link{create_treelog}}
}
\value{
nothing
}
\description{
Calls \code{stop} if an element is missing
}
\seealso{
Use \link{create_treelog} to create a valid \code{treelog}
}
\author{
Richèl J.C. Bilderbeek
}
