% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_treelog.R
\name{check_treelog_values}
\alias{check_treelog_values}
\title{Check if the treelog has the list elements with valid values
for being a valid treelog object.}
\usage{
check_treelog_values(treelog)
}
\arguments{
\item{treelog}{a \code{treelog},
as created by \link{create_treelog}}
}
\value{
nothing
}
\description{
Calls \code{stop} if a value is invalid
}
\seealso{
Use \link{create_treelog} to create a valid treelog
}
\author{
Richèl J.C. Bilderbeek
}
