% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_model_to_xml.R
\name{clock_model_to_brm_xml}
\alias{clock_model_to_brm_xml}
\title{Converts a clock model to the \code{<branchRateModel>} section of the
XML as text}
\usage{
clock_model_to_brm_xml(id, clock_model)
}
\arguments{
\item{id}{the ID of the alignment (can be extracted from
its FASTA filesname using \code{\link{get_id}})}

\item{clock_model}{a clock_model,
as created by \code{\link{create_clock_model}}}
}
\description{
Converts a clock model to the \code{<branchRateModel>} section of the
XML as text
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
