% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_model_to_xml.R
\name{clock_model_to_other_brm_xml}
\alias{clock_model_to_other_brm_xml}
\title{Creates the second or later clock models' section in the distribution section
of a BEAST2 XML parameter file}
\usage{
clock_model_to_other_brm_xml(clock_model)
}
\arguments{
\item{clock_model}{a clock_model,
as created by \code{\link{create_clock_model}}}

\item{id}{the ID of the alignment (can be extracted from
its FASTA filesname using \code{\link{get_id}})}
}
\description{
Creates the second or later clock models' section in the distribution section
of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
