% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_model_to_xml.R
\name{clock_model_to_prior_xml}
\alias{clock_model_to_prior_xml}
\title{Converts a clock model to the \code{<prior>} section of the
XML as text}
\usage{
clock_model_to_prior_xml(clock_model)
}
\arguments{
\item{clock_model}{a clock_model,
as created by \code{\link{create_clock_model}}}

\item{id}{the ID of the alignment (can be extracted from
its FASTA filesname using \code{\link{get_id}})}
}
\description{
Converts a clock model to the \code{<prior>} section of the
XML as text
}
\author{
Richel J.C. Bilderbeek
}
