% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_model_to_xml.R
\name{clock_model_to_xml_brm}
\alias{clock_model_to_xml_brm}
\title{Converts a clock model to the \code{branchRateModel} section of the
XML as text}
\usage{
clock_model_to_xml_brm(clock_model)
}
\arguments{
\item{clock_model}{a clock_model,
as created by \code{\link{create_clock_model}}}
}
\description{
Converts a clock model to the \code{branchRateModel} section of the
XML as text
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
