% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_model_to_xml_operators.R
\name{clock_model_to_xml_operators}
\alias{clock_model_to_xml_operators}
\title{Converts a clock model to the \code{operators} section of the
XML as text}
\usage{
clock_model_to_xml_operators(clock_model)
}
\arguments{
\item{clock_model}{a clock model,
as returned by \code{\link{create_clock_model}}}
}
\description{
Converts a clock model to the \code{operators} section of the
XML as text
}
\author{
Richel J.C. Bilderbeek
}
