% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_model_to_xml_treelogger.R
\name{clock_model_to_xml_treelogger}
\alias{clock_model_to_xml_treelogger}
\title{Convert a clock model to the XML of the \code{TreeLogger}}
\usage{
clock_model_to_xml_treelogger(clock_model)
}
\arguments{
\item{clock_model}{a clock model,
as returned by \code{\link{create_clock_model}}}
}
\value{
a character vector of XML strings
}
\description{
Convert a clock model to the XML of the \code{TreeLogger}
}
\note{
This is an internal function, so it should be marked with
  \code{@export}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
