% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_models_to_xml_operators.R
\name{clock_models_to_xml_operators}
\alias{clock_models_to_xml_operators}
\title{Create all clock models' operators' XML text}
\usage{
clock_models_to_xml_operators(...)
}
\value{
a character vector of XML strings
}
\description{
Create all clock models' operators' XML text
}
\author{
Richèl J.C. Bilderbeek
}
