% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_models_to_xml_state_parameters.R
\name{clock_models_to_xml_state_parameters}
\alias{clock_models_to_xml_state_parameters}
\title{Internal function}
\usage{
clock_models_to_xml_state_parameters(inference_model)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\value{
NULL or lines of XML text,
all in the form \code{<parameter id=[...]>}
}
\description{
Creates the XML of the \code{parameter} nodes of the \code{state} section
}
\author{
Richèl J.C. Bilderbeek
}
