% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_models_to_xml_tracelog.R
\name{clock_models_to_xml_tracelog}
\alias{clock_models_to_xml_tracelog}
\title{Creates the clock models' XML for the tracelog section}
\usage{
clock_models_to_xml_tracelog(clock_models, mrca_priors = NA)
}
\arguments{
\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}

\item{mrca_priors}{a list of one or more Most Recent Common Ancestor priors,
as returned by \code{\link{create_mrca_prior}}}
}
\value{
a character vector of XML strings
}
\description{
Creates the clock models' XML for the tracelog section
}
\examples{
# <logger id="tracelog" ...>
#'   # Here
# </logger>
}
\seealso{
the complete tracelog section is created
  by \code{\link{create_tracelog_xml}}
}
\author{
Richèl J.C. Bilderbeek
}
