% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_rate_param_to_xml.R
\name{clock_rate_param_to_xml}
\alias{clock_rate_param_to_xml}
\title{Internal function}
\usage{
clock_rate_param_to_xml(
  clock_rate_param,
  beauti_options = create_beauti_options()
)
}
\arguments{
\item{clock_rate_param}{a \code{clockRate} parameter,
a numeric value,
as created by \link{create_clock_rate_param}}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the parameter as XML text
}
\description{
Converts a \code{clockRate} parameter to XML
}
\author{
Richèl J.C. Bilderbeek
}
