% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_ids.R
\name{collect_ids}
\alias{collect_ids}
\title{Of each element of x, extract the \code{id}}
\usage{
collect_ids(x)
}
\arguments{
\item{one}{or more elements with an \code{id} element}
}
\value{
a vector of IDs
}
\description{
Of each element of x, extract the \code{id}
}
