% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastscriptr.R
\name{convert_fasta_file_to_sequences}
\alias{convert_fasta_file_to_sequences}
\title{Convert a FASTA file to a table of sequences}
\usage{
convert_fasta_file_to_sequences(fasta_filename)
}
\arguments{
\item{fasta_filename}{Name of an existing FASTA file}
}
\value{
a table of sequences
}
\description{
Convert a FASTA file to a table of sequences
}

