% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_phylos.R
\name{count_phylos}
\alias{count_phylos}
\title{Count the number of phylo objects in vector x}
\usage{
count_phylos(x)
}
\arguments{
\item{x}{a vector of phylo objects}
}
\value{
the number of phylo objects
}
\description{
Count the number of phylo objects in vector x
}
\note{
use \code{\link{get_nth_phylo}} to extract the nth phylo object
}
\author{
Richel J.C. Bilderbeek
}
