% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_trailing_spaces.R
\name{count_trailing_spaces}
\alias{count_trailing_spaces}
\title{Count the number of spaces before the first character}
\usage{
count_trailing_spaces(line)
}
\arguments{
\item{line}{line of text}
}
\value{
the number of spaces before the first character
}
\description{
Count the number of spaces before the first character
}
\author{
Richel J.C. Bilderbeek
}
