% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_alpha_param}
\alias{create_alpha_param}
\title{Create a parameter called alpha}
\usage{
create_alpha_param(id = NA, estimate = FALSE, value = 0)
}
\arguments{
\item{id}{the parameter's ID}

\item{estimate}{TRUE if this parameter alpha be estimated by BEAST2,
FALSE otherwise}

\item{value}{value of the parameter}
}
\value{
a parameter called alpha
}
\description{
Create a parameter called alpha
}
\seealso{
the function \code{\link{create_param}} contains a list
  of all parameters that can be created
}
\author{
Richel J.C. Bilderbeek
}
