% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_b_pop_sizes_param}
\alias{create_b_pop_sizes_param}
\alias{create_param_b_pop_sizes}
\title{Create a parameter called `b_pop_sizes`.}
\usage{
create_b_pop_sizes_param(id = NA, value = 1, upper = "380000.0")
}
\arguments{
\item{id}{the parameter's ID}

\item{value}{value of the parameter}

\item{upper}{upper value of the parameter}
}
\value{
a parameter called b_pop_sizes
}
\description{
Create a parameter called `b_pop_sizes`.
}
\note{
this parameter is used in a CBS model,
as created by \link{create_cbs_tree_prior}
}
\examples{
# Create the parameter
b_pop_sizes_param <- create_b_pop_sizes_param()
}
\seealso{
the function \code{\link{create_param}} contains a list
  of all parameters that can be created
}
\author{
Richèl J.C. Bilderbeek
}
