% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_beast_xml.R
\name{create_beast2_beast_xml}
\alias{create_beast2_beast_xml}
\title{Create the \code{<beast ...>} XML}
\usage{
create_beast2_beast_xml(beast2_version)
}
\description{
Create the \code{<beast ...>} XML
}
