% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_1_12.R
\name{create_beast2_input_1_12}
\alias{create_beast2_input_1_12}
\title{Create a BEAST2 XML input text, interface of v1.12}
\usage{
create_beast2_input_1_12(input_filenames,
  site_models = create_jc69_site_models(ids = get_ids(input_filenames)),
  clock_models = create_strict_clock_models(ids = get_ids(input_filenames)),
  tree_priors = create_yule_tree_priors(ids = get_ids(input_filenames)),
  mcmc = create_mcmc(), misc_options = create_misc_options(),
  fixed_crown_ages = rep(FALSE, times = length(input_filenames)),
  initial_phylogenies = rep(NA, length(input_filenames)))
}
\description{
Create a BEAST2 XML input text, interface of v1.12
}
\examples{
  create_beast2_input_file_1_12(
    input_filenames = get_fasta_filename(),
    "my_beast.xml"
  )
}
\seealso{
\code{\link{create_beast2_input_file}} shows more examples
}
\author{
Richel J.C. Bilderbeek
}
