% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_beast.R
\name{create_beast2_input_beast}
\alias{create_beast2_input_beast}
\title{Creates the beast section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_beast(input_fasta_filenames,
  site_models = create_site_model(name = "JC69"),
  clock_models = create_clock_model(name = "strict"),
  tree_priors = create_tree_prior(name = "yule"), mcmc_chainlength,
  fixed_crown_age, initial_phylogeny)
}
\arguments{
\item{input_fasta_filenames}{one ore more FASTA filenames}

\item{site_models}{one or more site models,
as returned by 'create_site_model'}

\item{clock_models}{On or more clock models,
as returned by 'create_clock_model'}

\item{tree_priors}{one ore more tree priors,
as returned from 'create_tree_prior'}

\item{mcmc_chainlength}{MCMC chain length}

\item{fixed_crown_age}{is the crown age fixed TRUE or FALSE}

\item{initial_phylogeny}{initial phylogeny or NA}
}
\description{
Creates the beast section of a BEAST2 XML parameter file
}
