% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_data.R
\name{create_beast2_input_data}
\alias{create_beast2_input_data}
\title{Creates the data section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_data(input_fasta_filenames,
  capitalize_first_char_id = FALSE, nucleotides_uppercase = FALSE)
}
\arguments{
\item{input_fasta_filenames}{One or more fasta filenames}

\item{capitalize_first_char_id}{capitalize the first character of the ID}

\item{nucleotides_uppercase}{are the nucleotides in uppercase?}
}
\description{
Creates the data section of a BEAST2 XML parameter file
}
\author{
Richel J.C. Bilderbeek
}
