% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_data_sequences.R
\name{create_beast2_input_data_sequences}
\alias{create_beast2_input_data_sequences}
\title{Creates the data section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_data_sequences(input_fasta_filename,
  nucleotides_uppercase = FALSE)
}
\arguments{
\item{input_fasta_filename}{one FASTA filename}

\item{nucleotides_uppercase}{are the nucleotides written in uppercase?
Yes if TRUE, no if FALSE}
}
\description{
Creates the data section of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
