% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distr.R
\name{create_beast2_input_distr_clock_models}
\alias{create_beast2_input_distr_clock_models}
\title{Creates the clock models section in the distribution section
of a BEAST2 XML parameter file}
\usage{
create_beast2_input_distr_clock_models(clock_model)
}
\arguments{
\item{clock_model}{a clock_model,
as created by \code{\link{create_clock_model}}}
}
\description{
Creates the clock models section in the distribution section
of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
