% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distr.R
\name{create_beast2_input_distr_prior_distr}
\alias{create_beast2_input_distr_prior_distr}
\title{Creates the distribution section in the prior section of the
distribution section of a BEAST2 XML parameter file.
These lines start with '<distribution id='}
\usage{
create_beast2_input_distr_prior_distr(tree_priors)
}
\arguments{
\item{tree_priors}{one or more tree priors,
as returned by \code{\link{create_tree_prior}}}
}
\description{
Creates the distribution section in the prior section of the
distribution section of a BEAST2 XML parameter file.
These lines start with '<distribution id='
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
