% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distribution.R
\name{create_beast2_input_distr_prior_prior_tree_prior_ccp_pop_size}
\alias{create_beast2_input_distr_prior_prior_tree_prior_ccp_pop_size}
\title{Creates the tree prior section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file
for a Coalescent Constant Population tree prior}
\usage{
create_beast2_input_distr_prior_prior_tree_prior_ccp_pop_size(
  ccp_pop_size_distr,
  id
)
}
\arguments{
\item{ccp_pop_size_distr}{a Coalescent Constant Population
population size distribution,
as created by \code{\link{create_distr}}}

\item{id}{the ID of the alignment (can be extracted from
its FASTA filesname using \code{\link{get_id}})}
}
\description{
Creates the tree prior section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file
for a Coalescent Constant Population tree prior
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
