% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distr.R
\name{create_beast2_input_distr_prior_prior_tree_prior_cep_growth_rate}
\alias{create_beast2_input_distr_prior_prior_tree_prior_cep_growth_rate}
\title{Creates the tree prior section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file
for a Coalescent Exponential Population tree prior}
\usage{
create_beast2_input_distr_prior_prior_tree_prior_cep_growth_rate(
  cep_growth_rate_distr,
  id
)
}
\arguments{
\item{cep_growth_rate_distr}{a Coalescent Exponential Population
growth rate distribution,
as created by \code{\link{create_distr}}}

\item{id}{the ID of the alignment (can be extracted from
its FASTA filesname using \code{\link{get_id}})}
}
\description{
Creates the tree prior section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file
for a Coalescent Exponential Population tree prior
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
