% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distribution.R
\name{create_beast2_input_distribution}
\alias{create_beast2_input_distribution}
\title{Creates the distribution section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_distribution(ids, site_models = create_site_model(name =
  "JC69"), clock_models = create_clock_model(name = "strict"),
  tree_priors = create_tree_prior(name = "yule"))
}
\arguments{
\item{ids}{the IDs of the alignments (can be extracted from
their FASTA filesnames using 'get_file_base_sans_ext')}

\item{site_models}{one or more site models,
as returned by 'create_site_model'}

\item{clock_models}{On or more clock models,
as returned by 'create_clock_model'}

\item{tree_priors}{one or more tree priors}
}
\description{
Creates the distribution section of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
