% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distribution.R
\name{create_beast2_input_distribution_distribution}
\alias{create_beast2_input_distribution_distribution}
\title{Creates the distribution section in the distribution section
of a BEAST2 XML parameter file}
\usage{
create_beast2_input_distribution_distribution(ids, tree_priors)
}
\arguments{
\item{ids}{the IDs of the alignments (can be extracted from
their FASTA filesnames using 'get_file_base_sans_ext')}

\item{tree_priors}{one or more tree priors}
}
\description{
Creates the distribution section in the distribution section
of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
