% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distribution.R
\name{create_beast2_input_distribution_gamma_site_models}
\alias{create_beast2_input_distribution_gamma_site_models}
\title{Creates the gamma site models section in the distribution section
of a BEAST2 XML parameter file}
\usage{
create_beast2_input_distribution_gamma_site_models(id, site_model)
}
\description{
Creates the gamma site models section in the distribution section
of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
