% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distribution.R
\name{create_beast2_input_distribution_prior_distribution}
\alias{create_beast2_input_distribution_prior_distribution}
\title{Creates the distribution section in the prior section of the
distribution section of a BEAST2 XML parameter file.
These lines start with '<distribution id='}
\usage{
create_beast2_input_distribution_prior_distribution(ids,
  tree_priors = create_yule_tree_priors(n = length(ids)))
}
\arguments{
\item{ids}{the IDs of the alignments (can be extracted from
their FASTA filesnames using \code{\link{get_id}})}

\item{tree_priors}{one or more tree priors,
as returned by \code{\link{create_tree_prior}}}
}
\description{
Creates the distribution section in the prior section of the
distribution section of a BEAST2 XML parameter file.
These lines start with '<distribution id='
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
