% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distribution.R
\name{create_beast2_input_distribution_prior_prior_tree_prior}
\alias{create_beast2_input_distribution_prior_prior_tree_prior}
\title{Creates the tree prior section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file}
\usage{
create_beast2_input_distribution_prior_prior_tree_prior(id, tree_prior)
}
\arguments{
\item{id}{the ID of the alignment (can be extracted from
its FASTA filesname using \code{\link{get_id}})}

\item{tree_prior}{a tree_prior, as created by \code{\link{create_tree_prior}}}
}
\description{
Creates the tree prior section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
