% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distribution.R
\name{create_beast2_input_distribution_prior_prior_tree_prior_bd_birth_rate}
\alias{create_beast2_input_distribution_prior_prior_tree_prior_bd_birth_rate}
\title{Creates the tree prior section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file
for a Birth-Death tree prior}
\usage{
create_beast2_input_distribution_prior_prior_tree_prior_bd_birth_rate(
  bd_birth_rate_distr,
  id
)
}
\arguments{
\item{bd_birth_rate_distr}{a Birth-Death birth rate distribution,
as created by \code{\link{create_distribution}}}

\item{id}{the ID of the alignment (can be extracted from
its FASTA filesname using \code{\link{get_id}})}
}
\description{
Creates the tree prior section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file
for a Birth-Death tree prior
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
