% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distribution.R
\name{create_beast2_input_distribution_prior_prior_tree_prior_ccp}
\alias{create_beast2_input_distribution_prior_prior_tree_prior_ccp}
\title{Creates the tree prior section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file for a
Coalescent Constant Population tree prior}
\usage{
create_beast2_input_distribution_prior_prior_tree_prior_ccp(id, ccp_tree_prior)
}
\arguments{
\item{id}{the ID of the alignment (can be extracted from
its FASTA filesname using \code{\link{get_id}})}

\item{ccp_tree_prior}{a Coalescent Constant Population tree prior,
as created by \code{\link{create_ccp_tree_prior}}}
}
\description{
Creates the tree prior section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file for a
Coalescent Constant Population tree prior
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
