% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distribution.R
\name{create_beast2_input_distribution_subst_model}
\alias{create_beast2_input_distribution_subst_model}
\title{Creates the substModel section in the distribution section
of a BEAST2 XML parameter file}
\usage{
create_beast2_input_distribution_subst_model(id, site_model)
}
\arguments{
\item{id}{alignment ID}

\item{site_model}{a site_model, as created by \code{\link{create_site_model}}}
}
\description{
Creates the substModel section in the distribution section
of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
