% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_file.R
\name{create_beast2_input_file_data}
\alias{create_beast2_input_file_data}
\title{Creates the data section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_file_data(filename_base, input_fasta_filenames)
}
\arguments{
\item{filename_base}{base of the filenames}

\item{input_fasta_filenames}{name of the FASTA file}
}
\description{
Creates the data section of a BEAST2 XML parameter file
}
