% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_file.R
\name{create_beast2_input_file_distribution}
\alias{create_beast2_input_file_distribution}
\title{Creates the distribution section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_file_distribution(filename_base, tree_priors)
}
\arguments{
\item{filename_base}{base of the filenames}

\item{tree_priors}{one or more tree priors}
}
\description{
Creates the distribution section of a BEAST2 XML parameter file
}
