% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_file.R
\name{create_beast2_input_file_init}
\alias{create_beast2_input_file_init}
\title{Creates the map section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_file_init(filename_base, initial_phylogeny)
}
\arguments{
\item{filename_base}{base of the filenames}

\item{initial_phylogeny}{initial phylogeny}
}
\description{
Creates the map section of a BEAST2 XML parameter file
}
