% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_file.R
\name{create_beast2_input_file_operators}
\alias{create_beast2_input_file_operators}
\title{Creates the operators section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_file_operators(filename_base, tree_priors, fixed_crown_age)
}
\arguments{
\item{filename_base}{base of the filenames}

\item{fixed_crown_age}{determines if the phylogeny its crown age is
fixed. If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}

\item{tree_prior}{The tree prior, can be 'birth_death' or
'coalescent_constant_population'}
}
\description{
Creates the operators section of a BEAST2 XML parameter file
}
