% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_file.R
\name{create_beast2_input_file_run}
\alias{create_beast2_input_file_run}
\title{Creates the state section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_file_run(filename_base, mcmc_chainlength, tree_priors,
  fixed_crown_age, initial_phylogeny)
}
\arguments{
\item{filename_base}{filename its base}

\item{mcmc_chainlength}{MCMC chain length}

\item{tree_priors}{one or more tree priors,
as returned by 'create_tree_prior'}

\item{fixed_crown_age}{is the crown age fixed TRUE or FALSE}

\item{initial_phylogeny}{initial phylogeny or NA}
}
\description{
Creates the state section of a BEAST2 XML parameter file
}
