% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_file.R
\name{create_beast2_input_file_xml}
\alias{create_beast2_input_file_xml}
\title{Creates the xml section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_file_xml()
}
\description{
Creates the xml section of a BEAST2 XML parameter file
}
