% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_from_model.R
\name{create_beast2_input_from_model}
\alias{create_beast2_input_from_model}
\title{Create a BEAST2 XML input text from an inference model}
\usage{
create_beast2_input_from_model()
}
\value{
a character vector of XML strings
}
\description{
Create a BEAST2 XML input text from an inference model
}
\examples{
  input_filename = get_fasta_filename()
  text <- create_beast2_input_from_model(
  )
  testit::assert(substr(text[1], 1, 5) == "<?xml")
  text[1]
  testit::assert(tail(text, n = 1) == "</beast>")
}
\seealso{
Use \link{create_beast2_input_file_from_model} to also save it to file.

\code{\link{create_beast2_input_file}} shows more examples
}
\author{
Richèl J.C. Bilderbeek
}
