% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_init.R
\name{create_beast2_input_init}
\alias{create_beast2_input_init}
\title{Creates the map section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_init(ids, initial_phylogeny)
}
\arguments{
\item{ids}{the IDs of the alignments (can be extracted from
their FASTA filesnames using 'get_file_base_sans_ext')}

\item{initial_phylogeny}{initial phylogeny}
}
\description{
Creates the map section of a BEAST2 XML parameter file
}
