% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_loggers.R
\name{create_beast2_input_loggers_clock_models}
\alias{create_beast2_input_loggers_clock_models}
\title{Creates the clock models part of the two logger sections
  of a BEAST2 XML parameter file}
\usage{
create_beast2_input_loggers_clock_models(clock_model)
}
\arguments{
\item{clock_model}{a clock model,
as returned by \code{\link{create_clock_model}}}
}
\description{
Creates the clock models part of the two logger sections
  of a BEAST2 XML parameter file
}
\author{
Richel J.C. Bilderbeek
}
