% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_loggers.R
\name{create_beast2_input_loggers_freq_parameter}
\alias{create_beast2_input_loggers_freq_parameter}
\title{Creates the freqParameter part of the log sections
  of a BEAST2 XML parameter file}
\usage{
create_beast2_input_loggers_freq_parameter(id, site_model)
}
\arguments{
\item{id}{the id of the alignments (can be extracted from
their FASTA filesnames using \code{\link{get_file_base_sans_ext}})}
}
\description{
Creates the freqParameter part of the log sections
  of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
