% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_loggers.R
\name{create_beast2_input_loggers_rates}
\alias{create_beast2_input_loggers_rates}
\title{Creates the first site models part of the two logger sections
  of a BEAST2 XML parameter file}
\usage{
create_beast2_input_loggers_rates(site_model)
}
\arguments{
\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}
}
\description{
Creates the first site models part of the two logger sections
  of a BEAST2 XML parameter file
}
\author{
Richel J.C. Bilderbeek
}
