% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_loggers.R
\name{create_beast2_input_loggers_tree_prior}
\alias{create_beast2_input_loggers_tree_prior}
\title{Creates the tree priors part of the two logger sections
  of a BEAST2 XML parameter file}
\usage{
create_beast2_input_loggers_tree_prior(tree_prior)
}
\arguments{
\item{tree_prior}{a tree prior,
as created by \code{\link{create_tree_prior}}}
}
\description{
Creates the tree priors part of the two logger sections
  of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
