% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_loggers.R
\name{create_beast2_input_loggers_tree_priors}
\alias{create_beast2_input_loggers_tree_priors}
\title{Creates the tree priors part of the two logger sections
  of a BEAST2 XML parameter file}
\usage{
create_beast2_input_loggers_tree_priors(id, tree_priors)
}
\arguments{
\item{id}{the id of the alignments (can be extracted from
their FASTA filesnames using \code{\link{get_file_base_sans_ext}})}

\item{tree_priors}{On or more tree priors,
as returned by \code{\link{create_tree_prior}}}
}
\description{
Creates the tree priors part of the two logger sections
  of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
