% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_map.R
\name{create_beast2_input_map}
\alias{create_beast2_input_map}
\title{Creates the map section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_map(beauti_options)
}
\arguments{
\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
lines of XML text
}
\description{
Creates the map section of a BEAST2 XML parameter file
}
\author{
Richèl J.C. Bilderbeek
}
