% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_operators.R
\name{create_beast2_input_operators_clock_models}
\alias{create_beast2_input_operators_clock_models}
\title{Creates the clock_models section in the operators section
of a BEAST2 XML parameter file}
\usage{
create_beast2_input_operators_clock_models(id,
  clock_model = create_strict_clock_model())
}
\arguments{
\item{id}{the id of the alignments (can be extracted from
their FASTA filesnames using \code{\link{get_file_base_sans_ext}})}
}
\description{
Creates the clock_models section in the operators section
of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
