% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_operators.R
\name{create_beast2_input_operators_rates}
\alias{create_beast2_input_operators_rates}
\title{Creates the site_models section in the operators section
of a BEAST2 XML parameter file}
\usage{
create_beast2_input_operators_rates(id, site_model = create_site_model(name =
  "JC69"))
}
\description{
Creates the site_models section in the operators section
of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
