% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_operators.R
\name{create_beast2_input_operators_tree_priors}
\alias{create_beast2_input_operators_tree_priors}
\title{Creates the first tree_priors section in the operators section
of a BEAST2 XML parameter file}
\usage{
create_beast2_input_operators_tree_priors(tree_prior, fixed_crown_age)
}
\arguments{
\item{tree_prior}{tree prior, as created by \code{\link{create_tree_prior}}}

\item{fixed_crown_age}{determines if the phylogeny its crown age is
fixed. If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}
}
\description{
Creates the first tree_priors section in the operators section
of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
