% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_operators.R
\name{create_beast2_input_operators_tree_priors_3}
\alias{create_beast2_input_operators_tree_priors_3}
\title{Creates the third tree_priors section in the operators section
of a BEAST2 XML parameter file}
\usage{
create_beast2_input_operators_tree_priors_3(id,
  tree_prior = create_tree_prior(name = "yule"))
}
\arguments{
\item{id}{the ID of the alignment (can be extracted from
its FASTA filesname using \code{\link{get_id}})}

\item{tree_prior}{tree prior, as created by \code{\link{create_tree_prior}}}
}
\description{
Creates the third tree_priors section in the operators section
of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
