% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_run.R
\name{create_beast2_input_run}
\alias{create_beast2_input_run}
\title{Creates the state section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_run(ids, site_models = create_site_model(name = "JC69"),
  mcmc_chainlength, tree_priors = create_tree_prior(name = "yule"),
  fixed_crown_age, initial_phylogeny)
}
\arguments{
\item{ids}{the IDs of the alignments (can be extracted from
their FASTA filesnames using 'get_file_base_sans_ext')}

\item{site_models}{one or more site models, as returned
by 'create_site_model'}

\item{mcmc_chainlength}{MCMC chain length}

\item{tree_priors}{one or more tree priors,
as returned by 'create_tree_prior'}

\item{fixed_crown_age}{is the crown age fixed TRUE or FALSE}

\item{initial_phylogeny}{initial phylogeny or NA}
}
\description{
Creates the state section of a BEAST2 XML parameter file
}
