% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_state.R
\name{create_beast2_input_state}
\alias{create_beast2_input_state}
\title{Creates the state section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_state(ids, site_models = create_site_model(name = "JC69"),
  clock_models = create_clock_model(name = "strict"),
  tree_priors = create_tree_prior(name = "yule"), initial_phylogeny)
}
\arguments{
\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}

\item{clock_models}{On or more clock models,
as returned by \code{\link{create_clock_model}}}

\item{tree_priors}{On or more tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{initial_phylogeny}{the MCMC chain its initial phylogeny. If
this is set to NA, BEAST2 will use a random phylogeny. Else
a phylogeny must be supplied of class ape::phylo.}
}
\description{
Creates the state section of a BEAST2 XML parameter file
}
\author{
Richel J.C. Bilderbeek
}
