% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_state.R
\name{create_beast2_input_state_gamma_site_models_gamma_shape}
\alias{create_beast2_input_state_gamma_site_models_gamma_shape}
\title{Creates the gamma_site_models part of the state section of a BEAST2
XML parameter file}
\usage{
create_beast2_input_state_gamma_site_models_gamma_shape(id, site_model)
}
\arguments{
\item{ids}{the IDs of the alignments (can be extracted from
their FASTA filesnames using \code{\link{get_file_base_sans_ext}})}
}
\description{
Creates the gamma_site_models part of the state section of a BEAST2
XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
