% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_state.R
\name{create_beast2_input_state_tree}
\alias{create_beast2_input_state_tree}
\title{Creates the tree part of the state section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_state_tree(ids, initial_phylogeny)
}
\arguments{
\item{ids}{the IDs of the alignments (can be extracted from
their FASTA filesnames using \code{\link{get_file_base_sans_ext}})}

\item{initial_phylogeny}{the MCMC chain its initial phylogeny. If
this is set to NA, BEAST2 will use a random phylogeny. Else
a phylogeny must be supplied of class ape::phylo.}
}
\description{
Creates the tree part of the state section of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
