% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_state.R
\name{create_beast2_input_state_tree_prior}
\alias{create_beast2_input_state_tree_prior}
\title{Creates the tree priors part of the state section of a BEAST2
XML parameter file}
\usage{
create_beast2_input_state_tree_prior(id, tree_prior)
}
\arguments{
\item{id}{the ID of the alignments (can be extracted from
their FASTA filesnames using \code{\link{get_file_base_sans_ext}})}
}
\description{
Creates the tree priors part of the state section of a BEAST2
XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
