% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_loggers.R
\name{create_beast2_input_treelogs}
\alias{create_beast2_input_treelogs}
\title{Creates the tracelog section of the logger section
of a BEAST2 XML parameter file}
\usage{
create_beast2_input_treelogs(ids,
  clock_models = create_strict_clock_models(ids = ids))
}
\arguments{
\item{ids}{one or more alignments' IDs.
IDs can be extracted from their FASTA filesnames
with \code{\link{get_ids}})}

\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}
}
\description{
Creates the tracelog section of the logger section
of a BEAST2 XML parameter file
}
\author{
Richel J.C. Bilderbeek
}
