% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beauti_options_v2_4.R
\name{create_beauti_options_v2_4}
\alias{create_beauti_options_v2_4}
\title{Function to create the BEAUti options for version 2.4.}
\usage{
create_beauti_options_v2_4()
}
\value{
a BEAUti options structure
}
\description{
Function to create the BEAUti options for version 2.4, by
calling \link{create_beauti_options}.
}
\examples{
beauti_options <- create_beauti_options_v2_4()
xml <- create_beast2_input(
  get_fasta_filename(),
  beauti_options = beauti_options
)

library(testthat)
expect_true(is.character(xml))
expect_true(length(xml) > 1)
}
\author{
Richèl J.C. Bilderbeek
}
