% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_beta_distr}
\alias{create_beta_distr}
\title{Create a beta distribution}
\usage{
create_beta_distr(id = NA, alpha = create_alpha_parameter(id = NA, estimate
  = NA, value = NA), beta = create_beta_parameter(id = NA, estimate = NA,
  value = NA))
}
\arguments{
\item{id}{the distribution's ID}

\item{alpha}{the alpha shape parameter}

\item{beta}{the beta shape parameter}
}
\value{
a beta distribution
}
\description{
Create a beta distribution
}
\seealso{
the function \code{\link{create_distribution}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
