% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_branch_rate_model_stuff_xml.R
\name{create_branch_rate_model_stuff_xml}
\alias{create_branch_rate_model_stuff_xml}
\title{Internal function called by \link{create_branch_rate_model_xml}}
\usage{
create_branch_rate_model_stuff_xml(inference_model)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\value{
lines of XML text
}
\description{
It generates the desired XML for some circumstances.
Yes, that is a vague description.
Would be nice if someone would untangle this :-)
}
\author{
Richèl J.C. Bilderbeek
}
