% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_branch_rate_model_xml.R
\name{create_branch_rate_model_xml}
\alias{create_branch_rate_model_xml}
\title{Converts a clock model to the \code{branchRateModel} section of the
XML as text.}
\usage{
create_branch_rate_model_xml(inference_model)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\value{
a character vector of XML strings
}
\description{
This function will be called only if there are no MRCA priors.
}
\examples{
 # <distribution id="posterior" spec="util.CompoundDistribution">
 #     <distribution id="prior" spec="util.CompoundDistribution">
 #     </distribution>
 #     <distribution id="likelihood" ...>
 #       HERE, where the ID of the distribution is 'likelihood'
 #     </distribution>
 # </distribution>
}
\author{
Richèl J.C. Bilderbeek
}
