% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clock_model.R
\name{create_clock_model_strict}
\alias{create_clock_model_strict}
\title{Alternative name for \code{\link{create_strict_clock_model}}, to help
the user find the function from a search tree. See
\code{\link{create_strict_clock_model}} for examples.}
\usage{
create_clock_model_strict(id = NA,
  clock_rate_param = create_clock_rate_param(),
  clock_rate_distr = create_uniform_distr())
}
\arguments{
\item{id}{an alignment's IDs.
An ID can be extracted from its FASTA filesname
with \code{\link{get_ids}})}

\item{clock_rate_param}{the clock rate's parameter,
as created by the \code{\link{create_clock_rate_param}} function}

\item{clock_rate_distr}{the clock rate's distribution,
as created by a \code{\link{create_distr}} function}
}
\value{
a strict clock_model
}
\description{
Alternative name for \code{\link{create_strict_clock_model}}, to help
the user find the function from a search tree. See
\code{\link{create_strict_clock_model}} for examples.
}
