% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clock_models.R
\name{create_clock_models}
\alias{create_clock_models}
\title{Creates all supported clock models}
\usage{
create_clock_models()
}
\value{
a list of clock_models
}
\description{
Creates all supported clock models
}
