% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_xml.R
\name{create_data_xml}
\alias{create_data_xml}
\title{Create the \code{<data ..>} XML}
\usage{
create_data_xml(id, beast2_version)
}
\value{
lines of XML text
}
\description{
Create the \code{<data ..>} XML
}
\author{
Richèl J.C. Bilderbeek
}
