% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_distr}
\alias{create_distr}
\title{General function to create a distribution.}
\usage{
create_distr(name, id, ...)
}
\arguments{
\item{name}{the distribution name. Valid
names can be found in \code{\link{get_distribution_names}}}

\item{id}{the distribution's ID}

\item{...}{specific distribution parameters}
}
\value{
a distribution
}
\description{
General function to create a distribution.
}
\note{
Prefer using the
  named functions
  \code{\link{create_uniform_distr}},
  \code{\link{create_normal_distr}},
  \code{\link{create_log_normal_distr}},
  \code{\link{create_one_div_x_distr}},
  \code{\link{create_exponential_distr}},
  \code{\link{create_gamma_distr}},
  \code{\link{create_beta_distr}},
  \code{\link{create_laplace_distr}},
  \code{\link{create_inv_gamma_distr}},
  and \code{\link{create_poisson_distr}}
}
\seealso{
use \code{\link{is_distribution}} to check if a
  distribution is valid
}
\author{
Richel J.C. Bilderbeek
}
