% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_distr_beta}
\alias{create_distr_beta}
\title{Alternative name for \code{\link{create_beta_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_beta_distr}} for examples.}
\usage{
create_distr_beta(id = NA, alpha = create_alpha_param(),
  beta = create_beta_param())
}
\arguments{
\item{id}{the distribution's ID}

\item{alpha}{the alpha shape parameter,
as returned by \code{\link{create_alpha_param}}. The value
of alpha must be at least 0.0.}

\item{beta}{the beta shape parameter,
as returned by \code{\link{create_beta_param}}. The value
of beta must be at least 1.0.}
}
\value{
a beta distribution
}
\description{
Alternative name for \code{\link{create_beta_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_beta_distr}} for examples.
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
