% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_distr_exp}
\alias{create_distr_exp}
\title{Alternative name for \code{\link{create_exp_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_exp_distr}} for examples.}
\usage{
create_distr_exp(id = NA, mean = create_mean_param())
}
\arguments{
\item{id}{the distribution's ID}

\item{mean}{the mean parameter,
as returned by \code{\link{create_mean_param}}}
}
\value{
an exponential distribution
}
\description{
Alternative name for \code{\link{create_exp_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_exp_distr}} for examples.
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
