% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_distr_gamma}
\alias{create_distr_gamma}
\title{Alternative name for \code{\link{create_gamma_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_gamma_distr}} for examples.}
\usage{
create_distr_gamma(id = NA, alpha = create_alpha_param(id = NA, estimate =
  FALSE, value = "0.5396"), beta = create_beta_param(id = NA, estimate =
  FALSE, value = "0.3819"))
}
\arguments{
\item{id}{the distribution's ID}

\item{alpha}{the alpha shape parameter,
as returned by \code{\link{create_alpha_param}}}

\item{beta}{the beta shape parameter,
as returned by \code{\link{create_beta_param}}}
}
\value{
a gamma distribution
}
\description{
Alternative name for \code{\link{create_gamma_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_gamma_distr}} for examples.
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
