% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_distr_laplace}
\alias{create_distr_laplace}
\title{Alternative name for \code{\link{create_laplace_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_laplace_distr}} for examples.}
\usage{
create_distr_laplace(id = NA, mu = create_mu_param(id = NA, estimate =
  FALSE, value = 0), scale = create_scale_param(id = NA, estimate = FALSE,
  value = 1))
}
\arguments{
\item{id}{the distribution's ID}

\item{mu}{the mu parameter,
as returned by \code{\link{create_mu_param}}}

\item{scale}{the scale parameter,
as returned by \code{\link{create_scale_param}}}
}
\value{
a Laplace distribution
}
\description{
Alternative name for \code{\link{create_laplace_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_laplace_distr}} for examples.
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
