% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_distr_log_normal}
\alias{create_distr_log_normal}
\title{Alternative name for \code{\link{create_log_normal_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_log_normal_distr}} for examples.}
\usage{
create_distr_log_normal(id = NA, m = create_m_param(),
  s = create_s_param())
}
\arguments{
\item{id}{the distribution's ID}

\item{m}{the m parameter,
as returned by \code{\link{create_m_param}}}

\item{s}{the s parameter,
as returned by \code{\link{create_s_param}}}
}
\value{
a log-normal distribution
}
\description{
Alternative name for \code{\link{create_log_normal_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_log_normal_distr}} for examples.
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
