% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_distr_normal}
\alias{create_distr_normal}
\title{Alternative name for \code{\link{create_normal_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_normal_distr}} for examples.}
\usage{
create_distr_normal(id = NA, mean = create_mean_param(),
  sigma = create_sigma_param())
}
\arguments{
\item{id}{the distribution's ID}

\item{mean}{the mean parameter
as returned by \code{\link{create_mean_param}}}

\item{sigma}{the sigma parameter
as returned by \code{\link{create_sigma_param}}}
}
\value{
a normal distribution
}
\description{
Alternative name for \code{\link{create_normal_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_normal_distr}} for examples.
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
