% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_distr_one_div_x}
\alias{create_distr_one_div_x}
\title{Alternative name for \code{\link{create_one_div_x_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_one_div_x_distr}} for examples.}
\usage{
create_distr_one_div_x(id = NA)
}
\arguments{
\item{id}{the distribution's ID}
}
\value{
a 1/x distribution
}
\description{
Alternative name for \code{\link{create_one_div_x_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_one_div_x_distr}} for examples.
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
