% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_distr_poisson}
\alias{create_distr_poisson}
\title{Alternative name for \code{\link{create_poisson_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_poisson_distr}} for examples.}
\usage{
create_distr_poisson(id = NA, lambda = create_lambda_param())
}
\arguments{
\item{id}{the distribution's ID}

\item{lambda}{the lambda parameter
as returned by \code{\link{create_lambda_param}}}
}
\value{
a Poisson distribution
}
\description{
Alternative name for \code{\link{create_poisson_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_poisson_distr}} for examples.
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
