% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_distr_uniform}
\alias{create_distr_uniform}
\title{Alternative name for \code{\link{create_uniform_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_uniform_distr}} for examples.}
\usage{
create_distr_uniform(id = NA, upper = Inf)
}
\arguments{
\item{id}{the distribution's ID}

\item{upper}{an upper limit of the uniform distribution.
If the upper limits needs to be infinity, set \code{upper} to \code{Inf}.}
}
\value{
a uniform distribution
}
\description{
Alternative name for \code{\link{create_uniform_distr}}, to help
the user find the function from a search tree. See
\code{\link{create_uniform_distr}} for examples.
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
