% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_distribution}
\alias{create_distribution}
\title{General function to create a distribution.}
\usage{
create_distribution(name, ...)
}
\arguments{
\item{name}{the distribution name. Valid
names can be found in \code{\link{get_distribution_names}}}

\item{...}{specific distribution parameters}
}
\value{
a distribution
}
\description{
General function to create a distribution.
}
\note{
Prefer using the
  named functions \code{\link{create_uniform_distribution}},
  and \code{\link{create_normal_distribution}}
}
\author{
Richel J.C. Bilderbeek
}
