% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_exponential_distr}
\alias{create_exponential_distr}
\title{Create an exponential distribution}
\usage{
create_exponential_distr(id = NA, mean = create_meanparam(id = NA, estimate
  = NA, value = NA))
}
\arguments{
\item{id}{the distribution's ID}

\item{mean}{the mean parameter,
as returned by \code{\link{create_meanparam}}}
}
\value{
an exponential distribution
}
\description{
Create an exponential distribution
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
