% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_exponential_distribution}
\alias{create_exponential_distribution}
\title{Create an exponential distribution}
\usage{
create_exponential_distribution()
}
\value{
an exponential distribution
}
\description{
Create an exponential distribution
}
