% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_freq_param}
\alias{create_freq_param}
\alias{create_param_freq}
\title{Create a parameter called freq}
\usage{
create_freq_param(
  id = NA,
  lower = "0.0",
  upper = "1.0",
  value = "0.25",
  estimate = TRUE,
  dimension = 4
)
}
\arguments{
\item{id}{the parameter's ID}

\item{lower}{lowest possible value of the parameter. If the parameter
is estimated, \code{lower} must be less than \code{value}}

\item{upper}{upper value of the parameter}

\item{value}{value of the parameter}

\item{estimate}{TRUE if this parameter is to be estimated by BEAST2,
FALSE otherwise}

\item{dimension}{the number of dimensions, for example, as used in
\link{create_freq_param}}
}
\value{
a parameter called freq
}
\description{
Create a parameter called freq
}
\author{
Richèl J.C. Bilderbeek
}
