% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_gamma_distribution}
\alias{create_gamma_distribution}
\title{Create a gamma distribution}
\usage{
create_gamma_distribution(id = NA)
}
\arguments{
\item{id}{the distribution's ID}
}
\value{
a gamma distribution
}
\description{
Create a gamma distribution
}
\author{
Richel J.C. Bilderbeek
}
