% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subst_model_xml.R
\name{create_hky_subst_model_xml}
\alias{create_hky_subst_model_xml}
\title{Converts a site model to XML,
  used in the \code{substModel} section}
\usage{
create_hky_subst_model_xml(site_model)
}
\arguments{
\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}
}
\value{
the site model as XML text
}
\description{
Converts a site model to XML,
  used in the \code{substModel} section
}
\author{
Richèl J.C. Bilderbeek
}
