% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inference_model.R
\name{create_inference_model}
\alias{create_inference_model}
\title{Create an MCMC inference model.}
\usage{
create_inference_model(site_model = create_jc69_site_model(),
  clock_model = create_strict_clock_model(),
  tree_prior = create_bd_tree_prior(), mrca_prior = NA,
  mcmc = create_mcmc(), misc_options = create_misc_options(),
  tipdates_filename = NA)
}
\arguments{
\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}

\item{clock_model}{a clock model,
as returned by \code{\link{create_clock_model}}}

\item{tree_prior}{a tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{mrca_prior}{a Most Recent Common Ancestor prior,
as returned by \code{\link{create_mrca_prior}}}

\item{mcmc}{one MCMC
as returned by \code{\link{create_mcmc}}}

\item{misc_options}{one misc_options object,
as returned by \code{\link{create_misc_options}}}

\item{tipdates_filename}{name of the file containing the tip dates.
This file is assumed to have two columns, separated by a tab.
The first column contains the taxa names, the second column contains
the date.}
}
\value{
an MCMC configuration
}
\description{
Create an MCMC configuration, as in the BEAUti MCMC tab.
The number of states that will be saved equals the chain
length (\code{chain_length}) divided by the number of
states between each sampling event (\code{store_every})
}
\examples{
  # Create an MCMC chain with 50 states
  inference_model <- create_inference_model(chain_length = 50000, store_every = 1000)

  create_beast2_input_file(
    get_fasta_filename(),
    "create_inference_model.xml",
    inference_model = inference_model
  )
  testit::assert(file.exists("create_inference_model.xml"))
}
\author{
Richel J.C. Bilderbeek
}
