% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_inv_gamma_distr}
\alias{create_inv_gamma_distr}
\title{Create an inverse gamma distribution}
\usage{
create_inv_gamma_distr(id = NA, alpha = create_alphaparam(),
  beta = create_betaparam())
}
\arguments{
\item{id}{the distribution's ID}

\item{alpha}{the alpha shape parameter,
as returned by \code{\link{create_alphaparam}}}

\item{beta}{the beta shape parameter,
as returned by \code{\link{create_betaparam}}}
}
\value{
an inverse gamma distribution
}
\description{
Create an inverse gamma distribution
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
