% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_site_models.R
\name{create_jc69_site_models}
\alias{create_jc69_site_models}
\title{Creates n jc69_site_models}
\usage{
create_jc69_site_models(n)
}
\arguments{
\item{n}{the number of jc69_site_models}
}
\value{
a list of site_models
}
\description{
Creates n jc69_site_models
}
\examples{
  m <- create_jc69_site_models(1)
  testthat::expect_equal(length(m), 1)
  testthat::expect_true(is_jc69_site_model(m[[1]]))

  m <- create_jc69_site_models(2)
  testthat::expect_equal(length(m), 2)
  testthat::expect_true(is_jc69_site_model(m[[1]]))
  testthat::expect_true(is_jc69_site_model(m[[2]]))
}
\author{
Richel J.C. Bilderbeek
}
