% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter.R
\name{create_lambda_parameter}
\alias{create_lambda_parameter}
\title{Create a parameter called lambda}
\usage{
create_lambda_parameter(id = NA, value = 0)
}
\arguments{
\item{id}{the parameter's ID}

\item{value}{value of the parameter}

\item{estimate}{TRUE if this parameter alpha be estimated by BEAST2,
FALSE otherwise}
}
\value{
a parameter called lambda
}
\description{
Create a parameter called lambda
}
\seealso{
the function \code{\link{create_parameter}} contains a list
  of all parameters that can be created
}
\author{
Richel J.C. Bilderbeek
}
