% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_lambdaparam}
\alias{create_lambdaparam}
\title{Create a parameter called lambda}
\usage{
create_lambdaparam(id = NA, value = 0)
}
\arguments{
\item{id}{the parameter's ID}

\item{value}{value of the parameter}
}
\value{
a parameter called lambda
}
\description{
Create a parameter called lambda
}
\note{
this parameter is used in a Poisson distribution
  (as returned by \code{\link{create_poisson_distr}})
}
\seealso{
the function \code{\link{create_param}} contains a list
  of all parameters that can be created
}
\author{
Richel J.C. Bilderbeek
}
