% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_laplace_distr}
\alias{create_laplace_distr}
\title{Create a Laplace distribution}
\usage{
create_laplace_distr(id = NA)
}
\arguments{
\item{id}{the distribution's ID}
}
\value{
a Laplace distribution
}
\description{
Create a Laplace distribution
}
\seealso{
the function \code{\link{create_distribution}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
