% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_log_normal_distribution}
\alias{create_log_normal_distribution}
\title{Create a log-normal distribution}
\usage{
create_log_normal_distribution(id = NA)
}
\arguments{
\item{id}{the distribution's ID}
}
\value{
a log-normal distribution
}
\description{
Create a log-normal distribution
}
\author{
Richel J.C. Bilderbeek
}
