% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter.R
\name{create_m_parameter}
\alias{create_m_parameter}
\title{Create a parameter called m}
\usage{
create_m_parameter(id = NA, estimate = FALSE, value = 0)
}
\arguments{
\item{id}{the parameter's ID}

\item{estimate}{TRUE if this parameter alpha be estimated by BEAST2,
FALSE otherwise}

\item{value}{value of the parameter}
}
\value{
a parameter called m
}
\description{
Create a parameter called m
}
\seealso{
the function \code{\link{create_parameter}} contains a list
  of all parameters that can be created
}
\author{
Richel J.C. Bilderbeek
}
