% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mcmc_nested_sampling.R
\name{create_mcmc_nested_sampling}
\alias{create_mcmc_nested_sampling}
\title{Create an MCMC object that does nested sampling}
\usage{
create_mcmc_nested_sampling(chain_length = 1e+07, store_every = -1,
  particle_count = 1, sub_chain_length = 5000, epsilon = "1e-12")
}
\arguments{
\item{particle_count}{number of particles}

\item{sub_chain_length}{sub-chain length}

\item{epsilon}{epsilon}
}
\value{
an MCMC object
}
\description{
Create an MCMC object that does nested sampling
}
\references{
* Maturana, Patricio, et al. "Model selection and parameter inference
    in phylogenetics using Nested Sampling."
    arXiv preprint arXiv:1703.05471 (2017).
}
\author{
Richel J.C. Bilderbeek
}
