% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_muparam}
\alias{create_muparam}
\title{Create a parameter called mu}
\usage{
create_muparam(id = NA, estimate = FALSE, value = 0)
}
\arguments{
\item{id}{the parameter's ID}

\item{estimate}{TRUE if this parameter mu be estimated by BEAST2,
FALSE otherwise}

\item{value}{value of the parameter}
}
\value{
a parameter called mu
}
\description{
Create a parameter called mu
}
\examples{
  muparam <- create_muparam()
  testit::assert(is_muparam(muparam))
}
\seealso{
the function \code{\link{create_param}} contains a list
  of all parameters that can be created
}
\author{
Richel J.C. Bilderbeek
}
