% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_normal_distribution}
\alias{create_normal_distribution}
\title{Create an normal distribution}
\usage{
create_normal_distribution()
}
\value{
a normal distribution
}
\description{
Create an normal distribution
}
