% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_one_div_x_distr}
\alias{create_one_div_x_distr}
\title{Create a 1/x distribution}
\usage{
create_one_div_x_distr(id = NA)
}
\arguments{
\item{id}{the distribution's ID}
}
\value{
a 1/x distribution
}
\description{
Create a 1/x distribution
}
\seealso{
the function \code{\link{create_distribution}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
