% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_param}
\alias{create_param}
\title{General function to create a parameter.}
\usage{
create_param(name, id, ...)
}
\arguments{
\item{name}{the parameters' name. Valid
names can be found in \code{\link{get_param_names}}}

\item{id}{the parameter's ID}

\item{...}{specific parameter parameters}
}
\value{
a parameter
}
\description{
General function to create a parameter.
}
\note{
Prefer using the
  named functions
  \code{\link{create_alpha_param}},
  \code{\link{create_beta_param}},
  \code{\link{create_clock_rate_param}},
  \code{\link{create_m_param}},
  \code{\link{create_mean_param}},
  \code{\link{create_mu_param}},
  \code{\link{create_s_param}},
  \code{\link{create_scale_param}},
  and \code{\link{create_sigma_param}}
}
\author{
Richel J.C. Bilderbeek
}
