% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_param_lambda}
\alias{create_param_lambda}
\title{Alternative name for \code{\link{create_lambda_param}}, to help
the user find the function from a search tree. See
\code{\link{create_lambda_param}} for examples.}
\usage{
create_param_lambda(id = NA, estimate = FALSE, value = 0)
}
\arguments{
\item{id}{the parameter's ID}

\item{value}{value of the parameter}
}
\value{
a parameter called lambda
}
\description{
Alternative name for \code{\link{create_lambda_param}}, to help
the user find the function from a search tree. See
\code{\link{create_lambda_param}} for examples.
}
\seealso{
the function \code{\link{create_param}} contains a list
  of all parameters that can be created
}
