% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_param_rate_ag}
\alias{create_param_rate_ag}
\title{Alternative name for \code{\link{create_rate_ag_param}}, to help
the user find the function from a search tree. See
\code{\link{create_rate_ag_param}} for examples.}
\usage{
create_param_rate_ag(id = NA, estimate = FALSE, value = 0)
}
\arguments{
\item{id}{the parameter's ID}

\item{estimate}{TRUE if this parameter mu be estimated by BEAST2,
FALSE otherwise}

\item{value}{value of the parameter}
}
\value{
a parameter called 'rate AG'
}
\description{
Alternative name for \code{\link{create_rate_ag_param}}, to help
the user find the function from a search tree. See
\code{\link{create_rate_ag_param}} for examples.
}
\seealso{
the function \code{\link{create_param}} contains a list
  of all parameters that can be created
}
