% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_poisson_distr}
\alias{create_poisson_distr}
\title{Create a Poisson distribution}
\usage{
create_poisson_distr(id = NA)
}
\arguments{
\item{id}{the distribution's ID}
}
\value{
a Poisson distribution
}
\description{
Create a Poisson distribution
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
