% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_poisson_distribution}
\alias{create_poisson_distribution}
\title{Create a Poisson distribution}
\usage{
create_poisson_distribution()
}
\value{
a Poisson distribution
}
\description{
Create a Poisson distribution
}
