% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_posterior.R
\name{create_posterior}
\alias{create_posterior}
\title{Create a BEAST2 posterior by running BEAST2 from a random alignment}
\usage{
create_posterior(n_taxa = 5, sequence_length = 10,
  mcmc_chainlength = 10000)
}
\description{
Create a BEAST2 posterior by running BEAST2 from a random alignment
}
\author{
Richel Bilderbeek
}
