% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_alignment.R
\name{create_random_alignment}
\alias{create_random_alignment}
\title{Create a random alignment}
\usage{
create_random_alignment(n_taxa, sequence_length, rate = 1)
}
\arguments{
\item{n_taxa}{The number of taxa}

\item{sequence_length}{The number of base pairs the alignment will have}

\item{rate}{mutation rate}
}
\value{
an alignment of class DNAbin
}
\description{
Create a random alignment
}
\author{
Richel J.C. Bilderbeek
}
