% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_random_exp_distr}
\alias{create_random_exp_distr}
\title{Create a random exponential distribution}
\usage{
create_random_exp_distr()
}
\description{
Create a random exponential distribution
}
\author{
Richel J.C. Bilderbeek
}
