% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_fasta.R
\name{create_random_fasta}
\alias{create_random_fasta}
\title{Create a random FASTA file}
\usage{
create_random_fasta(n_taxa, sequence_length, filename)
}
\arguments{
\item{n_taxa}{The number of taxa}

\item{sequence_length}{The number of base pairs the alignment will have}

\item{filename}{the name of the FASTA file created}

\item{rate}{mutation rate}
}
\value{
Nothing
}
\description{
Create a random FASTA file
}

