% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_random_hky_site_model}
\alias{create_random_hky_site_model}
\title{Create a random HKY site model}
\usage{
create_random_hky_site_model()
}
\description{
Create a random HKY site model
}
\author{
Richel J.C. Bilderbeek
}
