% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_random_laplace_distr}
\alias{create_random_laplace_distr}
\title{Create a random Laplace distribution}
\usage{
create_random_laplace_distr()
}
\description{
Create a random Laplace distribution
}
\author{
Richel J.C. Bilderbeek
}
