% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_random_m_param}
\alias{create_random_m_param}
\title{Create a random m parameter}
\usage{
create_random_m_param()
}
\description{
Create a random m parameter
}
\author{
Richel J.C. Bilderbeek
}
