% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_name.R
\name{create_random_name}
\alias{create_random_name}
\title{Create a random name, for example,
  to be used by \code{\link{create_mrca_prior}}}
\usage{
create_random_name()
}
\value{
a random name
}
\description{
Create a random name, for example,
  to be used by \code{\link{create_mrca_prior}}
}
\author{
Richel J.C. Bilderbeek
}
