% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_random_normal_distr}
\alias{create_random_normal_distr}
\title{Create a random normal distribution}
\usage{
create_random_normal_distr()
}
\description{
Create a random normal distribution
}
\author{
Richel J.C. Bilderbeek
}
