% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_random_one_div_x_distr}
\alias{create_random_one_div_x_distr}
\title{Create a random 1/x distribution}
\usage{
create_random_one_div_x_distr()
}
\description{
Create a random 1/x distribution
}
\author{
Richel J.C. Bilderbeek
}
