% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_random_rate_ac_param}
\alias{create_random_rate_ac_param}
\title{Create a random rate AC parameter}
\usage{
create_random_rate_ac_param()
}
\description{
Create a random rate AC parameter
}
\author{
Richel J.C. Bilderbeek
}
