% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_random_tree_prior}
\alias{create_random_tree_prior}
\title{Create a random tree prior}
\usage{
create_random_tree_prior()
}
\description{
Create a random tree prior
}
\author{
Richel J.C. Bilderbeek
}
