% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_rate_ac_param}
\alias{create_rate_ac_param}
\title{Create a parameter called 'rate AC'}
\usage{
create_rate_ac_param(id = NA, estimate = TRUE, value = "1.0",
  lower = "0.0")
}
\arguments{
\item{id}{the parameter's ID}

\item{estimate}{TRUE if this parameter mu be estimated by BEAST2,
FALSE otherwise}

\item{value}{value of the parameter}

\item{lower}{lowest possible value of the parameter}
}
\value{
a parameter called 'rate AC'
}
\description{
Create a parameter called 'rate AC'
}
\examples{
  rate_ac_param <- create_rate_ac_param()
  testit::assert(is_rate_ac_param(rate_ac_param))
}
\seealso{
the function \code{\link{create_param}} contains a list
  of all parameters that can be created
}
\author{
Richel J.C. Bilderbeek
}
