% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_rate_cg_param}
\alias{create_rate_cg_param}
\title{Create a parameter called 'rate CG'}
\usage{
create_rate_cg_param(id = NA, lower = "0.0", value = "1.0")
}
\arguments{
\item{id}{the parameter's ID}

\item{lower}{lowest possible value of the parameter}

\item{value}{value of the parameter}
}
\value{
a parameter called 'rate CG'
}
\description{
Create a parameter called 'rate CG'
}
\examples{
  rate_cg_param <- create_rate_cg_param()
  testit::assert(is_rate_cg_param(rate_cg_param))
}
\seealso{
the function \code{\link{create_param}} contains a list
  of all parameters that can be created
}
\author{
Richel J.C. Bilderbeek
}
