% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clock_model.R
\name{create_rln_clock_model}
\alias{create_rln_clock_model}
\title{Create a relaxed log-normal clock model}
\usage{
create_rln_clock_model()
}
\value{
a relaxed log-normal clock_model
}
\description{
Create a relaxed log-normal clock model
}
