% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_alpha_param}
\alias{create_rnd_alpha_param}
\title{Create a random alpha parameter}
\usage{
create_rnd_alpha_param()
}
\description{
Create a random alpha parameter
}
\author{
Richel J.C. Bilderbeek
}
