% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_bd_tree_prior}
\alias{create_rnd_bd_tree_prior}
\title{Create a random BD tree prior}
\usage{
create_rnd_bd_tree_prior()
}
\description{
Create a random BD tree prior
}
\author{
Richel J.C. Bilderbeek
}
