% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_beta_distr}
\alias{create_rnd_beta_distr}
\title{Create a random beta distribution}
\usage{
create_rnd_beta_distr()
}
\description{
Create a random beta distribution
}
\author{
Richel J.C. Bilderbeek
}
