% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_beta_param}
\alias{create_rnd_beta_param}
\title{Create a random beta parameter}
\usage{
create_rnd_beta_param()
}
\description{
Create a random beta parameter
}
\author{
Richel J.C. Bilderbeek
}
