% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_clock_rate_param}
\alias{create_rnd_clock_rate_param}
\title{Create a random clock rate parameter}
\usage{
create_rnd_clock_rate_param()
}
\description{
Create a random clock rate parameter
}
\author{
Richel J.C. Bilderbeek
}
