% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_estimate}
\alias{create_rnd_estimate}
\title{Create a random value for 'estimate',
which must be TRUE or FALSE}
\usage{
create_rnd_estimate()
}
\description{
Create a random value for 'estimate',
which must be TRUE or FALSE
}
\author{
Richel J.C. Bilderbeek
}
