% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_freq_equilibrium}
\alias{create_rnd_freq_equilibrium}
\title{Create a random value for 'freq equilibrium', which
can be 'estimated', 'empirical' or 'all_equal'}
\usage{
create_rnd_freq_equilibrium()
}
\description{
Create a random value for 'freq equilibrium', which
can be 'estimated', 'empirical' or 'all_equal'
}
\author{
Richel J.C. Bilderbeek
}
