% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_gamma_site_model}
\alias{create_rnd_gamma_site_model}
\title{Create a random gamma site model}
\usage{
create_rnd_gamma_site_model()
}
\description{
Create a random gamma site model
}
\author{
Richel J.C. Bilderbeek
}
