% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_inv_gamma_distr}
\alias{create_rnd_inv_gamma_distr}
\title{Create a random inverse-gamma distribution}
\usage{
create_rnd_inv_gamma_distr()
}
\description{
Create a random inverse-gamma distribution
}
\author{
Richel J.C. Bilderbeek
}
