% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_jc69_site_model}
\alias{create_rnd_jc69_site_model}
\title{Create a random JC69 distribution}
\usage{
create_rnd_jc69_site_model()
}
\description{
Create a random JC69 distribution
}
\author{
Richel J.C. Bilderbeek
}
