% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_kappa_2_param}
\alias{create_rnd_kappa_2_param}
\title{Create a random kappa 2 parameter}
\usage{
create_rnd_kappa_2_param()
}
\description{
Create a random kappa 2 parameter
}
\author{
Richel J.C. Bilderbeek
}
