% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_lambda_param}
\alias{create_rnd_lambda_param}
\title{Create a random lambda parameter}
\usage{
create_rnd_lambda_param()
}
\description{
Create a random lambda parameter
}
\author{
Richel J.C. Bilderbeek
}
