% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_log_normal_distr}
\alias{create_rnd_log_normal_distr}
\title{Create a random log-normal distribution}
\usage{
create_rnd_log_normal_distr()
}
\description{
Create a random log-normal distribution
}
\author{
Richel J.C. Bilderbeek
}
