% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_mean_param}
\alias{create_rnd_mean_param}
\title{Create a random mean parameter}
\usage{
create_rnd_mean_param()
}
\description{
Create a random mean parameter
}
\author{
Richel J.C. Bilderbeek
}
