% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_mrca_priors}
\alias{create_rnd_mrca_priors}
\title{Create a random MRCA prior}
\usage{
create_rnd_mrca_priors(fasta_filename)
}
\arguments{
\item{fasta_filename}{a FASTA filename}
}
\description{
Create a random MRCA prior
}
\author{
Richel J.C. Bilderbeek
}
