% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_poisson_distr}
\alias{create_rnd_poisson_distr}
\title{Create a random Poisson distribution}
\usage{
create_rnd_poisson_distr()
}
\description{
Create a random Poisson distribution
}
\author{
Richel J.C. Bilderbeek
}
