% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_rate_ag_param}
\alias{create_rnd_rate_ag_param}
\title{Create a random rate AG parameter}
\usage{
create_rnd_rate_ag_param()
}
\description{
Create a random rate AG parameter
}
\author{
Richel J.C. Bilderbeek
}
