% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_rate_cg_param}
\alias{create_rnd_rate_cg_param}
\title{Create a random rate CG parameter}
\usage{
create_rnd_rate_cg_param()
}
\description{
Create a random rate CG parameter
}
\author{
Richel J.C. Bilderbeek
}
