% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_rate_gt_param}
\alias{create_rnd_rate_gt_param}
\title{Create a random rate GT parameter}
\usage{
create_rnd_rate_gt_param()
}
\description{
Create a random rate GT parameter
}
\author{
Richel J.C. Bilderbeek
}
