% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_rln_clock_model}
\alias{create_rnd_rln_clock_model}
\title{Create a random RLN clock model}
\usage{
create_rnd_rln_clock_model()
}
\description{
Create a random RLN clock model
}
\author{
Richel J.C. Bilderbeek
}
