% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_scale_param}
\alias{create_rnd_scale_param}
\title{Create a random scale parameter}
\usage{
create_rnd_scale_param()
}
\description{
Create a random scale parameter
}
\author{
Richel J.C. Bilderbeek
}
