% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_sigma_param}
\alias{create_rnd_sigma_param}
\title{Create a random sigma parameter}
\usage{
create_rnd_sigma_param()
}
\description{
Create a random sigma parameter
}
\author{
Richel J.C. Bilderbeek
}
