% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_two_mrca_priors}
\alias{create_rnd_two_mrca_priors}
\title{Creates two MRCA priors, checked to be compatible}
\usage{
create_rnd_two_mrca_priors(fasta_filename)
}
\arguments{
\item{fasta_filename}{a FASTA filename}
}
\description{
Creates two MRCA priors, checked to be compatible
}
\author{
Richel J.C. Bilderbeek
}
