% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_uniform_distr}
\alias{create_rnd_uniform_distr}
\title{Create a random uniform distribution}
\usage{
create_rnd_uniform_distr()
}
\description{
Create a random uniform distribution
}
\author{
Richel J.C. Bilderbeek
}
