% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter.R
\name{create_scale_parameter}
\alias{create_scale_parameter}
\title{Create a parameter called scale}
\usage{
create_scale_parameter(id = NA, estimate = FALSE, value = 0)
}
\arguments{
\item{id}{the parameter's ID}

\item{estimate}{TRUE if this parameter scale be estimated by BEAST2,
FALSE otherwise}

\item{value}{value of the parameter}
}
\value{
a parameter called scale
}
\description{
Create a parameter called scale
}
\examples{
  scale_parameter <- create_scale_parameter()
  testit::assert(is_scale_parameter(scale_parameter))
}
\seealso{
the function \code{\link{create_parameter}} contains a list
  of all parameters that can be created
}
\author{
Richel J.C. Bilderbeek
}
