% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_scaleparam}
\alias{create_scaleparam}
\title{Create a parameter called scale}
\usage{
create_scaleparam(id = NA, estimate = FALSE, value = 0)
}
\arguments{
\item{id}{the parameter's ID}

\item{estimate}{TRUE if this parameter scale be estimated by BEAST2,
FALSE otherwise}

\item{value}{value of the parameter}
}
\value{
a parameter called scale
}
\description{
Create a parameter called scale
}
\examples{
  scaleparam <- create_scaleparam()
  testit::assert(is_scaleparam(scaleparam))
}
\seealso{
the function \code{\link{createparam}} contains a list
  of all parameters that can be created
}
\author{
Richel J.C. Bilderbeek
}
