% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_screenlog.R
\name{create_screenlog}
\alias{create_screenlog}
\title{Create a \code{screenlog} object}
\usage{
create_screenlog(filename = "")
}
\arguments{
\item{filename}{name of the file to store the screenlog to}
}
\description{
Create a \code{screenlog} object
}
