% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_site_model.R
\name{create_site_model}
\alias{create_site_model}
\title{Create a site model}
\usage{
create_site_model(name)
}
\arguments{
\item{name}{the site model name. Valid
names can be found in 'get_site_model_names'}
}
\value{
a site_model
}
\description{
Create a site model
}
