% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_site_model_from_name.R
\name{create_site_model_from_name}
\alias{create_site_model_from_name}
\title{Create a site model from name}
\usage{
create_site_model_from_name(site_model_name)
}
\arguments{
\item{site_model_name}{name of a site model,
must be a name as returned by \code{\link{get_site_model_names}}}
}
\value{
a site model
}
\description{
Create a site model from name
}
\examples{
site_model <- create_site_model_from_name(get_site_model_names()[1])
}
\seealso{
Use \link{create_site_model} to create a site model
}
\author{
Richèl J.C. Bilderbeek
}
