% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_site_model.R
\name{create_site_model_jc69}
\alias{create_site_model_jc69}
\title{Alternative name for \code{\link{create_jc69_site_model}}, to help
the user find the function from a search tree. See
\code{\link{create_jc69_site_model}} for examples.}
\usage{
create_site_model_jc69(id = NA,
  gamma_site_model = create_gamma_site_model())
}
\arguments{
\item{id}{the IDs of the alignment (can be extracted from
the FASTA filesname using \code{\link{get_id}})}

\item{gamma_site_model}{a gamma site model, as created
by \code{\link{create_gamma_site_model}}}
}
\value{
a JC69 site_model
}
\description{
Alternative name for \code{\link{create_jc69_site_model}}, to help
the user find the function from a search tree. See
\code{\link{create_jc69_site_model}} for examples.
}
