% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_site_model.R
\name{create_site_model_tn93}
\alias{create_site_model_tn93}
\title{Alternative name for \code{\link{create_tn93_site_model}}, to help
the user find the function from a search tree. See
\code{\link{create_tn93_site_model}} for examples.}
\usage{
create_site_model_tn93(id = NA,
  gamma_site_model = create_gamma_site_model(),
  kappa_1_param = create_kappa_1_param(),
  kappa_2_param = create_kappa_2_param(),
  kappa_1_prior_distr = create_log_normal_distr(m = create_m_param(id = NA,
  estimate = FALSE, value = "1.0"), s = create_s_param(id = NA, estimate =
  FALSE, value = "1.25")), kappa_2_prior_distr = create_log_normal_distr(m =
  create_m_param(id = NA, estimate = FALSE, value = "1.0"), s =
  create_s_param(id = NA, estimate = FALSE, value = "1.25")),
  freq_equilibrium = "estimated")
}
\arguments{
\item{id}{the IDs of the alignment (can be extracted from
the FASTA filesname using \code{\link{get_id}})}

\item{gamma_site_model}{a gamma site model, as created
by \code{\link{create_gamma_site_model}}}

\item{kappa_1_param}{the 'kappa 1' parameter,
as returned by \code{\link{create_kappa_1_param}}}

\item{kappa_2_param}{the 'kappa 2' parameter,
as returned by \code{\link{create_kappa_2_param}}}

\item{kappa_1_prior_distr}{the distribution of the kappa 1 prior,
which is a log-normal distribution
(as created by \code{\link{create_log_normal_distr}})
by default}

\item{kappa_2_prior_distr}{the distribution of the kappa 2 prior,
which is a log-normal distribution
(as created by \code{\link{create_log_normal_distr}})
by default}

\item{freq_equilibrium}{the frequency in which the rates are at equilibrium
are either \code{estimated}, \code{empirical} or \code{all_equal}.
\code{\link{get_freq_equilibrium_names}} returns the possible values
for \code{freq_equilibrium}}
}
\value{
a TN93 site_model
}
\description{
Alternative name for \code{\link{create_tn93_site_model}}, to help
the user find the function from a search tree. See
\code{\link{create_tn93_site_model}} for examples.
}
