% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_site_models.R
\name{create_site_models}
\alias{create_site_models}
\title{Creates all supported site models}
\usage{
create_site_models()
}
\value{
a list of site_models
}
\description{
Creates all supported site models
}
