% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_site_models_from_names.R
\name{create_site_models_from_names}
\alias{create_site_models_from_names}
\title{Create site models from their names}
\usage{
create_site_models_from_names(site_model_names)
}
\description{
Create site models from their names
}
\examples{
  names <- get_site_model_names()
  site_models <- create_site_models_from_names(names)

  for (i in seq_along(names)) {
    testthat::expect_equal(names[i], site_models[[i]]$name)
  }
}
\author{
Richel J.C. Bilderbeek
}
