% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clock_model.R
\name{create_strict_clock_model}
\alias{create_strict_clock_model}
\title{Create a strict clock model}
\usage{

  create_strict_clock_model(clock_rate_parameter = create_clock_rate_parameter())
}
\arguments{
\item{clock_rate_parameter}{a clock_rate parameter,
as created by a \code{\link{create_clock_rate_parameter}} function}
}
\value{
a strict clock_model
}
\description{
Create a strict clock model
}
