% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clock_models.R
\name{create_strict_clock_models}
\alias{create_strict_clock_models}
\title{Creates n strict clock_models}
\usage{
create_strict_clock_models(n)
}
\arguments{
\item{n}{the number of strict_clock_models}
}
\value{
a list of strict_clock objects
}
\description{
Creates n strict clock_models
}
\examples{
  m <- create_strict_clock_models(1)
  testthat::expect_equal(length(m), 1)
  testthat::expect_true(is_strict_clock_model(m[[1]]))

  m <- create_strict_clock_models(2)
  testthat::expect_equal(length(m), 2)
  testthat::expect_true(is_strict_clock_model(m[[1]]))
  testthat::expect_true(is_strict_clock_model(m[[2]]))
}
