% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_strict_clock_rate_scaler_operator_xml.R
\name{create_strict_clock_rate_scaler_operator_xml}
\alias{create_strict_clock_rate_scaler_operator_xml}
\title{Internal function}
\usage{
create_strict_clock_rate_scaler_operator_xml(inference_model)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\value{
the following XML:
\code{
  ...
}
}
\description{
Creates the \code{StrictClockRateScaler} operator
such as:
\code{
  ...
}
}
\examples{
create_strict_clock_rate_scaler_operator_xml(
  create_inference_model(
    clock_model = create_strict_clock_model(id = 314)
  )
)
}
\author{
Richèl J.C. Bilderbeek
}
