% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_temp_tracelog_filename.R
\name{create_temp_tracelog_filename}
\alias{create_temp_tracelog_filename}
\title{Create a filename for a temporary tracelog file}
\usage{
create_temp_tracelog_filename()
}
\description{
Create a filename for a temporary tracelog file
}
\seealso{
use \link{create_tracelog} to create a tracelog.
}
