% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_treelog.R
\name{create_test_treelog}
\alias{create_test_treelog}
\title{Create a \code{treelog} object}
\usage{
create_test_treelog(
  filename = create_temp_treelog_filename(),
  log_every = 1000,
  mode = "tree",
  sanitise_headers = FALSE,
  sort = "none"
)
}
\arguments{
\item{filename}{name of the file to store the posterior trees}

\item{log_every}{number of MCMC states between writing to file}

\item{mode}{mode how to log.
Valid values are the ones returned by \link{get_log_modes}}

\item{sanitise_headers}{set to \link{TRUE} to sanitise the headers of the
log file}

\item{sort}{how to sort the log.
Valid values are the ones returned by \link{get_log_sorts}}
}
\description{
Create a \code{treelog} object
}
