% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tracelog.R
\name{create_tracelog}
\alias{create_tracelog}
\title{Create a \code{tracelog} object}
\usage{
create_tracelog(filename = "tracelog.log")
}
\arguments{
\item{filename}{name of the file to store the \code{tracelog} to.
Defaults to \code{[alignment_name].log}}
}
\description{
Create a \code{tracelog} object
}
