% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tracelog_xml.R
\name{create_tracelog_xml}
\alias{create_tracelog_xml}
\title{Creates the \code{tracelog} section of the \code{logger} section
of a BEAST2 XML parameter file}
\usage{
create_tracelog_xml(input_filename, inference_model)
}
\arguments{
\item{input_filename}{A FASTA filename.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}

\item{inference_model}{an Bayesian phylogenetic inference model,
as can be created by \code{\link{create_inference_model}}.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.}
}
\description{
Creates the \code{tracelog} section of the \code{logger} section
of a BEAST2 XML parameter file
}
\author{
Richèl J.C. Bilderbeek
}
