% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trait_set_string.R
\name{create_trait_set_string}
\alias{create_trait_set_string}
\title{Create a trait set string.}
\usage{
create_trait_set_string(df)
}
\arguments{
\item{df}{a data frame with two columns}
}
\value{
the trait set string
}
\description{
For example, a data frame with row \code{A 1}
and another row \code{B 2}, the trait set string will be
\code{A=1,B=2}
}
\author{
Richèl J.C. Bilderbeek
}
