% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_prior.R
\name{create_tree_prior_cbs}
\alias{create_tree_prior_cbs}
\title{Alternative name for \code{\link{create_cbs_tree_prior}}, to help
the user find the function from a search tree. See
\code{\link{create_cbs_tree_prior}} for examples.}
\usage{
create_tree_prior_cbs(id = NA)
}
\arguments{
\item{id}{the ID of the alignment}
}
\value{
a Coalescent Bayesian Skyline tree_prior
}
\description{
Alternative name for \code{\link{create_cbs_tree_prior}}, to help
the user find the function from a search tree. See
\code{\link{create_cbs_tree_prior}} for examples.
}
\seealso{
An alignment ID can be extracted from
  its FASTA filesname using \code{\link{get_id}}
}
