% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_prior.R
\name{create_tree_prior_cep}
\alias{create_tree_prior_cep}
\title{Alternative name for \code{\link{create_cep_tree_prior}}, to help
the user find the function from a search tree. See
\code{\link{create_cep_tree_prior}} for examples.}
\usage{
create_tree_prior_cep(id = NA, pop_size_distr = create_one_div_x_distr(),
  growth_rate_distr = create_laplace_distr())
}
\arguments{
\item{id}{the ID of the alignment}

\item{pop_size_distr}{the population distribution,
as created by a \code{\link{create_distr}} function}

\item{growth_rate_distr}{the growth rate distribution,
as created by a \code{\link{create_distr}} function}
}
\value{
a Coalescent Exponential Population tree_prior
}
\description{
Alternative name for \code{\link{create_cep_tree_prior}}, to help
the user find the function from a search tree. See
\code{\link{create_cep_tree_prior}} for examples.
}
\seealso{
An alignment ID can be extracted from
  its FASTA filesname using \code{\link{get_id}}
}
