% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_prior.R
\name{create_tree_prior_yule}
\alias{create_tree_prior_yule}
\title{Alternative name for \code{\link{create_yule_tree_prior}}, to help
the user find the function from a search tree. See
\code{\link{create_yule_tree_prior}} for examples.}
\usage{
create_tree_prior_yule(id = NA, birth_rate_distr = create_uniform_distr())
}
\arguments{
\item{id}{the ID of the alignment}

\item{birth_rate_distr}{the birth rate distribution,
as created by a \code{\link{create_distr}} function}
}
\value{
a Yule tree_prior
}
\description{
Alternative name for \code{\link{create_yule_tree_prior}}, to help
the user find the function from a search tree. See
\code{\link{create_yule_tree_prior}} for examples.
}
\seealso{
An alignment ID can be extracted from
  its FASTA filesname using \code{\link{get_id}}
}
