% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_priors.R
\name{create_tree_priors}
\alias{create_tree_priors}
\title{Creates all supported tree priors}
\usage{
create_tree_priors()
}
\value{
a list of tree_priors
}
\description{
Creates all supported tree priors
}
