% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_uniform_distribution}
\alias{create_uniform_distribution}
\title{Create a uniform distribution}
\usage{
create_uniform_distribution(id = 1)
}
\arguments{
\item{id}{the distribution's ID}
}
\value{
a uniform distribution
}
\description{
Create a uniform distribution
}
\author{
Richel J.C. Bilderbeek
}
