% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_xml.R
\name{create_xml_declaration}
\alias{create_xml_declaration}
\title{Create the XML declaration as XML text}
\usage{
create_xml_declaration()
}
\value{
lines of XML text
}
\description{
Create the XML declaration as XML text
}
\author{
Richèl J.C. Bilderbeek
}
