% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_parameters_doc}
\alias{default_parameters_doc}
\title{Documentation of parameters (for example, \code{create_param}.
This function does nothing. It is intended to inherit documentation from.}
\usage{
default_parameters_doc(estimate, id, lower, name, upper, value, ...)
}
\arguments{
\item{estimate}{TRUE if this parameter is to be estimated by BEAST2,
FALSE otherwise}

\item{id}{the parameter's ID}

\item{lower}{lowest possible value of the parameter. If the parameter
is estimated, \code{lower} must be less than \code{value}}

\item{name}{the parameters' name. Valid
names can be found in \code{get_param_names}}

\item{upper}{upper value of the parameter}

\item{value}{value of the parameter}

\item{...}{specific parameter parameters}
}
\description{
Documentation of parameters (for example, \code{create_param}.
This function does nothing. It is intended to inherit documentation from.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
