% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(ccp_tree_prior, clock_model, clock_models, distr_id,
  fixed_crown_age, initial_phylogenies, input_fasta_filenames, mcmc,
  misc_options, output_xml_filename, param_id, rln_clock_model, site_model,
  site_models, strict_clock_model, tree_prior, tree_priors)
}
\arguments{
\item{ccp_tree_prior}{a Coalescent Constant Population tree prior,
as returned by \code{\link{create_ccp_tree_prior}}}

\item{clock_model}{a clock model,
as returned by \code{\link{create_clock_model}}}

\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}

\item{distr_id}{a distributions' ID}

\item{fixed_crown_age}{determines if the phylogeny its crown age is
fixed. If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}

\item{initial_phylogenies}{one or more MCMC chain's initial phylogenies.
Each one set to NA will result in BEAST2 using a random phylogeny. Else
the phylogeny is assumed to be of class ape::phylo.}

\item{input_fasta_filenames}{One or more FASTA filenames.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}

\item{mcmc}{one mcmc object,
as returned by \code{\link{create_mcmc}}}

\item{misc_options}{one misc_options object,
as returned by \code{\link{create_misc_options}}}

\item{output_xml_filename}{Name of the XML parameter file created by this
function. BEAST2 uses this file as input.}

\item{param_id}{a parameter's ID}

\item{rln_clock_model}{a Relaxed Log-Normal clock model,
as returned by \code{\link{create_rln_clock_model}}}

\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}

\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}

\item{strict_clock_model}{a strict clock model,
as returned by \code{\link{create_strict_clock_model}}}

\item{tree_prior}{a tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{tree_priors}{one or more tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{id}{an alignment's IDs.
An ID can be extracted from its FASTA filesname
with \code{\link{get_ids}})}

\item{ids}{one or more alignments' IDs.
IDs can be extracted from their FASTA filesnames
with \code{\link{get_ids}})}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\author{
Richel J.C. Bilderbeek
}
