% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_to_xml.R
\name{distr_to_xml_beta}
\alias{distr_to_xml_beta}
\title{Internal function}
\usage{
distr_to_xml_beta(distr, beauti_options = create_beauti_options())
}
\arguments{
\item{distr}{a beta distribution,
as created by \code{\link{create_beta_distr}})}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the distribution as XML text
}
\description{
Converts a beta distribution to XML
}
\author{
Richèl J.C. Bilderbeek
}
