% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_to_xml.R
\name{distr_to_xml_exp}
\alias{distr_to_xml_exp}
\title{Internal function}
\usage{
distr_to_xml_exp(distr, beauti_options = create_beauti_options())
}
\arguments{
\item{distr}{an exponential distribution,
as created by \code{\link{create_exp_distr}})}
}
\value{
the distribution as XML text
}
\description{
Converts an exponential distribution to XML
}
\author{
Richèl J.C. Bilderbeek
}
