% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_to_xml.R
\name{distr_to_xml_gamma}
\alias{distr_to_xml_gamma}
\title{Converts a gamma distribution to XML}
\usage{
distr_to_xml_gamma(distr)
}
\arguments{
\item{distr}{a gamma distribution,
as created by \code{\link{create_gamma_distr}})}
}
\value{
the distribution as XML text
}
\description{
Converts a gamma distribution to XML
}
\author{
Richèl J.C. Bilderbeek
}
