% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_to_xml.R
\name{distr_to_xml_inv_gamma}
\alias{distr_to_xml_inv_gamma}
\title{Converts a inv_gamma distribution to XML}
\usage{
distr_to_xml_inv_gamma(distr)
}
\arguments{
\item{distr}{a inv_gamma distibution,
as created by \code{\link{create_inv_gamma_distr}})}
}
\value{
the distribution as XML text
}
\description{
Converts a inv_gamma distribution to XML
}
\author{
Richel J.C. Bilderbeek
}
