% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_to_xml.R
\name{distr_to_xml_log_normal}
\alias{distr_to_xml_log_normal}
\title{Converts a log-normal distribution to XML}
\usage{
distr_to_xml_log_normal(distr)
}
\arguments{
\item{distr}{a log-normal distibution,
as created by \code{\link{create_log_normal_distr}})}
}
\value{
the distribution as XML text
}
\description{
Converts a log-normal distribution to XML
}
\author{
Richel J.C. Bilderbeek
}
