% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_to_xml.R
\name{distr_to_xml_normal}
\alias{distr_to_xml_normal}
\title{Converts a normal distribution to XML}
\usage{
distr_to_xml_normal(distr)
}
\arguments{
\item{distr}{a normal distribution,
as created by \code{\link{create_normal_distr}})}
}
\value{
the distribution as XML text
}
\description{
Converts a normal distribution to XML
}
\author{
Richèl J.C. Bilderbeek
}
