% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_to_xml.R
\name{distr_to_xml_poisson}
\alias{distr_to_xml_poisson}
\title{Internal function}
\usage{
distr_to_xml_poisson(distr, beauti_options = create_beauti_options())
}
\arguments{
\item{distr}{a Poisson distribution,
as created by \code{\link{create_poisson_distr}})}
}
\value{
the distribution as XML text
}
\description{
Converts a Poisson distribution to XML
}
\author{
Richèl J.C. Bilderbeek
}
