% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_to_xml.R
\name{distr_to_xml_uniform}
\alias{distr_to_xml_uniform}
\title{Converts a uniform distribution to XML}
\usage{
distr_to_xml_uniform(distr)
}
\arguments{
\item{distr}{a uniform distibution,
as created by \code{\link{create_uniform_distr}})}
}
\value{
the distribution as XML text
}
\description{
Converts a uniform distribution to XML
}
\author{
Richel J.C. Bilderbeek
}
