% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_to_xml.R
\name{distribution_to_xml}
\alias{distribution_to_xml}
\title{Converts a distribution to XML}
\usage{
distribution_to_xml(distribution, n_spaces)
}
\arguments{
\item{distribution}{a distibution,
as created by \code{\link{create_distribution}})}

\item{n_spaces}{the number of spaces to add before the text}
}
\value{
the distribution as XML text
}
\description{
Converts a distribution to XML
}
\author{
Richel J.C. Bilderbeek
}
