% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_to_xml.R
\name{distribution_to_xml_beta}
\alias{distribution_to_xml_beta}
\title{Converts a beta distribution to XML}
\usage{
distribution_to_xml_beta(distribution)
}
\arguments{
\item{distribution}{a beta distibution,
as created by \code{\link{create_beta_distr}})}
}
\value{
the distribution as XML text
}
\description{
Converts a beta distribution to XML
}
\author{
Richel J.C. Bilderbeek
}
