% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_to_xml.R
\name{distribution_to_xml_exponential}
\alias{distribution_to_xml_exponential}
\title{Converts an exponential distribution to XML}
\usage{
distribution_to_xml_exponential(distribution)
}
\arguments{
\item{distribution}{an exponential distibution,
as created by \code{\link{create_exponential_distribution}})}
}
\value{
the distribution as XML text
}
\description{
Converts an exponential distribution to XML
}
\author{
Richel J.C. Bilderbeek
}
