% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_to_xml.R
\name{distribution_to_xml_gamma}
\alias{distribution_to_xml_gamma}
\title{Converts a gamma distribution to XML}
\usage{
distribution_to_xml_gamma(distribution)
}
\arguments{
\item{distribution}{a gamma distibution,
as created by \code{\link{create_gamma_distribution}})}
}
\value{
the distribution as XML text
}
\description{
Converts a gamma distribution to XML
}
\author{
Richel J.C. Bilderbeek
}
