% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_to_xml.R
\name{distribution_to_xml_laplace}
\alias{distribution_to_xml_laplace}
\title{Converts a laplace distribution to XML}
\usage{
distribution_to_xml_laplace(distribution)
}
\arguments{
\item{distribution}{a laplace distibution,
as created by \code{\link{create_laplace_distr}})}
}
\value{
the distribution as XML text
}
\description{
Converts a laplace distribution to XML
}
\author{
Richel J.C. Bilderbeek
}
