% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_to_xml.R
\name{distribution_to_xml_normal}
\alias{distribution_to_xml_normal}
\title{Converts a normal distribution to XML}
\usage{
distribution_to_xml_normal(distribution)
}
\arguments{
\item{distribution}{a normal distibution,
as created by \code{\link{create_normal_distr}})}
}
\value{
the distribution as XML text
}
\description{
Converts a normal distribution to XML
}
\author{
Richel J.C. Bilderbeek
}
