% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_to_xml.R
\name{distribution_to_xml_one_div_x}
\alias{distribution_to_xml_one_div_x}
\title{Converts a 1/x distribution to XML}
\usage{
distribution_to_xml_one_div_x(distribution)
}
\arguments{
\item{distribution}{a 1/x distibution,
as created by \code{\link{create_one_div_x_distr}})}
}
\value{
the distribution as XML text
}
\description{
Converts a 1/x distribution to XML
}
\author{
Richel J.C. Bilderbeek
}
