% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_to_xml.R
\name{distribution_to_xml_poisson}
\alias{distribution_to_xml_poisson}
\title{Converts a poisson distribution to XML}
\usage{
distribution_to_xml_poisson(distribution)
}
\arguments{
\item{distribution}{a poisson distibution,
as created by \code{\link{create_poisson_distribution}})}
}
\value{
the distribution as XML text
}
\description{
Converts a poisson distribution to XML
}
\author{
Richel J.C. Bilderbeek
}
