% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_to_xml.R
\name{distribution_to_xml_uniform}
\alias{distribution_to_xml_uniform}
\title{Converts a uniform distribution to XML}
\usage{
distribution_to_xml_uniform(distribution)
}
\arguments{
\item{distribution}{a uniform distibution,
as created by \code{\link{create_uniform_distribution}})}
}
\value{
the distribution as XML text
}
\description{
Converts a uniform distribution to XML
}
\author{
Richel J.C. Bilderbeek
}
