% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_xml_loggers_from_lines.R
\name{extract_xml_loggers_from_lines}
\alias{extract_xml_loggers_from_lines}
\title{Extract everything between first loggers and last loggers line}
\usage{
extract_xml_loggers_from_lines(lines)
}
\arguments{
\item{lines}{lines of text}
}
\value{
lines of text from the first to and including the last operators line
}
\description{
Extract everything between first loggers and last loggers line
}
\author{
Richèl J.C. Bilderbeek
}
