% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_xml_section_from_lines.R
\name{extract_xml_section_from_lines}
\alias{extract_xml_section_from_lines}
\title{Get the lines of an XML section, including the section tags}
\usage{
extract_xml_section_from_lines(lines, section)
}
\arguments{
\item{lines}{lines of the XML text}
}
\description{
Get the lines of an XML section, including the section tags
}
\section{the XML section name}{
NA
}

