% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_to_phylo.R
\name{fasta_to_phylo}
\alias{fasta_to_phylo}
\title{Create a random phylogeny, with a specified crown age, from a FASTA file}
\usage{
fasta_to_phylo(fasta_filename, crown_age)
}
\value{
a phylogeny
}
\description{
Create a random phylogeny, with a specified crown age, from a FASTA file
}
\examples{
  fasta_filename <- beautier::get_fasta_filename()
  phylo <- beautier::fasta_to_phylo(fasta_filename, crown_age = 15)
  testit::assert(5 == length(phylo$tip.label))
}
