% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_exist.R
\name{files_exist}
\alias{files_exist}
\title{Multi-file file.exists function}
\usage{
files_exist(filenames)
}
\arguments{
\item{filenames}{one or more filename}
}
\value{
TRUE if all files exists, FALSE otherwise
}
\description{
Multi-file file.exists function
}
\examples{
library(testthat)

filenames <- c(get_fasta_filename(), get_fasta_filename())
expect_true(files_exist(filenames))

filenames <- c(get_fasta_filename(), tempfile())
expect_false(files_exist(filenames))
}
\author{
Richèl J.C. Bilderbeek
}
