% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_clock_model.R
\name{find_clock_model}
\alias{find_clock_model}
\title{Finds a clock model with a certain ID}
\usage{
find_clock_model(clock_models, id)
}
\arguments{
\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}

\item{id}{the ID of the clock model}
}
\value{
the clock models with the desired ID, NULL if such a clock model is
  absent
}
\description{
Finds a clock model with a certain ID
}
\author{
Richel J.C. Bilderbeek
}
