% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_first_regex_line.R
\name{find_first_regex_line}
\alias{find_first_regex_line}
\title{Find the first line that satisfies a regex}
\usage{
find_first_regex_line(lines, regex)
}
\arguments{
\item{lines}{lines of text}

\item{regex}{the regex as text}
}
\value{
index of the line
}
\description{
Find the first line that satisfies a regex
}
\author{
Richèl J.C. Bilderbeek
}
