% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_first_xml_opening_tag_line.R
\name{find_first_xml_opening_tag_line}
\alias{find_first_xml_opening_tag_line}
\title{Find the line number of the first section's opening tag}
\usage{
find_first_xml_opening_tag_line(lines, section)
}
\arguments{
\item{lines}{the lines of an XML text}

\item{section}{the name of the XML section}
}
\value{
the line number's index (which is 1 for the first line) if the
  opening tag is found, else NA
}
\description{
Find the line number of the first section's opening tag
}
\author{
Richel J.C. Bilderbeek
}
