% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_last_regex_line.R
\name{find_last_regex_line}
\alias{find_last_regex_line}
\title{Find the index of the last line that matches a regex}
\usage{
find_last_regex_line(lines, regex)
}
\arguments{
\item{lines}{lines of text}

\item{regex}{regex string}
}
\value{
index of the line
}
\description{
Find the index of the last line that matches a regex
}
\author{
Richèl J.C. Bilderbeek
}
