% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_non_jc69_site_model.R
\name{find_non_jc69_site_model}
\alias{find_non_jc69_site_model}
\title{Finds the first non-JC69 site model from a list of one or more site models}
\usage{
find_non_jc69_site_model(site_models)
}
\arguments{
\item{site_models}{a list of one or more site models,
as returned by \code{\link{create_site_model}}}
}
\value{
the first non-JC69 site model, NULL if such a site model is
  absent
}
\description{
Finds the first non-JC69 site model from a list of one or more site models
}
\author{
Richel J.C. Bilderbeek
}
