% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_non_yule_tree_prior.R
\name{find_non_yule_tree_prior}
\alias{find_non_yule_tree_prior}
\title{Finds the first non-Yule tree prior from a list of one or more tree priors}
\usage{
find_non_yule_tree_prior(tree_priors)
}
\arguments{
\item{tree_priors}{a list of one or more tree priors,
as returned by \code{\link{create_tree_prior}}}
}
\value{
the first non-Yule tree prior, NULL if such a tree prior is
  absent
}
\description{
Finds the first non-Yule tree prior from a list of one or more tree priors
}
\author{
Richel J.C. Bilderbeek
}
