% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_site_model.R
\name{find_site_model}
\alias{find_site_model}
\title{Finds a site model with a certain ID}
\usage{
find_site_model(site_models, id)
}
\arguments{
\item{site_models}{a list of one or more site models,
as returned by \code{\link{create_site_model}}}

\item{id}{the ID of the site model}
}
\value{
the site models with the desired ID, NULL if such a site model is
  absent
}
\description{
Finds a site model with a certain ID
}
\author{
Richel J.C. Bilderbeek
}
