% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_param_to_xml.R
\name{freq_param_to_xml}
\alias{freq_param_to_xml}
\title{Internal function}
\usage{
freq_param_to_xml(freq_param, beauti_options = create_beauti_options())
}
\arguments{
\item{freq_param}{a `freq` parameter,
as created by \link{create_freq_param}}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the parameter as XML text
}
\description{
Converts a `freq` parameter to XML
}
\examples{
check_empty_beautier_folder()

# The freq parameter must be initialized, i.e. have an ID
freq_param_to_xml(freq_param = create_freq_param(id = "1"))

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
