% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_distr_to_xml.R
\name{gamma_distr_to_xml}
\alias{gamma_distr_to_xml}
\title{Internal function}
\usage{
gamma_distr_to_xml(gamma_distr, beauti_options = create_beauti_options())
}
\arguments{
\item{gamma_distr}{a gamma distribution,
as created by \code{\link{create_gamma_distr}})}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the distribution as XML text
}
\description{
Converts a gamma distribution to XML
}
\examples{
check_empty_beautier_folder()

# gamma distribution must be initialized
gamma_distr_to_xml(
  gamma_distr = create_gamma_distr(
    id = "0",
    alpha = create_alpha_param(id = "2", value = "0.5396"),
    beta = create_beta_param(id = "3", value = "0.3819")
  )
)

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
