% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_site_model_to_xml_prior_distr.R
\name{gamma_site_model_to_xml_prior_distr}
\alias{gamma_site_model_to_xml_prior_distr}
\title{Creates the gamma site models section in the distribution section
of a BEAST2 XML parameter file}
\usage{
gamma_site_model_to_xml_prior_distr(site_model)
}
\arguments{
\item{site_model}{a site_model, as created by \code{\link{create_site_model}}}
}
\description{
Creates the gamma site models section in the distribution section
of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
