% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_site_model_to_xml_state.R
\name{gamma_site_model_to_xml_state}
\alias{gamma_site_model_to_xml_state}
\title{Converts a gamma site model to XML,
  used in the \code{state} section}
\usage{
gamma_site_model_to_xml_state(gamma_site_model, id)
}
\arguments{
\item{gamma_site_model}{a gamma site model,
as created by \code{\link{create_gamma_site_model}})}

\item{id}{the site model's ID}
}
\value{
the gamma_site model as XML text
}
\description{
Converts a gamma site model to XML,
  used in the \code{state} section
}
\author{
Richèl J.C. Bilderbeek
}
