% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_site_models_to_xml_prior_distr.R
\name{gamma_site_models_to_xml_prior_distr}
\alias{gamma_site_models_to_xml_prior_distr}
\title{Deprecated function}
\usage{
gamma_site_models_to_xml_prior_distr(
  site_models,
  beauti_options = create_beauti_options()
)
}
\arguments{
\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
lines of XML text
}
\description{
Internal function to
creates the gamma site models section in the distribution section
of a BEAST2 XML parameter file
}
\author{
Richèl J.C. Bilderbeek
}
