% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bd_birth_rate_distr.R
\name{get_bd_birth_rate_distr}
\alias{get_bd_birth_rate_distr}
\title{Extract the birth rate distribution from a Birth-Death tree prior}
\usage{
get_bd_birth_rate_distr(bd_tree_prior)
}
\arguments{
\item{bd_tree_prior}{a Birth-Death tree prior, as created
by \code{\link{create_bd_tree_prior}}}
}
\value{
the birth rate distribution
}
\description{
Extract the birth rate distribution from a Birth-Death tree prior
}
\author{
Richel J.C. Bilderbeek
}
