% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bd_death_rate_distribution.R
\name{get_bd_death_rate_distr}
\alias{get_bd_death_rate_distr}
\title{Extract the death rate distribution from a death-Death tree prior}
\usage{
get_bd_death_rate_distr(bd_tree_prior)
}
\arguments{
\item{bd_tree_prior}{a Yule tree prior, as created
by \code{\link{create_bd_tree_prior}}}
}
\value{
the death rate distribution
}
\description{
Extract the death rate distribution from a death-Death tree prior
}
\author{
Richel J.C. Bilderbeek
}
