% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beautier_folder.R
\name{get_beautier_folder}
\alias{get_beautier_folder}
\title{Get the path to the \link{beautier} temporary files folder}
\usage{
get_beautier_folder()
}
\value{
the path to the \link{beautier} temporary files folder
}
\description{
Get the path to the \link{beautier} temporary files folder
}
\examples{
get_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
