% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ccp_pop_size_distr.R
\name{get_ccp_pop_size_distr}
\alias{get_ccp_pop_size_distr}
\title{Extract the Coalescent Contant Population tree prior's
population size distribution}
\usage{
get_ccp_pop_size_distr(ccp_tree_prior)
}
\arguments{
\item{ccp_tree_prior}{a ccp tree prior, as created
by \code{\link{create_ccp_tree_prior}}}
}
\value{
the birth rate distribution
}
\description{
Extract the Coalescent Contant Population tree prior's
population size distribution
}
\author{
Richel J.C. Bilderbeek
}
