% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cep_growth_rate_distr.R
\name{get_cep_growth_rate_distr}
\alias{get_cep_growth_rate_distr}
\title{Extract the Coalescent Exponential Population tree prior's
growth rate distribution}
\usage{
get_cep_growth_rate_distr(cep_tree_prior)
}
\arguments{
\item{cep_tree_prior}{a cep tree prior,
as returned by \code{\link{create_cep_tree_prior}}}
}
\value{
the growth rate distribution
}
\description{
Extract the Coalescent Exponential Population tree prior's
growth rate distribution
}
\author{
Richel J.C. Bilderbeek
}
