% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cep_pop_size_distr.R
\name{get_cep_pop_size_distr}
\alias{get_cep_pop_size_distr}
\title{Extract the Coalescent Exponential Population tree prior's
population size distribution}
\usage{
get_cep_pop_size_distr(cep_tree_prior)
}
\arguments{
\item{cep_tree_prior}{a cep tree prior, as created
by \code{\link{create_cep_tree_prior}}}
}
\value{
the birth rate distribution
}
\description{
Extract the Coalescent Exponential Population tree prior's
population size distribution
}
\author{
Richel J.C. Bilderbeek
}
