% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clock_model_default_id.R
\name{get_clock_model_default_id}
\alias{get_clock_model_default_id}
\title{Obtain the default clock model's ID}
\usage{
get_clock_model_default_id()
}
\description{
Obtain the default clock model's ID
}
