% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clock_model_name.R
\name{get_clock_model_name}
\alias{get_clock_model_name}
\title{Get the BEAUti name for a clock model}
\usage{
get_clock_model_name(clock_model)
}
\arguments{
\item{clock_model}{a clock model,
as returned by \code{\link{create_clock_model}}}
}
\description{
Get the BEAUti name for a clock model
}
\examples{
  strict <- create_strict_clock_model()
  testit::assert(beautier:::get_clock_model_name(strict) == "StrictClock")
  rln <- create_rln_clock_model()
  testit::assert(beautier:::get_clock_model_name(rln) == "RelaxedClock")
}
\author{
Richel J.C. Bilderbeek
}
