% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clock_model_names.R
\name{get_clock_model_names}
\alias{get_clock_model_names}
\title{Get the clock model names}
\usage{
get_clock_model_names()
}
\value{
the clock model names
}
\description{
Get the clock model names
}
\examples{
  names <- get_clock_model_names()
  testthat::expect_true("relaxed_log_normal" \%in\% names)
  testthat::expect_true("strict" \%in\% names)
}
\author{
Richel J.C. Bilderbeek
}
