% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clock_model_rate.R
\name{get_clock_model_rate}
\alias{get_clock_model_rate}
\title{Extract the rate from a clock model}
\usage{
get_clock_model_rate(clock_models)
}
\arguments{
\item{clock_models}{one or more clock_models}
}
\value{
the kappa
}
\description{
Extract the rate from a clock model
}
