% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clock_models_ids.R
\name{get_clock_models_ids}
\alias{get_clock_models_ids}
\title{Collect the IDs of the list of clock models}
\usage{
get_clock_models_ids(clock_models)
}
\arguments{
\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}
}
\value{
IDs of the clock models
}
\description{
Collect the IDs of the list of clock models
}
\author{
Richèl J.C. Bilderbeek
}
