% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_beast_namespace.R
\name{get_default_beast_namespace}
\alias{get_default_beast_namespace}
\title{Get the default `namespace` element value of the `beast` XML tag.}
\usage{
get_default_beast_namespace()
}
\value{
the default `namespace` element value of the `beast` XML tag.
}
\description{
Get the default `namespace` element value of the `beast` XML tag.
}
\examples{
get_default_beast_namespace()
}
\author{
Richèl J.C. Bilderbeek
}
