% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_beast_namespace.R
\name{get_default_beast_namespace_v2_4}
\alias{get_default_beast_namespace_v2_4}
\title{Get the default `namespace` element value of the `beast` XML tag
for BEAST 2.4}
\usage{
get_default_beast_namespace_v2_4()
}
\value{
the default `namespace` element value of the `beast` XML tag
for BEAST 2.4
}
\description{
Get the default `namespace` element value of the `beast` XML tag
for BEAST 2.4
}
\examples{
get_default_beast_namespace_v2_4()
}
\author{
Richèl J.C. Bilderbeek
}
