% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clock_model_rate.R
\name{get_default_clock_model_rate}
\alias{get_default_clock_model_rate}
\title{Get the default strict clock model rate.
Used in, among others, create_strict_clock_model}
\usage{
get_default_clock_model_rate()
}
\description{
Get the default strict clock model rate.
Used in, among others, create_strict_clock_model
}
\note{
the value is returned as a string, to be sure the testing XMLs
  are recreated exactly, as rounding errors are prevented
}
