% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_gamma_site_model.R
\name{get_default_gamma_site_model}
\alias{get_default_gamma_site_model}
\title{Create the deafult gamma site model}
\usage{
get_default_gamma_site_model()
}
\value{
a default gamma site model
}
\description{
Create the deafult gamma site model
}
\author{
Richel J.C. Bilderbeek
}
