% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kappa.R
\name{get_default_kappa}
\alias{get_default_kappa}
\title{Get the default kappa for the HKY nucleotide
substitution model. Use in, among others, create_hky_site_model}
\usage{
get_default_kappa()
}
\description{
Get the default kappa for the HKY nucleotide
substitution model. Use in, among others, create_hky_site_model
}
\note{
the value is returned as a string, to be sure the testing XMLs
  are recreated exactly, as rounding errors are prevented
}
