% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_mcmc_chain_length.R
\name{get_default_mcmc_chain_length}
\alias{get_default_mcmc_chain_length}
\title{Create the deafult gamma site model}
\usage{
get_default_mcmc_chain_length()
}
\value{
a default gamma site model
}
\description{
Create the deafult gamma site model
}
\author{
Richel J.C. Bilderbeek
}
