% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_invariant.R
\name{get_default_prop_invariant}
\alias{get_default_prop_invariant}
\title{Get the default proportion invariant for the HKY nucleotide
substitution model. Use in, among others, create_hky_site_model}
\usage{
get_default_prop_invariant()
}
\description{
Get the default proportion invariant for the HKY nucleotide
substitution model. Use in, among others, create_hky_site_model
}
\note{
this function returns a string to circumvent rounding
  errors in recreating the testing XMLs.
}
