% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distribution_id.R
\name{get_distr_id}
\alias{get_distr_id}
\title{Get the distribution's ID}
\usage{
get_distr_id(distribution)
}
\arguments{
\item{distribution}{a distibution,
as created by \code{\link{create_distr}})}
}
\value{
the distribution's ID
}
\description{
Get the distribution's ID
}
\author{
Richel J.C. Bilderbeek
}
