% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distr_n_params.R
\name{get_distr_n_params}
\alias{get_distr_n_params}
\title{Get the number of parameters a distribution uses}
\usage{
get_distr_n_params(distribution)
}
\arguments{
\item{distribution}{a distribution,
as created by \code{\link{create_distr}} or (preferable)
its named functions}
}
\value{
the number of parameters that distribution uses
}
\description{
Get the number of parameters a distribution uses
}
\author{
Richel J.C. Bilderbeek
}
