% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distribution_names.R
\name{get_distribution_names}
\alias{get_distribution_names}
\title{Get the distribution names}
\usage{
get_distribution_names()
}
\value{
the distribution names
}
\description{
Get the distribution names
}
\examples{
  names <- get_distribution_names()
  testthat::expect_true("uniform" \%in\% names)
  testthat::expect_true("normal" \%in\% names)
  testthat::expect_true("one_div_x" \%in\% names)
  testthat::expect_true("log_normal" \%in\% names)
  testthat::expect_true("exponential" \%in\% names)
}
\author{
Richel J.C. Bilderbeek
}
