% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_duplicate_ids.R
\name{get_duplicate_ids}
\alias{get_duplicate_ids}
\title{Find duplicate RealParameter IDs}
\usage{
get_duplicate_ids(text)
}
\arguments{
\item{text}{the XML as text}
}
\value{
a vector of duplicate IDs, will be empty if all IDs are unique
}
\description{
Find duplicate RealParameter IDs
}
\seealso{
to see if all IDs are unqiue, use \code{\link{has_unique_ids}}
}
\author{
Richel J.C. Bilderbeek
}
