% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_filenames.R
\name{get_fasta_filename}
\alias{get_fasta_filename}
\title{Get the path of a FASTA file used in testing}
\usage{
get_fasta_filename()
}
\value{
the path of a FASTA file used in testing
}
\description{
Get the path of a FASTA file used in testing
}
\examples{
  filename <- beautier::get_fasta_filename()
  testit::assert(file.exists(filename))

  create_beast2_input_file(
    input_filename = filename,
    "my_beast.xml"
  )
}
\author{
Richèl J.C. Bilderbeek
}
