% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_base_sans_ext.R
\name{get_file_base_sans_ext}
\alias{get_file_base_sans_ext}
\title{Get the base of the filename base without extension}
\usage{
get_file_base_sans_ext(filename)
}
\arguments{
\item{filename}{A filename}
}
\value{
That filename without its full path and extension
}
\description{
The path need not to actually exist
}
\examples{
# Path need not exist, use UNIX path as example
# test
get_file_base_sans_ext("/home/homer/test.txt")
}
\author{
Richèl J.C. Bilderbeek
}
