% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_freq_equilibrium_names.R
\name{get_freq_equilibrium_names}
\alias{get_freq_equilibrium_names}
\title{Returns valid values for the \code{freq_equilibrium} argument}
\usage{
get_freq_equilibrium_names()
}
\description{
Returns valid values for the \code{freq_equilibrium} argument
}
\examples{
  names <- beautier:::get_freq_equilibrium_names()
  testit::assert("estimated" \%in\% names)
  testit::assert("empirical" \%in\% names)
  testit::assert("all_equal" \%in\% names)
}
\seealso{
the \code{freq_equilibrium} argument is used in
  \code{\link{create_gtr_site_model}},
  \code{\link{create_hky_site_model}},
  and \code{\link{create_tn93_site_model}}
}
\author{
Richel J.C. Bilderbeek
}
