% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gamma_cat_count.R
\name{get_gamma_cat_count}
\alias{get_gamma_cat_count}
\title{Extract the gamma category count from an HKY site model}
\usage{
get_gamma_cat_count(gamma_site_models)
}
\arguments{
\item{site_models}{one or more gamma_site_models, as created
by \code{\link{create_gamma_site_model}}}
}
\value{
the gamma category count
}
\description{
Extract the gamma category count from an HKY site model
}
