% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gamma_shape.R
\name{get_gamma_shape}
\alias{get_gamma_shape}
\title{Extract the gamma shape from an gamma site model}
\usage{
get_gamma_shape(gamma_site_models)
}
\arguments{
\item{gamma_site_models}{one or more gamma_site_models, as created
by \code{\link{create_gamma_site_model}}}
}
\value{
the gamma category count
}
\description{
Extract the gamma shape from an gamma site model
}
