% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_has_non_strict_clock_model.R
\name{get_has_non_strict_clock_model}
\alias{get_has_non_strict_clock_model}
\title{Determines if there is at least one non-strict clock model
in the list of one or more clock models}
\usage{
get_has_non_strict_clock_model(clock_models)
}
\arguments{
\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}
}
\description{
Determines if there is at least one non-strict clock model
in the list of one or more clock models
}
\author{
Richel J.C. Bilderbeek
}
