% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{Conclude the ID from a FASTA filename}
\usage{
get_id(fasta_filename)
}
\arguments{
\item{fasta_filename}{name of a FASTA file}
}
\value{
the ID
}
\description{
Conclude the ID from a FASTA filename
}
\examples{
  testit::assert(get_id("anthus_aco.fas") == "Anthus_aco")
}
\author{
Richel J.C. Bilderbeek
}
