% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids.R
\name{get_ids}
\alias{get_ids}
\title{Conclude the IDs from one or more FASTA filenames}
\usage{
get_ids(fasta_filenames, capitalize_first_char_id = FALSE)
}
\arguments{
\item{fasta_filenames}{one or more FASTA filenames}

\item{capitalize_first_char_id}{capitalize the first character of the IDs}
}
\value{
the ID
}
\description{
Conclude the IDs from one or more FASTA filenames
}
\author{
Richel J.C. Bilderbeek
}
