% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kappa.R
\name{get_kappa}
\alias{get_kappa}
\title{Extract the kappa from an HKY site model}
\usage{
get_kappa(site_models)
}
\arguments{
\item{site_models}{one or more site_models}
}
\value{
the kappa
}
\description{
Extract the kappa from an HKY site model
}
