% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log_sorts.R
\name{get_log_sorts}
\alias{get_log_sorts}
\title{Get the possible log sorts}
\usage{
get_log_sorts()
}
\value{
the possible log sorts
}
\description{
Get the possible log sorts
}
\examples{
get_log_sorts()
}
\author{
Richèl J.C. Bilderbeek
}
