% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcmc_chain_length.R
\name{get_mcmc_chain_length}
\alias{get_mcmc_chain_length}
\title{Extract the MCMC chain length from an mcmc object.}
\usage{
get_mcmc_chain_length(mcmc)
}
\arguments{
\item{mcmc}{one mcmc, as created
by \code{\link{create_mcmc}}}
}
\value{
the MCMC chain length
}
\description{
Extract the MCMC chain length from an mcmc object.
}
\author{
Richel J.C. Bilderbeek
}
