% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_taxa.R
\name{get_n_taxa}
\alias{get_n_taxa}
\title{Extract the number of taxa from a file}
\usage{
get_n_taxa(filename)
}
\arguments{
\item{filename}{name of a FASTA file}
}
\value{
the number of taxa
}
\description{
Extract the number of taxa from a file
}
\examples{
library(testthat)

fasta_filename <- get_beautier_path("test_output_5.fas")
expect_equal(get_n_taxa(fasta_filename), 5)
}
\author{
Richèl J.C. Bilderbeek
}
