% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_operator_id_pre.R
\name{get_operator_id_pre}
\alias{get_operator_id_pre}
\title{Get the prefix of operator IDs}
\usage{
get_operator_id_pre(tree_priors)
}
\arguments{
\item{tree_priors}{tree priors, as created by create_tree_prior}
}
\description{
Get the prefix of operator IDs
}
\examples{
  bd_pre <- get_operator_id_pre(
    tree_priors = create_tree_prior(name = "birth_death")
  )
  testthat::expect_equal(bd_pre, "BirthDeath")
}
\author{
Richel J.C. Bilderbeek
}
