% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_filenames.R
\name{get_output_xml_filename}
\alias{get_output_xml_filename}
\title{Get the path of a BEAST2 XML parameter file used in testing}
\usage{
get_output_xml_filename()
}
\value{
the path of a BEAST2 XML parameter file used in testing
}
\description{
Get the path of a BEAST2 XML parameter file used in testing
}
\examples{
  my_file <- beastscriptr::get_output_xml_filename()
  does_exist <- file.exists(my_file)
  testit::assert(does_exist)
}

