% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_id.R
\name{get_param_id}
\alias{get_param_id}
\title{Get the parameter's ID}
\usage{
get_param_id(parameter)
}
\arguments{
\item{parameter}{a parameter,
as created by \code{\link{create_param}})}
}
\value{
the parameter's ID
}
\description{
Get the parameter's ID
}
\author{
Richel J.C. Bilderbeek
}
