% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_names.R
\name{get_param_names}
\alias{get_param_names}
\title{Get the parameter names}
\usage{
get_param_names()
}
\value{
the parameter names
}
\description{
Get the parameter names
}
\examples{
  names <- get_param_names()
  testthat::expect_true("alpha" \%in\% names)
  testthat::expect_true("beta" \%in\% names)
  testthat::expect_true("clock_rate" \%in\% names)
  testthat::expect_true("lambda" \%in\% names)
  testthat::expect_true("m" \%in\% names)
  testthat::expect_true("mean" \%in\% names)
  testthat::expect_true("mu" \%in\% names)
  testthat::expect_true("s" \%in\% names)
  testthat::expect_true("scale" \%in\% names)
  testthat::expect_true("sigma" \%in\% names)
}
\author{
Richel J.C. Bilderbeek
}
