% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameter_id.R
\name{get_parameter_id}
\alias{get_parameter_id}
\title{Get the parameter's ID}
\usage{
get_parameter_id(parameter)
}
\arguments{
\item{parameter}{a distibution,
as created by \code{\link{create_parameter}})}
}
\value{
the parameter's ID
}
\description{
Get the parameter's ID
}
\author{
Richel J.C. Bilderbeek
}
