% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameter_names.R
\name{get_parameter_names}
\alias{get_parameter_names}
\title{Get the parameter names}
\usage{
get_parameter_names()
}
\value{
the parameter names
}
\description{
Get the parameter names
}
\examples{
  names <- get_parameter_names()
  testthat::expect_true("alpha" \%in\% names)
  testthat::expect_true("beta" \%in\% names)
  testthat::expect_true("mu" \%in\% names)
  testthat::expect_true("scale" \%in\% names)
}
\author{
Richel J.C. Bilderbeek
}
