% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path.R
\name{get_path}
\alias{get_path}
\title{Get the full path of a file in the 'inst/extdata' folder}
\usage{
get_path(filename)
}
\arguments{
\item{filename}{the file's name, without the path}
}
\value{
the filename's full path
}
\description{
Get the full path of a file in the 'inst/extdata' folder
}
\author{
Richel J.C. Bilderbeek
}
