% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paths.R
\name{get_paths}
\alias{get_paths}
\title{Get the full paths of files in the 'inst/extdata' folder}
\usage{
get_paths(filenames)
}
\arguments{
\item{filenames}{the files' names, without the path}
}
\value{
the filenames' full path
}
\description{
Get the full paths of files in the 'inst/extdata' folder
}
\author{
Richel J.C. Bilderbeek
}
