% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phylo_crown_age.R
\name{get_phylo_crown_age}
\alias{get_phylo_crown_age}
\title{Obtain the crown age of a phylony.}
\usage{
get_phylo_crown_age(phylogeny)
}
\arguments{
\item{phylogeny}{The phylogeny to obtain the crown age of}
}
\value{
the crown age of the phylogeny
}
\description{
The crown age of a phylogeny is the time between
the present and the moment of at which the first
diversification (resulting in two lineages) happened.
}
\examples{
  phylogeny <- ape::read.tree(text = "(a:15,b:15):1;")
  created <- beautier:::get_phylo_crown_age(phylogeny = phylogeny)
  testit::assert(created == 15)
}
\author{
Richel J.C. Bilderbeek
}
