% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_invariant.R
\name{get_prop_invariant}
\alias{get_prop_invariant}
\title{Extract the proporion invariants from an HKY site model}
\usage{
get_prop_invariant(site_models)
}
\arguments{
\item{site_models}{one or more site_models}
}
\value{
the proporion invariants
}
\description{
Extract the proporion invariants from an HKY site model
}
