% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rln_ucldstdev_distribution.R
\name{get_rln_ucldstdev_distr}
\alias{get_rln_ucldstdev_distr}
\title{Extract the ucldstdev distribution a Relaxed
log-normal clock model}
\usage{
get_rln_ucldstdev_distr(rln_clock_model)
}
\arguments{
\item{rln_clock_model}{a Yule tree prior, as created
by \code{\link{create_rln_clock_model}}}
}
\value{
the uclstdev distribution
}
\description{
Extract the ucldstdev distribution a Relaxed
log-normal clock model
}
\author{
Richel J.C. Bilderbeek
}
