% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rln_ucldstdev_distribution.R
\name{get_rln_ucldstdev_distribution}
\alias{get_rln_ucldstdev_distribution}
\title{Extract the Yule birth rate distribution a Yule tree prior}
\usage{
get_rln_ucldstdev_distribution(rln_clock_model)
}
\arguments{
\item{rln_clock_model}{a Yule tree prior, as created
by \code{\link{create_rln_clock_model}}}
}
\value{
the uclstdev distribution
}
\description{
Extract the Yule birth rate distribution a Yule tree prior
}
\author{
Richel J.C. Bilderbeek
}
