% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_model_default_id.R
\name{get_site_model_default_id}
\alias{get_site_model_default_id}
\title{Obtain the default site model's ID}
\usage{
get_site_model_default_id()
}
\description{
Obtain the default site model's ID
}
