% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_model_names.R
\name{get_site_model_names}
\alias{get_site_model_names}
\title{Get the site model names}
\usage{
get_site_model_names()
}
\value{
the site model names
}
\description{
Get the site model names
}
\examples{
  names <- get_site_model_names()
  testthat::expect_true("JC69" \%in\% names)
  testthat::expect_true("HKY" \%in\% names)
  testthat::expect_true("TN93" \%in\% names)
  testthat::expect_true("GTR" \%in\% names)
}
\author{
Richel J.C. Bilderbeek
}
