% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_models_n_params.R
\name{get_site_models_n_params}
\alias{get_site_models_n_params}
\title{Get the number of distributions one or more site models have}
\usage{
get_site_models_n_params(site_models)
}
\arguments{
\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}
}
\value{
the number of parameters the site models have
}
\description{
Get the number of distributions one or more site models have
}
\examples{
check_empty_beautier_folder()

# Ten
get_site_models_n_params(list(create_gtr_site_model()))

# Two
get_site_models_n_params(list(create_hky_site_model()))

# Zero
get_site_models_n_params(list(create_jc69_site_model()))

# Four
get_site_models_n_params(list(create_tn93_site_model()))

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
