% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_prior_default_id.R
\name{get_tree_prior_default_id}
\alias{get_tree_prior_default_id}
\title{Obtain the default tree prior's ID}
\usage{
get_tree_prior_default_id()
}
\description{
Obtain the default tree prior's ID
}
