% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_prior_n_distrs.R
\name{get_tree_prior_n_distrs}
\alias{get_tree_prior_n_distrs}
\title{Get the number of distributions a tree prior has}
\usage{
get_tree_prior_n_distrs(tree_prior)
}
\arguments{
\item{tree_prior}{a tree priors,
as returned by \code{\link{create_tree_prior}}}
}
\value{
the number of distributions a tree prior has
}
\description{
Get the number of distributions a tree prior has
}
\examples{
 # birth_rate_distr and death_rate_distr
 testit::assert(
   beautier:::get_tree_prior_n_distrs(create_bd_tree_prior()) == 2
 )

 # none
 testit::assert(
   beautier:::get_tree_prior_n_distrs(create_cbs_tree_prior()) == 0
 )

 # pop_size_distr
 testit::assert(
   beautier:::get_tree_prior_n_distrs(create_ccp_tree_prior()) == 1
 )

 # pop_size_distr and growth_rate_distr
 testit::assert(
   beautier:::get_tree_prior_n_distrs(create_cep_tree_prior()) == 2
 )

 # birth_rate_distr
 testit::assert(
   beautier:::get_tree_prior_n_distrs(create_yule_tree_prior()) == 1
 )
}
\author{
Richel J.C. Bilderbeek
}
