% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_prior_names.R
\name{get_tree_prior_names}
\alias{get_tree_prior_names}
\title{Get the tree prior names}
\usage{
get_tree_prior_names()
}
\value{
the tree prior names
}
\description{
Get the tree prior names
}
\examples{
  names <- get_tree_prior_names()
  testthat::expect_true("birth_death" \%in\% names)
  testthat::expect_true("coalescent_bayesian_skyline" \%in\% names)
  testthat::expect_true("coalescent_constant_population" \%in\% names)
  testthat::expect_true("coalescent_exp_population" \%in\% names)
  testthat::expect_true("yule" \%in\% names)
}
\author{
Richel J.C. Bilderbeek
}
