% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_priors_n_distrs.R
\name{get_tree_priors_n_distrs}
\alias{get_tree_priors_n_distrs}
\title{Get the number of distributions a tree prior has}
\usage{
get_tree_priors_n_distrs(tree_priors)
}
\arguments{
\item{tree_prior}{a tree_prior,
as created by \code{\link{create_tree_prior}}}
}
\value{
the number of distributions a tree prior has
}
\description{
Get the number of distributions a tree prior has
}
\examples{
 testit::assert(
   get_tree_prior_n_distrs(
     list(
       create_bd_tree_prior(), # has two distributions
       create_ccp_tree_prior() # has one distribution
     )
   ) == 3)
}
\author{
Richel J.C. Bilderbeek
}
