% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_priors_n_params.R
\name{get_tree_priors_n_params}
\alias{get_tree_priors_n_params}
\title{Get the number of parameters a list of tree priors has}
\usage{
get_tree_priors_n_params(tree_priors)
}
\arguments{
\item{tree_priors}{one or more tree priors,
as returned by \code{\link{create_tree_prior}}}
}
\value{
the number of parameters the tree priors have
}
\description{
Get the number of parameters a list of tree priors has
}
\examples{
# Two
get_tree_priors_n_params(
  list(
    create_bd_tree_prior(), # zero
    create_cep_tree_prior() # two
  )
)
}
\author{
Richèl J.C. Bilderbeek
}
