% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unlinked_clock_models.R
\name{get_unlinked_clock_models}
\alias{get_unlinked_clock_models}
\title{Get the first clock model of each ID}
\usage{
get_unlinked_clock_models(clock_models)
}
\arguments{
\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}
}
\description{
Get the first clock model of each ID
}
\author{
Richel J.C. Bilderbeek
}
