% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unlinked_tree_priors.R
\name{get_unlinked_tree_priors}
\alias{get_unlinked_tree_priors}
\title{Get the first tree prior of each ID}
\usage{
get_unlinked_tree_priors(tree_priors)
}
\arguments{
\item{tree_priors}{a list of one or more tree priors}
}
\description{
Get the first tree prior of each ID
}
