% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yule_birth_rate_distribution.R
\name{get_yule_birth_rate_distr}
\alias{get_yule_birth_rate_distr}
\title{Extract the Yule birth rate distribution a Yule tree prior}
\usage{
get_yule_birth_rate_distr(yule_tree_prior)
}
\arguments{
\item{yule_tree_prior}{a Yule tree prior, as created
by \code{\link{create_yule_tree_prior}}}
}
\value{
the birth rate distribution
}
\description{
Extract the Yule birth rate distribution a Yule tree prior
}
\author{
Richel J.C. Bilderbeek
}
