% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_id.R
\name{getparam_id}
\alias{getparam_id}
\title{Get the parameter's ID}
\usage{
getparam_id(parameter)
}
\arguments{
\item{parameter}{a distibution,
as created by \code{\link{createparam}})}
}
\value{
the parameter's ID
}
\description{
Get the parameter's ID
}
\author{
Richel J.C. Bilderbeek
}
