% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_mrca_prior.R
\name{has_mrca_prior}
\alias{has_mrca_prior}
\title{Determines if the inference model has an MRCA prior}
\usage{
has_mrca_prior(inference_model)
}
\description{
Determines if the inference model has an MRCA prior
}
\author{
Richel J.C. Bilderbeek
}
