% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_mrca_prior_with_distr.R
\name{has_mrca_prior_with_distr}
\alias{has_mrca_prior_with_distr}
\title{See if the inference model has one MRCA prior with a distribution}
\usage{
has_mrca_prior_with_distr(inference_model)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\value{
TRUE if the inference model has one MRCA prior with a distribution,
  FALSE otherwise
}
\description{
See if the inference model has one MRCA prior with a distribution
}
\author{
Richèl J.C. Bilderbeek
}
