% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_unique_ids.R
\name{has_unique_ids}
\alias{has_unique_ids}
\title{Determine if the XML text has unique distribution IDs and unique
  parameter IDs}
\usage{
has_unique_ids(text)
}
\arguments{
\item{text}{the XML as text}
}
\value{
TRUE if all IDs are unique, FALSE otherwise
}
\description{
Determine if the XML text has unique distribution IDs and unique
  parameter IDs
}
\author{
Richel J.C. Bilderbeek
}
