% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_xml_closing_tag.R
\name{has_xml_closing_tag}
\alias{has_xml_closing_tag}
\title{Is an XML opening tag with value 'section' present amongst the lines of
  the text?}
\usage{
has_xml_closing_tag(lines, section)
}
\arguments{
\item{lines}{lines of the XML text}

\item{section}{the XML section}
}
\description{
Is an XML opening tag with value 'section' present amongst the lines of
  the text?
}
