% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_xml_short_closing_tag.R
\name{has_xml_short_closing_tag}
\alias{has_xml_short_closing_tag}
\title{Is an XML closing tag with short closing text at the end of the text?}
\usage{
has_xml_short_closing_tag(lines)
}
\arguments{
\item{lines}{lines of an XML text}
}
\description{
Is an XML closing tag with short closing text at the end of the text?
}
\examples{
  testit::assert(beautier:::has_xml_short_closing_tag("<my_tag id=1/>"))
  testit::assert(
    !beautier:::has_xml_short_closing_tag(
      "<my_tag id=1>text</my_tag>"
    )
  )
}
\author{
Richel J.C. Bilderbeek
}
