% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hky_site_model_to_xml_prior_distr.R
\name{hky_site_model_to_xml_prior_distr}
\alias{hky_site_model_to_xml_prior_distr}
\title{Internal function}
\usage{
hky_site_model_to_xml_prior_distr(site_model, beauti_options)
}
\arguments{
\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the site model as XML text
}
\description{
Converts an HKY site model to XML,
  used in the \code{prior} section
}
\examples{
hky_site_model_to_xml_prior_distr(
  site_model = create_hky_site_model(
    id = 1,
    kappa_prior_distr = create_uniform_distr(id = 2)
  ),
  beauti_options = create_beauti_options()
)
}
\author{
Richèl J.C. Bilderbeek
}
