% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent.R
\name{indent}
\alias{indent}
\title{Determine if x consists out of IDs}
\usage{
indent(text, n_spaces)
}
\arguments{
\item{text}{the text to indent}

\item{n_spaces}{the number of spaces to add before the text}
}
\value{
the indented text
}
\description{
Determine if x consists out of IDs
}
\author{
Richel J.C. Bilderbeek
}
