% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_tree_priors.R
\name{init_bd_tree_prior}
\alias{init_bd_tree_prior}
\title{Initializes a Birth-Death tree prior}
\usage{
init_bd_tree_prior(bd_tree_prior, distr_id, param_id)
}
\value{
an initialized Birth-Death tree prior
}
\description{
Initializes a Birth-Death tree prior
}
\author{
Richel J.C. Bilderbeek
}
