% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_distr.R
\name{init_beta_distr}
\alias{init_beta_distr}
\title{Initializes a beta distribution}
\usage{
init_beta_distr(beta_distr, distr_id = 0, param_id = 0)
}
\arguments{
\item{beta_distr}{a beta distribution,
using \link{create_beta_distr}}

\item{distr_id}{the first distribution's ID}

\item{param_id}{the first parameter's ID}
}
\value{
an initialized beta distribution
}
\description{
Initializes a beta distribution
}
\author{
Richèl J.C. Bilderbeek
}
