% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_tree_priors.R
\name{init_ccp_tree_prior}
\alias{init_ccp_tree_prior}
\title{Initializes a Coalescent Constant Population tree prior}
\usage{
init_ccp_tree_prior(ccp_tree_prior, distr_id, param_id)
}
\value{
an initialized Coalescent Constant Population tree prior
}
\description{
Initializes a Coalescent Constant Population tree prior
}
\author{
Richel J.C. Bilderbeek
}
