% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_clock_models.R
\name{init_clock_models}
\alias{init_clock_models}
\title{Initializes all clock models}
\usage{
init_clock_models(fasta_filenames, clock_models, distr_id = 0, param_id = 0)
}
\arguments{
\item{fasta_filenames}{the one ore more names of the FASTA files that
contain the one or more alignments}

\item{clock_models}{a list of one or more clock models to be initialized.
Clock priors can be created using \code{\link{create_clock_model}}}

\item{distr_id}{the first distributions' ID}

\item{param_id}{the first parameter's ID}
}
\value{
a list of initialized clock models
}
\description{
Initializes all clock models
}
\author{
Richel J.C. Bilderbeek
}
