% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_distr.R
\name{init_distr}
\alias{init_distr}
\title{Initializes a distribution}
\usage{
init_distr(distr, distr_id = 0, param_id = 0)
}
\arguments{
\item{distr}{a distribution,
using \code{\link{create_distr}}}

\item{distr_id}{the first distribution's ID}

\item{param_id}{the first parameter's ID}
}
\value{
an initialized distribution
}
\description{
Initializes a distribution
}
\author{
Richèl J.C. Bilderbeek
}
