% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_distr.R
\name{init_exp_distr}
\alias{init_exp_distr}
\title{Initializes an exponential distribution}
\usage{
init_exp_distr(exp_distr, distr_id = 0, param_id = 0)
}
\arguments{
\item{exp_distr}{a exponential distribution,
using \link{create_exp_distr}}

\item{distr_id}{the first distribution's ID}

\item{param_id}{the first parameter's ID}
}
\value{
an initialized exponential distribution
}
\description{
Initializes an exponential distribution
}
\author{
Richèl J.C. Bilderbeek
}
