% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_distr.R
\name{init_gamma_distr}
\alias{init_gamma_distr}
\title{Initializes a gamma distribution}
\usage{
init_gamma_distr(gamma_distr, distr_id = 0, param_id = 0)
}
\arguments{
\item{gamma_distr}{a gamma distribution,
using \link{create_gamma_distr}}

\item{distr_id}{the first distribution's ID}

\item{param_id}{the first parameter's ID}
}
\value{
an initialized gamma distribution
}
\description{
Initializes a gamma distribution
}
\author{
Richèl J.C. Bilderbeek
}
