% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_site_models.R
\name{init_gtr_site_model}
\alias{init_gtr_site_model}
\title{Initializes a GTR site model}
\usage{
init_gtr_site_model(gtr_site_model, distr_id, param_id)
}
\arguments{
\item{gtr_site_model}{a GTR site model,
as returned by \code{\link{create_gtr_site_model}}}

\item{distr_id}{the first distributions' ID}

\item{param_id}{the first parameter's ID}
}
\value{
an initialized GTR site model
}
\description{
Initializes a GTR site model
}
\author{
Richel J.C. Bilderbeek
}
