% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_inference_model.R
\name{init_inference_model}
\alias{init_inference_model}
\title{Initialize an inference model}
\usage{
init_inference_model(input_filename, inference_model)
}
\arguments{
\item{input_filename}{A FASTA filename.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}

\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\description{
Initialize an inference model
}
