% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_distr.R
\name{init_inv_gamma_distr}
\alias{init_inv_gamma_distr}
\title{Initializes an inverse gamma distribution}
\usage{
init_inv_gamma_distr(inv_gamma_distr, distr_id = 0, param_id = 0)
}
\arguments{
\item{inv_gamma_distr}{an inverse gamma distribution,
using \link{create_inv_gamma_distr}}

\item{distr_id}{the first distribution's ID}

\item{param_id}{the first parameter's ID}
}
\value{
an initialized inverse gamma distribution
}
\description{
Initializes an inverse gamma distribution
}
\author{
Richèl J.C. Bilderbeek
}
