% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_site_models.R
\name{init_jc69_site_model}
\alias{init_jc69_site_model}
\title{Initializes a JC69 site model}
\usage{
init_jc69_site_model(jc69_site_model, distr_id = 0, param_id = 0)
}
\value{
an initialized HKY site model
}
\description{
Initializes a JC69 site model
}
\examples{
  hky_site_model <- create_hky_site_model()
  testit::assert(!beautier:::is_init_hky_site_model(hky_site_model))
  hky_site_model <- beautier:::init_hky_site_model(hky_site_model)
  testit::assert(beautier:::is_init_hky_site_model(hky_site_model))
}
\author{
Richel J.C. Bilderbeek
}
