% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_distr.R
\name{init_laplace_distr}
\alias{init_laplace_distr}
\title{Initializes an Laplace distribution}
\usage{
init_laplace_distr(laplace_distr, distr_id = 0, param_id = 0)
}
\arguments{
\item{laplace_distr}{a Laplace distribution,
using \link{create_laplace_distr}}

\item{distr_id}{the first distribution's ID}

\item{param_id}{the first parameter's ID}
}
\value{
an initialized Laplace distribution
}
\description{
Initializes an Laplace distribution
}
\author{
Richèl J.C. Bilderbeek
}
