% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_distr.R
\name{init_log_normal_distr}
\alias{init_log_normal_distr}
\title{Initializes an log-normal distribution}
\usage{
init_log_normal_distr(log_normal_distr, distr_id = 0, param_id = 0)
}
\arguments{
\item{log_normal_distr}{a log-normal distribution,
using \link{create_log_normal_distr}}

\item{distr_id}{the first distribution's ID}

\item{param_id}{the first parameter's ID}
}
\value{
an initialized log-normal distribution
}
\description{
Initializes an log-normal distribution
}
\author{
Richèl J.C. Bilderbeek
}
