% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_distr.R
\name{init_normal_distr}
\alias{init_normal_distr}
\title{Initializes an normal distribution}
\usage{
init_normal_distr(normal_distr, distr_id = 0, param_id = 0)
}
\arguments{
\item{normal_distr}{a normal distribution,
using \link{create_normal_distr}}

\item{distr_id}{the first distribution's ID}

\item{param_id}{the first parameter's ID}
}
\value{
an initialized normal distribution
}
\description{
Initializes an normal distribution
}
\author{
Richèl J.C. Bilderbeek
}
