% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_distr.R
\name{init_one_div_x_distr}
\alias{init_one_div_x_distr}
\title{Initializes an one-divided-by-x distribution}
\usage{
init_one_div_x_distr(one_div_x_distr, distr_id = 0)
}
\arguments{
\item{one_div_x_distr}{a one-divided-by-x distribution,
using \link{create_one_div_x_distr}}

\item{distr_id}{the first distribution's ID}
}
\value{
an initialized one-divided-by-x distribution
}
\description{
Initializes an one-divided-by-x distribution
}
\author{
Richèl J.C. Bilderbeek
}
