% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_param.R
\name{init_param}
\alias{init_param}
\title{Initializes a parameter}
\usage{
init_param(param, id)
}
\arguments{
\item{param}{a parameter,
using \code{\link{create_param}}}

\item{id}{the parameter's ID. Will be ignored if the parameter already
has an ID}
}
\value{
an initialized parameter
}
\description{
Initializes a parameter
}
\author{
Richèl J.C. Bilderbeek
}
