% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_distr.R
\name{init_poisson_distr}
\alias{init_poisson_distr}
\title{Initializes an Poisson distribution}
\usage{
init_poisson_distr(poisson_distr, distr_id = 0, param_id = 0)
}
\arguments{
\item{poisson_distr}{a Poisson distribution,
using \link{create_poisson_distr}}

\item{distr_id}{the first distribution's ID}

\item{param_id}{the first parameter's ID}
}
\value{
an initialized Poisson distribution
}
\description{
Initializes an Poisson distribution
}
\author{
Richèl J.C. Bilderbeek
}
