% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_clock_models.R
\name{init_rln_clock_model}
\alias{init_rln_clock_model}
\title{Initializes a Relaxed Log-Normal clock model}
\usage{
init_rln_clock_model(rln_clock_model, distr_id, param_id)
}
\arguments{
\item{rln_clock_model}{a Relaxed Log-Normal clock model,
as returned by \code{\link{create_rln_clock_model}}}

\item{distr_id}{a distributions' ID}

\item{param_id}{a parameter's ID}
}
\value{
an initialized Relaxed Log-Normal clock model
}
\description{
Initializes a Relaxed Log-Normal clock model
}
\author{
Richel J.C. Bilderbeek
}
