% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_site_models.R
\name{init_site_models}
\alias{init_site_models}
\title{Initializes all site models}
\usage{
init_site_models(site_models, ids, distr_id = 0, param_id = 0)
}
\arguments{
\item{site_models}{a list of one or more site models to be initialized.
Clock priors can be created using \code{\link{create_site_model}}}

\item{ids}{the IDs of the alignments (can be extracted from
their FASTA filesnames using \code{\link{get_ids}})}

\item{distr_id}{the first distributions' ID}

\item{param_id}{the first parameter's ID}
}
\value{
a list of initialized site models
}
\description{
Initializes all site models
}
\author{
Richel J.C. Bilderbeek
}
