% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_clock_models.R
\name{init_strict_clock_model}
\alias{init_strict_clock_model}
\title{Initializes a strict clock model}
\usage{
init_strict_clock_model(strict_clock_model, id)
}
\arguments{
\item{strict_clock_model}{a strict clock model,
as returned by \code{\link{create_strict_clock_model}}}

\item{id}{an alignment's IDs.
An ID can be extracted from its FASTA filesname
with \code{\link{get_ids}})}
}
\value{
an initialized strict clock model
}
\description{
Initializes a strict clock model
}
\author{
Richel J.C. Bilderbeek
}
