% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_site_models.R
\name{init_strict_site_model}
\alias{init_strict_site_model}
\title{Initializes a strict site model}
\usage{
init_strict_site_model(strict_site_model, id)
}
\arguments{
\item{strict_site_model}{a strict site model,
as returned by \code{\link{create_strict_site_model}}}

\item{id}{the ID of the alignments this site model is associated with,
as can be extracted from its FASTA filesnames using \code{\link{get_id}})}
}
\value{
an initialized strict site model
}
\description{
Initializes a strict site model
}
\author{
Richel J.C. Bilderbeek
}
