% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_site_models.R
\name{init_tn93_site_model}
\alias{init_tn93_site_model}
\title{Initializes a TN93 site model}
\usage{
init_tn93_site_model(tn93_site_model, distr_id, param_id)
}
\arguments{
\item{tn93_site_model}{a TN93 site model,
as returned by \code{\link{create_tn93_site_model}}}

\item{distr_id}{the first distributions' ID}

\item{param_id}{the first parameter's ID}
}
\value{
an initialized TN93 site model
}
\description{
Initializes a TN93 site model
}
\author{
Richel J.C. Bilderbeek
}
