% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_distr.R
\name{init_uniform_distr}
\alias{init_uniform_distr}
\title{Initializes a uniform distribution}
\usage{
init_uniform_distr(uniform_distr, distr_id = 0)
}
\arguments{
\item{uniform_distr}{a uniform distribution,
using \link{create_uniform_distr}}

\item{distr_id}{the first distribution's ID}
}
\value{
an initialized uniform distribution
}
\description{
Initializes a uniform distribution
}
\author{
Richèl J.C. Bilderbeek
}
