% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_tree_priors.R
\name{init_yule_tree_prior}
\alias{init_yule_tree_prior}
\title{Initializes a Yule tree prior}
\usage{
init_yule_tree_prior(yule_tree_prior, distr_id, param_id)
}
\value{
an initialized Yule tree prior
}
\description{
Initializes a Yule tree prior
}
\author{
Richel J.C. Bilderbeek
}
