% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_tree_priors.R
\name{initialize_bd_tree_prior}
\alias{initialize_bd_tree_prior}
\title{Initializes a Birth-Death tree prior}
\usage{
initialize_bd_tree_prior(bd_tree_prior, distr_id, param_id)
}
\arguments{
\item{bd_tree_prior}{a Birth-Death tree prior,
as returned by \code{\link{create_bd_tree_prior}}}

\item{distr_id}{the first distributions' ID}

\item{param_id}{the first parameter's ID}
}
\value{
an initialized Birth-Death tree prior
}
\description{
Initializes a Birth-Death tree prior
}
\author{
Richel J.C. Bilderbeek
}
