% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_tree_priors.R
\name{initialize_ccp_tree_prior}
\alias{initialize_ccp_tree_prior}
\title{Initializes a Coalescent Constant Population tree prior}
\usage{
initialize_ccp_tree_prior(ccp_tree_prior, distr_id, param_id)
}
\arguments{
\item{ccp_tree_prior}{a Coalescent Constant Population tree prior,
as returned by \code{\link{create_ccp_tree_prior}}}

\item{id}{the index of the first distribution}
}
\value{
an initialized Coalescent Constant Population tree prior
}
\description{
Initializes a Coalescent Constant Population tree prior
}
\author{
Richel J.C. Bilderbeek
}
