% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_tree_priors.R
\name{initialize_cep_tree_prior}
\alias{initialize_cep_tree_prior}
\title{Initializes a Coalescent Exponential Population tree prior}
\usage{
initialize_cep_tree_prior(cep_tree_prior, distr_id, param_id)
}
\arguments{
\item{cep_tree_prior}{a Coalescent Exponential Population tree prior,
as returned by \code{\link{create_cep_tree_prior}}}
}
\value{
an initialized Coalescent Exponential Population tree prior
}
\description{
Initializes a Coalescent Exponential Population tree prior
}
\author{
Richel J.C. Bilderbeek
}
