% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_clock_models.R
\name{initialize_clock_models}
\alias{initialize_clock_models}
\title{Initializes all clock models}
\usage{
initialize_clock_models(clock_models, id = 0)
}
\arguments{
\item{clock_models}{a list of one or more clock models to be initialized.
Clock priors can be created using \code{\link{create_clock_model}}}

\item{id}{the first distributions' ID}
}
\value{
a list of initialized clock models
}
\description{
Initializes all clock models
}
\author{
Richel J.C. Bilderbeek
}
