% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_distr.R
\name{initialize_distr}
\alias{initialize_distr}
\title{Initializes a distribution}
\usage{
initialize_distr(distribution, distr_id = 0, param_id = 0)
}
\arguments{
\item{distribution}{a distribution,
using \code{\link{create_distribution}}}

\item{distr_id}{the first distribution's ID}

\item{param_id}{the first parameter's ID}
}
\value{
an initialized distribution
}
\description{
Initializes a distribution
}
\author{
Richel J.C. Bilderbeek
}
