% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_clock_models.R
\name{initialize_rln_clock_model}
\alias{initialize_rln_clock_model}
\title{Initializes a Birth-Death clock model}
\usage{
initialize_rln_clock_model(rln_clock_model, id)
}
\arguments{
\item{rln_clock_model}{a Birth-Death clock model,
as returned by \code{\link{create_rln_clock_model}}}

\item{id}{the index of the first distribution}
}
\value{
an initialized Birth-Death clock model
}
\description{
Initializes a Birth-Death clock model
}
\author{
Richel J.C. Bilderbeek
}
