% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_clock_models.R
\name{initialize_strict_clock_model}
\alias{initialize_strict_clock_model}
\title{Initializes a strict clock model}
\usage{
initialize_strict_clock_model(strict_clock_model, id)
}
\arguments{
\item{strict_clock_model}{a strict clock model,
as returned by \code{\link{create_strict_clock_model}}}

\item{id}{the ID of the alignments this clock model is associated with,
as can be extracted from its FASTA filesnames using \code{\link{get_id}})}
}
\value{
an initialized strict clock model
}
\description{
Initializes a strict clock model
}
\author{
Richel J.C. Bilderbeek
}
