% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_tree_priors.R
\name{initialize_tree_priors}
\alias{initialize_tree_priors}
\title{Initializes all tree priors}
\usage{
initialize_tree_priors(tree_priors)
}
\arguments{
\item{tree_priors}{the one or more tree priors to be initialized,
as returned by \code{\link{create_tree_prior}}}
}
\value{
initialized tree priors
}
\description{
Initializes all tree priors
}
\author{
Richel J.C. Bilderbeek
}
