% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_tree_priors.R
\name{initialize_yule_tree_prior}
\alias{initialize_yule_tree_prior}
\title{Initializes a Yule tree prior}
\usage{
initialize_yule_tree_prior(yule_tree_prior, id)
}
\arguments{
\item{yule_tree_prior}{a Yule tree prior,
as returned by \code{\link{create_yule_tree_prior}}}

\item{id}{the index of the first distribution}
}
\value{
an initialized Yule tree prior
}
\description{
Initializes a Yule tree prior
}
\author{
Richel J.C. Bilderbeek
}
