% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_alpha_parameter.R
\name{is_alpha_parameter}
\alias{is_alpha_parameter}
\title{Determine if the object is a valid
alpha parameter}
\usage{
is_alpha_parameter(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
alpha parameter}
}
\value{
TRUE if x is a valid alpha parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid
alpha parameter
}
\author{
Richel J.C. Bilderbeek
}
