% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_beast2_input_file_with_tipdates.R
\name{is_beast2_input_file_with_tipdates}
\alias{is_beast2_input_file_with_tipdates}
\title{Determine if the file is a BEAST2 input file that has tip dating}
\usage{
is_beast2_input_file_with_tipdates(beast2_input_file)
}
\value{
TRUE if x is the BEAST2 input file has tip dating, FALSE otherwise
}
\description{
Determine if the file is a BEAST2 input file that has tip dating
}
\examples{
check_empty_beautier_folder()

is_beast2_input_file_with_tipdates(
  get_beautier_path("babette_issue_109_expected_v2_6.xml")
) # TRUE

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
