% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_beta_distribution.R
\name{is_beta_distribution}
\alias{is_beta_distribution}
\title{Determine if the object is a valid
beta distribution}
\usage{
is_beta_distribution(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
beta distribution}
}
\value{
TRUE if x is a valid beta distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid
beta distribution
}
\seealso{
use \code{\link{is_distribution}} to see if x is any
  distribution
}
\author{
Richel J.C. Bilderbeek
}
