% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_beta_param.R
\name{is_beta_param}
\alias{is_beta_param}
\title{Determine if the object is a valid
beta parameter}
\usage{
is_beta_param(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
beta parameter}
}
\value{
TRUE if x is a valid beta parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid
beta parameter
}
\author{
Richel J.C. Bilderbeek
}
