% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cep_tree_prior.R
\name{is_cep_tree_prior}
\alias{is_cep_tree_prior}
\title{Determine if the object is a valid
coalescent exponential population tree prior
  as returned by \code{\link{create_cep_tree_prior}}}
\usage{
is_cep_tree_prior(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
constant coalescent exponential population tree prior}
}
\value{
TRUE if x is a valid coalescent exponentialpopulation tree prior,
  FALSE otherwise
}
\description{
Determine if the object is a valid
coalescent exponential population tree prior
  as returned by \code{\link{create_cep_tree_prior}}
}
\author{
Richel J.C. Bilderbeek
}
