% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_clock_model.R
\name{is_clock_model}
\alias{is_clock_model}
\title{Determine if the object is a valid clock_model}
\usage{
is_clock_model(x)
}
\arguments{
\item{x}{an object, to be determined if it is a clock_model}
}
\value{
TRUE if the clock_model is a valid clock_model, FALSE otherwise
}
\description{
Determine if the object is a valid clock_model
}
\examples{
check_empty_beautier_folder()

# TRUE
is_clock_model(create_strict_clock_model())
is_clock_model(create_rln_clock_model())

# FALSE
is_clock_model(NA)
is_clock_model(NULL)
is_clock_model("nonsense")
is_clock_model(create_jc69_site_model())
is_clock_model(create_mcmc())

check_empty_beautier_folder()
}
\seealso{
see \code{\link{create_clock_model}} for an overview of functions
  to create valid clock model
}
\author{
Richèl J.C. Bilderbeek
}
