% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_clock_model_name.R
\name{is_clock_model_name}
\alias{is_clock_model_name}
\title{Determines if the name is a valid clock model name}
\usage{
is_clock_model_name(name)
}
\arguments{
\item{name}{the name to be tested}
}
\value{
TRUE if the name is a valid clock_model name, FALSE otherwise
}
\description{
Determines if the name is a valid clock model name
}
\examples{
# TRUE
is_clock_model_name("relaxed_log_normal")
is_clock_model_name("strict")
}
\author{
Richèl J.C. Bilderbeek
}
