% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_clock_rate_param.R
\name{is_clock_rate_param}
\alias{is_clock_rate_param}
\title{Determine if the object is a valid
clock_rate parameter}
\usage{
is_clock_rate_param(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
clock_rate parameter}
}
\value{
TRUE if x is a valid clock_rate parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid
clock_rate parameter
}
\author{
Richel J.C. Bilderbeek
}
