% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_distr_name.R
\name{is_distr_name}
\alias{is_distr_name}
\title{Determines if the name is a valid distribution name}
\usage{
is_distr_name(name)
}
\arguments{
\item{name}{the name to be tested}
}
\value{
TRUE if the name is a valid distribution name, FALSE otherwise
}
\description{
Determines if the name is a valid distribution name
}
