% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_distr.R
\name{is_exp_distr}
\alias{is_exp_distr}
\title{Determine if the object is a valid
exponential distribution
as created by \code{\link{create_exp_distr}}}
\usage{
is_exp_distr(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
exponential distribution}
}
\value{
TRUE if x is a valid exponential distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid
exponential distribution
as created by \code{\link{create_exp_distr}}
}
\examples{
check_empty_beautier_folder()

# TRUE
is_exp_distr(create_exp_distr())
# FALSE
is_exp_distr(create_gamma_distr())
is_exp_distr(NA)
is_exp_distr(NULL)
is_exp_distr("nonsense")

check_empty_beautier_folder()
}
\seealso{
use \code{\link{is_distr}} to see if x is any
  distribution
}
\author{
Richèl J.C. Bilderbeek
}
