% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_exponential_distribution.R
\name{is_exponential_distribution}
\alias{is_exponential_distribution}
\title{Determine if the object is a valid
exponential distribution}
\usage{
is_exponential_distribution(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
exponential distribution}
}
\value{
TRUE if x is a valid exponential distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid
exponential distribution
}
\seealso{
use \code{\link{is_distribution}} to see if x is any
  distribution
}
\author{
Richel J.C. Bilderbeek
}
