% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gamma_site_model.R
\name{is_gamma_site_model}
\alias{is_gamma_site_model}
\title{Is object x a gamma site model?}
\usage{
is_gamma_site_model(x)
}
\arguments{
\item{x}{the object to be determined if it is a valid gamma site object}
}
\value{
TRUE if x is a valid gamma site object, FALSE otherwise
}
\description{
Is object x a gamma site model?
}
\examples{
library(testthat)

expect_true(is_gamma_site_model(create_gamma_site_model()))

expect_false(is_gamma_site_model("nonsense"))
expect_false(is_gamma_site_model(NA))
expect_false(is_gamma_site_model(NULL))
expect_false(is_gamma_site_model(""))
expect_false(is_gamma_site_model(c()))
}
\author{
Richèl J.C. Bilderbeek
}
