% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gtr_site_model.R
\name{is_gtr_site_model}
\alias{is_gtr_site_model}
\title{Determine if the object is a valid GTR site model}
\usage{
is_gtr_site_model(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid GTR site model}
}
\value{
TRUE if x is a valid GTR site model, FALSE otherwise
}
\description{
Determine if the object is a valid GTR site model
}
\author{
Richel J.C. Bilderbeek
}
