% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_hky_site_model.R
\name{is_hky_site_model}
\alias{is_hky_site_model}
\title{Determine if the object is a valid HKY site model,
as created by \code{\link{create_hky_site_model}}}
\usage{
is_hky_site_model(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid HKY site model}
}
\value{
TRUE if x is a valid HKY site model, FALSE otherwise
}
\description{
Determine if the object is a valid HKY site model,
as created by \code{\link{create_hky_site_model}}
}
\examples{
  hky_site_model <- create_hky_site_model()
  testit::assert(beautier:::is_hky_site_model(hky_site_model))
}
\author{
Richel J.C. Bilderbeek
}
