% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_in_patterns.R
\name{is_in_patterns}
\alias{is_in_patterns}
\title{Is there at least one regular expression having a match with the line?}
\usage{
is_in_patterns(line, patterns)
}
\arguments{
\item{line}{a line of text}

\item{patterns}{one or more regular expression patterns}
}
\value{
TRUE if there is at least one match found
}
\description{
Is there at least one regular expression having a match with the line?
}
\author{
Richèl J.C. Bilderbeek
}
