% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_index_of_non_first_shared_clock_model.R
\name{is_index_of_non_first_shared_clock_model}
\alias{is_index_of_non_first_shared_clock_model}
\title{Is the index the first of a series of clock models that are shared?}
\usage{
is_index_of_non_first_shared_clock_model(i, clock_models)
}
\arguments{
\item{i}{the index of the clock model within \code{clock_models}}

\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}
}
\description{
Is the index the first of a series of clock models that are shared?
}
\author{
Richel J.C. Bilderbeek
}
