% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_inference_model.R
\name{is_inference_model}
\alias{is_inference_model}
\title{Determine if the input is an inference model}
\usage{
is_inference_model(x)
}
\arguments{
\item{x}{object to be determined of if it is an inference model}
}
\value{
TRUE if the object is an inference model
}
\description{
Determine if the input is an inference model
}
