% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_tree_prior.R
\name{is_init_cep_tree_prior}
\alias{is_init_cep_tree_prior}
\title{Determine if x is an initialized Coalescent Exponential Population
  tree_prior object}
\usage{
is_init_cep_tree_prior(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized Coalescent Exponential Population tree prior object}
}
\value{
TRUE if x is an initialized Coalescent Exponential Population
  tree prior object
}
\description{
Determine if x is an initialized Coalescent Exponential Population
  tree_prior object
}
\author{
Richel J.C. Bilderbeek
}
