% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_clock_model.R
\name{is_init_clock_model}
\alias{is_init_clock_model}
\title{Determine if x is an initialized clock_model object,
as created by \code{\link{create_clock_model}}}
\usage{
is_init_clock_model(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized clock_models object}
}
\value{
TRUE if x is an initialized clock_model object
}
\description{
Determine if x is an initialized clock_model object,
as created by \code{\link{create_clock_model}}
}
\author{
Richel J.C. Bilderbeek
}
