% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_distr.R
\name{is_init_exp_distr}
\alias{is_init_exp_distr}
\title{Determine if x is an initialized exponential distribution object
  as created by \code{\link{create_exp_distr}}}
\usage{
is_init_exp_distr(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized exponential distribution object}
}
\value{
TRUE if x is an initialized exponential distribution object
}
\description{
Determine if x is an initialized exponential distribution object
  as created by \code{\link{create_exp_distr}}
}
\author{
Richèl J.C. Bilderbeek
}
