% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_distr.R
\name{is_init_gamma_distr}
\alias{is_init_gamma_distr}
\title{Determine if x is an initialized gamma distribution object}
\usage{
is_init_gamma_distr(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized gamma distribution object}
}
\value{
TRUE if x is an initialized gamma distribution object
}
\description{
Determine if x is an initialized gamma distribution object
}
\author{
Richel J.C. Bilderbeek
}
