% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_gamma_site_model.R
\name{is_init_gamma_site_model}
\alias{is_init_gamma_site_model}
\title{Determine if x is an initialized gamma site model,
as created by \code{\link{create_gamma_site_model}}}
\usage{
is_init_gamma_site_model(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized gamma site_models object}
}
\value{
TRUE if x is an initialized gamma site model
}
\description{
Determine if x is an initialized gamma site model,
as created by \code{\link{create_gamma_site_model}}
}
\author{
Richel J.C. Bilderbeek
}
