% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_site_model.R
\name{is_init_gtr_site_model}
\alias{is_init_gtr_site_model}
\title{Determine if x is an initialized GTR site model
as created by \code{\link{create_gtr_site_model}}}
\usage{
is_init_gtr_site_model(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized GTR site model}
}
\value{
TRUE if x is an initialized GTR site model
}
\description{
Determine if x is an initialized GTR site model
as created by \code{\link{create_gtr_site_model}}
}
\examples{
library(testthat)

gtr_site_model <- create_gtr_site_model()
expect_false(is_init_gtr_site_model(gtr_site_model))
gtr_site_model <- init_gtr_site_model(gtr_site_model)
expect_true(is_init_gtr_site_model(gtr_site_model))
}
\author{
Richèl J.C. Bilderbeek
}
