% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_distr.R
\name{is_init_log_normal_distr}
\alias{is_init_log_normal_distr}
\title{Determine if x is an initialized log_normal distribution object
  as created by \code{\link{create_log_normal_distr}}}
\usage{
is_init_log_normal_distr(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized log_normal distribution object}
}
\value{
TRUE if x is an initialized log_normal distribution object
}
\description{
Determine if x is an initialized log_normal distribution object
  as created by \code{\link{create_log_normal_distr}}
}
\author{
Richèl J.C. Bilderbeek
}
