% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_mrca_prior.R
\name{is_init_mrca_prior}
\alias{is_init_mrca_prior}
\title{Determine if x is an initialized mrca_prior objects}
\usage{
is_init_mrca_prior(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized mrca_priors object}
}
\value{
TRUE if x is an initialized mrca_prior object
}
\description{
Determine if x is an initialized mrca_prior objects
}
\author{
Richel J.C. Bilderbeek
}
