% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_distr.R
\name{is_init_one_div_x_distr}
\alias{is_init_one_div_x_distr}
\title{Determine if x is an initialized one_div_x distribution object
  as created by \code{\link{create_one_div_x_distr}}}
\usage{
is_init_one_div_x_distr(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized one_div_x distribution object}
}
\value{
TRUE if x is an initialized one_div_x distribution object
}
\description{
Determine if x is an initialized one_div_x distribution object
  as created by \code{\link{create_one_div_x_distr}}
}
\author{
Richel J.C. Bilderbeek
}
