% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_param.R
\name{is_init_param}
\alias{is_init_param}
\title{Determine if x is an initialized parameter,
  as created by \link{create_param}}
\usage{
is_init_param(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized parameter}
}
\value{
\link{TRUE} if x is an initialized parameter,
\link{FALSE} otherwise
}
\description{
Determine if x is an initialized parameter,
  as created by \link{create_param}
}
\author{
Richèl J.C. Bilderbeek
}
