% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_distr.R
\name{is_init_poisson_distr}
\alias{is_init_poisson_distr}
\title{Determine if x is an initialized poisson distribution object
  as created by \code{\link{create_poisson_distr}}}
\usage{
is_init_poisson_distr(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized poisson distribution object}
}
\value{
TRUE if x is an initialized poisson distribution object
}
\description{
Determine if x is an initialized poisson distribution object
  as created by \code{\link{create_poisson_distr}}
}
\author{
Richel J.C. Bilderbeek
}
