% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_clock_model.R
\name{is_init_rln_clock_model}
\alias{is_init_rln_clock_model}
\title{Determine if x is an initialized relaxed log-normal clock_model object}
\usage{
is_init_rln_clock_model(rln_clock_model)
}
\arguments{
\item{rln_clock_model}{a Relaxed Log-Normal clock model,
as returned by \code{\link{create_rln_clock_model}}}
}
\value{
TRUE if x is an initialized relaxed log-normal clock_model object,
  FALSE otherwise
}
\description{
Determine if x is an initialized relaxed log-normal clock_model object
}
\author{
Richèl J.C. Bilderbeek
}
