% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_clock_model.R
\name{is_init_strict_clock_model}
\alias{is_init_strict_clock_model}
\title{Determine if x is an initialized strict clock_model object}
\usage{
is_init_strict_clock_model(strict_clock_model)
}
\arguments{
\item{strict_clock_model}{a strict clock model,
as returned by \code{\link{create_strict_clock_model}}}
}
\value{
TRUE if x is an initialized strict clock_model object
}
\description{
Determine if x is an initialized strict clock_model object
}
\author{
Richèl J.C. Bilderbeek
}
