% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_site_model.R
\name{is_init_tn93_site_model}
\alias{is_init_tn93_site_model}
\title{Determine if x is an initialized tn93 site model
as created by \code{\link{create_tn93_site_model}}}
\usage{
is_init_tn93_site_model(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized TN93 site model}
}
\value{
TRUE if x is an initialized TN93 site model
}
\description{
Determine if x is an initialized tn93 site model
as created by \code{\link{create_tn93_site_model}}
}
\examples{
  tn93_site_model <- create_tn93_site_model()
  testit::assert(!beautier:::is_init_tn93_site_model(tn93_site_model))
  tn93_site_model <- beautier:::init_tn93_site_model(tn93_site_model)
  testit::assert(beautier:::is_init_tn93_site_model(tn93_site_model))
}
\author{
Richèl J.C. Bilderbeek
}
