% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_tree_prior.R
\name{is_init_tree_prior}
\alias{is_init_tree_prior}
\title{Determine if x is an initialized tree_prior objects}
\usage{
is_init_tree_prior(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized tree_priors object}
}
\value{
TRUE if x is an initialized tree_prior object
}
\description{
Determine if x is an initialized tree_prior objects
}
\author{
Richèl J.C. Bilderbeek
}
