% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_initialized_clock_model.R
\name{is_initialized_clock_model}
\alias{is_initialized_clock_model}
\title{Determine if x is an initialized clock_model objects}
\usage{
is_initialized_clock_model(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized clock_models object}
}
\value{
TRUE if x is an initialized clock_model object
}
\description{
Determine if x is an initialized clock_model objects
}
\author{
Richel J.C. Bilderbeek
}
