% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_initialized_distribution.R
\name{is_initialized_distribution}
\alias{is_initialized_distribution}
\title{Determine if x is an initialized distribution object
  as created by \code{\link{create_distribution}}}
\usage{
is_initialized_distribution(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized distribution object}
}
\value{
TRUE if x is an initialized distribution object
}
\description{
Determine if x is an initialized distribution object
  as created by \code{\link{create_distribution}}
}
\author{
Richel J.C. Bilderbeek
}
