% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_initialized_distribution.R
\name{is_initialized_exponential_distribution}
\alias{is_initialized_exponential_distribution}
\title{Determine if x is an initialized exponential distribution object
  as created by \code{\link{create_exponential_distr}}}
\usage{
is_initialized_exponential_distribution(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized exponential distribution object}
}
\value{
TRUE if x is an initialized exponential distribution object
}
\description{
Determine if x is an initialized exponential distribution object
  as created by \code{\link{create_exponential_distr}}
}
\author{
Richel J.C. Bilderbeek
}
