% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_initialized_distribution.R
\name{is_initialized_log_normal_distribution}
\alias{is_initialized_log_normal_distribution}
\title{Determine if x is an initialized log_normal distribution object
  as created by \code{\link{create_log_normal_distr}}}
\usage{
is_initialized_log_normal_distribution(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized log_normal distribution object}
}
\value{
TRUE if x is an initialized log_normal distribution object
}
\description{
Determine if x is an initialized log_normal distribution object
  as created by \code{\link{create_log_normal_distr}}
}
\author{
Richel J.C. Bilderbeek
}
