% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_initialized_tree_prior.R
\name{is_initialized_yule_tree_prior}
\alias{is_initialized_yule_tree_prior}
\title{Determine if x is an initialized Yule tree_prior object}
\usage{
is_initialized_yule_tree_prior(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized Yule tree prior object}
}
\value{
TRUE if x is an initialized tree_prior object
}
\description{
Determine if x is an initialized Yule tree_prior object
}
\author{
Richel J.C. Bilderbeek
}
