% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_inv_gamma_distr.R
\name{is_inv_gamma_distr}
\alias{is_inv_gamma_distr}
\title{Determine if the object is a valid
inv_gamma distribution,
as created by \code{\link{create_inv_gamma_distr}}}
\usage{
is_inv_gamma_distr(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
inv_gamma distribution}
}
\value{
TRUE if x is a valid inv_gamma distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid
inv_gamma distribution,
as created by \code{\link{create_inv_gamma_distr}}
}
\seealso{
use \code{\link{is_distr}} to see if x is any
  distribution
}
\author{
Richel J.C. Bilderbeek
}
