% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_inv_gamma_distribution.R
\name{is_inv_gamma_distribution}
\alias{is_inv_gamma_distribution}
\title{Determine if the object is a valid
inv_gamma distribution}
\usage{
is_inv_gamma_distribution(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
inv_gamma distribution}
}
\value{
TRUE if x is a valid inv_gamma distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid
inv_gamma distribution
}
\author{
Richel J.C. Bilderbeek
}
