% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_kappa_1_param.R
\name{is_kappa_1_param}
\alias{is_kappa_1_param}
\title{Determine if the object is a valid kappa 1 parameter}
\usage{
is_kappa_1_param(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
kappa 1 parameter}
}
\value{
TRUE if x is a valid kappa 1 parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid kappa 1 parameter
}
\examples{
  kappa_1_param <- create_kappa_1_param()
  testit::assert(beautier:::is_kappa_1_param(kappa_1_param))
}
\seealso{
kappa 1 parameters are returned by
  \code{\link{create_kappa_1_param}}
}
\author{
Richel J.C. Bilderbeek
}
