% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_kappa_2_param.R
\name{is_kappa_2_param}
\alias{is_kappa_2_param}
\title{Determine if the object is a valid kappa 2 parameter}
\usage{
is_kappa_2_param(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
kappa 2 parameter}
}
\value{
TRUE if x is a valid kappa_2 parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid kappa 2 parameter
}
\examples{
  kappa_2_param <- create_kappa_2_param()
  testit::assert(beautier:::is_kappa_2_param(kappa_2_param))
}
\seealso{
kappa 2 parameters are returned by
  \code{\link{create_kappa_2_param}}
}
\author{
Richel J.C. Bilderbeek
}
