% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_lambda_param.R
\name{is_lambda_param}
\alias{is_lambda_param}
\title{Determine if the object is a valid
lambda parameter}
\usage{
is_lambda_param(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
lambda parameter}
}
\value{
TRUE if x is a valid lambda parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid
lambda parameter
}
\author{
Richel J.C. Bilderbeek
}
