% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_laplace_distr.R
\name{is_laplace_distr}
\alias{is_laplace_distr}
\title{Determine if the object is a valid
laplace distribution,
as created by \code{\link{create_laplace_distr}}}
\usage{
is_laplace_distr(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
laplace distribution}
}
\value{
TRUE if x is a valid laplace distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid
laplace distribution,
as created by \code{\link{create_laplace_distr}}
}
\examples{
  laplace_distr <- create_laplace_distr()
  testit::assert(beautier:::is_laplace_distr(laplace_distr))
}
\seealso{
use \code{\link{is_distr}} to see if x is any
  distribution
}
\author{
Richel J.C. Bilderbeek
}
