% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_log_normal_distribution.R
\name{is_log_normal_distribution}
\alias{is_log_normal_distribution}
\title{Determine if the object is a valid
log-normal distribution}
\usage{
is_log_normal_distribution(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
log-normal distribution}
}
\value{
TRUE if x is a valid log-normal distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid
log-normal distribution
}
\author{
Richel J.C. Bilderbeek
}
