% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_m_parameter.R
\name{is_m_parameter}
\alias{is_m_parameter}
\title{Determine if the object is a valid
m parameter}
\usage{
is_m_parameter(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
m parameter}
}
\value{
TRUE if x is a valid m parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid
m parameter
}
\author{
Richel J.C. Bilderbeek
}
