% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mcmc.R
\name{is_mcmc}
\alias{is_mcmc}
\title{Determine if the object is a valid MCMC}
\usage{
is_mcmc(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid MCMC}
}
\value{
TRUE if x is a valid MCMC, FALSE otherwise
}
\description{
Determine if the object is a valid MCMC
}
\examples{
check_empty_beautier_folder()

# Returns TRUE
is_mcmc(create_mcmc())
is_mcmc(create_ns_mcmc())

# Returns FALSE
is_mcmc("nonsense")
is_mcmc(NULL)
is_mcmc(NA)
is_mcmc("")
is_mcmc(c())

check_empty_beautier_folder()
}
\seealso{
Use \code{\link{create_mcmc}} to create an MCMC
}
\author{
Richèl J.C. Bilderbeek
}
