% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mean_parameter.R
\name{is_mean_parameter}
\alias{is_mean_parameter}
\title{Determine if the object is a valid
mean parameter}
\usage{
is_mean_parameter(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
mean parameter}
}
\value{
TRUE if x is a valid mean parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid
mean parameter
}
\author{
Richel J.C. Bilderbeek
}
