% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_misc_options.R
\name{is_misc_options}
\alias{is_misc_options}
\title{Determine if the object is a valid misc_options}
\usage{
is_misc_options(x)
}
\arguments{
\item{x}{an object, to be determined if it is a misc_options}
}
\value{
TRUE if the object is a valid misc_options, FALSE otherwise
}
\description{
Determine if the object is a valid misc_options
}
\examples{
  misc_options <- create_misc_options()
  testit::assert(beautier:::is_misc_options(misc_options))
}
\author{
Richel J.C. Bilderbeek
}
