% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mrca_align_ids_in_fastas.R
\name{is_mrca_align_ids_in_fastas}
\alias{is_mrca_align_ids_in_fastas}
\title{Determine if an MRCA prior's alignment IDs are present in the FASTA files}
\usage{
is_mrca_align_ids_in_fastas(mrca_prior, fasta_filenames)
}
\arguments{
\item{mrca_prior}{a Most Recent Common Ancestor prior,
as returned by \code{\link{create_mrca_prior}}}

\item{fasta_filenames}{One or more FASTA filenames.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}
}
\value{
TRUE if the MRCA prior's alignment IDs
  is present in the FASTA files.
  Returns FALSE otherwise
}
\description{
Determine if an MRCA prior's alignment IDs are present in the FASTA files
}
\author{
Richèl J.C. Bilderbeek
}
