% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mrca_prior.R
\name{is_mrca_prior}
\alias{is_mrca_prior}
\title{Determine of the object is an MRCA prior,
  as returned by \code{\link{create_mrca_prior}}}
\usage{
is_mrca_prior(x)
}
\arguments{
\item{x}{object to be determined if it is an MRCA prior}
}
\value{
TRUE if \code{x} is an MRCA prior, FALSE otherwise
}
\description{
Determine of the object is an MRCA prior,
  as returned by \code{\link{create_mrca_prior}}
}
\author{
Richel J.C. Bilderbeek
}
