% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mrca_prior_with_distr.R
\name{is_mrca_prior_with_distr}
\alias{is_mrca_prior_with_distr}
\title{See if x is one MRCA prior with a distribution}
\usage{
is_mrca_prior_with_distr(x)
}
\arguments{
\item{x}{the object to be tested}
}
\value{
TRUE if x is one MRCA prior with a distribution,
  FALSE otherwise
}
\description{
See if x is one MRCA prior with a distribution
}
\author{
Richèl J.C. Bilderbeek
}
