% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_normal_distribution.R
\name{is_normal_distribution}
\alias{is_normal_distribution}
\title{Determine if the object is a valid
normal distribution
as created by \code{\link{create_normal_distr}}}
\usage{
is_normal_distribution(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
normal distribution}
}
\value{
TRUE if x is a valid normal distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid
normal distribution
as created by \code{\link{create_normal_distr}}
}
\seealso{
use \code{\link{is_distribution}} to see if x is any
  distribution
}
\author{
Richel J.C. Bilderbeek
}
