% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_one_bool.R
\name{is_one_bool}
\alias{is_one_bool}
\title{Check if the argument is one boolean}
\usage{
is_one_bool(x)
}
\arguments{
\item{x}{the argument to be tested to be boolean}
}
\description{
Check if the argument is one boolean
}
\examples{
check_empty_beautier_folder()

# TRUE
is_one_bool(TRUE)
is_one_bool(FALSE)

# FALSE
is_one_bool(NULL)
is_one_bool(NA)
is_one_bool(c())
is_one_bool("nonsense")
is_one_bool(is_one_bool)
is_one_bool(c(TRUE, FALSE))

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
