% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_distr.R
\name{is_one_div_x_distr}
\alias{is_one_div_x_distr}
\title{Determine if the object is a valid
1/x distribution,
as created by \code{\link{create_one_div_x_distr}}}
\usage{
is_one_div_x_distr(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
1/x distribution}
}
\value{
TRUE if x is a valid 1/x distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid
1/x distribution,
as created by \code{\link{create_one_div_x_distr}}
}
\examples{
# TRUE
is_one_div_x_distr(create_one_div_x_distr())
# FALSE
is_one_div_x_distr(create_poisson_distr())
is_one_div_x_distr(NA)
is_one_div_x_distr(NULL)
is_one_div_x_distr("nonsense")
}
\seealso{
use \code{\link{is_distr}} to see if x is any
  distribution
}
\author{
Richèl J.C. Bilderbeek
}
