% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_one_double.R
\name{is_one_double}
\alias{is_one_double}
\title{Determines if the argument is a double}
\usage{
is_one_double(x)
}
\arguments{
\item{x}{the object to be determined of if it is one double}
}
\description{
Determines if the argument is a double
}
\examples{
check_empty_beautier_folder()

# TRUE
is_one_double(314)
is_one_double(0)
is_one_double(-314)
is_one_double(3.14)

# FALSE
is_one_double(NULL)
is_one_double(NA)
is_one_double(Inf)
is_one_double("nonsense")
is_one_double(is_one_double)
is_one_double(c())
is_one_double(c(1, 2))

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
