% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_one_empty_string.R
\name{is_one_empty_string}
\alias{is_one_empty_string}
\title{Determine if an object is one empty string}
\usage{
is_one_empty_string(x)
}
\arguments{
\item{x}{the object that may be one string that may be empty}
}
\value{
TRUE is `x` is one string that is empty
}
\description{
Determine if an object is one empty string
}
\examples{
# TRUE
is_one_empty_string("")

# FALSE
is_one_empty_string("3.14")
is_one_empty_string(c("", ""))
is_one_empty_string(42)
is_one_empty_string("nonsense")
}
\author{
Richèl J.C. Bilderbeek
}
