% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_one_na.R
\name{is_one_na}
\alias{is_one_na}
\title{Determines if x is one NA}
\usage{
is_one_na(x)
}
\arguments{
\item{x}{the object to be determined if it is one NA}
}
\value{
TRUE if x is one NA, FALSE otherwise
}
\description{
Determines if x is one NA
}
\examples{
  testit::assert(is_one_na(NA))
  testit::assert(!is_one_na(NULL))
  testit::assert(!is_one_na(42))
  testit::assert(!is_one_na("Hello"))
  testit::assert(!is_one_na(3.14))
  testit::assert(!is_one_na(c(NA, NA)))
}
\author{
Richèl J.C. Bilderbeek
}
