% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_one_string.R
\name{is_one_string}
\alias{is_one_string}
\title{Check if the argument is one string}
\usage{
is_one_string(x)
}
\arguments{
\item{x}{the argument to be tested to be one string}
}
\description{
Check if the argument is one string
}
\examples{
library(testthat)
expect_true(is_one_string("hello"))
expect_true(is_one_string("hello world"))
expect_false(is_one_string(NULL))
expect_false(is_one_string(NA))
expect_false(is_one_string(TRUE))
expect_false(is_one_string(c()))
expect_false(is_one_string(c("hello", "world")))
}
\author{
Richèl J.C. Bilderbeek
}
