% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_one_string_that_is_a_number.R
\name{is_one_string_that_is_a_number}
\alias{is_one_string_that_is_a_number}
\title{General function to create a distribution.}
\usage{
is_one_string_that_is_a_number(x)
}
\arguments{
\item{x}{the object that may be one string that may be a number}
}
\value{
TRUE is `x` is one string that is a number
}
\description{
General function to create a distribution.
}
\examples{
# TRUE
is_one_string_that_is_a_number("3.14")

# FALSE
is_one_string_that_is_a_number(c("3.14", "42"))
is_one_string_that_is_a_number("")
is_one_string_that_is_a_number(42)
is_one_string_that_is_a_number("nonsense")
}
\author{
Richèl J.C. Bilderbeek
}
