% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_parameter.R
\name{is_parameter}
\alias{is_parameter}
\title{Determine if the object is a valid parameter}
\usage{
is_parameter(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
parameter}
}
\value{
TRUE if x is a valid parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid parameter
}
\author{
Richel J.C. Bilderbeek
}
