% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_phylo.R
\name{is_phylo}
\alias{is_phylo}
\title{Checks if the input is a phylogeny}
\usage{
is_phylo(x)
}
\arguments{
\item{x}{input to be checked}
}
\value{
TRUE or FALSE
}
\description{
Checks if the input is a phylogeny
}
\author{
Richel J.C. Bilderbeek
}
