% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_poisson_distr.R
\name{is_poisson_distr}
\alias{is_poisson_distr}
\title{Determine if the object is a valid
poisson distribution
as created by \code{\link{create_poisson_distr}}}
\usage{
is_poisson_distr(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
poisson distribution}
}
\value{
TRUE if x is a valid poisson distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid
poisson distribution
as created by \code{\link{create_poisson_distr}}
}
\seealso{
use \code{\link{is_distr}} to see if x is any
  distribution
}
\author{
Richel J.C. Bilderbeek
}
