% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_poisson_distribution.R
\name{is_poisson_distribution}
\alias{is_poisson_distribution}
\title{Determine if the object is a valid
poisson distribution}
\usage{
is_poisson_distribution(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
poisson distribution}
}
\value{
TRUE if x is a valid poisson distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid
poisson distribution
}
\author{
Richel J.C. Bilderbeek
}
