% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_param.R
\name{is_rate_ac_param}
\alias{is_rate_ac_param}
\title{Determine if the object is a valid
'rate AC' parameter}
\usage{
is_rate_ac_param(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
'rate AC' parameter}
}
\value{
TRUE if x is a valid 'rate AC' parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid
'rate AC' parameter
}
\examples{

is_rate_ac_param(create_alpha_param())
is_rate_ac_param(create_beta_param())
is_rate_ac_param(create_clock_rate_param())
is_rate_ac_param(create_kappa_1_param())
is_rate_ac_param(create_kappa_2_param())
is_rate_ac_param(create_lambda_param())
is_rate_ac_param(create_m_param())
is_rate_ac_param(create_mean_param())
is_rate_ac_param(create_mu_param())
is_rate_ac_param(create_rate_ac_param())
is_rate_ac_param(create_rate_ag_param())
is_rate_ac_param(create_rate_at_param())
is_rate_ac_param(create_rate_cg_param())
is_rate_ac_param(create_rate_ct_param())
is_rate_ac_param(create_rate_gt_param())
is_rate_ac_param(create_s_param())
is_rate_ac_param(create_scale_param())
is_rate_ac_param(create_sigma_param())

is_rate_ac_param(NA)
is_rate_ac_param(NULL)
is_rate_ac_param("nonsense")
is_rate_ac_param(create_jc69_site_model())
is_rate_ac_param(create_strict_clock_model())
is_rate_ac_param(create_yule_tree_prior())
is_rate_ac_param(create_mcmc())
}
\seealso{
\code{\link{create_rate_ac_param}} creates a 'rate AC' parameter
}
\author{
Richèl J.C. Bilderbeek
}
