% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rate_ag_param.R
\name{is_rate_ag_param}
\alias{is_rate_ag_param}
\title{Determine if the object is a valid
'rate AG' parameter}
\usage{
is_rate_ag_param(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
'rate AG' parameter}
}
\value{
TRUE if x is a valid 'rate AG' parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid
'rate AG' parameter
}
\examples{
  rate_ag_param <- create_rate_ag_param()
}
\seealso{
\code{\link{create_rate_ag_param}} creates a 'rate AG' parameter
}
\author{
Richel J.C. Bilderbeek
}
