% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rate_at_param.R
\name{is_rate_at_param}
\alias{is_rate_at_param}
\title{Determine if the object is a valid
'rate AT' parameter}
\usage{
is_rate_at_param(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
'rate AT' parameter}
}
\value{
TRUE if x is a valid 'rate AT' parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid
'rate AT' parameter
}
\examples{
  rate_at_param <- create_rate_at_param()
  testit::assert(beautier:::is_rate_at_param(rate_at_param))
}
\seealso{
\code{\link{create_rate_at_param}} creates a 'rate AT' parameter
}
\author{
Richel J.C. Bilderbeek
}
