% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rate_ct_param.R
\name{is_rate_ct_param}
\alias{is_rate_ct_param}
\title{Determine if the object is a valid
'rate CT' parameter}
\usage{
is_rate_ct_param(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
'rate CT' parameter}
}
\value{
TRUE if x is a valid 'rate CG' parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid
'rate CT' parameter
}
\examples{
  rate_ct_param <- create_rate_ct_param()
  testit::assert(beautier:::is_rate_ct_param(rate_ct_param))
}
\seealso{
\code{\link{create_rate_ct_param}} creates a 'rate CT' parameter
}
\author{
Richel J.C. Bilderbeek
}
