% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rate_gt_param.R
\name{is_rate_gt_param}
\alias{is_rate_gt_param}
\title{Determine if the object is a valid
'rate GT' parameter}
\usage{
is_rate_gt_param(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
'rate GT' parameter}
}
\value{
TRUE if x is a valid 'rate GT' parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid
'rate GT' parameter
}
\examples{
  rate_gt_param <- create_rate_gt_param()
  testit::assert(beautier:::is_rate_gt_param(rate_gt_param))
}
\seealso{
\code{\link{create_rate_gt_param}} creates a 'rate GT' parameter
}
\author{
Richel J.C. Bilderbeek
}
