% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_relaxed_log_normal_clock_model.R
\name{is_rln_clock_model}
\alias{is_rln_clock_model}
\title{Determine if the object is a valid relaxed log normal clock model}
\usage{
is_rln_clock_model(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid relaxed log normal clock model}
}
\value{
TRUE if x is a valid relaxed log normal clock model, FALSE otherwise
}
\description{
Determine if the object is a valid relaxed log normal clock model
}
\author{
Richel Bilderbeek
}
