% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_s_parameter.R
\name{is_s_parameter}
\alias{is_s_parameter}
\title{Determine if the object is a valid
s parameter}
\usage{
is_s_parameter(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
s parameter}
}
\value{
TRUE if x is a valid s parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid
s parameter
}
\author{
Richel J.C. Bilderbeek
}
