% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_scale_param.R
\name{is_scale_param}
\alias{is_scale_param}
\title{Determine if the object is a valid
scale parameter}
\usage{
is_scale_param(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
scale parameter}
}
\value{
TRUE if x is a valid scale parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid
scale parameter
}
\author{
Richel J.C. Bilderbeek
}
