% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sigma_param.R
\name{is_sigmaparam}
\alias{is_sigmaparam}
\title{Determine if the object is a valid
sigma parameter}
\usage{
is_sigmaparam(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
sigma parameter}
}
\value{
TRUE if x is a valid sigma parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid
sigma parameter
}
\author{
Richel J.C. Bilderbeek
}
