% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_site_model.R
\name{is_site_model}
\alias{is_site_model}
\title{Determine if the object is a valid site_model}
\usage{
is_site_model(x)
}
\arguments{
\item{x}{an object, to be determined if it is a site_model}
}
\value{
TRUE if the site_model is a valid site_model, FALSE otherwise
}
\description{
Determine if the object is a valid site_model
}
