% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_site_model_name.R
\name{is_site_model_name}
\alias{is_site_model_name}
\title{Determines if the name is a valid site_model name}
\usage{
is_site_model_name(name)
}
\arguments{
\item{name}{the name to be tested}
}
\value{
TRUE if the name is a valid site_model name, FALSE otherwise
}
\description{
Determines if the name is a valid site_model name
}
\examples{
# TRUE
is_site_model_name("JC69")
is_site_model_name("HKY")
is_site_model_name("TN93")
is_site_model_name("GTR")
# FALSE
is_site_model_name("nonsense")
}
\author{
Richèl J.C. Bilderbeek
}
