% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_strict_clock_model.R
\name{is_strict_clock_model}
\alias{is_strict_clock_model}
\title{Determine if the object is a valid strict clock model}
\usage{
is_strict_clock_model(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid strict clock model}
}
\value{
TRUE if x is a valid strict clock model, FALSE otherwise
}
\description{
Determine if the object is a valid strict clock model
}
\author{
Richel Bilderbeek
}
