% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_distr.R
\name{is_uniform_distr}
\alias{is_uniform_distr}
\title{Determine if the object is a valid
uniform distribution
as created by \code{\link{create_uniform_distr}}}
\usage{
is_uniform_distr(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
uniform distribution}
}
\value{
TRUE if x is a valid uniform distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid
uniform distribution
as created by \code{\link{create_uniform_distr}}
}
\examples{
check_empty_beautier_folder()

# TRUE
is_uniform_distr(create_uniform_distr())
# FALSE
is_uniform_distr(create_beta_distr())
is_uniform_distr(NA)
is_uniform_distr(NULL)
is_uniform_distr("nonsense")

check_empty_beautier_folder()
}
\seealso{
use \code{\link{is_distr}} to see if x is any
  distribution
}
\author{
Richèl J.C. Bilderbeek
}
