% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_uniform_distribution.R
\name{is_uniform_distribution}
\alias{is_uniform_distribution}
\title{Determine if the object is a valid
uniform distribution}
\usage{
is_uniform_distribution(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
uniform distribution}
}
\value{
TRUE if x is a valid uniform distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid
uniform distribution
}
\seealso{
use \code{\link{is_distribution}} to see if x is any
  distribution
}
\author{
Richel J.C. Bilderbeek
}
