% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_beast2_input_file.R
\name{is_valid_beast2_input_file}
\alias{is_valid_beast2_input_file}
\title{Deprecated name for function \code{\link{is_beast2_input_file}}}
\usage{
is_valid_beast2_input_file(filename)
}
\arguments{
\item{filename}{name of the BEAST2 XML input file}
}
\value{
TRUE if the file is valid, FALSE if not
}
\description{
Deprecated name for function \code{\link{is_beast2_input_file}}
}
