% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_site_model.R
\name{is_valid_site_model}
\alias{is_valid_site_model}
\title{Create a random alignment}
\usage{
is_valid_site_model(site_model)
}
\arguments{
\item{site_model}{a site_model, as created by create_site_model}
}
\value{
TRUE if the site_model is a valid site_model, FALSE otherwise
}
\description{
Create a random alignment
}
