% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_tree_prior.R
\name{is_valid_tree_prior}
\alias{is_valid_tree_prior}
\title{Determine if an object is a valid tree_prior}
\usage{
is_valid_tree_prior(x)
}
\arguments{
\item{x}{an object}
}
\value{
TRUE if x is a valid tree_prior, FALSE otherwise
}
\description{
Determine if an object is a valid tree_prior
}
