% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_param_name.R
\name{isparam_name}
\alias{isparam_name}
\title{Determines if the name is a valid parameter name}
\usage{
isparam_name(name)
}
\arguments{
\item{name}{the name to be tested}
}
\value{
TRUE if the name is a valid parameter name, FALSE otherwise
}
\description{
Determines if the name is a valid parameter name
}
