% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_param_to_xml.R
\name{kappa_param_to_xml}
\alias{kappa_param_to_xml}
\title{Internal function}
\usage{
kappa_param_to_xml(kappa_param, beauti_options = create_beauti_options())
}
\arguments{
\item{kappa_param}{a kappa parameter,
as created by \link{create_kappa_param}}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the parameter as XML text
}
\description{
Converts an kappa parameter to XML
}
\examples{
check_empty_beautier_folder()

# The kappa parameter must be initialized, i.e. have an ID
kappa_param_to_xml(kappa_param = create_kappa_param(id = "1"))

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
