% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_parameter_to_xml.R
\name{kappa_parameter_to_xml}
\alias{kappa_parameter_to_xml}
\title{Internal function}
\usage{
kappa_parameter_to_xml(
  kappa_parameter,
  beauti_options = create_beauti_options()
)
}
\arguments{
\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the parameter as XML text
}
\description{
Converts an kappa parameter to XML
}
\examples{
check_empty_beautier_folder()

# The kappa parameter must be initialized, i.e. have an ID
kappa_parameter_to_xml(kappa_parameter = create_kappa_param(id = "1"))

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
