% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_param_to_xml.R
\name{m_param_to_xml}
\alias{m_param_to_xml}
\title{Internal function}
\usage{
m_param_to_xml(m_param, beauti_options = create_beauti_options())
}
\arguments{
\item{m_param}{an m parameter,
as created by \link{create_m_param}}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the parameter as XML text
}
\description{
Converts an m parameter to XML
}
\author{
Richèl J.C. Bilderbeek
}
