% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_to_xml_run.R
\name{mcmc_to_xml_run}
\alias{mcmc_to_xml_run}
\title{Converts an MCMC object to the run section's XML}
\usage{
mcmc_to_xml_run(mcmc)
}
\arguments{
\item{mcmc}{one mcmc object,
as returned by \code{\link{create_mcmc}}}
}
\value{
the XML as text
}
\description{
Converts an MCMC object to the run section's XML
}
\examples{
  xml <- beautier:::mcmc_to_xml_run(create_mcmc())
  testit::assert(xml ==
    "<run id=\\"mcmc\\" spec=\\"MCMC\\" chainLength=\\"10000000\\">"
  )
}
\author{
Richel J.C. Bilderbeek
}
