% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrca_prior_to_xml_lh_distr.R
\name{mrca_prior_to_xml_lh_distr}
\alias{mrca_prior_to_xml_lh_distr}
\title{Converts an MRCA prior to the \code{branchRateModel} section of the
XML as text}
\usage{
mrca_prior_to_xml_lh_distr(mrca_prior, has_non_strict_clock_model = FALSE)
}
\arguments{
\item{mrca_prior}{a Most Recent Common Ancestor prior,
as returned by \code{\link{create_mrca_prior}}}

\item{has_non_strict_clock_model}{boolean to indicate that the is
already at least one non-strict (i.e. relaxed log-normal) clock model}
}
\description{
Converts an MRCA prior to the \code{branchRateModel} section of the
XML as text
}
\examples{
 # <distribution id="posterior" spec="util.CompoundDistribution">
 #     <distribution id="prior" spec="util.CompoundDistribution">
 #     </distribution>
 #     <distribution id="likelihood" ...>
 #       HERE, where the ID of the distribution is 'likelihood'
 #     </distribution>
 # </distribution>
}
\author{
Richel J.C. Bilderbeek
}
