% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrca_prior_to_xml_state.R
\name{mrca_prior_to_xml_state}
\alias{mrca_prior_to_xml_state}
\title{Creates the XML of an MRCA prior,
  as used in the \code{state} section}
\usage{
mrca_prior_to_xml_state(mrca_prior)
}
\value{
the tree prior as XML text
}
\description{
Creates the XML of an MRCA prior,
  as used in the \code{state} section
}
\author{
Richel J.C. Bilderbeek
}
