% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrca_priors_to_xml_state.R
\name{mrca_priors_to_xml_state}
\alias{mrca_priors_to_xml_state}
\title{Converts one or more MRCA priors to the \code{state} section of the
XML as text}
\usage{
mrca_priors_to_xml_state(mrca_priors)
}
\arguments{
\item{mrca_priors}{a list of one or more Most Recent Common Ancestor priors,
as returned by \code{\link{create_mrca_prior}}}
}
\value{
lines of XML text, without indentation nor \code{state}
  tags
}
\description{
Converts one or more MRCA priors to the \code{state} section of the
XML as text
}
\author{
Richel J.C. Bilderbeek
}
