% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needs_trait_set_str.R
\name{needs_trait_set_str}
\alias{needs_trait_set_str}
\title{Is it needed to add the tip dates as a string to the BEAST2 XML input file?}
\usage{
needs_trait_set_str(inference_model)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\value{
TRUE if this is needed
}
\description{
The hypothesis is that this is needed when the tree
prior is a coalescent model.
}
\details{
Here are the notes I concluded this from:

Site model|Clock model|Tree prior|Needed|Other notes
----------|-----------|----------|------|----------------
GTR       |RLN        |CBS       |Yes   |#108
HKY       |RLN        |CCP       |Yes   |#109
JC69      |RLN        |Yule      |No    |#99
JC69      |Strict     |Yule      |No    |tipdates_2_6.xml
JC69      |RLN        |Yule      |No    |tipdates_2_6.xml
}
\author{
Richèl J.C. Bilderbeek
}
