% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-obj-type.R
\name{obj_type_friendly}
\alias{obj_type_friendly}
\title{Return English-friendly type}
\usage{
obj_type_friendly(x, value = TRUE)
}
\arguments{
\item{x}{Any R object.}

\item{value}{Whether to describe the value of `x`. Special values
like `NA` or `""` are always described.}
}
\value{
A string describing the type. Starts with an indefinite
  article, e.g. "an integer vector".
}
\description{
Return English-friendly type
}
\note{
adapted from
[https://github.com/r-lib/rlang](https://github.com/r-lib/rlang)
file `R/standalone-obj-type.R`
}
\author{
[`olivroy`](https://github.com/olivroy)
}
