% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml_alpha}
\alias{parameter_to_xml_alpha}
\title{Converts an alpha parameter to XML}
\usage{
parameter_to_xml_alpha(parameter)
}
\arguments{
\item{parameter}{an alpha parameter,
as created by \code{\link{create_alpha_param}})}
}
\value{
the parameter as XML text
}
\description{
Converts an alpha parameter to XML
}
\author{
Richel J.C. Bilderbeek
}
