% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml_beta}
\alias{parameter_to_xml_beta}
\title{Converts a beta parameter to XML}
\usage{
parameter_to_xml_beta(parameter)
}
\arguments{
\item{parameter}{a beta parameter,
a numeric value.
For advanced usage, use the structure
as created by \code{\link{create_beta_param}})}
}
\value{
the parameter as XML text
}
\description{
Converts a beta parameter to XML
}
\author{
Richèl J.C. Bilderbeek
}
