% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml_clock_rate}
\alias{parameter_to_xml_clock_rate}
\title{Converts a \code{clockRate} parameter to XML}
\usage{
parameter_to_xml_clock_rate(parameter)
}
\arguments{
\item{parameter}{a \code{clockRate} parameter,
as created by \code{\link{create_clock_rate_param}})}
}
\value{
the parameter as XML text
}
\description{
Converts a \code{clockRate} parameter to XML
}
\author{
Richel J.C. Bilderbeek
}
