% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml_kappa_1}
\alias{parameter_to_xml_kappa_1}
\title{Converts a kappa 1 parameter to XML}
\usage{
parameter_to_xml_kappa_1(parameter)
}
\arguments{
\item{parameter}{a kappa 1 parameter,
a numeric value.
For advanced usage, use the structure
as created by \code{\link{create_kappa_1_param}})}
}
\value{
the parameter as XML text
}
\description{
Converts a kappa 1 parameter to XML
}
\author{
Richèl J.C. Bilderbeek
}
