% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml_m}
\alias{parameter_to_xml_m}
\title{Internal function}
\usage{
parameter_to_xml_m(parameter, beauti_options = create_beauti_options())
}
\arguments{
\item{parameter}{a m parameter,
a numeric value.
For advanced usage, use the structure
as created by \code{\link{create_m_param}})}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the parameter as XML text
}
\description{
Converts a m parameter to XML
}
\author{
Richèl J.C. Bilderbeek
}
