% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml_mean}
\alias{parameter_to_xml_mean}
\title{Converts a mean parameter to XML}
\usage{
parameter_to_xml_mean(parameter)
}
\arguments{
\item{parameter}{a mean parameter,
as created by \code{\link{create_mean_param}})}
}
\value{
the parameter as XML text
}
\description{
Converts a mean parameter to XML
}
\author{
Richel J.C. Bilderbeek
}
