% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml_rate_ag}
\alias{parameter_to_xml_rate_ag}
\title{Converts a 'rate AG' parameter to XML}
\usage{
parameter_to_xml_rate_ag(parameter)
}
\arguments{
\item{parameter}{a 'rate AG' parameter,
as created by \code{\link{create_rate_ac_param}})}
}
\value{
the parameter as XML text
}
\description{
Converts a 'rate AG' parameter to XML
}
\author{
Richel J.C. Bilderbeek
}
