% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml_rate_at}
\alias{parameter_to_xml_rate_at}
\title{Converts a 'rate AT' parameter to XML}
\usage{
parameter_to_xml_rate_at(parameter)
}
\arguments{
\item{parameter}{a 'rate AT' parameter,
as created by \code{\link{create_rate_at_param}})}
}
\value{
the parameter as XML text
}
\description{
Converts a 'rate AT' parameter to XML
}
\author{
Richel J.C. Bilderbeek
}
