% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml_s}
\alias{parameter_to_xml_s}
\title{Converts a s parameter to XML}
\usage{
parameter_to_xml_s(parameter)
}
\arguments{
\item{parameter}{a s parameter,
a numeric value.
For advanced usage, use the structure
as created by \code{\link{create_s_param}})}
}
\value{
the parameter as XML text
}
\description{
Converts a s parameter to XML
}
\author{
Richèl J.C. Bilderbeek
}
