% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_to_xml_state.R
\name{phylo_to_xml_state}
\alias{phylo_to_xml_state}
\title{Creates the XML of a phylogeny,
  as used in the \code{state} section}
\usage{
phylo_to_xml_state(
  id = "irrelevant",
  inference_model = "irrelevant",
  tipdates_filename = "deprecated"
)
}
\arguments{
\item{id}{the ID of the alignment}

\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}

\item{tipdates_filename}{name of the file containing the tip dates.
This file is assumed to have two columns, separated by a tab.
The first column contains the taxa names, the second column contains
the date.}
}
\value{
the random phylogeny as XML text
}
\description{
Creates the XML of a phylogeny,
  as used in the \code{state} section
}
\author{
Richèl J.C. Bilderbeek
}
