% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_species_tree_to_xml.R
\name{random_species_tree_to_xml}
\alias{random_species_tree_to_xml}
\title{Creates the XML of a random species tree,
  as used in the \code{init} section}
\usage{
random_species_tree_to_xml(id)
}
\arguments{
\item{id}{the ID of the alignment (can be extracted from
its FASTA filesname using \code{\link{get_id}})}
}
\value{
the parameter as XML text
}
\description{
Creates the XML of a random species tree,
  as used in the \code{init} section
}
\author{
Richel J.C. Bilderbeek
}
