% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_species_tree_to_xml_state.R
\name{random_species_tree_to_xml_state}
\alias{random_species_tree_to_xml_state}
\title{Creates the XML of a random species tree,
  as used in the \code{state} section}
\usage{
random_species_tree_to_xml_state(id)
}
\arguments{
\item{id}{the ID of the alignment (can be extracted from
its FASTA filesname using \code{\link{get_id}})}
}
\value{
the random species tree as XML text
}
\description{
Creates the XML of a random species tree,
  as used in the \code{state} section
}
\author{
Richel J.C. Bilderbeek
}
