% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tipdates_file.R
\name{read_tipdates_file}
\alias{read_tipdates_file}
\title{Read a tip dates file}
\usage{
read_tipdates_file(tipdates_filename)
}
\arguments{
\item{tipdates_filename}{name of the file containing the tip dates.
This file is assumed to have two columns, separated by a tab.
The first column contains the taxa names, the second column contains
the date.}
}
\value{
the tip dates as a tibble, with column names `taxon`
and `time`.  The `time` column creates strings instead
of numbers, to allow using the literally same values from the
tip dates file in the BEAST2 XML input file (i.e. avoid
rounding errors)
}
\description{
Read a tip dates file
}
\examples{
read_tipdates_file(get_beautier_path("babette_issue_109.tsv"))
}
\author{
Richèl J.C. Bilderbeek
}
