% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_beautier_folder.R
\name{remove_beautier_folder}
\alias{remove_beautier_folder}
\title{Check there are no files in the default \link{beautier} folder}
\usage{
remove_beautier_folder()
}
\value{
No return value, called for side effects.
}
\description{
Check there are no files in the default \link{beautier} folder.
The goal is to make sure no temporary files are left undeleted.
Will \link{stop} if there are files in the \link{beautier} folder.
}
\examples{
check_empty_beautier_folder()

remove_beautier_folder()

check_empty_beautier_folder()
}
\seealso{
use \link{remove_beautier_folder} to remove the default
`beautier` folder
}
\author{
Richèl J.C. Bilderbeek
}
