% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty_lines.R
\name{remove_empty_lines}
\alias{remove_empty_lines}
\title{Remove all lines that are only whitespace}
\usage{
remove_empty_lines(lines, trim = FALSE)
}
\arguments{
\item{lines}{character vector with text}

\item{trim}{FALSE if indentation must be preserved,
TRUE will remove all surrounding whitespace}
}
\value{
the lines with text
}
\description{
Remove all lines that are only whitespace
}
\author{
Richèl J.C. Bilderbeek
}
