% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastscriptr.R
\name{remove_file_extension}
\alias{remove_file_extension}
\title{Remove a file extension}
\usage{
remove_file_extension(filename)
}
\arguments{
\item{filename}{A filename}
}
\value{
That filename without its extension
}
\description{
Remove a file extension
}

