% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_multiline.R
\name{remove_multiline}
\alias{remove_multiline}
\title{Remove consequetive lines}
\usage{
remove_multiline(text, lines_to_remove)
}
\arguments{
\item{text}{lines of characters}

\item{lines_to_remove}{lines of character that need to be removed from text}
}
\description{
Remove consequetive lines
}
\author{
Richel J.C. Bilderbeek
}
