% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_inference_model_filenames.R
\name{rename_inference_model_filenames}
\alias{rename_inference_model_filenames}
\title{Rename the filenames in an inference model}
\usage{
rename_inference_model_filenames(inference_model, rename_fun)
}
\arguments{
\item{inference_model}{an Bayesian phylogenetic inference model,
as can be created by \code{\link{create_inference_model}}.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.}

\item{rename_fun}{a function to rename a filename,
as can be checked by \link{check_rename_fun}. This function should
have one argument, which will be a filename or \link{NA}. The
function should \link{return} one filename (when passed one filename) or
one \link{NA} (when passed one \link{NA}).
Example rename functions are:
\itemize{
  \item \link{get_remove_dir_fun} function that removes the directory
    paths from the filenames, in effect turning these into local files
  \item \link{get_replace_dir_fun} function that replaces the directory
    paths from the filenames
}}
}
\description{
Rename the filenames in an inference model
}
