% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_mcmc_filenames.R
\name{rename_mcmc_filenames}
\alias{rename_mcmc_filenames}
\title{Rename the filenames within an MCMC}
\usage{
rename_mcmc_filenames(mcmc, rename_fun)
}
\arguments{
\item{mcmc}{one MCMC.
Use \link{create_mcmc} to create an MCMC.
Use \link{create_ns_mcmc} to create an MCMC for a Nested Sampling run.
Use \link{check_mcmc} to check if an MCMC is valid.
Use \link{rename_mcmc_filenames} to rename the filenames in an MCMC.}
}
\description{
Rename the filenames within an MCMC
}
\examples{
library(testthat)

# Create an MCMC with local filenames
mcmc <- create_mcmc()
mcmc$tracelog$filename <- "trace.log"
mcmc$screenlog$filename <- "screen.log"
mcmc$treelog$filename <- "tree.log"

# Nah, files should be put in '/home/john' folder
mcmc <- rename_mcmc_filenames(
  mcmc = mcmc,
  rename_fun = get_replace_dir_fun("/home/john")
)

expect_equal(mcmc$tracelog$filename, "/home/john/trace.log")
expect_equal(mcmc$screenlog$filename, "/home/john/screen.log")
expect_equal(mcmc$treelog$filename, "/home/john/tree.log")

# Nah, files should be put in '/home/doe' folder instead
mcmc <- rename_mcmc_filenames(
  mcmc = mcmc,
  rename_fun = get_replace_dir_fun("/home/doe")
)

expect_equal(mcmc$tracelog$filename, "/home/doe/trace.log")
expect_equal(mcmc$screenlog$filename, "/home/doe/screen.log")
expect_equal(mcmc$treelog$filename, "/home/doe/tree.log")

# Nah, files should be put in local folder instead
mcmc <- rename_mcmc_filenames(
  mcmc = mcmc,
  rename_fun = get_remove_dir_fun()
)

expect_equal(mcmc$tracelog$filename, "trace.log")
expect_equal(mcmc$screenlog$filename, "screen.log")
expect_equal(mcmc$treelog$filename, "tree.log")
}
