% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rln_clock_model_to_xml_mean_rate_prior.R
\name{rln_clock_model_to_xml_mean_rate_prior}
\alias{rln_clock_model_to_xml_mean_rate_prior}
\title{Used by \code{\link{clock_models_to_xml_prior_distr}}}
\usage{
rln_clock_model_to_xml_mean_rate_prior(rln_clock_model)
}
\arguments{
\item{rln_clock_model}{a Relaxed Log-Normal clock model,
as returned by \code{\link{create_rln_clock_model}}}
}
\value{
lines of XML text
}
\description{
Used by \code{\link{clock_models_to_xml_prior_distr}}
}
\author{
Richèl J.C. Bilderbeek
}
