% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rln_clock_model_to_xml_state.R
\name{rln_clock_model_to_xml_state}
\alias{rln_clock_model_to_xml_state}
\title{Internal function}
\usage{
rln_clock_model_to_xml_state(inference_model)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\value{
lines of XML text, without indentation nor \code{state}
  tags
}
\description{
Converts an RLN clock model to the `state` section of the
XML as text
}
\author{
Richèl J.C. Bilderbeek
}
