% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnd_phylo_to_xml_state.R
\name{rnd_phylo_to_xml_state}
\alias{rnd_phylo_to_xml_state}
\title{Creates the XML of a random phylogeny,
  as used in the \code{state} section}
\usage{
rnd_phylo_to_xml_state(id)
}
\arguments{
\item{id}{an alignment's IDs.
An ID can be extracted from its FASTA filesname
with \code{\link{get_ids}})}
}
\value{
the random phylogeny as XML text
}
\description{
Creates the XML of a random phylogeny,
  as used in the \code{state} section
}
\author{
Richel J.C. Bilderbeek
}
