% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnd_phylogeny_xml_init.R
\name{rnd_phylogeny_to_xml_init}
\alias{rnd_phylogeny_to_xml_init}
\title{Creates the XML of a random species tree,
  as used in the \code{init} section}
\usage{
rnd_phylogeny_to_xml_init(id)
}
\arguments{
\item{id}{the ID of the alignment (can be extracted from
its FASTA filesname using \code{\link{get_id}})}
}
\value{
the parameter as XML text
}
\description{
Creates the XML of a random species tree,
  as used in the \code{init} section
}
\author{
Richel J.C. Bilderbeek
}
