% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lines.R
\name{save_lines}
\alias{save_lines}
\title{Save text (a container of strings) to a file}
\usage{
save_lines(filename, lines)
}
\arguments{
\item{filename}{filename of the file to have the text written to}

\item{lines}{lines of text to be written to file}
}
\value{
Nothing
}
\description{
Save text (a container of strings) to a file
}
\author{
Richel J.C. Bilderbeek
}
