% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_text.R
\name{save_text}
\alias{save_text}
\title{Save text (a container of strings) to a file}
\usage{
save_text(filename, text)
}
\arguments{
\item{filename}{filename of the file to have the text written to}

\item{text}{text to be written to file}
}
\value{
Nothing
}
\description{
Save text (a container of strings) to a file
}
\author{
Richel Bilderbeek
}
