% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_model_to_xml.R
\name{site_model_to_xml_rates}
\alias{site_model_to_xml_rates}
\title{Converts a site model to XML,
  used in the \code{rates} section}
\usage{
site_model_to_xml_rates(site_model)
}
\arguments{
\item{site_model}{a site model,
as created by \code{\link{create_site_model}})}

\item{id}{a site model's ID}
}
\value{
the site model as XML text
}
\description{
Converts a site model to XML,
  used in the \code{rates} section
}
\author{
Richel J.C. Bilderbeek
}
