% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_model_to_xml_site_model.R
\name{site_model_to_xml_site_model}
\alias{site_model_to_xml_site_model}
\title{Converts a site model to XML,
  used in the \code{siteModel} section}
\usage{
site_model_to_xml_site_model(site_model)
}
\arguments{
\item{site_model}{a site model,
as created by \code{\link{create_site_model}})}
}
\value{
the site model as XML text
}
\description{
Converts a site model to XML,
  used in the \code{siteModel} section
}
\author{
Richel J.C. Bilderbeek
}
