% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_model_to_xml_state.R
\name{site_model_to_xml_state}
\alias{site_model_to_xml_state}
\title{Converts a site model to XML,
  used in the \code{state} section}
\usage{
site_model_to_xml_state(site_model)
}
\arguments{
\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}
}
\value{
the site model as XML text
}
\description{
Converts a site model to XML,
  used in the \code{state} section
}
\author{
Richèl J.C. Bilderbeek
}
