% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_model_to_xml_subst_model.R
\name{site_model_to_xml_subst_model}
\alias{site_model_to_xml_subst_model}
\title{Converts a site model to XML,
  used in the \code{substModel} section}
\usage{
site_model_to_xml_subst_model(inference_model, site_model = "deprecated")
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}

\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}
}
\value{
the site model as XML text
}
\description{
Converts a site model to XML,
  used in the \code{substModel} section
}
\author{
Richèl J.C. Bilderbeek
}
