% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_model_to_xml_tracelog.R
\name{site_model_to_xml_tracelog}
\alias{site_model_to_xml_tracelog}
\title{Creates the site model's XML for the tracelog section}
\usage{
site_model_to_xml_tracelog(site_model)
}
\arguments{
\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}
}
\description{
Creates the site model's XML for the tracelog section
}
\examples{
# <logger id="tracelog" ...>
#'   # Here
# </logger>
}
\seealso{
all site models' tracelog section is created
  by \code{\link{site_models_to_xml_tracelog}}
}
\author{
Richel J.C. Bilderbeek
}
