% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_models_to_xml_operators.R
\name{site_models_to_xml_operators}
\alias{site_models_to_xml_operators}
\title{Write the XML \code{operators} section from the site models.}
\usage{
site_models_to_xml_operators(site_models)
}
\arguments{
\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}
}
\value{
lines of XML text
}
\description{
Write the XML \code{operators} section from the site models.
}
\author{
Richèl J.C. Bilderbeek
}
