% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_models_to_xml_prior_distr.R
\name{site_models_to_xml_prior_distr}
\alias{site_models_to_xml_prior_distr}
\title{Represent the site models as XML}
\usage{
site_models_to_xml_prior_distr(site_models)
}
\arguments{
\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}
}
\value{
lines of XML text
}
\description{
Represent the site models as XML
}
\examples{
check_empty_beautier_folder()

 # <distribution id="posterior" spec="util.CompoundDistribution">
 #     <distribution id="prior" spec="util.CompoundDistribution">
 #       HERE, where the ID of the distribution is 'prior'
 #     </distribution>
 #     <distribution id="likelihood" ...>
 #     </distribution>
 # </distribution>

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
