% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_models_to_xml_state.R
\name{site_models_to_xml_state}
\alias{site_models_to_xml_state}
\title{Converts one or more clock models to the \code{state} section of the
XML as text}
\usage{
site_models_to_xml_state(site_models)
}
\arguments{
\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}
}
\value{
lines of XML text, without indentation nor \code{state}
  tags
}
\description{
Converts one or more clock models to the \code{state} section of the
XML as text
}
\author{
Richel J.C. Bilderbeek
}
