% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-obj-type.R
\name{stop_input_type}
\alias{stop_input_type}
\title{Stops on the input type.}
\usage{
stop_input_type(
  x,
  what,
  ...,
  allow_na = FALSE,
  allow_null = FALSE,
  show_value = TRUE,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{The object type which does not conform to `what`. Its
`obj_type_friendly()` is taken and mentioned in the error message.}

\item{what}{The friendly expected type as a string. Can be a
character vector of expected types, in which case the error
message mentions all of them in an "or" enumeration.}

\item{...}{Arguments passed to \link[rlang]{abort}.}

\item{allow_na}{allow NA to be valid}

\item{allow_null}{allow NULL to be valid}

\item{show_value}{set to TRUE to show the value}

\item{arg}{No idea what this does}

\item{call}{No idea what this does}
}
\description{
Stops on the input type.
}
\note{
adapted from
[`r-lib`](https://github.com/r-lib/rlang/blob/main/R/standalone-obj-type.R)
}
\author{
[`olivroy`](https://github.com/olivroy)
}
