% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strict_clock_model_to_xml_prior_distr.R
\name{strict_clock_model_to_xml_prior_distr}
\alias{strict_clock_model_to_xml_prior_distr}
\title{Internal function}
\usage{
strict_clock_model_to_xml_prior_distr(inference_model)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\value{
a character vector of XML strings
}
\description{
Internal function to converts a strict clock model
to the \code{prior} section of the XML as text
}
\examples{
check_empty_beautier_folder()
 # <distribution id="posterior" spec="util.CompoundDistribution">
 #     <distribution id="prior" spec="util.CompoundDistribution">
 #       HERE, where the ID of the distribution is 'prior'
 #     </distribution>
 #     <distribution id="likelihood" ...>
 #     </distribution>
 # </distribution>
strict_clock_model_to_xml_prior_distr(
  inference_model = create_inference_model()
)
check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
