% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strict_clock_model_to_xml_tracelog.R
\name{strict_clock_model_to_xml_tracelog}
\alias{strict_clock_model_to_xml_tracelog}
\title{Internal function}
\usage{
strict_clock_model_to_xml_tracelog(inference_model)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\value{
a character vector of XML strings
}
\description{
Creates a strict clock model's XML for the tracelog section
}
\examples{
check_empty_beautier_folder()

# <logger id="tracelog" ...>
#'   # Here
# </logger>

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
