% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tn93_site_model_to_xml_prior_distr.R
\name{tn93_site_model_to_xml_prior_distr}
\alias{tn93_site_model_to_xml_prior_distr}
\title{Internal function}
\usage{
tn93_site_model_to_xml_prior_distr(site_model, beauti_options)
}
\arguments{
\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the site model as XML text
}
\description{
Converts a TN93 site model to XML,
  used in the \code{prior} section
}
\examples{
tn93_site_model_to_xml_prior_distr(
  site_model = create_tn93_site_model(
    id = 1,
    kappa_1_prior_distr = create_uniform_distr(id = 2),
    kappa_2_prior_distr = create_uniform_distr(id = 3)
  ),
  beauti_options = create_beauti_options()
)
}
\author{
Richèl J.C. Bilderbeek
}
