% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tn93_site_model_to_xml_state.R
\name{tn93_site_model_to_xml_state}
\alias{tn93_site_model_to_xml_state}
\title{Converts a site model to XML,
  used in the \code{state} section}
\usage{
tn93_site_model_to_xml_state(
  site_model,
  beauti_options = create_beauti_options()
)
}
\arguments{
\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the site model as XML text
}
\description{
Converts a site model to XML,
  used in the \code{state} section
}
\author{
Richèl J.C. Bilderbeek
}
