% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_model_to_xml_subst_model.R
\name{tn93_site_model_to_xml_subst_model}
\alias{tn93_site_model_to_xml_subst_model}
\title{Converts a TN93 site model to XML,
  used in the \code{substModel} section}
\usage{
tn93_site_model_to_xml_subst_model(site_model)
}
\arguments{
\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}
}
\value{
the site model as XML text
}
\description{
Converts a TN93 site model to XML,
  used in the \code{substModel} section
}
\author{
Richèl J.C. Bilderbeek
}
