% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_do.R
\name{to_do}
\alias{to_do}
\title{Indicates that something needs to be done}
\usage{
to_do()
}
\description{
Indicates that something needs to be done
}
\examples{
  x <- 4
  # Not  yet in the mood to respond to values of x below zero yet
  testit::assert(x > 0 && to_do())
}
