% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_prior_to_xml_operators.R
\name{tree_prior_to_xml_operators}
\alias{tree_prior_to_xml_operators}
\title{Internal function}
\usage{
tree_prior_to_xml_operators(
  inference_model,
  tree_prior = "deprecated",
  fixed_crown_age = "deprecated"
)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}

\item{tree_prior}{a tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{fixed_crown_age}{determines if the phylogeny's crown age is
fixed. If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}
}
\value{
the tree prior as XML text
}
\description{
Creates the XML of a tree prior,
  as used in the \code{operators} section
}
\author{
Richèl J.C. Bilderbeek
}
