% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_prior_to_xml_prior.R
\name{tree_prior_to_xml_prior}
\alias{tree_prior_to_xml_prior}
\title{Creates the XML of a tree prior,
  as used in the \code{prior} section}
\usage{
tree_prior_to_xml_prior(tree_prior)
}
\arguments{
\item{tree_prior}{a tree prior,
as returned by \code{\link{create_tree_prior}})}
}
\value{
the tree prior as XML text
}
\description{
Creates the XML of a tree prior,
  as used in the \code{prior} section
}
\author{
Richel J.C. Bilderbeek
}
