% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_prior_to_xml_prior_distr.R
\name{tree_prior_to_xml_prior_distr}
\alias{tree_prior_to_xml_prior_distr}
\title{Creates the distribution section in the prior section of the
distribution section of a BEAST2 XML parameter file.
These lines start with '<distribution id='}
\usage{
tree_prior_to_xml_prior_distr(tree_prior)
}
\description{
Creates the distribution section in the prior section of the
distribution section of a BEAST2 XML parameter file.
These lines start with '<distribution id='
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
