% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_prior_to_xml_state.R
\name{tree_prior_to_xml_state}
\alias{tree_prior_to_xml_state}
\title{Creates the XML of a tree prior,
  as used in the \code{state} section}
\usage{
tree_prior_to_xml_state(inference_model)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\value{
the tree prior as XML text
}
\description{
Creates the XML of a tree prior,
  as used in the \code{state} section
}
\author{
Richèl J.C. Bilderbeek
}
