% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_prior_to_xml_tracelog.R
\name{tree_prior_to_xml_tracelog}
\alias{tree_prior_to_xml_tracelog}
\title{Creates the tree priors part of the two logger sections
  of a BEAST2 XML parameter file}
\usage{
tree_prior_to_xml_tracelog(tree_prior)
}
\arguments{
\item{tree_prior}{a tree prior,
as created by \code{\link{create_tree_prior}}}
}
\description{
Creates the tree priors part of the two logger sections
  of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
