% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_priors_to_xml_operators.R
\name{tree_priors_to_xml_operators}
\alias{tree_priors_to_xml_operators}
\title{Creates the XML of a list of one or more tree priors,
  as used in the \code{operators} section}
\usage{
tree_priors_to_xml_operators(tree_priors, fixed_crown_ages = rep(FALSE, times
  = length(tree_priors)))
}
\arguments{
\item{tree_priors}{one or more tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{fixed_crown_ages}{one or more booleans to determine if the
phylogenies' crown ages are fixed.
If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}
}
\value{
the tree priors as XML text
}
\description{
Creates the XML of a list of one or more tree priors,
  as used in the \code{operators} section
}
\author{
Richel J.C. Bilderbeek
}
