% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_priors_to_xml_distr.R
\name{tree_priors_to_xml_prior_distr}
\alias{tree_priors_to_xml_prior_distr}
\title{Creates the distribution section in the prior section of the
distribution section of a BEAST2 XML parameter file.}
\usage{
tree_priors_to_xml_prior_distr(tree_priors)
}
\arguments{
\item{tree_priors}{one or more tree priors,
as returned by \code{\link{create_tree_prior}}}
}
\value{
lines of XML text
}
\description{
These lines start with '<distribution id='
}
\examples{
check_empty_beautier_folder()

 # <distribution id="posterior" spec="util.CompoundDistribution">
 #     <distribution id="prior" spec="util.CompoundDistribution">
 #       HERE, where the ID of the distribution is 'prior'
 #     </distribution>
 #     <distribution id="likelihood" ...>
 #     </distribution>
 # </distribution>

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
