% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_priors_to_xml_state.R
\name{tree_priors_to_xml_state}
\alias{tree_priors_to_xml_state}
\title{Converts one or more tree priors to the \code{state} section of the
XML as text}
\usage{
tree_priors_to_xml_state(tree_priors = "deprecated")
}
\arguments{
\item{tree_priors}{one or more tree priors,
as returned by \code{\link{create_tree_prior}}}
}
\value{
lines of XML text, without indentation nor \code{state}
  tags
}
\description{
Converts one or more tree priors to the \code{state} section of the
XML as text
}
\author{
Richèl J.C. Bilderbeek
}
