% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unindent.R
\name{unindent}
\alias{unindent}
\title{Unindents text}
\usage{
unindent(text)
}
\arguments{
\item{text}{one or more lines of text}
}
\value{
unindented lines of text
}
\description{
Unindents text
}
\author{
Richèl J.C. Bilderbeek
}
