% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yule_tree_prior_to_xml_operators.R
\name{yule_tree_prior_to_xml_operators}
\alias{yule_tree_prior_to_xml_operators}
\title{Internal function}
\usage{
yule_tree_prior_to_xml_operators(inference_model)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\value{
the tree prior as XML text
}
\description{
Creates the XML of a Yule tree prior,
  as used in the \code{operators} section
}
\author{
Richèl J.C. Bilderbeek
}
