% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_clock_models.R
\name{are_clock_models}
\alias{are_clock_models}
\title{Determine if x consists out of clock_models objects}
\usage{
are_clock_models(x)
}
\arguments{
\item{x}{the object to check if it consists out of clock_models objects}
}
\value{
TRUE if x, or all elements of x, are clock_model objects
}
\description{
Determine if x consists out of clock_models objects
}
\author{
Richel J.C. Bilderbeek
}
