% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_equivalent_xmls.R
\name{are_equivalent_xml_files}
\alias{are_equivalent_xml_files}
\title{Internal function}
\usage{
are_equivalent_xml_files(filename_1, filename_2, section = NA)
}
\arguments{
\item{filename_1}{name of a first XML file}

\item{filename_2}{name of a second XML file}

\item{section}{the name of the XML section, use NA to check the whole file}
}
\value{
TRUE if the two XML files result in equivalent trees,
  FALSE otherwise
}
\description{
Internal function used for debugging to
determine if XML files result in equivalent trees
}
\examples{
check_empty_beautier_folder()

are_equivalent_xml_files(
  filename_1 = get_beautier_path("2_4.xml"),
  filename_2 = get_beautier_path("2_6_0.xml")
)

check_empty_beautier_folder()
}
\seealso{
to check for equality, use \code{are_equal_xml_files}
}
\author{
Richèl J.C. Bilderbeek
}
