% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_mrca_priors.R
\name{are_mrca_priors}
\alias{are_mrca_priors}
\title{Determine if x consists out of MRCA priors}
\usage{
are_mrca_priors(mrca_priors)
}
\arguments{
\item{mrca_priors}{a list of one or more Most Recent Common Ancestor priors,
as returned by \code{\link{create_mrca_prior}}}
}
\value{
TRUE if x, or all elements of x, are MRCA priors.
  Returns FALSE otherwise
}
\description{
Determine if x consists out of MRCA priors
}
\author{
Richèl J.C. Bilderbeek
}
