% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_mrca_priors.R
\name{are_mrca_priors}
\alias{are_mrca_priors}
\title{Determine if x consists out of mrca_priors objects}
\usage{
are_mrca_priors(x)
}
\arguments{
\item{x}{the object to check if it consists out of mrca_priors objects}
}
\value{
TRUE if x, or all elements of x, are mrca_prior objects
}
\description{
Determine if x consists out of mrca_priors objects
}
\author{
Richel J.C. Bilderbeek
}
