% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_tree_priors.R
\name{are_tree_priors}
\alias{are_tree_priors}
\title{Determine if x consists out of tree_priors objects}
\usage{
are_tree_priors(x)
}
\arguments{
\item{x}{the object to check if it consists out of tree_priors objects}
}
\value{
TRUE if x, or all elements of x, are tree_prior objects
}
\description{
Determine if x consists out of tree_priors objects
}
\examples{
  yule_tree_prior <- create_yule_tree_prior()
  bd_tree_prior <- create_bd_tree_prior()
  both_tree_priors <- list(yule_tree_prior, bd_tree_prior)
  testit::assert(are_tree_priors(yule_tree_prior))
  testit::assert(are_tree_priors(bd_tree_prior))
  testit::assert(are_tree_priors(both_tree_priors))
}
\seealso{
Use \link{create_yule_tree_prior} to create a Yule tree prior
}
\author{
Richèl J.C. Bilderbeek
}
