% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastscriptr.R
\name{beast_scriptr}
\alias{beast_scriptr}
\title{Create a BEAST2 XML parameter file}
\usage{
beast_scriptr(input_fasta_filename, mcmc_chainlength, tree_prior,
  output_xml_filename)
}
\arguments{
\item{input_fasta_filename}{Filename of a fasta file}

\item{mcmc_chainlength}{Length of MCMC chain}

\item{tree_prior}{The tree prior, can be 'birth_death' or
'coalescent_constant_population'}

\item{output_xml_filename}{Filename of the XML parameter file created}
}
\description{
Create a BEAST2 XML parameter file
}
