% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beast_scriptr.R
\name{beast_scriptr}
\alias{beast_scriptr}
\title{Deprecated interface for \code{\link{create_beast2_input_file}},
  present for backwards-compatibility}
\usage{
beast_scriptr(input_fasta_filename, output_xml_filename,
  mcmc_chainlength = 1e+07, tree_prior = "yule", fixed_crown_age = FALSE,
  initial_phylogenies = rep(NA, length(input_fasta_filename)))
}
\arguments{
\item{input_fasta_filename}{One or more fasta filename}

\item{output_xml_filename}{Name of the XML parameter file created by this
function. BEAST2 uses this file as input.}

\item{mcmc_chainlength}{Length of MCMC chain}

\item{tree_prior}{One or more tree prior names,
must be a name in \code{\link{get_tree_prior_names}}}

\item{fixed_crown_age}{determines if the phylogeny its crown age is
fixed. If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}

\item{initial_phylogenies}{the MCMC chain its initial phylogenies. If
this is set to NA, BEAST2 will use a random phylogeny. Else
a phylogeny must be supplied of class ape::phylo.}
}
\description{
Deprecated interface for \code{\link{create_beast2_input_file}},
  present for backwards-compatibility
}
\examples{
  # The file created by beastscriptr, a BEAST2 input file
  output_xml_filename <- "example_bd.xml"

  # Birth-Death tree prior, crown age is estimated
  beast_scriptr(
    input_fasta_filename = get_input_fasta_filename(),
    output_xml_filename = output_xml_filename,
    mcmc_chainlength = 10000000,
    tree_prior = "birth_death"
  )
  testthat::expect_true(file.exists(output_xml_filename))

  # The file created by beastscriptr, a BEAST2 input file
  output_xml_filename_fixed <- "example_bd_fixed.xml"

  # Birth-Death tree prior, crown age is fixed at 15 time units
  beast_scriptr(
    input_fasta_filename = get_input_fasta_filename(),
    output_xml_filename = output_xml_filename_fixed,
    mcmc_chainlength = 10000000,
    tree_prior = "birth_death",
    fixed_crown_age = TRUE,
    initial_phylogenies = beastscriptr::fasta_to_phylo(
      fasta_filename = get_input_fasta_filename(),
      crown_age = 15)
  )
  testthat::expect_true(file.exists(output_xml_filename_fixed))
}
\author{
Richel J.C. Bilderbeek
}
