% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastscriptr.R
\name{beast_scriptr_distribution}
\alias{beast_scriptr_distribution}
\title{Creates the distribution section of a BEAST2 XML parameter file}
\usage{
beast_scriptr_distribution(filename_base, tree_prior)
}
\arguments{
\item{filename_base}{base of the filenames}

\item{tree_prior}{the tree prior}
}
\description{
Creates the distribution section of a BEAST2 XML parameter file
}
