% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastscriptr.R
\name{beast_scriptr_map}
\alias{beast_scriptr_map}
\title{Creates the map section of a BEAST2 XML parameter file}
\usage{
beast_scriptr_map()
}
\description{
Creates the map section of a BEAST2 XML parameter file
}
