% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_clock_model.R
\name{check_clock_model}
\alias{check_clock_model}
\title{Check if the clock model is a valid clock model.}
\usage{
check_clock_model(clock_model)
}
\arguments{
\item{clock_model}{a clock model,
as returned by \code{\link{create_clock_model}}}
}
\value{
TRUE if \code{clock_model} is a valid clock model
}
\description{
Calls \code{stop} if the clock model is invalid
}
\examples{
check_empty_beautier_folder()

check_clock_model(create_strict_clock_model())
check_clock_model(create_rln_clock_model())

check_empty_beautier_folder()
}
\seealso{
Use \link{create_clock_model} to create a valid clock model
}
\author{
Richèl J.C. Bilderbeek
}
