% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_true_false.R
\name{check_false}
\alias{check_false}
\title{Determine if `x` is `FALSE`}
\usage{
check_false(
  x,
  ...,
  allow_na = FALSE,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{x}{the object to be determined to be `FALSE`}
}
\value{
Nothing. Will raise an exception if the value is not `FALSE`
}
\description{
Determine if `x` is `FALSE`
}
\note{
Created by modifying check_bool from import-standalone-type-check.
}
\examples{
check_false(FALSE)
}
\author{
[`olivroy`](https://github.com/olivroy)
}
