% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_exists.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Function to check if a file exists.
Calls \code{stop} if the file is absent}
\usage{
check_file_exists(filename, filename_description = NA)
}
\arguments{
\item{filename}{name of the file}

\item{filename_description}{description of the filename}
}
\value{
nothing. Will \code{stop} if the file is absent,
  with a proper error message
}
\description{
Function to check if a file exists.
Calls \code{stop} if the file is absent
}
\examples{
check_file_exists(get_beautier_path("anthus_aco_sub.fas"))
}
\author{
Richèl J.C. Bilderbeek
}
