% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filename.R
\name{check_filename}
\alias{check_filename}
\title{Check if the `filename` is valid}
\usage{
check_filename(filename, allow_empty_str = FALSE, allow_na = FALSE)
}
\arguments{
\item{filename}{a filename, as can be checked by \link{check_filename}}

\item{allow_empty_str}{allow a string to be empty}

\item{allow_na}{allow \link{NA}}
}
\value{
The filename (invisibly)
}
\description{
Calls \code{stop} if the filename is invalid
}
\examples{
check_empty_beautier_folder()

check_filename("trace.log")
check_filename("my.trees")

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
