% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gamma_site_model.R
\name{check_gamma_site_model}
\alias{check_gamma_site_model}
\title{Checks if the parameter is a valid gamma site model}
\usage{
check_gamma_site_model(gamma_site_model)
}
\arguments{
\item{gamma_site_model}{a site model's gamma site model,
as returned by \code{\link{create_gamma_site_model}}}
}
\value{
nothing. Will call \code{stop} if the argument is not a valid
  gamma site model
}
\description{
Checks if the parameter is a valid gamma site model
}
\examples{
library(testthat)

expect_silent(
  check_gamma_site_model(
    create_gamma_site_model()
  )
)
expect_error(
  check_gamma_site_model(
    "not a gamma site model"
  )
)
}
\author{
Richèl J.C. Bilderbeek
}
