% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mcmc.R
\name{check_mcmc_values}
\alias{check_mcmc_values}
\title{Check if the MCMC has the list elements with valid values
for being a valid MCMC object.}
\usage{
check_mcmc_values(mcmc)
}
\arguments{
\item{mcmc}{one MCMC.
Use \code{\link{create_mcmc}} to create an MCMC.
Use \code{\link{create_ns_mcmc}} to create an MCMC
  for a Nested Sampling run.
Use \code{\link{check_mcmc}} to check if an MCMC is valid.
Use \code{\link{rename_mcmc_filenames}} to rename the filenames in an MCMC.}
}
\value{
nothing
}
\description{
Calls \code{stop} if a value is invalid
}
\seealso{
Use \code{\link{create_mcmc}} to create a valid MCMC
}
\author{
Richèl J.C. Bilderbeek
}
