% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mcmc_nested_sampling.R
\name{check_ns_mcmc}
\alias{check_ns_mcmc}
\alias{check_mcmc_nested_sampling}
\alias{check_nested_sampling_mcmc}
\title{Check if this an MCMC that uses Nested Sampling
to estimate a marginal likelihood.}
\usage{
check_ns_mcmc(mcmc)
}
\arguments{
\item{mcmc}{one MCMC
as returned by \code{\link{create_mcmc}}}
}
\description{
Will \link{stop} if not, else will do nothing
}
\seealso{
use \link{create_ns_mcmc}
to create an MCMC that uses Nested Sampling
to estimate a marginal likelihood
}
\author{
Richèl J.C. Bilderbeek
}
