% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_string.R
\name{check_string}
\alias{check_string}
\title{Determine if `x` is one string}
\usage{
check_string(x, allow_na = FALSE, allow_empty = TRUE)
}
\arguments{
\item{x}{the object to be determined to be one string}

\item{allow_na}{allow \link{NA}}
}
\value{
Nothing. Will raise an exception if the value is not one string
}
\description{
Determine if `x` is one string
}
\examples{
check_string("This is a string :-)")
}
\author{
Richèl J.C. Bilderbeek
}
