% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_treelog.R
\name{check_treelog}
\alias{check_treelog}
\title{Check if a \code{treelog} is valid.}
\usage{
check_treelog(treelog)
}
\arguments{
\item{treelog}{a \code{treelog},
as created by \link{create_treelog}}
}
\value{
No return value, called for side effects
}
\description{
Will call \link{stop} if not.
}
\examples{
check_empty_beautier_folder()

check_treelog(create_test_treelog())

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
