% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_true_false.R
\name{check_true}
\alias{check_true}
\title{Determine if `x` is `TRUE`}
\usage{
check_true(x)
}
\arguments{
\item{x}{the object to be determined to be `TRUE`}
}
\value{
Nothing. Will raise an exception if the value is not `TRUE`
}
\description{
Determine if `x` is `TRUE`
}
\examples{
check_true(TRUE)
}
\author{
Richèl J.C. Bilderbeek
}
