% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_model_to_xml_lh_distr.R
\name{clock_model_to_xml_lh_distr}
\alias{clock_model_to_xml_lh_distr}
\title{Converts a clock model to the \code{branchRateModel} section of the
XML as text}
\usage{
clock_model_to_xml_lh_distr(clock_model, is_first = TRUE,
  is_non_first_shared = TRUE)
}
\arguments{
\item{clock_model}{a clock model,
as returned by \code{\link{create_clock_model}}}

\item{is_first}{boolean to indicate if this is the first
out of many (one, two or more)}

\item{is_non_first_shared}{is this clock model not the first of
multiple shared ones?}
}
\description{
Converts a clock model to the \code{branchRateModel} section of the
XML as text
}
\examples{
 # <distribution id="posterior" spec="util.CompoundDistribution">
 #     <distribution id="prior" spec="util.CompoundDistribution">
 #     </distribution>
 #     <distribution id="likelihood" ...>
 #       HERE, where the ID of the distribution is 'likelihood'
 #     </distribution>
 # </distribution>
}
\author{
Richel J.C. Bilderbeek
}
