% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_model_to_xml_operators.R
\name{clock_model_to_xml_operators}
\alias{clock_model_to_xml_operators}
\title{Converts a clock model to the \code{operators} section of the
XML as text}
\usage{
clock_model_to_xml_operators(
  inference_model,
  clock_model = "deprecated",
  mrca_priors = "deprecated",
  tipdates_filename = "deprecated"
)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}

\item{clock_model}{a clock model,
as returned by \code{\link{create_clock_model}}}

\item{mrca_priors}{a list of one or more Most Recent Common Ancestor priors,
as returned by \code{\link{create_mrca_prior}}}

\item{tipdates_filename}{name of the file containing the tip dates.
This file is assumed to have two columns, separated by a tab.
The first column contains the taxa names, the second column contains
the date.}
}
\value{
a character vector of XML strings
}
\description{
Converts a clock model to the \code{operators} section of the
XML as text
}
\author{
Richèl J.C. Bilderbeek
}
