% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_model_to_xml_prior_distr.R
\name{clock_model_to_xml_prior_distr}
\alias{clock_model_to_xml_prior_distr}
\title{Converts a clock model to the \code{prior} section of the
XML as text}
\usage{
clock_model_to_xml_prior_distr(clock_model)
}
\arguments{
\item{clock_model}{a clock model,
as returned by \code{\link{create_clock_model}}}
}
\description{
Converts a clock model to the \code{prior} section of the
XML as text
}
\author{
Richel J.C. Bilderbeek
}
