% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_model_to_xml_state.R
\name{clock_model_to_xml_state}
\alias{clock_model_to_xml_state}
\title{Converts a clock model to the \code{state} section of the
XML as text}
\usage{
clock_model_to_xml_state(clock_model, has_tip_dating = FALSE)
}
\arguments{
\item{clock_model}{a clock model,
as returned by \code{\link{create_clock_model}}}

\item{has_tip_dating}{TRUE if the user has supplied tip dates,
FALSE otherwise}
}
\value{
lines of XML text, without indentation nor \code{state}
  tags
}
\description{
Converts a clock model to the \code{state} section of the
XML as text
}
\author{
Richèl J.C. Bilderbeek
}
