% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_model_to_xml_tracelog.R
\name{clock_model_to_xml_tracelog}
\alias{clock_model_to_xml_tracelog}
\title{Creates the clock model's XML for the tracelog section}
\usage{
clock_model_to_xml_tracelog(clock_model, is_first)
}
\arguments{
\item{clock_model}{a clock model,
as returned by \code{\link{create_clock_model}}}

\item{is_first}{boolean to indicate if this is the first
out of many (one, two or more)}
}
\description{
Creates the clock model's XML for the tracelog section
}
\examples{
# <logger id="tracelog" ...>
#'   # Here
# </logger>
}
\seealso{
all clock models' tracelog section is created
  by \code{\link{clock_model_to_xml_tracelog}}
}
\author{
Richel J.C. Bilderbeek
}
