% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_models_to_xml_operators.R
\name{clock_models_to_xml_operators}
\alias{clock_models_to_xml_operators}
\title{Deprecated}
\usage{
clock_models_to_xml_operators()
}
\description{
Deprecated
}
\author{
Richèl J.C. Bilderbeek
}
