% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_models_to_xml_operators.R
\name{clock_models_to_xml_operators}
\alias{clock_models_to_xml_operators}
\title{Create all clock models' operators' XML text}
\usage{
clock_models_to_xml_operators(clock_models)
}
\arguments{
\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}
}
\description{
Create all clock models' operators' XML text
}
\author{
Richel J.C. Bilderbeek
}
