% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_models_to_xml_prior_distr.R
\name{clock_models_to_xml_prior_distr}
\alias{clock_models_to_xml_prior_distr}
\title{Represent the clock models as XML}
\usage{
clock_models_to_xml_prior_distr(clock_models)
}
\arguments{
\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}
}
\description{
Represent the clock models as XML
}
\examples{
 # <distribution id="posterior" spec="util.CompoundDistribution">
 #     <distribution id="prior" spec="util.CompoundDistribution">
 #       HERE, where the ID of the distribution is 'prior'
 #     </distribution>
 #     <distribution id="likelihood" ...>
 #     </distribution>
 # </distribution>
}
\author{
Richel J.C. Bilderbeek
}
