% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_models_to_xml_state.R
\name{clock_models_to_xml_state}
\alias{clock_models_to_xml_state}
\title{Converts one or more clock models to the \code{state} section of the
XML as text}
\usage{
clock_models_to_xml_state(clock_models)
}
\arguments{
\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}
}
\value{
lines of XML text, without indentation nor \code{state}
  tags
}
\description{
Converts one or more clock models to the \code{state} section of the
XML as text
}
\author{
Richel J.C. Bilderbeek
}
