% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_models_to_xml_tracelog.R
\name{clock_models_to_xml_tracelog}
\alias{clock_models_to_xml_tracelog}
\title{Creates the clock models' XML for the tracelog section}
\usage{
clock_models_to_xml_tracelog(clock_models)
}
\arguments{
\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}
}
\description{
Creates the clock models' XML for the tracelog section
}
\examples{
# <logger id="tracelog" ...>
#'   # Here
# </logger>
}
\seealso{
the complete tracelog section is created
  by \code{\link{create_beast2_input_tracelog}}
}
\author{
Richel J.C. Bilderbeek
}
