% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_lines.R
\name{compare_lines}
\alias{compare_lines}
\title{Internal debug function to compare the actually created
lines to expected lines.}
\usage{
compare_lines(lines, expected)
}
\description{
Internal debug function to compare the actually created
lines to expected lines.
}
