% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_lines.R
\name{compare_lines}
\alias{compare_lines}
\title{Internal debug function to compare the actually created
lines to expected lines using any diff tool}
\usage{
compare_lines(lines, expected, section = NA,
  created_lines_filename = "~/created.xml",
  expected_lines_filename = "~/expected.xml")
}
\arguments{
\item{lines}{the created lines}

\item{expected}{the expected/goal/target lines}

\item{section}{the XML section. Leave at NA to compare all lines}

\item{created_lines_filename}{name of the file where the (section of
the) created lines are stored}

\item{expected_lines_filename}{name of the file where the (section of
the) expected lines are stored}
}
\description{
Internal debug function to compare the actually created
lines to expected lines using any diff tool
}
