% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_trailing_spaces.R
\name{count_trailing_spaces}
\alias{count_trailing_spaces}
\title{Count the number of spaces before the first character}
\usage{
count_trailing_spaces(line)
}
\arguments{
\item{line}{line of text}
}
\value{
the number of spaces before the first character
}
\description{
Count the number of spaces before the first character
}
\examples{
library(testthat)

expect_equal(count_trailing_spaces("x"), 0)
expect_equal(count_trailing_spaces(" y"), 1)
expect_equal(count_trailing_spaces("  <"), 2)
expect_equal(count_trailing_spaces(""), 0)
expect_equal(count_trailing_spaces(" "), 1)
expect_equal(count_trailing_spaces("  "), 2)
}
\author{
Richèl J.C. Bilderbeek
}
