% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_trailing_spaces.R
\name{count_trailing_spaces}
\alias{count_trailing_spaces}
\title{Count the number of spaces before the first character}
\usage{
count_trailing_spaces(line)
}
\arguments{
\item{line}{line of text}
}
\value{
the number of spaces before the first character
}
\description{
Count the number of spaces before the first character
}
\examples{
check_empty_beautier_folder()

# 0
count_trailing_spaces("x")
# 1
count_trailing_spaces(" y")
# 2
count_trailing_spaces("  <")
# 0
count_trailing_spaces("")
# 1
count_trailing_spaces(" ")
# 2
count_trailing_spaces("  ")

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
