% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_alpha_param}
\alias{create_alpha_param}
\title{Create a parameter called alpha}
\usage{
create_alpha_param(id = NA, estimate = FALSE, value = 0)
}
\arguments{
\item{id}{the parameter's ID}

\item{estimate}{TRUE if this parameter alpha be estimated by BEAST2,
FALSE otherwise}

\item{value}{value of the parameter}
}
\value{
a parameter called alpha
}
\description{
Create a parameter called alpha
}
\note{
this parameter is used in a beta distribution
  (as returned by \code{\link{create_beta_distr}})
and gamma distribution
  (as returned by \code{\link{create_gamma_distr}})
and inverse gamma distribution
  (as returned by \code{\link{create_inv_gamma_distr}})
}
\examples{
  # Create the parameter
  alpha_param <- create_alpha_param()
  testit::assert(is_alpha_param(alpha_param))

  # Use the parameter in a distribution
  beta_distr <- create_beta_distr(
    alpha = alpha_param
  )
  testit::assert(is_beta_distr(beta_distr))

  # Use the distribution to create a BEAST2 input file
  input_fasta_filename <- system.file(
    "extdata", "anthus_aco.fas", package = "beautier"
  )
  create_beast2_input_file(
    input_fasta_filenames = input_fasta_filename,
    "my_beast.xml",
    tree_priors = create_yule_tree_prior(
      birth_rate_distr = beta_distr
    )
  )
  testit::assert(file.exists("my_beast.xml"))
}
\seealso{
the function \code{\link{create_param}} contains a list
  of all parameters that can be created
}
\author{
Richel J.C. Bilderbeek
}
