% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_prior.R
\name{create_bd_tree_prior}
\alias{create_bd_tree_prior}
\title{Create a Birth-Death tree prior}
\usage{
create_bd_tree_prior()
}
\value{
a Birth-Death tree_prior
}
\description{
Create a Birth-Death tree prior
}
