% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_prior.R
\name{create_bd_tree_prior}
\alias{create_bd_tree_prior}
\title{Create a Birth-Death tree prior}
\usage{
create_bd_tree_prior(
  birth_rate_distr = beautier::create_uniform_distr(),
  death_rate_distr = beautier::create_uniform_distr()
)
}
\arguments{
\item{birth_rate_distr}{the birth rate distribution,
as created by a \code{\link{create_distr}} function}

\item{death_rate_distr}{the death rate distribution,
as created by a \code{\link{create_distr}} function}
}
\value{
a Birth-Death tree_prior

a Birth-Death tree_prior
}
\description{
Create a Birth-Death tree prior
}
\examples{
  bd_tree_prior <- create_bd_tree_prior()
  testit::assert(is_bd_tree_prior(bd_tree_prior))

  input_fasta_filename <- system.file(
    "extdata", "anthus_aco.fas", package = "beautier"
  )
  create_beast2_input_file(
    input_fasta_filenames = input_fasta_filename,
    "my_beast.xml",
    tree_priors = bd_tree_prior
  )
  testit::assert(file.exists("my_beast.xml"))
}
\author{
Richel J.C. Bilderbeek
}
