% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_beast_xml.R
\name{create_beast2_beast_xml}
\alias{create_beast2_beast_xml}
\title{Create the \code{<beast ...>} XML}
\usage{
create_beast2_beast_xml(beast2_version, required = "")
}
\arguments{
\item{beast2_version}{BEAST2 version}

\item{required}{requirement}
}
\value{
the XML
}
\description{
Create the \code{<beast ...>} XML
}
\author{
Richèl J.C. Bilderbeek
}
