% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input.R
\name{create_beast2_input}
\alias{create_beast2_input}
\title{Create a BEAST2 XML input text}
\usage{
create_beast2_input(input_fasta_filenames,
  site_models = create_site_model(name = "JC69"), mcmc_chainlength = 1e+07,
  tree_priors = create_tree_prior(name = "birth_death"),
  fixed_crown_age = FALSE, initial_phylogeny = NA)
}
\arguments{
\item{input_fasta_filenames}{One or more fasta filenames}

\item{site_models}{one or more site models,
as returned by 'create_site_models'}

\item{mcmc_chainlength}{Length of MCMC chain}

\item{tree_priors}{On or more tree priors,
as returned by 'create_tree_prior'}

\item{fixed_crown_age}{determines if the phylogeny its crown age is
fixed. If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}

\item{initial_phylogeny}{the MCMC chain its initial phylogeny. If
this is set to NA, BEAST2 will use a random phylogeny. Else
a phylogeny must be supplied of class ape::phylo.}
}
\description{
Create a BEAST2 XML input text
}
\examples{
  # Create a BEAST2 input file's text from th example FASTA file
  xml <- create_beast2_input(
    input_fasta_filenames = get_input_fasta_filename()
  )
}
\author{
Richel Bilderbeek
}
