% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input.R
\name{create_beast2_input}
\alias{create_beast2_input}
\title{Create a BEAST2 XML input text}
\usage{
create_beast2_input(input_fasta_filenames,
  site_models = create_jc69_site_models(n = length(input_fasta_filenames)),
  clock_models = create_strict_clock_models(n =
  length(input_fasta_filenames)), tree_priors = rep(create_yule_tree_prior(),
  length(input_fasta_filenames)), mcmc_chainlength = 1e+07,
  misc_options = create_misc_options(), fixed_crown_age = FALSE,
  initial_phylogenies = rep(NA, length(input_fasta_filenames)))
}
\arguments{
\item{input_fasta_filenames}{One or more fasta filenames}

\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}

\item{clock_models}{On or more clock models,
as returned by \code{\link{create_clock_model}}}

\item{tree_priors}{On or more tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{mcmc_chainlength}{Length of MCMC chain}

\item{misc_options}{one misc_options object,
as returned by \code{\link{create_misc_options}}}

\item{fixed_crown_age}{determines if the phylogeny its crown age is
fixed. If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}

\item{initial_phylogenies}{one or more MCMC chain's initial phylogenies.
Each one set to NA will result in BEAST2 using a random phylogeny. Else
the phylogeny is assumed to be of class ape::phylo.}
}
\description{
Create a BEAST2 XML input text
}
\examples{
  # Create a BEAST2 input file's text from th example FASTA file
  xml <- create_beast2_input(
    input_fasta_filenames = get_input_fasta_filename()
  )
}
\author{
Richel J.C. Bilderbeek
}
