% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_beast.R
\name{create_beast2_input_beast}
\alias{create_beast2_input_beast}
\title{Creates the beast section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_beast(input_fasta_filenames,
  site_models = create_site_model(name = "JC69"),
  clock_models = create_clock_model(name = "strict"),
  tree_priors = create_tree_prior(name = "yule"), mcmc_chainlength = 1e+07,
  fixed_crown_age = FALSE, initial_phylogeny = NA)
}
\arguments{
\item{input_fasta_filenames}{One or more fasta filenames}

\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}

\item{clock_models}{On or more clock models,
as returned by \code{\link{create_clock_model}}}

\item{tree_priors}{On or more tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{mcmc_chainlength}{Length of MCMC chain}

\item{fixed_crown_age}{determines if the phylogeny its crown age is
fixed. If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}

\item{initial_phylogeny}{the MCMC chain its initial phylogeny. If
this is set to NA, BEAST2 will use a random phylogeny. Else
a phylogeny must be supplied of class ape::phylo.}
}
\description{
Creates the beast section of a BEAST2 XML parameter file
}
