% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_data.R
\name{create_beast2_input_data}
\alias{create_beast2_input_data}
\title{Creates the data section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_data(input_filenames,
  misc_options = create_misc_options())
}
\arguments{
\item{input_filenames}{One or more FASTA filenames.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}

\item{misc_options}{one misc_options object,
as returned by \code{\link{create_misc_options}}}
}
\description{
Creates the data section of a BEAST2 XML parameter file
}
\author{
Richel J.C. Bilderbeek
}
