% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_data_sequences.R
\name{create_beast2_input_data_sequences}
\alias{create_beast2_input_data_sequences}
\title{Creates the data section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_data_sequences(
  input_fasta_filename,
  beauti_options = beautier::create_beauti_options()
)
}
\arguments{
\item{input_fasta_filename}{one FASTA filename}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
lines of XML text
}
\description{
Creates the data section of a BEAST2 XML parameter file
}
\author{
Richèl J.C. Bilderbeek
}
