% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distribution.R
\name{create_beast2_input_distribution}
\alias{create_beast2_input_distribution}
\title{Creates the distribution section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_distribution(ids, site_models = create_jc69_site_models(n
  = length(ids)), clock_models = create_strict_clock_models(n = length(ids)),
  tree_priors = create_yule_tree_priors(n = length(ids)))
}
\arguments{
\item{ids}{the IDs of the alignments (can be extracted from
their FASTA filesnames using \code{\link{get_id}})}

\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}

\item{clock_models}{On or more clock models,
as returned by \code{\link{create_clock_model}}}

\item{tree_priors}{On or more tree priors,
as returned by \code{\link{create_tree_prior}}}
}
\description{
Creates the distribution section of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
