% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_file.R
\name{create_beast2_input_file}
\alias{create_beast2_input_file}
\title{Create a BEAST2 input file}
\usage{
create_beast2_input_file(input_fasta_filenames, output_xml_filename,
  site_models = create_site_model(name = "JC69"),
  clock_models = create_clock_model(name = "strict"),
  tree_priors = create_tree_prior(name = "yule"), mcmc_chainlength = 1e+07,
  fixed_crown_age = FALSE, initial_phylogeny = NA)
}
\arguments{
\item{input_fasta_filenames}{One or more fasta filenames}

\item{output_xml_filename}{Name of the XML parameter file created by this
function. BEAST2 uses this file as input.}

\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}

\item{clock_models}{On or more clock models,
as returned by \code{\link{create_clock_model}}}

\item{tree_priors}{On or more tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{mcmc_chainlength}{Length of MCMC chain}

\item{fixed_crown_age}{determines if the phylogeny its crown age is
fixed. If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}

\item{initial_phylogeny}{the MCMC chain its initial phylogeny. If
this is set to NA, BEAST2 will use a random phylogeny. Else
a phylogeny must be supplied of class ape::phylo.}
}
\description{
Create a BEAST2 input file
}
\examples{
  # Get the filename of an example FASTA file
  input_fasta_filename <- get_input_fasta_filename()
  testit::assert(file.exists(input_fasta_filename))

  # The file created by beastscriptr, a BEAST2 input file
  output_xml_filename <- "example_bd.xml"

  # Birth-Death tree prior, crown age is estimated
  create_beast2_input_file(
    input_fasta_filenames = get_input_fasta_filename(),
    output_xml_filename = output_xml_filename
  )
  testit::assert(file.exists(output_xml_filename))

  # The file created by beastscriptr, a BEAST2 input file
  output_xml_filename_fixed <- "example_bd_fixed.xml"

  # Birth-Death tree prior, crown age is fixed at 15 time units
  create_beast2_input_file(
    input_fasta_filenames = get_input_fasta_filename(),
    output_xml_filename = output_xml_filename_fixed,
    fixed_crown_age = TRUE,
    initial_phylogeny = beastscriptr::fasta_to_phylo(
      input_fasta_filename, crown_age = 15)
  )
  testit::assert(file.exists(output_xml_filename_fixed))
}
\author{
Richel J.C. Bilderbeek
}
