% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_from_model.R
\name{create_beast2_input_from_model}
\alias{create_beast2_input_from_model}
\title{Create a BEAST2 XML input text from an inference model}
\usage{
create_beast2_input_from_model(input_filename, inference_model)
}
\arguments{
\item{input_filename}{A FASTA filename.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}

\item{inference_model}{an Bayesian phylogenetic inference model,
as can be created by \code{\link{create_inference_model}}.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.}
}
\value{
a character vector of XML strings
}
\description{
Create a BEAST2 XML input text from an inference model
}
\examples{

  text <- create_beast2_input_from_model(
    input_filename = get_fasta_filename(),
    inference_model = create_inference_model()
  )
  testit::assert(substr(text[1], 1, 5) == "<?xml")
  text[1]
  testit::assert(tail(text, n = 1) == "</beast>")
}
\seealso{
Use \link{create_beast2_input_file_from_model} to also save it to file.
}
\author{
Richèl J.C. Bilderbeek
}
