% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_init.R
\name{create_beast2_input_init}
\alias{create_beast2_input_init}
\title{Creates the map section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_init(fasta_filenames, initial_phylogeny)
}
\arguments{
\item{fasta_filenames}{the FASTA filenames}

\item{initial_phylogeny}{initial phylogeny}
}
\description{
Creates the map section of a BEAST2 XML parameter file
}
