% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_init.R
\name{create_beast2_input_init}
\alias{create_beast2_input_init}
\title{Creates the map section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_init(ids, initial_phylogeny)
}
\arguments{
\item{initial_phylogeny}{the MCMC chain its initial phylogeny. If
this is set to NA, BEAST2 will use a random phylogeny. Else
a phylogeny must be supplied of class ape::phylo.}
}
\description{
Creates the map section of a BEAST2 XML parameter file
}
