% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_init.R
\name{create_beast2_input_init}
\alias{create_beast2_input_init}
\title{Creates the map section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_init(ids, initial_phylogenies)
}
\arguments{
\item{ids}{one or more alignments' IDs.
IDs can be extracted from their FASTA filesnames
with \code{\link{get_ids}})}

\item{initial_phylogenies}{one or more MCMC chain's initial phylogenies.
Each one set to NA will result in BEAST2 using a random phylogeny. Else
the phylogeny is assumed to be of class ape::phylo.}
}
\description{
Creates the map section of a BEAST2 XML parameter file
}
\author{
Richel J.C. Bilderbeek
}
