% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_init.R
\name{create_beast2_input_init}
\alias{create_beast2_input_init}
\title{Creates the \code{init} section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_init(ids)
}
\arguments{
\item{ids}{one or more alignments' IDs.
IDs can be extracted from their FASTA filenames
with \code{\link{get_alignment_ids_from_fasta_filenames}})}
}
\value{
lines of XML text
}
\description{
Creates the \code{init} section of a BEAST2 XML parameter file
}
\author{
Richèl J.C. Bilderbeek
}
