% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_init.R
\name{create_beast2_input_init}
\alias{create_beast2_input_init}
\title{Creates the \code{init} section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_init(inference_model)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\value{
lines of XML text
}
\description{
Creates the \code{init} section of a BEAST2 XML parameter file
}
\examples{
check_empty_beautier_folder()

inference_model <- init_inference_model(
  input_filename = get_fasta_filename(),
  inference_model = create_test_inference_model()
)
xml <- create_beast2_input_init(
  inference_model = inference_model
)

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
