% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_loggers.R
\name{create_beast2_input_loggers}
\alias{create_beast2_input_loggers}
\title{Creates the two logger sections of a BEAST2 XML parameter file}
\usage{
create_beast2_input_loggers(ids, tree_priors = create_tree_prior(name =
  "yule"))
}
\arguments{
\item{ids}{the IDs of the alignments (can be extracted from
their FASTA filesnames using 'get_file_base_sans_ext')}

\item{tree_priors}{one or more tree priors}
}
\description{
Creates the two logger sections of a BEAST2 XML parameter file
}
