% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_loggers.R
\name{create_beast2_input_loggers}
\alias{create_beast2_input_loggers}
\title{Creates the two logger sections of a BEAST2 XML parameter file}
\usage{
create_beast2_input_loggers(ids, site_models = create_jc69_site_models(n =
  length(ids)), clock_models = create_strict_clock_models(n = length(ids)),
  tree_priors = create_yule_tree_priors(n = length(ids)))
}
\arguments{
\item{ids}{the IDs of the alignments (can be extracted from
their FASTA filesnames using \code{\link{get_ids}})}

\item{site_models}{a list of one or more site models,
as returned by \code{\link{create_site_model}}}

\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}

\item{tree_priors}{a list of one or more tree priors,
as returned by \code{\link{create_tree_prior}}}
}
\description{
Creates the two logger sections of a BEAST2 XML parameter file
}
\author{
Richel J.C. Bilderbeek
}
