% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_loggers.R
\name{create_beast2_input_loggers_clock_models}
\alias{create_beast2_input_loggers_clock_models}
\title{Creates the clock models part of the two logger sections
  of a BEAST2 XML parameter file}
\usage{
create_beast2_input_loggers_clock_models(ids, clock_models)
}
\arguments{
\item{ids}{the IDs of the alignments (can be extracted from
their FASTA filesnames using \code{\link{get_file_base_sans_ext}})}

\item{clock_models}{On or more clock models,
as returned by \code{\link{create_clock_model}}}
}
\description{
Creates the clock models part of the two logger sections
  of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
