% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_operators.R
\name{create_beast2_input_operators}
\alias{create_beast2_input_operators}
\title{Creates the operators section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_operators(ids, site_models = create_site_model(name =
  "JC69"), clock_models = create_clock_model(name = "strict"),
  tree_priors = create_tree_prior(name = "yule"), fixed_crown_age)
}
\arguments{
\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}

\item{clock_models}{On or more clock models,
as returned by \code{\link{create_clock_model}}}

\item{tree_priors}{On or more tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{fixed_crown_age}{determines if the phylogeny its crown age is
fixed. If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}
}
\description{
Creates the operators section of a BEAST2 XML parameter file
}
