% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_operators.R
\name{create_beast2_input_operators}
\alias{create_beast2_input_operators}
\title{Creates the operators section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_operators(inference_model)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\value{
lines of XML text
}
\description{
Creates the operators section of a BEAST2 XML parameter file
}
\author{
Richèl J.C. Bilderbeek
}
