% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_operators.R
\name{create_beast2_input_operators_rates}
\alias{create_beast2_input_operators_rates}
\title{Creates the site_models section in the operators section
of a BEAST2 XML parameter file}
\usage{
create_beast2_input_operators_rates(ids, site_models = create_site_model(name
  = "JC69"))
}
\arguments{
\item{ids}{the IDs of the alignments (can be extracted from
their FASTA filesnames using \code{\link{get_file_base_sans_ext}})}

\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}
}
\description{
Creates the site_models section in the operators section
of a BEAST2 XML parameter file
}
\note{
this function is not intended for regular use, thus its
  long name length is accepted
}
\author{
Richel J.C. Bilderbeek
}
