% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_loggers.R
\name{create_beast2_input_screenlog}
\alias{create_beast2_input_screenlog}
\title{Creates the \code{screenlog} section of the \code{logger} section
of a BEAST2 XML parameter file}
\usage{
create_beast2_input_screenlog()
}
\value{
the XML text
}
\description{
Creates the \code{screenlog} section of the \code{logger} section
of a BEAST2 XML parameter file
}
\author{
Richèl J.C. Bilderbeek
}
