% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_state.R
\name{create_beast2_input_state}
\alias{create_beast2_input_state}
\title{Creates the state section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_state(ids, site_models = create_jc69_site_models(n =
  length(ids)), clock_models = create_strict_clock_models(n = length(ids)),
  tree_priors = create_tree_prior(name = "yule"),
  initial_phylogenies = rep(NA, length(ids)))
}
\arguments{
\item{ids}{the IDs of the alignments (can be extracted from
their FASTA filesnames using \code{\link{get_file_base_sans_ext}})}

\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}

\item{clock_models}{On or more clock models,
as returned by \code{\link{create_clock_model}}}

\item{tree_priors}{On or more tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{initial_phylogenies}{initial phylogenies, can be NAs if random
phylogenies are desired}
}
\description{
Creates the state section of a BEAST2 XML parameter file
}
\author{
Richel J.C. Bilderbeek
}
