% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_file.R
\name{create_beast2_input_state}
\alias{create_beast2_input_state}
\title{Creates the state section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_state(filename_base, tree_priors, initial_phylogeny)
}
\arguments{
\item{filename_base}{filename its base}

\item{tree_priors}{one or more tree priors, as returned
by 'create_tree_prior'}

\item{initial_phylogeny}{initial phylogeny or NA}
}
\description{
Creates the state section of a BEAST2 XML parameter file
}
