% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_state.R
\name{create_beast2_input_state}
\alias{create_beast2_input_state}
\title{Creates the state section of a BEAST2 XML parameter file}
\usage{
create_beast2_input_state(fasta_filenames,
  site_models = create_site_model(name = "JC69"),
  tree_priors = create_tree_prior(name = "yule"), initial_phylogeny)
}
\arguments{
\item{fasta_filenames}{the fasta filenames}

\item{site_models}{one or more site models, as returned
by 'create_site_model'}

\item{tree_priors}{one or more tree priors, as returned
by 'create_tree_prior'}

\item{initial_phylogeny}{initial phylogeny or NA}
}
\description{
Creates the state section of a BEAST2 XML parameter file
}
