% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_loggers.R
\name{create_beast2_input_tracelog}
\alias{create_beast2_input_tracelog}
\title{Creates the tracelog section of the logger section
of a BEAST2 XML parameter file}
\usage{
create_beast2_input_tracelog(ids, site_models = create_jc69_site_models(ids =
  ids), clock_models = create_strict_clock_models(ids = ids),
  tree_priors = create_yule_tree_priors(ids = ids), mrca_priors = NA)
}
\description{
Creates the tracelog section of the logger section
of a BEAST2 XML parameter file
}
\author{
Richel J.C. Bilderbeek
}
