% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_loggers.R
\name{create_beast2_input_tracelog}
\alias{create_beast2_input_tracelog}
\title{Creates the tracelog section of the logger section
of a BEAST2 XML parameter file}
\usage{
create_beast2_input_tracelog(ids, site_models = create_jc69_site_models(ids =
  ids), clock_models = create_strict_clock_models(ids = ids),
  tree_priors = create_yule_tree_priors(ids = ids), mrca_priors = NA)
}
\arguments{
\item{ids}{one or more alignments' IDs.
IDs can be extracted from their FASTA filesnames
with \code{\link{get_ids}})}

\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}

\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}

\item{tree_priors}{one or more tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{mrca_priors}{a list of one or more Most Recent Common Ancestor priors,
as returned by \code{\link{create_mrca_prior}}}
}
\description{
Creates the tracelog section of the logger section
of a BEAST2 XML parameter file
}
\author{
Richel J.C. Bilderbeek
}
