% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_loggers.R
\name{create_beast2_input_treelogs}
\alias{create_beast2_input_treelogs}
\title{Creates the tracelog section of the logger section
of a BEAST2 XML parameter file}
\usage{
create_beast2_input_treelogs(clock_models)
}
\description{
Creates the tracelog section of the logger section
of a BEAST2 XML parameter file
}
\author{
Richel J.C. Bilderbeek
}
