% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_loggers.R
\name{create_beast2_input_treelogs}
\alias{create_beast2_input_treelogs}
\title{Creates the \code{tracelog} section of the \code{logger} section
of a BEAST2 XML parameter file}
\usage{
create_beast2_input_treelogs(inference_model)
}
\arguments{
\item{inference_model}{an Bayesian phylogenetic inference model,
as can be created by \code{\link{create_inference_model}}.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.}
}
\description{
Creates the \code{tracelog} section of the \code{logger} section
of a BEAST2 XML parameter file
}
\author{
Richèl J.C. Bilderbeek
}
