% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_beta_distr}
\alias{create_beta_distr}
\title{Create a beta distribution}
\usage{
create_beta_distr(id = NA)
}
\arguments{
\item{id}{the distribution's ID}
}
\value{
a beta distribution
}
\description{
Create a beta distribution
}
\seealso{
the function \code{\link{create_distribution}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
