% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_beta_distr}
\alias{create_beta_distr}
\title{Create a beta distribution}
\usage{
create_beta_distr(id = NA, alpha = create_alpha_param(id = NA, estimate =
  NA, value = NA), beta = create_beta_param(id = NA, estimate = NA, value =
  NA))
}
\arguments{
\item{id}{the distribution's ID}

\item{alpha}{the alpha shape parameter,
as returned by \code{\link{create_alpha_param}}}

\item{beta}{the beta shape parameter,
as returned by \code{\link{create_beta_param}}}
}
\value{
a beta distribution
}
\description{
Create a beta distribution
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
