% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_beta_distribution}
\alias{create_beta_distribution}
\title{Create a beta distribution}
\usage{
create_beta_distribution()
}
\value{
a beta distribution
}
\description{
Create a beta distribution
}
