% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_prior.R
\name{create_cbs_tree_prior}
\alias{create_cbs_tree_prior}
\title{Create a Coalescent Bayesian Skyline tree prior}
\usage{
create_cbs_tree_prior(id = NA)
}
\arguments{
\item{id}{the ID of the alignment}
}
\value{
a Coalescent Bayesian Skyline tree_prior
}
\description{
Create a Coalescent Bayesian Skyline tree prior
}
\examples{
  cbs_tree_prior <- create_cbs_tree_prior()
  testit::assert(is_cbs_tree_prior(cbs_tree_prior))

  input_fasta_filename <- beautier::get_path("anthus_aco.fas")
  create_beast2_input_file(
    input_fasta_filenames = input_fasta_filename,
    "my_beast.xml",
    tree_priors = cbs_tree_prior
  )
  testit::assert(file.exists("my_beast.xml"))
}
\seealso{
An alignment ID can be extracted from
  its FASTA filesname using \code{\link{get_id}}
}
\author{
Richel J.C. Bilderbeek
}
