% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_prior.R
\name{create_ccp_tree_prior}
\alias{create_ccp_tree_prior}
\title{Create a Coalescent Constant Population tree prior}
\usage{
create_ccp_tree_prior()
}
\value{
a Coalescent Constant Population tree_prior
}
\description{
Create a Coalescent Constant Population tree prior
}
