% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_prior.R
\name{create_ccp_tree_prior}
\alias{create_ccp_tree_prior}
\title{Create a Coalescent Constant Population tree prior}
\usage{

  create_ccp_tree_prior(pop_size_distribution = beautier::create_one_div_x_distr())
}
\arguments{
\item{pop_size_distribution}{the population distribution,
as created by a \code{\link{create_distribution}} function}
}
\value{
a Coalescent Constant Population tree_prior
}
\description{
Create a Coalescent Constant Population tree prior
}
