% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_prior.R
\name{create_cep_tree_prior}
\alias{create_cep_tree_prior}
\title{Create a Coalescent Exponential Population tree prior}
\usage{

  create_cep_tree_prior(pop_size_distribution = create_one_div_x_distribution(),
  growth_rate_distribution = create_laplace_distribution())
}
\value{
a Coalescent Exponential Population tree_prior
}
\description{
Create a Coalescent Exponential Population tree prior
}
