% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_prior.R
\name{create_cep_tree_prior}
\alias{create_cep_tree_prior}
\title{Create a Coalescent Exponential Population tree prior}
\usage{
create_cep_tree_prior(id = get_tree_prior_default_id(),
  pop_size_distr = create_one_div_x_distr(),
  growth_rate_distr = create_laplace_distr())
}
\arguments{
\item{id}{the ID of the alignment}

\item{pop_size_distr}{the population distribution,
as created by a \code{\link{create_distr}} function}

\item{growth_rate_distr}{the growth rate distribution,
as created by a \code{\link{create_distr}} function}
}
\value{
a Coalescent Exponential Population tree_prior
}
\description{
Create a Coalescent Exponential Population tree prior
}
\examples{
  cep_tree_prior <- create_cep_tree_prior()
  testit::assert(is_cep_tree_prior(cep_tree_prior))

  input_fasta_filename <- system.file(
    "extdata", "anthus_aco.fas", package = "beautier"
  )
  create_beast2_input_file(
    input_fasta_filenames = input_fasta_filename,
    "my_beast.xml",
    tree_priors = cep_tree_prior
  )
  testit::assert(file.exists("my_beast.xml"))
}
\seealso{
An alignment ID can be extracted from
  its FASTA filesname using \code{\link{get_id}}
}
\author{
Richel J.C. Bilderbeek
}
