% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clock_model.R
\name{create_clock_model}
\alias{create_clock_model}
\title{Create a clock model}
\usage{
create_clock_model(name)
}
\arguments{
\item{name}{the clock model name. Valid
names can be found in 'get_clock_model_names'}
}
\value{
a clock_model
}
\description{
Create a clock model
}
