% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clock_model.R
\name{create_clock_model}
\alias{create_clock_model}
\title{General function to create a clock model.
Prefer using 'create_rln_clock_model' and 'create_strict_clock_model'}
\usage{
create_clock_model(name, ...)
}
\arguments{
\item{name}{the clock model name. Valid
names can be found in 'get_clock_model_names'}

\item{...}{specific clock model parameters}
}
\value{
a clock_model
}
\description{
General function to create a clock model.
Prefer using 'create_rln_clock_model' and 'create_strict_clock_model'
}
