% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clock_model_from_name.R
\name{create_clock_model_from_name}
\alias{create_clock_model_from_name}
\title{Create a clock model from name}
\usage{
create_clock_model_from_name(clock_model_name)
}
\arguments{
\item{clock_model_name}{name of a clock model,
must be a name as returned by \code{\link{get_clock_model_names}}}
}
\value{
a clock model,
  as can be created by \link{create_clock_model}
}
\description{
Create a clock model from name
}
\examples{
create_clock_model_from_name(get_clock_model_names()[1])
}
\seealso{
Use \code{\link{create_clock_model}} to create a clock model
}
\author{
Richèl J.C. Bilderbeek
}
