% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clock_models.R
\name{create_clock_models}
\alias{create_clock_models}
\title{Creates all supported clock models,
  which is just a list of the types returned by
  \code{\link{create_rln_clock_model}},
  and \code{\link{create_strict_clock_model}}}
\usage{
create_clock_models()
}
\value{
a list of site_models
}
\description{
Creates all supported clock models,
  which is just a list of the types returned by
  \code{\link{create_rln_clock_model}},
  and \code{\link{create_strict_clock_model}}
}
\examples{
 clock_models <- beautier:::create_clock_models()
 testit::assert(beautier:::is_rln_clock_model(clock_models[[1]]))
 testit::assert(beautier:::is_strict_clock_model(clock_models[[2]]))
}
\author{
Richel J.C. Bilderbeek
}
