% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clock_models.R
\name{create_clock_models}
\alias{create_clock_models}
\title{Creates all supported clock models,
  which is a list of the types returned by
  \code{\link{create_rln_clock_model}},
  and \code{\link{create_strict_clock_model}}}
\usage{
create_clock_models()
}
\value{
a list of site_models
}
\description{
Creates all supported clock models,
  which is a list of the types returned by
  \code{\link{create_rln_clock_model}},
  and \code{\link{create_strict_clock_model}}
}
\examples{
check_empty_beautier_folder()

clock_models <- create_clock_models()
is_rln_clock_model(clock_models[[1]])
is_strict_clock_model(clock_models[[2]])

check_empty_beautier_folder()
}
\seealso{
Use \link{create_clock_model} to create a clock model
}
\author{
Richèl J.C. Bilderbeek
}
