% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_clock_rate_param}
\alias{create_clock_rate_param}
\title{Create a parameter called \code{clock_rate},
  as needed by \code{\link{create_strict_clock_model}}}
\usage{
create_clock_rate_param(id = NA, estimate = FALSE, value = "1.0")
}
\arguments{
\item{id}{the alignment id}

\item{estimate}{TRUE if this parameter is estimated by BEAST2,
FALSE otherwise}

\item{value}{value of the parameter}
}
\value{
a parameter called rate
}
\description{
Create a parameter called \code{clock_rate},
  as needed by \code{\link{create_strict_clock_model}}
}
\examples{
  clock_rateparam <- create_clock_rate_param(
    id = "anthus_aco", estimate = FALSE, value = 1.0
  )
  testit::assert(is_clock_rate_param(clock_rateparam))
}
\seealso{
the function \code{\link{create_param}} contains a list
  of all parameters that can be created
}
\author{
Richel J.C. Bilderbeek
}
