% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_exponential_distr}
\alias{create_exponential_distr}
\title{Create an exponential distribution}
\usage{
create_exponential_distr(id = NA)
}
\arguments{
\item{id}{the distribution's ID}
}
\value{
an exponential distribution
}
\description{
Create an exponential distribution
}
\seealso{
the function \code{\link{create_distribution}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
