% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_gamma_distr}
\alias{create_gamma_distr}
\title{Create a gamma distribution}
\usage{
create_gamma_distr(id = NA, alpha = create_alpha_parameter(id = NA, estimate
  = FALSE, value = "0.5396"), beta = create_beta_parameter(id = NA, estimate =
  FALSE, value = "0.3819"))
}
\arguments{
\item{id}{the distribution's ID}

\item{alpha}{the alpha shape parameter}

\item{beta}{the beta shape parameter}
}
\value{
a gamma distribution
}
\description{
Create a gamma distribution
}
\seealso{
the function \code{\link{create_distribution}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
