% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_gamma_distr}
\alias{create_gamma_distr}
\title{Create a gamma distribution}
\usage{
create_gamma_distr(id = NA, alpha = create_alpha_param(id = NA, estimate =
  FALSE, value = "0.5396"), beta = create_beta_param(id = NA, estimate =
  FALSE, value = "0.3819"))
}
\arguments{
\item{id}{the distribution's ID}

\item{alpha}{the alpha shape parameter,
as returned by \code{\link{create_alpha_param}}}

\item{beta}{the beta shape parameter,
as returned by \code{\link{create_beta_param}}}
}
\value{
a gamma distribution
}
\description{
Create a gamma distribution
}
\examples{
  gamma_distr <- create_gamma_distr()
  testit::assert(is_gamma_distr(gamma_distr))

  input_fasta_filename <- beautier::get_path("anthus_aco.fas")
  create_beast2_input_file(
    input_fasta_filenames = input_fasta_filename,
    "my_beast.xml",
    tree_priors = create_yule_tree_prior(
      birth_rate_distr = gamma_distr
    )
  )
  testit::assert(file.exists("my_beast.xml"))
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
