% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_site_model.R
\name{create_gtr_site_model}
\alias{create_gtr_site_model}
\title{Create a GTR site model}
\usage{
create_gtr_site_model(gamma_site_model = get_default_gamma_site_model())
}
\arguments{
\item{gamma_site_model}{a gamma site model, as created
by \code{\link{create_gamma_site_model}}}
}
\value{
a GTR site_model
}
\description{
Create a GTR site model
}
