% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_site_model.R
\name{create_hky_site_model}
\alias{create_hky_site_model}
\title{Create an HKY site model}
\usage{
create_hky_site_model(kappa = get_default_kappa(),
  gamma_site_model = get_default_gamma_site_model())
}
\arguments{
\item{kappa}{the kappa}

\item{gamma_site_model}{a gamma site model, as created
by \code{\link{create_gamma_site_model}}}
}
\value{
an HKY site_model
}
\description{
Create an HKY site model
}
