% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_site_model.R
\name{create_hky_site_model}
\alias{create_hky_site_model}
\title{Create an HKY site model}
\usage{
create_hky_site_model(kappa = 2)
}
\arguments{
\item{kappa}{the kappa}
}
\value{
an HKY site_model
}
\description{
Create an HKY site model
}
