% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_inv_gamma_distr}
\alias{create_inv_gamma_distr}
\title{Create an inverse gamma distribution}
\usage{
create_inv_gamma_distr(id = NA)
}
\arguments{
\item{id}{the distribution's ID}
}
\value{
an inverse gamma distribution
}
\description{
Create an inverse gamma distribution
}
\seealso{
the function \code{\link{create_distribution}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
