% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_inv_gamma_distribution}
\alias{create_inv_gamma_distribution}
\title{Create an inverse gamma distribution}
\usage{
create_inv_gamma_distribution(id = NA)
}
\arguments{
\item{id}{the distribution's ID}
}
\value{
an inverse gamma distribution
}
\description{
Create an inverse gamma distribution
}
\author{
Richel J.C. Bilderbeek
}
