% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_kappa_2_param}
\alias{create_kappa_2_param}
\alias{create_param_kappa_2}
\title{Create a parameter called kappa 2}
\usage{
create_kappa_2_param(id = NA, lower = "0.0", value = "2.0")
}
\arguments{
\item{id}{the parameter's ID}

\item{lower}{lowest possible value of the parameter}

\item{value}{value of the parameter}
}
\value{
a parameter called kappa 2
}
\description{
Create a parameter called kappa 2
}
\author{
Richel J.C. Bilderbeek
}
