% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_laplace_distr}
\alias{create_laplace_distr}
\title{Create a Laplace distribution}
\usage{
create_laplace_distr(id = NA, mu = create_muparam(id = NA, estimate = FALSE,
  value = 0), scale = create_scaleparam(id = NA, estimate = FALSE, value = 1))
}
\arguments{
\item{id}{the distribution's ID}

\item{mu}{the mu parameter,
as returned by \code{\link{create_muparam}}}

\item{scale}{the scale parameter,
as returned by \code{\link{create_scaleparam}}}
}
\value{
a Laplace distribution
}
\description{
Create a Laplace distribution
}
\examples{
  laplace_distr <- create_laplace_distr()
  testit::assert(is_laplace_distr(laplace_distr))
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
