% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_laplace_distr}
\alias{create_laplace_distr}
\title{Create a Laplace distribution}
\usage{
create_laplace_distr(id = NA, mu = create_mu_parameter(id = NA, estimate =
  FALSE, value = 0), scale = create_scale_parameter(id = NA, estimate = FALSE,
  value = 1))
}
\arguments{
\item{id}{the distribution's ID}

\item{mu}{the mu parameter}

\item{scale}{the scale parameter}
}
\value{
a Laplace distribution
}
\description{
Create a Laplace distribution
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
