% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_laplace_distribution}
\alias{create_laplace_distribution}
\title{Create a Laplace distribution}
\usage{
create_laplace_distribution(id = NA)
}
\arguments{
\item{id}{the distribution's ID}
}
\value{
a Laplace distribution
}
\description{
Create a Laplace distribution
}
\author{
Richel J.C. Bilderbeek
}
