% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_log_normal_distr}
\alias{create_log_normal_distr}
\title{Create a log-normal distribution}
\usage{
create_log_normal_distr(id = NA, m = create_m_parameter(),
  s = create_s_parameter())
}
\arguments{
\item{id}{the distribution's ID}

\item{m}{the m parameter}

\item{s}{the s parameter}
}
\value{
a log-normal distribution
}
\description{
Create a log-normal distribution
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
