% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mcmc.R
\name{create_mcmc}
\alias{create_mcmc}
\title{Function to create the MCMC options, as in the BEAUti MCMC tab.}
\usage{
create_mcmc(chain_length = get_default_mcmc_chain_length())
}
\arguments{
\item{chain_length}{the MCMC's chain length}
}
\value{
an mcmc
}
\description{
Function to create the MCMC options, as in the BEAUti MCMC tab.
}
\author{
Richel J.C. Bilderbeek
}
