% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mcmc.R
\name{create_mcmc}
\alias{create_mcmc}
\title{Create an MCMC configuration.}
\usage{
create_mcmc(chain_length = 1e+07, store_every = -1)
}
\value{
an MCMC configuration
}
\description{
Create an MCMC configuration, as in the BEAUti MCMC tab.
The number of states that will be saved equals the chain
length (\code{chain_length}) divided by the number of
states between each sampling event (\code{store_every})
}
\examples{
  # Create an MCMC chain with 50 states
  mcmc <- create_mcmc(chain_length = 50000, store_every = 1000)

  create_beast2_input_file(
    get_fasta_filename(),
    "create_mcmc.xml",
    mcmc = mcmc
  )
  testit::assert(file.exists("create_mcmc.xml"))
}
\author{
Richel J.C. Bilderbeek
}
