% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mcmc_nested_sampling.R
\name{create_mcmc_nested_sampling}
\alias{create_mcmc_nested_sampling}
\title{Create an MCMC object that does nested sampling}
\usage{
create_mcmc_nested_sampling(chain_length = 1e+07, store_every = -1,
  particle_count = 1, sub_chain_length = 5000, epsilon = "1e-12")
}
\arguments{
\item{chain_length}{length of the MCMC chain}

\item{store_every}{number of states the MCMC will process
before the posterior's state will be saved to file.
Use -1 or NA to use the default frequency.}

\item{particle_count}{number of particles}

\item{sub_chain_length}{sub-chain length}

\item{epsilon}{epsilon}
}
\value{
an MCMC object
}
\description{
Create an MCMC object that does nested sampling
}
\examples{
  # Create an MCMC chain with 50 states
  mcmc <- create_mcmc_nested_sampling(
    chain_length = 50000,
    store_every = 1000,
    particle_count = 1,
    sub_chain_length = 5000,
    epsilon = 1e-12
  )

  create_beast2_input_file(
    get_fasta_filename(),
    "create_mcmc_nested_sampling.xml",
    mcmc = mcmc
  )
  testit::assert(file.exists("create_mcmc_nested_sampling.xml"))
}
\references{
* [1] Maturana, P., Brewer, B. J., Klaere, S., & Bouckaert, R. (2017).
    Model selection and parameter inference in phylogenetics
    using Nested Sampling. arXiv preprint arXiv:1703.05471.
}
\author{
Richel J.C. Bilderbeek
}
