% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mrca_prior.R
\name{create_mrca_prior}
\alias{create_mrca_prior}
\title{Create a Most Recent Common Ancestor prior}
\usage{
create_mrca_prior(taxa_names)
}
\arguments{
\item{taxa_names}{names of the taxa,
as returned by \code{\link{get_taxa_names}}}
}
\value{
an MRCA prior
}
\description{
Create a Most Recent Common Ancestor prior
}
\author{
Richel J.C. Bilderbeek
}
