% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mrca_prior.R
\name{create_mrca_prior}
\alias{create_mrca_prior}
\title{Create a Most Recent Common Ancestor prior}
\usage{
create_mrca_prior(name, alignment_id, taxa_names, mrca_distr,
  clock_prior_distr_id = NA)
}
\arguments{
\item{name}{the unique name of the MRCA prior, for example a genus, family,
order or even class name}

\item{alignment_id}{ID of the alignment,
as returned by \code{\link{get_alignment_id}}}

\item{taxa_names}{names of the taxa,
as returned by \code{\link{get_taxa_names}}}

\item{mrca_distr}{the MRCA distribution,
as returned by \code{\link{create_distr}}}

\item{clock_prior_distr_id}{ID of an MRCA clock model's distribution.
Keep at NA to have it initialized automatically}
}
\value{
an MRCA prior
}
\description{
Create a Most Recent Common Ancestor prior
}
\author{
Richel J.C. Bilderbeek
}
