% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{create_normal_distr}
\alias{create_normal_distr}
\title{Create an normal distribution}
\usage{
create_normal_distr(id = NA)
}
\arguments{
\item{id}{the distribution's ID}
}
\value{
a normal distribution
}
\description{
Create an normal distribution
}
\author{
Richel J.C. Bilderbeek
}
