% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_normal_distr}
\alias{create_normal_distr}
\title{Create an normal distribution}
\usage{
create_normal_distr(id = NA, mean = create_meanparam(),
  sigma = create_sigmaparam())
}
\arguments{
\item{id}{the distribution's ID}

\item{mean}{the mean parameter
as returned by \code{\link{create_meanparam}}}

\item{sigma}{the sigma parameter
as returned by \code{\link{create_sigmaparam}}}
}
\value{
a normal distribution
}
\description{
Create an normal distribution
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
