% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_param}
\alias{create_param}
\title{General function to create a parameter.}
\usage{
create_param(name, id, ...)
}
\arguments{
\item{name}{the parameters' name. Valid
names can be found in \code{\link{get_param_names}}}

\item{id}{the parameter's ID}

\item{...}{specific parameter parameters}
}
\value{
a parameter
}
\description{
General function to create a parameter.
}
\note{
Prefer using the
  named functions
  \code{\link{create_alphaparam}},
  \code{\link{create_betaparam}},
  \code{\link{create_clock_rateparam}},
  \code{\link{create_mparam}},
  \code{\link{create_meanparam}},
  \code{\link{create_muparam}},
  \code{\link{create_sparam}},
  \code{\link{create_scaleparam}},
  and \code{\link{create_sigmaparam}}
}
\author{
Richel J.C. Bilderbeek
}
