% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter.R
\name{create_parameter}
\alias{create_parameter}
\title{General function to create a parameter.}
\usage{
create_parameter(name, id, ...)
}
\arguments{
\item{name}{the parameters' name. Valid
names can be found in \code{\link{get_parameter_names}}}

\item{id}{the parameter's ID}

\item{...}{specific parameter parameters}
}
\value{
a parameter
}
\description{
General function to create a parameter.
}
\note{
Prefer using the
  named functions
  \code{\link{create_parameter_alpha}},
  and \code{\link{create_parameter_beta}}
}
\author{
Richel J.C. Bilderbeek
}
