% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_poisson_distr}
\alias{create_poisson_distr}
\alias{create_distr_poisson}
\title{Create a Poisson distribution}
\usage{
create_poisson_distr(id = NA, lambda = create_lambda_param())
}
\arguments{
\item{id}{the distribution's ID}

\item{lambda}{the lambda parameter
as returned by \code{\link{create_lambda_param}}}
}
\value{
a Poisson distribution
}
\description{
Create a Poisson distribution
}
\examples{
  poisson_distr <- create_poisson_distr()

  create_beast2_input_file(
    input_filenames = get_fasta_filename(),
    "create_poisson_distr.xml",
    tree_priors = create_yule_tree_prior(
      birth_rate_distr = poisson_distr
    )
  )
  testit::assert(file.exists("create_poisson_distr.xml"))
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
