% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_posterior.R
\name{create_posterior}
\alias{create_posterior}
\title{Create a BEAST2 posterior by running BEAST2 from a random alignment}
\usage{
create_posterior(n_taxa, sequence_length,
  tree_priors = create_yule_tree_prior(), mcmc = create_mcmc(),
  crown_ages = NA, verbose = FALSE)
}
\arguments{
\item{n_taxa}{number of taxa in the simulated phylogeny}

\item{sequence_length}{a DNA sequence length, in base pairs}

\item{tree_priors}{one or more tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{mcmc}{one mcmc object,
as returned by \code{\link{create_mcmc}}}

\item{crown_ages}{the crown ages of the phylogenies. Set to NA
if the crown age needs to be estimated}
}
\description{
Create a BEAST2 posterior by running BEAST2 from a random alignment
}
\author{
Richel J.C. Bilderbeek
}
