% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_rate_cg_param}
\alias{create_rate_cg_param}
\alias{create_param_rate_cg}
\title{Create a parameter called 'rate CG'}
\usage{
create_rate_cg_param(id = NA, estimate = TRUE, value = "1.0",
  lower = "0.0")
}
\arguments{
\item{id}{the parameter's ID}

\item{estimate}{TRUE if this parameter mu be estimated by BEAST2,
FALSE otherwise}

\item{value}{value of the parameter}

\item{lower}{lowest possible value of the parameter}
}
\value{
a parameter called 'rate CG'
}
\description{
Create a parameter called 'rate CG'
}
\examples{
  # Create parameter
  rate_cg_param <- create_rate_cg_param(value = 1, estimate = FALSE)

  # Use the parameter to create a BEAST2 input file
  create_beast2_input_file(
    input_filenames = get_fasta_filename(),
    "create_rate_cg_param.xml",
    site_models = create_gtr_site_model(
      rate_cg_param = rate_cg_param
    )
  )
  testit::assert(file.exists("create_rate_cg_param.xml"))
}
\seealso{
the function \code{\link{create_param}} contains a list
  of all parameters that can be created
}
\author{
Richel J.C. Bilderbeek
}
