% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clock_model.R
\name{create_rln_clock_model}
\alias{create_rln_clock_model}
\title{Create a relaxed log-normal clock model}
\usage{
create_rln_clock_model(id = NA, uclstdev_distr = create_gamma_distr(),
  mparam_id = NA, mean_clock_rate = 1, n_rate_categories = -1,
  normalize_mean_clock_rate = FALSE)
}
\arguments{
\item{id}{a clock model's ID}

\item{uclstdev_distr}{the uclstdev distribution,
as created by a \code{\link{create_distr}} function}

\item{mparam_id}{the ID of the M paramater in the branchRateModel,
set to NA to have it initialized}

\item{mean_clock_rate}{the mean clock rate, 1.0 by default
(is called 'ucld_stdev' in XML, where 'ucld_stdev' is always 0.1)}

\item{n_rate_categories}{the number of rate categories.
-1 is default,
0 denotes as much rates as branches}

\item{normalize_mean_clock_rate}{normalize the mean clock rate}
}
\value{
a relaxed log-normal clock_model
}
\description{
Create a relaxed log-normal clock model
}
\examples{
  rln_clock_model <- create_rln_clock_model()
  testit::assert(is_rln_clock_model(rln_clock_model))

  input_fasta_filenames <- system.file(
    "extdata", "test_output_0.fas", package = "beautier"
  )
  create_beast2_input_file(
    input_fasta_filenames = input_fasta_filenames,
    "my_beast.xml",
    clock_models = rln_clock_model
  )
}
\author{
Richel J.C. Bilderbeek
}
