% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_rnd_mrca_priors}
\alias{create_rnd_mrca_priors}
\title{Creates a random MRCA prior option. This is either zero, one
or more MRCA priors. For zero MRCA priors, NA is used. When two
or more MRCA priors (see \link{create_rnd_two_mrca_priors}) are created,
those are checked to be compatible}
\usage{
create_rnd_mrca_priors(fasta_filename)
}
\arguments{
\item{fasta_filename}{a FASTA filename}
}
\description{
Creates a random MRCA prior option. This is either zero, one
or more MRCA priors. For zero MRCA priors, NA is used. When two
or more MRCA priors (see \link{create_rnd_two_mrca_priors}) are created,
those are checked to be compatible
}
\author{
Richel J.C. Bilderbeek
}
