% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_screenlog.R
\name{create_screenlog}
\alias{create_screenlog}
\title{Create a \code{screenlog} object}
\usage{
create_screenlog(
  filename = "",
  log_every = 1000,
  mode = "autodetect",
  sanitise_headers = FALSE,
  sort = "none"
)
}
\arguments{
\item{filename}{name of the file to store the posterior screens
phylogenies to. By default, this is \code{$(screen).screens}}

\item{log_every}{number of MCMC states between writing to file}

\item{mode}{mode how to log.
Valid values are the ones returned by \link{get_log_modes}}

\item{sanitise_headers}{set to \link{TRUE} to sanitise the headers of the
log file}

\item{sort}{how to sort the log.
Valid values are the ones returned by \link{get_log_sorts}}
}
\value{
a \code{screenlog} object
}
\description{
Create a \code{screenlog} object
}
\examples{
check_empty_beautier_folder()

create_screenlog()

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
