% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_screenlog_xml.R
\name{create_screenlog_xml}
\alias{create_screenlog_xml}
\title{Creates the \code{screenlog} section of the \code{logger} section
of a BEAST2 XML parameter file}
\usage{
create_screenlog_xml(inference_model = beautier::create_inference_model())
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\value{
the XML text
}
\description{
Creates the \code{screenlog} section of the \code{logger} section
of a BEAST2 XML parameter file
}
\author{
Richèl J.C. Bilderbeek
}
