% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_site_model.R
\name{create_site_model}
\alias{create_site_model}
\title{Create a site model}
\usage{
create_site_model(name)
}
\arguments{
\item{name}{the site model name. Can be 'JC69', 'HKY', 'TN93' or 'GTR'}
}
\value{
a site_model
}
\description{
Create a site model
}
