% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_site_model.R
\name{create_site_model}
\alias{create_site_model}
\title{General function to create a site model.}
\usage{
create_site_model(name, ...)
}
\arguments{
\item{name}{the site model name. Valid
names can be found in \code{\link{get_site_model_names}}}

\item{...}{specific site model parameters}
}
\value{
a site_model
}
\description{
General function to create a site model.
}
\note{
Prefer using the
  named functions \code{\link{create_jc69_site_model}},
  \code{\link{create_hky_site_model}},
  \code{\link{create_tn93_site_model}},
  and \code{\link{create_gtr_site_model}}
}
\author{
Richel J.C. Bilderbeek
}
