% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_site_models.R
\name{create_site_models}
\alias{create_site_models}
\title{Creates all supported site models
  which is just a list of the types returned by
  \code{\link{create_gtr_site_model}},
  \code{\link{create_hky_site_model}},
  \code{\link{create_jc69_site_model}}
  and \code{\link{create_tn93_site_model}}}
\usage{
create_site_models()
}
\value{
a list of site_models
}
\description{
Creates all supported site models
  which is just a list of the types returned by
  \code{\link{create_gtr_site_model}},
  \code{\link{create_hky_site_model}},
  \code{\link{create_jc69_site_model}}
  and \code{\link{create_tn93_site_model}}
}
\examples{
 site_models <- beautier:::create_site_models()
 testit::assert(beautier:::is_gtr_site_model(site_models[[1]]))
 testit::assert(beautier:::is_hky_site_model(site_models[[2]]))
 testit::assert(beautier:::is_jc69_site_model(site_models[[3]]))
 testit::assert(beautier:::is_tn93_site_model(site_models[[4]]))
}
\author{
Richel J.C. Bilderbeek
}
