% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_site_models_from_names.R
\name{create_site_models_from_names}
\alias{create_site_models_from_names}
\title{Create site models from their names}
\usage{
create_site_models_from_names(site_model_names)
}
\arguments{
\item{site_model_names}{one or more names of a site model,
must be name among those returned by \code{\link{get_site_model_names}}}
}
\value{
one or more site models
}
\description{
Create site models from their names
}
\examples{
check_empty_beautier_folder()

create_site_models_from_names(get_site_model_names())

check_empty_beautier_folder()
}
\seealso{
Use \link{create_site_model} to create a site model
}
\author{
Richèl J.C. Bilderbeek
}
