% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clock_model.R
\name{create_strict_clock_model}
\alias{create_strict_clock_model}
\title{Create a strict clock model}
\usage{

  create_strict_clock_model(clock_rate_parameter = create_clock_rate_parameter())
}
\arguments{
\item{clock_rate_parameter}{a clock_rate parameter,
as created by a \code{\link{create_clock_rate_parameter}} function}
}
\value{
a strict clock_model
}
\description{
Create a strict clock model
}
\examples{
  strict_clock_model <- create_strict_clock_model()
  testit::assert(is_strict_clock_model(strict_clock_model))

  input_fasta_filenames <- system.file(
    "extdata", "test_output_0.fas", package = "beautier"
  )
  create_beast2_input(
    input_fasta_filenames = input_fasta_filenames,
    clock_models = strict_clock_model
  )
}
