% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clock_model.R
\name{create_strict_clock_model}
\alias{create_strict_clock_model}
\title{Create a strict clock model}
\usage{
create_strict_clock_model(id = get_clock_model_default_id(),
  clock_rate_param = create_clock_rate_param())
}
\arguments{
\item{id}{a clock model's ID}

\item{clock_rate_param}{a clock_rate parameter,
as created by a \code{\link{create_clock_rate_param}} function}
}
\value{
a strict clock_model
}
\description{
Create a strict clock model
}
\examples{
  strict_clock_model <- create_strict_clock_model()
  testit::assert(is_strict_clock_model(strict_clock_model))

  input_fasta_filenames <- system.file(
    "extdata", "test_output_0.fas", package = "beautier"
  )
  create_beast2_input_file(
    input_fasta_filenames = input_fasta_filenames,
    "my_beast.xml",
    clock_models = strict_clock_model
  )
}
\author{
Richel J.C. Bilderbeek
}
