% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clock_model.R
\name{create_strict_clock_model}
\alias{create_strict_clock_model}
\title{Create a strict clock model}
\usage{
create_strict_clock_model(rate = get_default_clock_model_rate())
}
\value{
a strict clock_model
}
\description{
Create a strict clock model
}
