% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_temp_screenlog_filename.R
\name{create_temp_screenlog_filename}
\alias{create_temp_screenlog_filename}
\title{Create a filename for a temporary screenlog file}
\usage{
create_temp_screenlog_filename()
}
\description{
Create a filename for a temporary screenlog file
}
\seealso{
use \link{create_screenlog} to create a screenlog.
}
