% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tracelog_xml.R
\name{create_tracelog_xml}
\alias{create_tracelog_xml}
\title{Internal function}
\usage{
create_tracelog_xml(input_filename, inference_model)
}
\arguments{
\item{input_filename}{A FASTA filename.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}

\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\description{
Creates the \code{tracelog} section of the \code{logger} section
of a BEAST2 XML parameter file
}
\author{
Richèl J.C. Bilderbeek
}
