% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_prior.R
\name{create_tree_prior}
\alias{create_tree_prior}
\title{Create a tree prior}
\usage{
create_tree_prior(name)
}
\arguments{
\item{name}{the tree prior name. Can be 'birth_death'
or 'coalescent_constant_population'}
}
\value{
a site_model
}
\description{
Create a tree prior
}
