% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_priors.R
\name{create_tree_priors}
\alias{create_tree_priors}
\title{Creates all supported tree priors,
  which is just a list of the types returned by
  \code{\link{create_bd_tree_prior}},
  \code{\link{create_cbs_tree_prior}},
  \code{\link{create_ccp_tree_prior}},
  \code{\link{create_cep_tree_prior}}
  and \code{\link{create_yule_tree_prior}}}
\usage{
create_tree_priors()
}
\value{
a list of tree_priors
}
\description{
Creates all supported tree priors,
  which is just a list of the types returned by
  \code{\link{create_bd_tree_prior}},
  \code{\link{create_cbs_tree_prior}},
  \code{\link{create_ccp_tree_prior}},
  \code{\link{create_cep_tree_prior}}
  and \code{\link{create_yule_tree_prior}}
}
\examples{
  tree_priors <- create_tree_priors()
  testit::assert(is_bd_tree_prior(tree_priors[[1]]))
  testit::assert(is_cbs_tree_prior(tree_priors[[2]]))
  testit::assert(is_ccp_tree_prior(tree_priors[[3]]))
  testit::assert(is_cep_tree_prior(tree_priors[[4]]))
  testit::assert(is_yule_tree_prior(tree_priors[[5]]))
}
\author{
Richel J.C. Bilderbeek
}
