% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_priors.R
\name{create_tree_priors}
\alias{create_tree_priors}
\title{Creates all supported tree priors,
  which is a list of the types returned by
  \code{\link{create_bd_tree_prior}},
  \code{\link{create_cbs_tree_prior}},
  \code{\link{create_ccp_tree_prior}},
  \code{\link{create_cep_tree_prior}}
  and \code{\link{create_yule_tree_prior}}}
\usage{
create_tree_priors()
}
\value{
a list of tree_priors
}
\description{
Creates all supported tree priors,
  which is a list of the types returned by
  \code{\link{create_bd_tree_prior}},
  \code{\link{create_cbs_tree_prior}},
  \code{\link{create_ccp_tree_prior}},
  \code{\link{create_cep_tree_prior}}
  and \code{\link{create_yule_tree_prior}}
}
\examples{
library(testthat)

tree_priors <- create_tree_priors()
expect_true(is_bd_tree_prior(tree_priors[[1]]))
expect_true(is_cbs_tree_prior(tree_priors[[2]]))
expect_true(is_ccp_tree_prior(tree_priors[[3]]))
expect_true(is_cep_tree_prior(tree_priors[[4]]))
expect_true(is_yule_tree_prior(tree_priors[[5]]))
}
\author{
Richèl J.C. Bilderbeek
}
