% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_treelog.R
\name{create_treelog}
\alias{create_treelog}
\title{Create a \code{treelog} object}
\usage{
create_treelog(filename = "$(tree).trees", log_every = 1000,
  mode = "tree", sanitize_headers = FALSE, sort = "none")
}
\arguments{
\item{filename}{name of the file to store the posterior trees}

\item{log_every}{number of MCMC states between the logging of that state
phylogenies to. By default, this is \code{$(tree).trees}}

\item{mode}{mode how to log.
Valid are \code{tree}, \code{autodetect} and \code{compound}}
}
\description{
Create a \code{treelog} object
}
