% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_treelog.R
\name{create_treelog}
\alias{create_treelog}
\title{Create a \code{treelog} object}
\usage{
create_treelog(
  filename = "$(tree).trees",
  log_every = 1000,
  mode = "tree",
  sanitise_headers = FALSE,
  sort = "none"
)
}
\arguments{
\item{filename}{name of the file to store the posterior trees}

\item{log_every}{number of MCMC states between writing to file}

\item{mode}{mode how to log.
Valid values are the ones returned by \link{get_log_modes}}

\item{sanitise_headers}{set to \link{TRUE} to sanitise the headers of the
log file}

\item{sort}{how to sort the log.
Valid values are the ones returned by \link{get_log_sorts}}
}
\value{
a `treelog`, as can be checked by \link{check_treelog}
}
\description{
Create a \code{treelog} object
}
\examples{
check_empty_beautier_folder()

create_treelog()

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
