% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_uniform_distr}
\alias{create_uniform_distr}
\title{Create a uniform distribution}
\usage{
create_uniform_distr(id = NA, upper = Inf)
}
\arguments{
\item{id}{the distribution's ID}

\item{upper}{an upper limit of the uniform distribution}
}
\value{
a uniform distribution
}
\description{
Create a uniform distribution
}
\examples{
  uniform_distr <- create_uniform_distr()
  testit::assert(is_uniform_distr(uniform_distr))

  input_fasta_filename <- system.file(
    "extdata", "anthus_aco.fas", package = "beautier"
  )
  create_beast2_input_file(
    input_fasta_filenames = input_fasta_filename,
    "my_beast.xml",
    tree_priors = create_yule_tree_prior(
      birth_rate_distr = uniform_distr
    )
  )
  testit::assert(file.exists("my_beast.xml"))
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
