% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distr.R
\name{create_uniform_distr}
\alias{create_uniform_distr}
\title{Create a uniform distribution}
\usage{
create_uniform_distr(id = NA, upper = Inf)
}
\arguments{
\item{id}{the distribution's ID}

\item{upper}{an upper limit of the uniform distribution}
}
\value{
a uniform distribution
}
\description{
Create a uniform distribution
}
\seealso{
the function \code{\link{create_distr}} shows an overview
  of all supported distributions
}
\author{
Richel J.C. Bilderbeek
}
