% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_xml.R
\name{create_xml_declaration}
\alias{create_xml_declaration}
\title{Create the XML declaration of the BEAST2 XML input file}
\usage{
create_xml_declaration()
}
\value{
one line of XML text
}
\description{
Create the XML declaration of the BEAST2 XML input file
}
\examples{
check_empty_beautier_folder()

create_xml_declaration()

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
