% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_prior.R
\name{create_yule_tree_prior}
\alias{create_yule_tree_prior}
\title{Create a Yule tree prior}
\usage{
create_yule_tree_prior(
  id = NA,
  birth_rate_distr = create_uniform_distr()
)
}
\arguments{
\item{id}{the ID of the alignment}

\item{birth_rate_distr}{the birth rate distribution,
as created by a \code{\link{create_distr}} function}
}
\value{
a Yule tree_prior
}
\description{
Create a Yule tree prior
}
\examples{
  yule_tree_prior <- create_yule_tree_prior()
  testit::assert(is_yule_tree_prior(yule_tree_prior))

  input_fasta_filename <- system.file(
    "extdata", "anthus_aco.fas", package = "beautier"
  )
  create_beast2_input_file(
    input_fasta_filenames = input_fasta_filename,
    "my_beast.xml",
    tree_priors = yule_tree_prior
  )
  testit::assert(file.exists("my_beast.xml"))
}
\seealso{
An alignment ID can be extracted from
  its FASTA filesname using \code{\link{get_id}}
}
\author{
Richel J.C. Bilderbeek
}
