% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_prior.R
\name{create_yule_tree_prior}
\alias{create_yule_tree_prior}
\title{Create a Yule tree prior}
\usage{
create_yule_tree_prior(
  birth_rate_distribution = beautier::create_uniform_distr()
)
}
\arguments{
\item{birth_rate_distribution}{the birth rate distribution,
as created by a \code{\link{create_distribution}} function}
}
\value{
a Yule tree_prior
}
\description{
Create a Yule tree prior
}
\author{
Richel J.C. Bilderbeek
}
