% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree_prior.R
\name{create_yule_tree_prior}
\alias{create_yule_tree_prior}
\alias{create_tree_prior_yule}
\title{Create a Yule tree prior}
\usage{
create_yule_tree_prior(
  id = NA,
  birth_rate_distr = create_uniform_distr()
)
}
\arguments{
\item{id}{the ID of the alignment}

\item{birth_rate_distr}{the birth rate distribution,
as created by a \code{\link{create_distr}} function}
}
\value{
a Yule tree_prior
}
\description{
Create a Yule tree prior
}
\examples{
 yule_tree_prior <- create_yule_tree_prior()

 beast2_input_file <- get_beautier_tempfilename()
 create_beast2_input_file(
   input_filename = get_fasta_filename(),
   beast2_input_file,
   tree_prior = yule_tree_prior
)
file.remove(beast2_input_file)
}
\seealso{
An alignment ID can be extracted from
  its FASTA filename using \code{\link{get_alignment_id}}
}
\author{
Richèl J.C. Bilderbeek
}
