% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_yule_tree_priors.R
\name{create_yule_tree_priors}
\alias{create_yule_tree_priors}
\title{Creates n Yule tree priors}
\usage{
create_yule_tree_priors(ids)
}
\arguments{
\item{ids}{the alignment IDs}
}
\value{
a list of Yule tree_prior objects
}
\description{
Creates n Yule tree priors
}
\examples{
  m <- create_yule_tree_priors(ids = "some_id")
  testthat::expect_equal(length(m), 1)
  testthat::expect_true(is_yule_tree_prior(m[[1]]))

  m <- create_yule_tree_priors(ids = c("a", "b"))
  testthat::expect_equal(length(m), 2)
  testthat::expect_true(is_yule_tree_prior(m[[1]]))
  testthat::expect_true(is_yule_tree_prior(m[[2]]))
}
