% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_yule_tree_priors.R
\name{create_yule_tree_priors}
\alias{create_yule_tree_priors}
\title{Creates n Yule tree priors}
\usage{
create_yule_tree_priors(ids)
}
\arguments{
\item{ids}{the alignment IDs}
}
\value{
a list of Yule tree_prior objects
}
\description{
Creates n Yule tree priors
}
\examples{
  fasta_filenames <- get_beautier_paths(
    c("anthus_aco.fas", "anthus_nd2.fas")
  )

  tree_priors <- create_yule_tree_priors(ids = get_ids(fasta_filenames))

  create_beast2_input_file(
    fasta_filenames,
    "create_yule_tree_priors.xml",
    tree_priors = tree_priors
  )
  testit::assert(file.exists("create_yule_tree_priors.xml"))
}
