% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_to_xml.R
\name{distr_to_xml}
\alias{distr_to_xml}
\title{Converts a distribution to XML}
\usage{
distr_to_xml(distr)
}
\arguments{
\item{distr}{a distibution,
as created by \code{\link{create_distr}})}
}
\value{
the distribution as XML text
}
\description{
Converts a distribution to XML
}
\examples{
  xml <- beautier:::distr_to_xml(create_uniform_distr(id = 1))
  testit::assert(is.character(xml))
  testit::assert(length(xml) == 1)
  testit::assert(nchar(xml) > 1)
}
\author{
Richel J.C. Bilderbeek
}
