% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_to_xml.R
\name{distr_to_xml}
\alias{distr_to_xml}
\title{Internal function}
\usage{
distr_to_xml(distr, beauti_options = create_beauti_options())
}
\arguments{
\item{distr}{a distribution,
as created by \code{\link{create_distr}})}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the distribution as XML text
}
\description{
Converts a distribution to XML
}
\examples{
distr_to_xml(create_uniform_distr(id = 1))
}
\author{
Richèl J.C. Bilderbeek
}
