% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_to_xml.R
\name{distr_to_xml_beta}
\alias{distr_to_xml_beta}
\title{Converts a beta distribution to XML}
\usage{
distr_to_xml_beta(distr)
}
\arguments{
\item{distr}{a beta distribution,
as created by \code{\link{create_beta_distr}})}
}
\value{
the distribution as XML text
}
\description{
Converts a beta distribution to XML
}
\author{
Richèl J.C. Bilderbeek
}
