% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_to_xml.R
\name{distr_to_xml_exp}
\alias{distr_to_xml_exp}
\title{Converts an exponential distribution to XML}
\usage{
distr_to_xml_exp(distr)
}
\arguments{
\item{distr}{an exponential distibution,
as created by \code{\link{create_exp_distr}})}
}
\value{
the distribution as XML text
}
\description{
Converts an exponential distribution to XML
}
\author{
Richel J.C. Bilderbeek
}
