% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_to_xml.R
\name{distr_to_xml_inv_gamma}
\alias{distr_to_xml_inv_gamma}
\title{Converts an inverse-gamma distribution to XML}
\usage{
distr_to_xml_inv_gamma(distr)
}
\arguments{
\item{distr}{an inverse-gamma distribution,
as created by \code{\link{create_inv_gamma_distr}})}
}
\value{
the distribution as XML text
}
\description{
Converts an inverse-gamma distribution to XML
}
\author{
Richèl J.C. Bilderbeek
}
