% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_to_xml.R
\name{distr_to_xml_laplace}
\alias{distr_to_xml_laplace}
\title{Internal function}
\usage{
distr_to_xml_laplace(distr, beauti_options)
}
\arguments{
\item{distr}{a Laplace distribution
as created by \code{\link{create_laplace_distr}})}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the distribution as XML text
}
\description{
Converts a Laplace distribution to XML
}
\author{
Richèl J.C. Bilderbeek
}
