% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_to_xml.R
\name{distr_to_xml_log_normal}
\alias{distr_to_xml_log_normal}
\title{Internal function}
\usage{
distr_to_xml_log_normal(distr, beauti_options)
}
\arguments{
\item{distr}{a log-normal distribution,
as created by \code{\link{create_log_normal_distr}})}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the distribution as XML text
}
\description{
Converts a log-normal distribution to XML
}
\author{
Richèl J.C. Bilderbeek
}
