% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_to_xml.R
\name{distr_to_xml_one_div_x}
\alias{distr_to_xml_one_div_x}
\title{Converts a 1/x distribution to XML}
\usage{
distr_to_xml_one_div_x(distr)
}
\arguments{
\item{distr}{a 1/x distribution,
as created by \code{\link{create_one_div_x_distr}})}
}
\value{
the distribution as XML text
}
\description{
Converts a 1/x distribution to XML
}
\author{
Richèl J.C. Bilderbeek
}
