% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_xml_section_from_lines.R
\name{extract_xml_section_from_lines}
\alias{extract_xml_section_from_lines}
\title{Get the lines of an XML section, including the section tags}
\usage{
extract_xml_section_from_lines(lines, section)
}
\arguments{
\item{lines}{lines of the XML text}

\item{section}{the XML section name}
}
\value{
the section's lines of XML text, including the tags
}
\description{
Get the lines of an XML section, including the section tags
}
\author{
Richel J.C. Bilderbeek
}
