% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_to_sequences.R
\name{fasta_file_to_sequences}
\alias{fasta_file_to_sequences}
\title{Convert a FASTA file to a table of sequences}
\usage{
fasta_file_to_sequences(fasta_filename)
}
\arguments{
\item{fasta_filename}{One existing FASTA filenames}
}
\value{
a table of sequences
}
\description{
Convert a FASTA file to a table of sequences
}
\author{
Richèl J.C. Bilderbeek
}
