% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_to_phylo.R
\name{fasta_to_phylo}
\alias{fasta_to_phylo}
\title{Create a random phylogeny, with a specified crown age, from a FASTA file}
\usage{
fasta_to_phylo(fasta_filename, crown_age)
}
\arguments{
\item{fasta_filename}{a FASTA filename.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}

\item{crown_age}{the crown age of the phylogeny}
}
\value{
a phylogeny
}
\description{
Create a random phylogeny, with a specified crown age, from a FASTA file
}
\examples{
  # Create a random phylogy, with
  # - the same taxa names as the FASTA
  # - the desired crown age
  fasta_filename <- get_fasta_filename()
  initial_phylogeny <- fasta_to_phylo(
    fasta_filename,
    crown_age = 15
   )

  # Crown age fixed to the crown age of the phylogeny
  create_beast2_input_file(
    input_fasta_filenames = fasta_filename,
    "fasta_to_phylo.xml",
    fixed_crown_age = TRUE,
    initial_phylogenies = initial_phylogeny
  )
  testthat::expect_true(file.exists("fasta_to_phylo.xml"))
}
