% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_site_model_to_xml_prior_distr.R
\name{gamma_site_model_to_xml_prior_distr}
\alias{gamma_site_model_to_xml_prior_distr}
\title{Internal function.}
\usage{
gamma_site_model_to_xml_prior_distr(inference_model)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\value{
lines of XML text
}
\description{
Creates the gamma site models section in the distribution section
of a BEAST2 XML parameter file
}
\author{
Richèl J.C. Bilderbeek
}
