% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alignment_id.R
\name{get_alignment_id}
\alias{get_alignment_id}
\title{Get the alignment ID from a FASTA filename}
\usage{
get_alignment_id(fasta_filename)
}
\arguments{
\item{fasta_filename}{a FASTA filename.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}
}
\description{
Get the alignment ID from a FASTA filename
}
\examples{
  created <- get_alignment_id("/home/homer/anthus_aco_sub.fas")
  expected <- "anthus_aco_sub"
  testit::assert(created == expected)
}
\author{
Richel J.C. Bilderbeek
}
