% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alignment_ids.R
\name{get_alignment_ids}
\alias{get_alignment_ids}
\title{Get the alignment ID from one or more FASTA filenames}
\usage{
get_alignment_ids(fasta_filenames)
}
\arguments{
\item{fasta_filenames}{One or more FASTA filenames.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}
}
\description{
Get the alignment ID from one or more FASTA filenames
}
\author{
Richel J.C. Bilderbeek
}
