% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beautier_paths.R
\name{get_beautier_paths}
\alias{get_beautier_paths}
\title{Get the full paths of files in the \code{inst/extdata} folder}
\usage{
get_beautier_paths(filenames)
}
\arguments{
\item{filenames}{the files' names, without the path}
}
\value{
the filenames' full paths
}
\description{
Get the full paths of files in the \code{inst/extdata} folder
}
\examples{
  testit::assert(
    length(
      get_beautier_paths(
        c("test_output_0.fas", "anthus_aco.fas", "anthus_nd2.fas")
      )
     ) == 3
   )
}
\seealso{
Use \link{get_beautier_path} to get the path of one file

for one file, use \code{\link{get_beautier_path}}
}
\author{
Richèl J.C. Bilderbeek
}
