% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clock_model_name.R
\name{get_clock_model_name}
\alias{get_clock_model_name}
\title{Get the BEAUti name for a clock model}
\usage{
get_clock_model_name(clock_model)
}
\arguments{
\item{clock_model}{a clock model,
as returned by \code{\link{create_clock_model}}}
}
\value{
name of the clock model
}
\description{
Will \link{stop} if the clock model is an invalid clock model
}
\examples{
# StrictClock
get_clock_model_name(create_strict_clock_model())

# RelaxedClock
get_clock_model_name(create_rln_clock_model())
}
\author{
Richèl J.C. Bilderbeek
}
