% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clock_model_names.R
\name{get_clock_model_names}
\alias{get_clock_model_names}
\title{Get the clock model names}
\usage{
get_clock_model_names()
}
\value{
the clock model names
}
\description{
Get the clock model names
}
\examples{
check_empty_beautier_folder()

get_clock_model_names()

check_empty_beautier_folder()
}
\seealso{
Use \link{create_clock_models} to create a list
  with all clock models
}
\author{
Richèl J.C. Bilderbeek
}
