% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gamma_cat_count.R
\name{get_default_gamma_cat_count}
\alias{get_default_gamma_cat_count}
\title{Get the default gamma category count}
\usage{
get_default_gamma_cat_count()
}
\description{
Get the default gamma category count
}
