% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distr_n_params.R
\name{get_distr_n_params}
\alias{get_distr_n_params}
\title{Get the number of parameters a distribution uses}
\usage{
get_distr_n_params(distr)
}
\arguments{
\item{distr}{a distribution,
as created by \code{\link{create_distr}} or (preferable)
its named functions}
}
\value{
the number of parameters that distribution uses
}
\description{
Get the number of parameters a distribution uses
}
\examples{
check_empty_beautier_folder()

get_distr_n_params(create_beta_distr())
get_distr_n_params(create_exp_distr())
get_distr_n_params(create_gamma_distr())
get_distr_n_params(create_inv_gamma_distr())
get_distr_n_params(create_laplace_distr())
get_distr_n_params(create_log_normal_distr())
get_distr_n_params(create_normal_distr())
get_distr_n_params(create_one_div_x_distr())
get_distr_n_params(create_poisson_distr())
get_distr_n_params(create_uniform_distr())

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
