% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distr_names.R
\name{get_distr_names}
\alias{get_distr_names}
\title{Get the distribution names}
\usage{
get_distr_names()
}
\value{
the distribution names
}
\description{
Get the distribution names
}
\examples{
  names <- beautier:::get_distr_names()
  testit::assert("uniform" \%in\% names)
  testit::assert("normal" \%in\% names)
  testit::assert("one_div_x" \%in\% names)
  testit::assert("log_normal" \%in\% names)
  testit::assert("exponential" \%in\% names)
  testit::assert("gamma" \%in\% names)
  testit::assert("beta" \%in\% names)
  testit::assert("laplace" \%in\% names)
  testit::assert("inv_gamma" \%in\% names)
  testit::assert("poisson" \%in\% names)
}
\author{
Richel J.C. Bilderbeek
}
