% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distr_names.R
\name{get_distr_names}
\alias{get_distr_names}
\title{Get the distribution names}
\usage{
get_distr_names()
}
\value{
the distribution names
}
\description{
Get the distribution names
}
\examples{
check_empty_beautier_folder()

get_distr_names()

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
