% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_filenames.R
\name{get_fasta_filename}
\alias{get_fasta_filename}
\title{Get the path of a FASTA file used in testing}
\usage{
get_fasta_filename()
}
\value{
the path of a FASTA file used in testing
}
\description{
Get the path of a FASTA file used in testing
}
\examples{
input_filename <- beautier::get_fasta_filename()
output_filename <- get_beautier_tempfilename()

create_beast2_input_file(
  input_filename = input_filename,
  output_filename = output_filename
)

file.remove(output_filename)
}
\author{
Richèl J.C. Bilderbeek
}
