% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_base_sans_ext.R
\name{get_file_base_sans_ext}
\alias{get_file_base_sans_ext}
\title{Get the filename's base without extension}
\usage{
get_file_base_sans_ext(filename)
}
\arguments{
\item{filename}{A filename}
}
\value{
That filename without its full path and extension
}
\description{
Get the filename's base without extension
}
\examples{
  testit::assert(
    beautier:::get_file_base_sans_ext("/home/richel/test.txt")
    == "test"
 )
}
