% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_taxa.R
\name{get_n_taxa}
\alias{get_n_taxa}
\title{Extract the number of taxa from a file}
\usage{
get_n_taxa(filename)
}
\arguments{
\item{filename}{name of a FASTA file}
}
\value{
the number of taxa
}
\description{
Extract the number of taxa from a file
}
\author{
Richel J.C. Bilderbeek
}
