% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_operator_id_pre.R
\name{get_operator_id_pre}
\alias{get_operator_id_pre}
\title{Get the prefix of operator IDs}
\usage{
get_operator_id_pre(tree_prior)
}
\arguments{
\item{tree_prior}{a tree priors,
as returned by \code{\link{create_tree_prior}}}
}
\value{
the prefix of operator IDs, similar to the name of a tree prior
}
\description{
Get the prefix of operator IDs
}
\examples{
check_empty_beautier_folder()

# BirthDeath
get_operator_id_pre(
  tree_prior = create_bd_tree_prior()
)

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
