% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_names.R
\name{get_param_names}
\alias{get_param_names}
\title{Get the parameter names}
\usage{
get_param_names()
}
\value{
the parameter names
}
\description{
Get the parameter names
}
\examples{
check_empty_beautier_folder()

names <- get_param_names()

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
