% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_names.R
\name{get_param_names}
\alias{get_param_names}
\title{Get the parameter names}
\usage{
get_param_names()
}
\value{
the parameter names
}
\description{
Get the parameter names
}
\examples{
  names <- beautier:::get_param_names()
  testit::assert("alpha" \%in\% names)
  testit::assert("beta" \%in\% names)
  testit::assert("clock_rate" \%in\% names)
  testit::assert("kappa_1" \%in\% names)
  testit::assert("kappa_2" \%in\% names)
  testit::assert("lambda" \%in\% names)
  testit::assert("m" \%in\% names)
  testit::assert("mean" \%in\% names)
  testit::assert("mu" \%in\% names)
  testit::assert("rate_ac" \%in\% names)
  testit::assert("rate_ag" \%in\% names)
  testit::assert("rate_at" \%in\% names)
  testit::assert("rate_cg" \%in\% names)
  testit::assert("rate_ct" \%in\% names)
  testit::assert("rate_gt" \%in\% names)
  testit::assert("s" \%in\% names)
  testit::assert("scale" \%in\% names)
  testit::assert("sigma" \%in\% names)
}
\author{
Richel J.C. Bilderbeek
}
