% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_invariant.R
\name{get_prop_invariant}
\alias{get_prop_invariant}
\title{Extract the proporion invariants from an HKY site model}
\usage{
get_prop_invariant(gamma_site_models)
}
\arguments{
\item{gamma_site_models}{one or more gamma_site_models,
as created by \code{\link{create_gamma_site_model}}}
}
\value{
the proporion invariants
}
\description{
Extract the proporion invariants from an HKY site model
}
