% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_model_n_distrs.R
\name{get_site_model_n_distrs}
\alias{get_site_model_n_distrs}
\title{Get the number of distributions a site model has}
\usage{
get_site_model_n_distrs(site_model)
}
\arguments{
\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}
}
\value{
the number of distributions a site model has
}
\description{
Get the number of distributions a site model has
}
\examples{
library(testthat)

# gamma site model, rates AC, AG, AT, CG and GT
expect_equal(
  get_site_model_n_distrs(create_gtr_site_model()),
  5
)

# gamma site model, kappa
expect_equal(
  get_site_model_n_distrs(create_hky_site_model()),
  1
)

# gamma site model
expect_equal(
  get_site_model_n_distrs(create_jc69_site_model()),
  0
)

# gamma site model, kappa 1 and kappa 2
expect_equal(
  get_site_model_n_distrs(create_tn93_site_model()),
  2
)
}
\author{
Richèl J.C. Bilderbeek
}
