% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_model_n_distrs.R
\name{get_site_model_n_distrs}
\alias{get_site_model_n_distrs}
\title{Get the number of distributions a site model has}
\usage{
get_site_model_n_distrs(site_model)
}
\arguments{
\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}
}
\value{
the number of distributions a site model has
}
\description{
Get the number of distributions a site model has
}
\examples{
check_empty_beautier_folder()

# 5: rates AC, AG, AT, CG and GT
get_site_model_n_distrs(create_gtr_site_model())

# 1: kappa
get_site_model_n_distrs(create_hky_site_model())

# 0: npne
get_site_model_n_distrs(create_jc69_site_model())

# 2: kappa 1 and kappa 2
get_site_model_n_distrs(create_tn93_site_model())

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
