% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_model_n_params.R
\name{get_site_model_n_params}
\alias{get_site_model_n_params}
\title{Get the number of distributions a site model has}
\usage{
get_site_model_n_params(site_model)
}
\arguments{
\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}
}
\value{
the number of parameters a site model has
}
\description{
Get the number of distributions a site model has
}
\examples{
check_empty_beautier_folder()

# Ten
get_site_model_n_params(create_gtr_site_model())

# Two
get_site_model_n_params(create_hky_site_model())

# Zero
get_site_model_n_params(create_jc69_site_model())

# Four
get_site_model_n_params(create_tn93_site_model())

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
