% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_model_names.R
\name{get_site_model_names}
\alias{get_site_model_names}
\title{Get the site model names}
\usage{
get_site_model_names(name)
}
\value{
the site model names
}
\description{
Get the site model names
}
