% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_model_names.R
\name{get_site_model_names}
\alias{get_site_model_names}
\title{Get the site models' names}
\usage{
get_site_model_names()
}
\value{
the site model names
}
\description{
Get the site models' names
}
\examples{
check_empty_beautier_folder()

get_site_model_names()

check_empty_beautier_folder()
}
\seealso{
Use \link{create_site_models} to get all site models
}
\author{
Richèl J.C. Bilderbeek
}
