% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_model_names.R
\name{get_site_model_names}
\alias{get_site_model_names}
\title{Get the site models' names}
\usage{
get_site_model_names()
}
\value{
the site model names
}
\description{
Get the site models' names
}
\examples{
  names <- beautier:::get_site_model_names()
  testit::assert("JC69" \%in\% names)
  testit::assert("HKY" \%in\% names)
  testit::assert("TN93" \%in\% names)
  testit::assert("GTR" \%in\% names)
}
\author{
Richel J.C. Bilderbeek
}
