% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_prior_names.R
\name{get_tree_prior_names}
\alias{get_tree_prior_names}
\title{Get the tree prior names}
\usage{
get_tree_prior_names(name)
}
\value{
the tree prior names
}
\description{
Get the tree prior names
}
