% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_prior_names.R
\name{get_tree_prior_names}
\alias{get_tree_prior_names}
\title{Get the tree prior names}
\usage{
get_tree_prior_names()
}
\value{
the tree prior names
}
\description{
Get the tree prior names
}
\examples{
library(testthat)

names <- get_tree_prior_names()
expect_true("birth_death" \%in\% names)
expect_true("coalescent_bayesian_skyline" \%in\% names)
expect_true("coalescent_constant_population" \%in\% names)
expect_true("coalescent_exp_population" \%in\% names)
expect_true("yule" \%in\% names)
}
\seealso{
Use \link{create_tree_priors} to get all tree priors
}
\author{
Richèl J.C. Bilderbeek
}
