% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_prior_names.R
\name{get_tree_prior_names}
\alias{get_tree_prior_names}
\title{Get the tree prior names}
\usage{
get_tree_prior_names()
}
\value{
the tree prior names
}
\description{
Get the tree prior names
}
\examples{
check_empty_beautier_folder()

get_tree_prior_names()

check_empty_beautier_folder()
}
\seealso{
Use \link{create_tree_priors} to get all tree priors
}
\author{
Richèl J.C. Bilderbeek
}
