% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unlinked_site_models.R
\name{get_unlinked_site_models}
\alias{get_unlinked_site_models}
\title{Get the first site model of each ID}
\usage{
get_unlinked_site_models(site_models)
}
\arguments{
\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}
}
\description{
Get the first site model of each ID
}
\author{
Richel J.C. Bilderbeek
}
