% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unlinked_tree_priors.R
\name{get_unlinked_tree_priors}
\alias{get_unlinked_tree_priors}
\title{Get the first tree prior of each ID}
\usage{
get_unlinked_tree_priors(tree_priors)
}
\arguments{
\item{tree_priors}{one or more tree priors,
as returned by \code{\link{create_tree_prior}}}
}
\description{
Get the first tree prior of each ID
}
\author{
Richel J.C. Bilderbeek
}
