% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xml_closing_tag.R
\name{get_xml_closing_tag}
\alias{get_xml_closing_tag}
\title{Get the XML closing tag}
\usage{
get_xml_closing_tag(text)
}
\arguments{
\item{text}{lines of XML to extract the XML closing tag from}
}
\value{
the closing tag if found, else NA
}
\description{
Get the XML closing tag
}
\examples{

# my_tag
get_xml_closing_tag("<my_tag text=something></my_tag>")

# Will return NA
get_xml_closing_tag("<my_tag text=something/>")
get_xml_closing_tag("no_xml")
}
\author{
Richèl J.C. Bilderbeek
}
