% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_mrca_prior.R
\name{has_mrca_prior}
\alias{has_mrca_prior}
\title{Determines if the inference model has an MRCA prior}
\usage{
has_mrca_prior(inference_model)
}
\arguments{
\item{inference_model}{an Bayesian phylogenetic inference model,
as can be created by \code{\link{create_inference_model}}.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.}
}
\description{
Determines if the inference model has an MRCA prior
}
\author{
Richel J.C. Bilderbeek
}
