% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_shared_rln_clock_models.R
\name{has_shared_rln_clock_models}
\alias{has_shared_rln_clock_models}
\title{Are there any shared RLN clock models?}
\usage{
has_shared_rln_clock_models(clock_models)
}
\arguments{
\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}
}
\description{
Are there any shared RLN clock models?
}
\note{
this must be FALSE, as BEAUti rejects this as well
}
\author{
Richel J.C. Bilderbeek
}
