% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_xml_closing_tag.R
\name{has_xml_closing_tag}
\alias{has_xml_closing_tag}
\title{Is an XML closing tag with the value of \code{section}
  present among the lines of
  the text?}
\usage{
has_xml_closing_tag(lines, section)
}
\arguments{
\item{lines}{lines of the XML text}

\item{section}{the XML section}
}
\value{
TRUE if there is an XML closing tag with the value of
  \code{section} present. FALSE otherwise
}
\description{
Is an XML closing tag with the value of \code{section}
  present among the lines of
  the text?
}
\author{
Richèl J.C. Bilderbeek
}
