% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_xml_opening_tag.R
\name{has_xml_opening_tag}
\alias{has_xml_opening_tag}
\title{Is an XML opening tag with value 'section' present amongst the lines of
  the text?}
\usage{
has_xml_opening_tag(lines, section = NA)
}
\arguments{
\item{lines}{lines of an XML text}

\item{section}{if NA, this function returns TRUE if there is any
XML opening tag. If \code{section} is set to a certain word,
this function returns TRUE if that tag matches \code{section}}
}
\description{
Is an XML opening tag with value 'section' present amongst the lines of
  the text?
}
