% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent.R
\name{indent}
\alias{indent}
\title{Determine if x consists out of IDs}
\usage{
indent(text, n_spaces)
}
\arguments{
\item{x}{the object to check if it consists out of IDs}
}
\value{
TRUE if x, or all elements of x, are IDs
}
\description{
Determine if x consists out of IDs
}
\author{
Richel J.C. Bilderbeek
}
