% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent.R
\name{indent}
\alias{indent}
\title{Indent text for a certain number of spaces.
If the text is only whitespace, leave it as such}
\usage{
indent(text, n_spaces = 4)
}
\arguments{
\item{text}{the text to indent}

\item{n_spaces}{the number of spaces to add before the text. BEAUti uses
four spaces by default}
}
\value{
the indented text
}
\description{
Indent text for a certain number of spaces.
If the text is only whitespace, leave it as such
}
\author{
Richèl J.C. Bilderbeek
}
