% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_tree_priors.R
\name{init_bd_tree_prior}
\alias{init_bd_tree_prior}
\title{Initializes a Birth-Death tree prior}
\usage{
init_bd_tree_prior(bd_tree_prior, distr_id, param_id)
}
\arguments{
\item{bd_tree_prior}{a Birth-Death tree prior, as created
by \code{\link{create_bd_tree_prior}}}

\item{distr_id}{a distributions' ID}

\item{param_id}{a parameter's ID}
}
\value{
an initialized Birth-Death tree prior
}
\description{
Initializes a Birth-Death tree prior
}
\author{
Richèl J.C. Bilderbeek
}
