% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_tree_priors.R
\name{init_ccp_tree_prior}
\alias{init_ccp_tree_prior}
\title{Initializes a Coalescent Constant Population tree prior}
\usage{
init_ccp_tree_prior(ccp_tree_prior, distr_id, param_id)
}
\arguments{
\item{ccp_tree_prior}{a Coalescent Constant Population tree prior,
as returned by \code{\link{create_ccp_tree_prior}}}

\item{distr_id}{a distributions' ID}

\item{param_id}{a parameter's ID}
}
\value{
an initialized Coalescent Constant Population tree prior
}
\description{
Initializes a Coalescent Constant Population tree prior
}
\author{
Richel J.C. Bilderbeek
}
