% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_tree_priors.R
\name{init_cep_tree_prior}
\alias{init_cep_tree_prior}
\title{Initializes a Coalescent Exponential Population tree prior}
\usage{
init_cep_tree_prior(cep_tree_prior, distr_id, param_id)
}
\arguments{
\item{cep_tree_prior}{a Coalescent Exponential Population tree prior,
as returned by \code{\link{create_cep_tree_prior}}}

\item{distr_id}{a distributions' ID}

\item{param_id}{a parameter's ID}
}
\value{
an initialized Coalescent Exponential Population tree prior
}
\description{
Initializes a Coalescent Exponential Population tree prior
}
\author{
Richèl J.C. Bilderbeek
}
