% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_clock_models.R
\name{init_clock_models}
\alias{init_clock_models}
\title{Initializes all clock models}
\usage{
init_clock_models(fasta_filenames, clock_models, distr_id = 0, param_id = 0)
}
\arguments{
\item{fasta_filenames}{One or more FASTA filenames.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}

\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}

\item{distr_id}{the first distributions' ID}

\item{param_id}{the first parameter's ID}
}
\value{
a list of initialized clock models
}
\description{
Initializes all clock models
}
\author{
Richèl J.C. Bilderbeek
}
