% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_gamma_site_model.R
\name{init_gamma_site_model}
\alias{init_gamma_site_model}
\title{Initializes all site models}
\usage{
init_gamma_site_model(gamma_site_model, distr_id = 0, param_id = 0)
}
\arguments{
\item{gamma_site_model}{a site model's gamma site model,
as returned by \code{\link{create_gamma_site_model}}}

\item{distr_id}{the first distributions' ID}

\item{param_id}{the first parameter's ID}
}
\value{
a list of initialized site models
}
\description{
Initializes all site models
}
\author{
Richel J.C. Bilderbeek
}
