% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_gamma_site_model.R
\name{init_gamma_site_model}
\alias{init_gamma_site_model}
\title{Initializes a gamma site model}
\usage{
init_gamma_site_model(gamma_site_model, distr_id = 0, param_id = 0)
}
\arguments{
\item{gamma_site_model}{a site model's gamma site model,
as returned by \code{\link{create_gamma_site_model}}}

\item{distr_id}{the first distributions' ID}

\item{param_id}{the first parameter's ID}
}
\value{
an initialized gamma site model
}
\description{
Initializes a gamma site model
}
\examples{
gamma_site_model <- create_gamma_site_model(
  gamma_cat_count = 2,
  gamma_shape_prior_distr = create_one_div_x_distr(id = NA)
)
# FALSE: not yet initialized
is_init_gamma_site_model(gamma_site_model)
gamma_site_model <- init_gamma_site_model(gamma_site_model)
# TRUE: now it is initialized
is_init_gamma_site_model(gamma_site_model)
}
\author{
Richèl J.C. Bilderbeek
}
