% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_site_models.R
\name{init_gtr_site_model}
\alias{init_gtr_site_model}
\title{Initializes a GTR site model}
\usage{
init_gtr_site_model(gtr_site_model, distr_id = 0, param_id = 0)
}
\value{
an initialized GTR site model
}
\description{
Initializes a GTR site model
}
\examples{
  gtr_site_model <- create_gtr_site_model()
  testit::assert(!beautier:::is_init_gtr_site_model(gtr_site_model))
  gtr_site_model <- beautier:::init_gtr_site_model(gtr_site_model)
  testit::assert(beautier:::is_init_gtr_site_model(gtr_site_model))
}
\author{
Richel J.C. Bilderbeek
}
