% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_site_models.R
\name{init_hky_site_model}
\alias{init_hky_site_model}
\title{Initializes an HKY site model}
\usage{
init_hky_site_model(hky_site_model, distr_id, param_id)
}
\arguments{
\item{hky_site_model}{an HKY site model,
as returned by \code{\link{create_hky_site_model}}}

\item{distr_id}{the first distributions' ID}

\item{param_id}{the first parameter's ID}
}
\value{
an initialized HKY site model
}
\description{
Initializes an HKY site model
}
\author{
Richel J.C. Bilderbeek
}
