% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_site_models.R
\name{init_hky_site_model}
\alias{init_hky_site_model}
\title{Initializes an HKY site model}
\usage{
init_hky_site_model(hky_site_model, distr_id = 0, param_id = 0)
}
\arguments{
\item{hky_site_model}{an HKY site model,
as returned by \code{\link{create_hky_site_model}}}

\item{distr_id}{a distributions' ID}

\item{param_id}{a parameter's ID}
}
\value{
an initialized HKY site model
}
\description{
Initializes an HKY site model
}
\examples{
check_empty_beautier_folder()

hky_site_model <- create_hky_site_model()
is_init_hky_site_model(hky_site_model)
hky_site_model <- init_hky_site_model(hky_site_model)
is_init_hky_site_model(hky_site_model)

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
