% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_site_models.R
\name{init_jc69_site_model}
\alias{init_jc69_site_model}
\title{Initializes a JC69 site model}
\usage{
init_jc69_site_model(jc69_site_model, distr_id = 0, param_id = 0)
}
\arguments{
\item{jc69_site_model}{a JC69 site model,
as returned by \code{\link{create_jc69_site_model}}}

\item{distr_id}{a distributions' ID}

\item{param_id}{a parameter's ID}
}
\value{
an initialized HKY site model
}
\description{
Initializes a JC69 site model
}
\examples{
library(testthat)

hky_site_model <- create_hky_site_model()
expect_false(is_init_hky_site_model(hky_site_model))
hky_site_model <- init_hky_site_model(hky_site_model)
expect_true(is_init_hky_site_model(hky_site_model))
}
\author{
Richèl J.C. Bilderbeek
}
