% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_mrca_priors.R
\name{init_mrca_priors}
\alias{init_mrca_priors}
\title{Initializes all MRCA priors}
\usage{
init_mrca_priors(mrca_priors, distr_id = 0, param_id = 0)
}
\arguments{
\item{mrca_priors}{a list of one or more Most Recent Common Ancestor priors,
as returned by \code{\link{create_mrca_prior}}}

\item{distr_id}{the first distributions' ID}

\item{param_id}{the first parameter's ID}
}
\value{
a list of initialized MRCA priors
}
\description{
Initializes all MRCA priors
}
\author{
Richèl J.C. Bilderbeek
}
