% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_site_models.R
\name{init_site_models}
\alias{init_site_models}
\title{Initializes all site models}
\usage{
init_site_models(site_models, ids, distr_id = 0, param_id = 0)
}
\arguments{
\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}

\item{ids}{one or more alignments' IDs.
IDs can be extracted from their FASTA filesnames
with \code{\link{get_ids}})}

\item{distr_id}{the first distributions' ID}

\item{param_id}{the first parameter's ID}
}
\value{
a list of initialized site models
}
\description{
Initializes all site models
}
\author{
Richel J.C. Bilderbeek
}
