% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_clock_models.R
\name{init_strict_clock_model}
\alias{init_strict_clock_model}
\title{Initializes a strict clock model}
\usage{
init_strict_clock_model(strict_clock_model, distr_id, param_id)
}
\value{
an initialized strict clock model
}
\description{
Initializes a strict clock model
}
\examples{
  strict_clock_model <- create_strict_clock_model()
}
\author{
Richel J.C. Bilderbeek
}
