% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_tree_priors.R
\name{init_tree_priors}
\alias{init_tree_priors}
\title{Initializes all tree priors}
\usage{
init_tree_priors(tree_priors, ids, distr_id = 0, param_id = 0)
}
\arguments{
\item{tree_priors}{one or more tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{ids}{one or more alignments' IDs.
IDs can be extracted from their FASTA filesnames
with \code{\link{get_ids}})}

\item{distr_id}{the first distributions' ID}

\item{param_id}{the first parameter's ID}
}
\value{
a list of initialized tree priors
}
\description{
Initializes all tree priors
}
\author{
Richel J.C. Bilderbeek
}
