% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_tree_priors.R
\name{init_yule_tree_prior}
\alias{init_yule_tree_prior}
\title{Initializes a Yule tree prior}
\usage{
init_yule_tree_prior(yule_tree_prior, distr_id, param_id)
}
\arguments{
\item{yule_tree_prior}{a Yule tree_prior,
as created by \code{\link{create_yule_tree_prior}}}

\item{distr_id}{a distributions' ID}

\item{param_id}{a parameter's ID}
}
\value{
an initialized Yule tree prior
}
\description{
Initializes a Yule tree prior
}
\author{
Richèl J.C. Bilderbeek
}
