% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tree_prior.R
\name{is_bd_tree_prior}
\alias{is_bd_tree_prior}
\title{Determine if the object is a valid Birth Death tree prior}
\usage{
is_bd_tree_prior(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid birth death tree prior}
}
\value{
TRUE if x is a valid birth death tree prior, FALSE otherwise
}
\description{
Determine if the object is a valid Birth Death tree prior
}
\examples{
check_empty_beautier_folder()

  is_bd_tree_prior(create_bd_tree_prior())
  !is_bd_tree_prior(create_cbs_tree_prior())
  !is_bd_tree_prior(create_ccp_tree_prior())
  !is_bd_tree_prior(create_cep_tree_prior())
  !is_bd_tree_prior(create_yule_tree_prior())

check_empty_beautier_folder()
}
\seealso{
Use \code{\link{create_bd_tree_prior}} to create a valid
  Birth-Death tree prior
}
\author{
Richèl J.C. Bilderbeek
}
