% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_beauti_options.R
\name{is_beauti_options}
\alias{is_beauti_options}
\title{Determine if the object is a valid \code{beauti_options}}
\usage{
is_beauti_options(x)
}
\arguments{
\item{x}{an object, to be determined if it is a \code{beauti_options}}
}
\value{
\link{TRUE} if the object is a valid \code{beauti_options},
  \link{FALSE} otherwise
}
\description{
Determine if the object is a valid \code{beauti_options}
}
\examples{
library(testthat)

expect_true(is_beauti_options(create_beauti_options()))

expect_false(is_beauti_options("nonsense"))
expect_false(is_beauti_options(NA))
expect_false(is_beauti_options(NULL))
expect_false(is_beauti_options(""))
expect_false(is_beauti_options(c()))
}
\seealso{
use \link{create_beauti_options} to create a valid
\code{beauti_options} object
}
\author{
Richèl J.C. Bilderbeek
}
