% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_beta_distr.R
\name{is_beta_distr}
\alias{is_beta_distr}
\title{Determine if the object is a valid
beta distribution,
as created by \code{\link{create_beta_distr}}}
\usage{
is_beta_distr(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
beta distribution,}
}
\value{
TRUE if x is a valid beta distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid
beta distribution,
as created by \code{\link{create_beta_distr}}
}
\seealso{
use \code{\link{is_distr}} to see if x is any
  distribution
}
\author{
Richel J.C. Bilderbeek
}
