% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cbs_tree_prior.R
\name{is_cbs_tree_prior}
\alias{is_cbs_tree_prior}
\title{Determine if the object is a valid constant coalescent Bayesian skyline prior
  as returned by \code{\link{create_cbs_tree_prior}}}
\usage{
is_cbs_tree_prior(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid constant coalescent
Bayesian skyline prior}
}
\value{
TRUE if x is a valid constant coalescent Bayesian skyline prior,
  FALSE otherwise
}
\description{
Determine if the object is a valid constant coalescent Bayesian skyline prior
  as returned by \code{\link{create_cbs_tree_prior}}
}
\author{
Richel J.C. Bilderbeek
}
