% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tree_prior.R
\name{is_cbs_tree_prior}
\alias{is_cbs_tree_prior}
\title{Determine if the object is a valid constant coalescent Bayesian skyline prior}
\usage{
is_cbs_tree_prior(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid constant coalescent
Bayesian skyline prior}
}
\value{
TRUE if x is a valid constant coalescent Bayesian skyline prior,
  FALSE otherwise
}
\description{
Determine if the object is a valid constant coalescent Bayesian skyline prior
}
\examples{
check_empty_beautier_folder()

# TRUE
is_cbs_tree_prior(create_cbs_tree_prior())

# FALSE
is_cbs_tree_prior(create_bd_tree_prior())
is_cbs_tree_prior(create_ccp_tree_prior())
is_cbs_tree_prior(create_cep_tree_prior())
is_cbs_tree_prior(create_yule_tree_prior())

check_empty_beautier_folder()
}
\seealso{
Use \code{\link{create_cbs_tree_prior}} to create a valid
  coalescent Bayes skyline tree prior
}
\author{
Richèl J.C. Bilderbeek
}
