% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tree_prior.R
\name{is_ccp_tree_prior}
\alias{is_ccp_tree_prior}
\title{Determine if the object is a valid
  constant coalescence population tree prior}
\usage{
is_ccp_tree_prior(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
constant coalescence population tree prior}
}
\value{
TRUE if x is a valid constant coalescence population tree prior,
  FALSE otherwise
}
\description{
Determine if the object is a valid
  constant coalescence population tree prior
}
\examples{
check_empty_beautier_folder()
  testit::assert(!is_ccp_tree_prior(create_bd_tree_prior()))
  testit::assert(!is_ccp_tree_prior(create_cbs_tree_prior()))
  testit::assert( is_ccp_tree_prior(create_ccp_tree_prior()))
  testit::assert(!is_ccp_tree_prior(create_cep_tree_prior()))
  testit::assert(!is_ccp_tree_prior(create_yule_tree_prior()))
check_empty_beautier_folder()
}
\seealso{
Use \code{\link{create_ccp_tree_prior}} to create a valid
  constant coalescence population tree prior
}
\author{
Richèl J.C. Bilderbeek
}
