% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_clock_model.R
\name{is_clock_model}
\alias{is_clock_model}
\title{Determine if the object is a valid clock_model}
\usage{
is_clock_model(x)
}
\arguments{
\item{x}{an object, to be determined if it is a clock_model}
}
\value{
TRUE if the clock_model is a valid clock_model, FALSE otherwise
}
\description{
Determine if the object is a valid clock_model
}
\seealso{
see \code{\link{create_clock_model}} for an overview of functions
  to create valid clock model
}
\author{
Richel J.C. Bilderbeek
}
