% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_default_mcmc.R
\name{is_default_mcmc}
\alias{is_default_mcmc}
\title{Determine if the MCMC is a default MCMC}
\usage{
is_default_mcmc(mcmc)
}
\value{
TRUE if the MCMC is a default MCMC
}
\description{
Determine if the MCMC is a default MCMC
}
\examples{
  # An MCMC created by 'create_mcmc' is default
  testthat::expect_true(beautier:::is_default_mcmc(
    create_mcmc()
  ))

  # An MCMC created by 'create_mcmc_nested_sampling' is not
  testthat::expect_false(beautier:::is_default_mcmc(
    create_mcmc_nested_sampling()
  ))
}
\author{
Richèl J.C. Bilderbeek
}
