% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_distr_name.R
\name{is_distr_name}
\alias{is_distr_name}
\title{Determines if the name is a valid distribution name}
\usage{
is_distr_name(name)
}
\arguments{
\item{name}{the name to be tested}
}
\value{
TRUE if the name is a valid distribution name, FALSE otherwise
}
\description{
Determines if the name is a valid distribution name
}
\examples{
  testit::assert(beautier:::is_distr_name("uniform"))
  testit::assert(beautier:::is_distr_name("normal"))
  testit::assert(beautier:::is_distr_name("one_div_x"))
  testit::assert(beautier:::is_distr_name("log_normal"))
  testit::assert(beautier:::is_distr_name("exponential"))
  testit::assert(beautier:::is_distr_name("gamma"))
  testit::assert(beautier:::is_distr_name("beta"))
  testit::assert(beautier:::is_distr_name("laplace"))
  testit::assert(beautier:::is_distr_name("inv_gamma"))
  testit::assert(beautier:::is_distr_name("poisson"))
}
\author{
Richel J.C. Bilderbeek
}
