% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_distribution.R
\name{is_distribution}
\alias{is_distribution}
\title{Determine if the object is a valid distribution}
\usage{
is_distribution(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
distribution}
}
\value{
TRUE if x is a valid distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid distribution
}
\author{
Richel J.C. Bilderbeek
}
