% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gamma_distribution.R
\name{is_gamma_distribution}
\alias{is_gamma_distribution}
\title{Determine if the object is a valid
gamma distribution,
as created by \code{\link{create_gamma_distr}}}
\usage{
is_gamma_distribution(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
gamma distribution}
}
\value{
TRUE if x is a valid gamma distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid
gamma distribution,
as created by \code{\link{create_gamma_distr}}
}
\seealso{
use \code{\link{is_distribution}} to see if x is any
  distribution
}
\author{
Richel J.C. Bilderbeek
}
