% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gamma_site_model.R
\name{is_gamma_site_model}
\alias{is_gamma_site_model}
\title{Is object x a gamma site model?}
\usage{
is_gamma_site_model(x)
}
\arguments{
\item{x}{the object to be determined if it is a valid gamma site object}
}
\value{
TRUE if x is a valid gamma site object, FALSE otherwise
}
\description{
Is object x a gamma site model?
}
\examples{
check_empty_beautier_folder()

# TRUE
is_gamma_site_model(create_gamma_site_model())

# FALSE
is_gamma_site_model("nonsense")
is_gamma_site_model(NA)
is_gamma_site_model(NULL)
is_gamma_site_model("")
is_gamma_site_model(c())

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
