% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gtr_site_model.R
\name{is_gtr_site_model}
\alias{is_gtr_site_model}
\title{Determine if the object is a valid GTR site model,
as created by \code{\link{create_gtr_site_model}}}
\usage{
is_gtr_site_model(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid GTR site model}
}
\value{
TRUE if x is a valid GTR site model, FALSE otherwise
}
\description{
Determine if the object is a valid GTR site model,
as created by \code{\link{create_gtr_site_model}}
}
\examples{
  gtr_site_model <- create_gtr_site_model()
  testit::assert(beautier:::is_gtr_site_model(gtr_site_model))
}
\author{
Richel J.C. Bilderbeek
}
