% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_id.R
\name{is_id}
\alias{is_id}
\title{Determine if the object is a valid ID}
\usage{
is_id(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid ID}
}
\value{
TRUE if x is a valid ID, FALSE otherwise
}
\description{
Determine if the object is a valid ID
}
\seealso{
to check multiple IDs, use \code{\link{are_ids}}
}
\author{
Richel J.C. Bilderbeek
}
