% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_in_patterns.R
\name{is_in_patterns}
\alias{is_in_patterns}
\title{Create a random beta distribution}
\usage{
is_in_patterns(line, patterns)
}
\arguments{
\item{line}{a line of text}

\item{patterns}{one or more regular expression patterns}
}
\value{
TRUE if there is at least one match found
}
\description{
Create a random beta distribution
}
\author{
Richel J.C. Bilderbeek
}
