% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_distr.R
\name{is_init_laplace_distr}
\alias{is_init_laplace_distr}
\title{Determine if x is an initialized laplace distribution object
  as created by \code{\link{create_laplace_distr}}}
\usage{
is_init_laplace_distr(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized laplace distribution object}
}
\value{
TRUE if x is an initialized laplace distribution object
}
\description{
Determine if x is an initialized laplace distribution object
  as created by \code{\link{create_laplace_distr}}
}
\author{
Richel J.C. Bilderbeek
}
