% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_init_mrca_prior.R
\name{is_init_mrca_prior}
\alias{is_init_mrca_prior}
\title{Determine if x is an initialized MRCA prior}
\usage{
is_init_mrca_prior(x)
}
\arguments{
\item{x}{the object to check if it is an
initialized MRCA prior}
}
\value{
TRUE if x is an initialized MRCA prior
}
\description{
Determine if x is an initialized MRCA prior
}
\author{
Richèl J.C. Bilderbeek
}
