% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_jc69_site_model.R
\name{is_jc69_site_model}
\alias{is_jc69_site_model}
\title{Determine if the object is a valid JC69 site model}
\usage{
is_jc69_site_model(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid JC69 site model}
}
\value{
TRUE if x is a valid JC69 site model, FALSE otherwise
}
\description{
Determine if the object is a valid JC69 site model
}
\examples{
  jc69_site_model <- create_jc69_site_model()
  testit::assert(beautier:::is_jc69_site_model(jc69_site_model))
}
\author{
Richel J.C. Bilderbeek
}
