% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_laplace_distribution.R
\name{is_laplace_distribution}
\alias{is_laplace_distribution}
\title{Determine if the object is a valid
laplace distribution,
as created by \code{\link{create_laplace_distr}}}
\usage{
is_laplace_distribution(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
laplace distribution}
}
\value{
TRUE if x is a valid laplace distribution,
  FALSE otherwise
}
\description{
Determine if the object is a valid
laplace distribution,
as created by \code{\link{create_laplace_distr}}
}
\seealso{
use \code{\link{is_distribution}} to see if x is any
  distribution
}
\author{
Richel J.C. Bilderbeek
}
