% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mcmc.R
\name{is_mcmc}
\alias{is_mcmc}
\title{Determine if the object is a valid mcmc object}
\usage{
is_mcmc(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid mcmc object}
}
\value{
TRUE if x is a valid mcmc object, FALSE otherwise
}
\description{
Determine if the object is a valid mcmc object
}
\seealso{
Use \code{\link{create_mcmc}}
}
\author{
Richel J.C. Bilderbeek
}
