% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mu_param.R
\name{is_mu_param}
\alias{is_mu_param}
\title{Determine if the object is a valid
mu parameter}
\usage{
is_mu_param(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
mu parameter}
}
\value{
TRUE if x is a valid mu parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid
mu parameter
}
\examples{
  mu_param <- create_mu_param()
  testit::assert(beautier:::is_mu_param(mu_param))
}
\seealso{
\code{\link{create_mu_param}} creates a mu parameter
}
\author{
Richel J.C. Bilderbeek
}
