% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_one_double.R
\name{is_one_double}
\alias{is_one_double}
\title{Determines if the argument is a double}
\usage{
is_one_double(x)
}
\arguments{
\item{x}{the object to be determined of if it is one double}
}
\description{
Determines if the argument is a double
}
\examples{
library(testthat)

expect_true(is_one_double(314))
expect_true(is_one_double(0))
expect_true(is_one_double(-314))
expect_true(is_one_double(3.14))
expect_false(is_one_double(NULL))
expect_false(is_one_double(NA))
expect_false(is_one_double(Inf))
expect_false(is_one_double("nonsense"))
expect_false(is_one_double(is_one_double))
expect_false(is_one_double(c()))
expect_false(is_one_double(c(1, 2)))
}
\author{
Richèl J.C. Bilderbeek
}
