% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_one_int.R
\name{is_one_int}
\alias{is_one_int}
\title{Determines if the argument is a whole number}
\usage{
is_one_int(x, tolerance = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{the object to be determined of if it is one integer}

\item{tolerance}{tolerance to rounding errors}
}
\description{
Determines if the argument is a whole number
}
\examples{
check_empty_beautier_folder()

# TRUE
is_one_int(314)
is_one_int(0)
is_one_int(-314)
# FALSE
is_one_int(3.14)
is_one_int(NULL)
is_one_int(NA)
is_one_int(Inf)
is_one_int("nonsense")
is_one_int(c())
is_one_int(c(1, 2))

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
