% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_one_string.R
\name{is_one_string}
\alias{is_one_string}
\title{Determines if the argument is one string}
\usage{
is_one_string(x)
}
\arguments{
\item{x}{the object to be determined of if it is one string}
}
\value{
TRUE if the argument is one string, FALSE otherwise
}
\description{
Determines if the argument is one string
}
\examples{
check_empty_beautier_folder()

# TRUE
is_one_string("This is one string")

# FALSE
is_one_string(NULL)
is_one_string(NA)
is_one_string(Inf)
is_one_string(314)
is_one_string(0)
is_one_string(-314)
is_one_string(3.14)
is_one_string(c("a", "b"))
is_one_string(is_one_string)
is_one_string(c())
is_one_string(c(1, 2))

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
