% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_param.R
\name{is_param}
\alias{is_param}
\title{Determine if the object is a valid parameter}
\usage{
is_param(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid parameter,
as created by \code{\link{create_param}})}
}
\value{
TRUE if x is a valid parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid parameter
}
\author{
Richel J.C. Bilderbeek
}
