% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_param_name.R
\name{is_param_name}
\alias{is_param_name}
\title{Determines if the name is a valid parameter name}
\usage{
is_param_name(name)
}
\arguments{
\item{name}{the name to be tested}
}
\value{
TRUE if the name is a valid parameter name, FALSE otherwise
}
\description{
Determines if the name is a valid parameter name
}
\examples{
  testit::assert(beautier:::is_param_name("alpha"))
  testit::assert(beautier:::is_param_name("beta"))
  testit::assert(beautier:::is_param_name("clock_rate"))
  testit::assert(beautier:::is_param_name("kappa_1"))
  testit::assert(beautier:::is_param_name("kappa_2"))
  testit::assert(beautier:::is_param_name("lambda"))
  testit::assert(beautier:::is_param_name("m"))
  testit::assert(beautier:::is_param_name("mean"))
  testit::assert(beautier:::is_param_name("mu"))
  testit::assert(beautier:::is_param_name("rate_ac"))
  testit::assert(beautier:::is_param_name("rate_ag"))
  testit::assert(beautier:::is_param_name("rate_at"))
  testit::assert(beautier:::is_param_name("rate_cg"))
  testit::assert(beautier:::is_param_name("rate_ct"))
  testit::assert(beautier:::is_param_name("rate_gt"))
  testit::assert(beautier:::is_param_name("s"))
  testit::assert(beautier:::is_param_name("scale"))
  testit::assert(beautier:::is_param_name("sigma"))
}
\author{
Richel J.C. Bilderbeek
}
