% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_phylo.R
\name{is_phylo}
\alias{is_phylo}
\title{Checks if the input is a phylogeny}
\usage{
is_phylo(x)
}
\arguments{
\item{x}{input to be checked}
}
\value{
TRUE or FALSE
}
\description{
Checks if the input is a phylogeny
}
\examples{
  phylogeny <- ape::read.tree(text = "(a:15,b:15):1;")
  testit::assert(is_phylo(phylogeny))

  testit::assert(!is_phylo("nonsense"))
  testit::assert(!is_phylo(NA))
  testit::assert(!is_phylo(NULL))
}
\seealso{
Use \link{check_phylogeny} to check for a phylogeny
}
\author{
Richèl J.C. Bilderbeek
}
