% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_phylo.R
\name{is_phylo}
\alias{is_phylo}
\title{Checks if the input is a phylogeny}
\usage{
is_phylo(x)
}
\arguments{
\item{x}{input to be checked}
}
\value{
TRUE or FALSE
}
\description{
Checks if the input is a phylogeny
}
\examples{
check_empty_beautier_folder()

# TRUE
phylogeny <- ape::read.tree(text = "(a:15,b:15):1;")
is_phylo(phylogeny)

# FALSE
is_phylo("nonsense")
is_phylo(NA)
is_phylo(NULL)

check_empty_beautier_folder()
}
\seealso{
Use \link{check_phylogeny} to check for a phylogeny
}
\author{
Richèl J.C. Bilderbeek
}
