% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rate_ac_param.R
\name{is_rate_ac_param}
\alias{is_rate_ac_param}
\title{Determine if the object is a valid
'rate AC' parameter}
\usage{
is_rate_ac_param(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
'rate AC' parameter}
}
\value{
TRUE if x is a valid 'rate AC' parameter,
  FALSE otherwise
}
\description{
Determine if the object is a valid
'rate AC' parameter
}
\examples{
  rate_ac_param <- create_rate_ac_param()
  testit::assert(beautier:::is_rate_ac_param(rate_ac_param))
}
\seealso{
\code{\link{create_rate_ac_param}} creates a 'rate AC' parameter
}
\author{
Richel J.C. Bilderbeek
}
