% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rln_clock_model.R
\name{is_rln_clock_model}
\alias{is_rln_clock_model}
\title{Determine if the object is a valid relaxed log normal clock model}
\usage{
is_rln_clock_model(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
relaxed log normal clock model,
as created by \code{\link{create_rln_clock_model}})}
}
\value{
TRUE if x is a valid relaxed log normal clock model, FALSE otherwise
}
\description{
Determine if the object is a valid relaxed log normal clock model
}
\examples{
  rln_clock_model <- create_rln_clock_model()
  testit::assert(beautier:::is_rln_clock_model(rln_clock_model))

  strict_clock_model <- create_strict_clock_model()
  testit::assert(beautier:::is_strict_clock_model(strict_clock_model))
}
\seealso{
\code{\link{create_clock_model}} shows an overview of
  functions to create a clock model
}
\author{
Richel J.C. Bilderbeek
}
