% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_clock_model.R
\name{is_rln_clock_model}
\alias{is_rln_clock_model}
\title{Determine if the object is a valid relaxed log normal clock model}
\usage{
is_rln_clock_model(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid
relaxed log normal clock model,
as created by \code{\link{create_rln_clock_model}})}
}
\value{
TRUE if x is a valid relaxed log normal clock model, FALSE otherwise
}
\description{
Determine if the object is a valid relaxed log normal clock model
}
\examples{

is_rln_clock_model(create_strict_clock_model())
is_rln_clock_model(create_rln_clock_model())

is_rln_clock_model(NA)
is_rln_clock_model(NULL)
is_rln_clock_model("nonsense")
is_rln_clock_model(create_jc69_site_model())
is_rln_clock_model(create_mcmc())
}
\seealso{
\code{\link{create_clock_model}} shows an overview of
  functions to create a clock model
}
\author{
Richèl J.C. Bilderbeek
}
