% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_site_model.R
\name{is_site_model}
\alias{is_site_model}
\title{Determine if the object is a valid site_model}
\usage{
is_site_model(x)
}
\arguments{
\item{x}{an object, to be determined if it is a site_model}
}
\value{
TRUE if the site_model is a valid site_model, FALSE otherwise
}
\description{
Determine if the object is a valid site_model
}
\examples{
  # site models
  testit::assert(is_site_model(create_gtr_site_model()))
  testit::assert(is_site_model(create_hky_site_model()))
  testit::assert(is_site_model(create_jc69_site_model()))
  testit::assert(is_site_model(create_tn93_site_model()))

  # other models
  testit::assert(!is_site_model(create_strict_clock_model()))
  testit::assert(!is_site_model(create_bd_tree_prior()))
  testit::assert(!is_site_model(create_mcmc()))
}
\seealso{
A site model can be created using \code{\link{create_site_model}}
}
