% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tn93_site_model.R
\name{is_tn93_site_model}
\alias{is_tn93_site_model}
\title{Determine if the object is a valid TN93 site model,}
\usage{
is_tn93_site_model(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid TN93 site model,
as created by \code{\link{create_tn93_site_model}}}
}
\value{
TRUE if x is a valid TN93 site model, FALSE otherwise
}
\description{
Determine if the object is a valid TN93 site model,
}
\examples{
 create_beast2_input_file(
   input_filenames = get_fasta_filename(),
   "beast.xml",
   site_models = create_tn93_site_model()
 )
}
\author{
Richel J.C. Bilderbeek
}
