% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tree_prior.R
\name{is_tree_prior}
\alias{is_tree_prior}
\title{Determine if an object is a valid tree prior}
\usage{
is_tree_prior(x)
}
\arguments{
\item{x}{an object}
}
\value{
TRUE if x is a valid tree_prior, FALSE otherwise
}
\description{
Determine if an object is a valid tree prior
}
\examples{
check_empty_beautier_folder()

is_tree_prior(create_bd_tree_prior())
is_tree_prior(create_yule_tree_prior())
!is_tree_prior("nonsense")

check_empty_beautier_folder()
}
\seealso{
tree priors can be created by \code{\link{create_tree_prior}}
}
\author{
Richèl J.C. Bilderbeek
}
