% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tree_prior_name.R
\name{is_tree_prior_name}
\alias{is_tree_prior_name}
\title{Determines if the name is a valid tree prior name}
\usage{
is_tree_prior_name(name)
}
\arguments{
\item{name}{the name to be tested}
}
\value{
TRUE if the name is a valid tree_prior name, FALSE otherwise
}
\description{
Determines if the name is a valid tree prior name
}
\examples{
check_empty_beautier_folder()

# TRUE
is_tree_prior_name("birth_death")
is_tree_prior_name("coalescent_bayesian_skyline")
is_tree_prior_name("coalescent_constant_population")
is_tree_prior_name("coalescent_exp_population")
is_tree_prior_name("yule")
# FALSE
is_tree_prior_name("nonsense")

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
