% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tree_prior_name.R
\name{is_tree_prior_name}
\alias{is_tree_prior_name}
\title{Determines if the name is a valid tree prior name}
\usage{
is_tree_prior_name(name)
}
\arguments{
\item{name}{the name to be tested}
}
\value{
TRUE if the name is a valid tree_prior name, FALSE otherwise
}
\description{
Determines if the name is a valid tree prior name
}
\examples{
  testit::assert(
    beautier:::is_tree_prior_name("birth_death")
  )
  testit::assert(
    beautier:::is_tree_prior_name("coalescent_bayesian_skyline")
  )
  testit::assert(
    beautier:::is_tree_prior_name("coalescent_constant_population")
  )
  testit::assert(
    beautier:::is_tree_prior_name("coalescent_exp_population")
  )
  testit::assert(
    beautier:::is_tree_prior_name("yule")
  )
}
\author{
Richel J.C. Bilderbeek
}
