% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tree_prior.R
\name{is_yule_tree_prior}
\alias{is_yule_tree_prior}
\title{Determine if the object is a valid Yule tree prior,}
\usage{
is_yule_tree_prior(x)
}
\arguments{
\item{x}{an object, to be determined if it is a valid Yule tree prior}
}
\value{
TRUE if x is a valid Yule tree prior, FALSE otherwise
}
\description{
Determine if the object is a valid Yule tree prior,
}
\examples{
  testit::assert(!is_yule_tree_prior(create_bd_tree_prior()))
  testit::assert(!is_yule_tree_prior(create_cbs_tree_prior()))
  testit::assert(!is_yule_tree_prior(create_ccp_tree_prior()))
  testit::assert(!is_yule_tree_prior(create_cep_tree_prior()))
  testit::assert( is_yule_tree_prior(create_yule_tree_prior()))
}
\seealso{
Use \code{\link{create_yule_tree_prior}} to create a valid
  Yule tree prior
}
\author{
Richèl J.C. Bilderbeek
}
