% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_to_xml_run.R
\name{mcmc_to_xml_run}
\alias{mcmc_to_xml_run}
\title{Converts an MCMC object to the run section's XML}
\usage{
mcmc_to_xml_run(mcmc)
}
\arguments{
\item{mcmc}{one MCMC.
Use \code{\link{create_mcmc}} to create an MCMC.
Use \code{\link{create_ns_mcmc}} to create an MCMC
  for a Nested Sampling run.
Use \code{\link{check_mcmc}} to check if an MCMC is valid.
Use \code{\link{rename_mcmc_filenames}} to rename the filenames in an MCMC.}
}
\value{
the XML as text
}
\description{
Converts an MCMC object to the run section's XML
}
\examples{
if (is_on_ci()) {

  check_empty_beautier_folder()

  # <run id=\"mcmc\" spec=\"MCMC\" chainLength=\"1e+07\">
  mcmc_to_xml_run(create_mcmc())

  check_empty_beautier_folder()
}
}
\author{
Richèl J.C. Bilderbeek
}
