% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrca_prior_to_xml_prior_distr.R
\name{mrca_prior_to_xml_prior_distr}
\alias{mrca_prior_to_xml_prior_distr}
\title{Creates the distribution section in the prior section of the
distribution section of a BEAST2 XML parameter file.}
\usage{
mrca_prior_to_xml_prior_distr(
  inference_model,
  mrca_prior = "deprecated",
  has_non_strict_clock_model = "deprecated",
  taxa_names_with_ids = "deprecated"
)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}

\item{mrca_prior}{a Most Recent Common Ancestor prior,
as returned by \code{\link{create_mrca_prior}}}

\item{has_non_strict_clock_model}{boolean to indicate that the is
already at least one non-strict (i.e. relaxed log-normal) clock model}

\item{taxa_names_with_ids}{taxa names that already have received
an ID. Causes the XML to \code{idref} these}
}
\value{
lines of XML text
}
\description{
These lines start with '<distribution id='
}
\examples{
 # <distribution id="posterior" spec="util.CompoundDistribution">
 #     <distribution id="prior" spec="util.CompoundDistribution">
 #       HERE, where the ID of the distribution is 'prior'
 #     </distribution>
 #     <distribution id="likelihood" ...>
 #     </distribution>
 # </distribution>
}
\author{
Richèl J.C. Bilderbeek
}
