% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrca_prior_to_xml_state.R
\name{mrca_prior_to_xml_state}
\alias{mrca_prior_to_xml_state}
\title{Creates the XML of an MRCA prior,
  as used in the \code{state} section}
\usage{
mrca_prior_to_xml_state(mrca_prior, has_non_strict_clock_model = FALSE)
}
\arguments{
\item{mrca_prior}{a Most Recent Common Ancestor prior,
as returned by \code{\link{create_mrca_prior}}}

\item{has_non_strict_clock_model}{boolean to indicate that the is
already at least one non-strict (i.e. relaxed log-normal) clock model}
}
\value{
the tree prior as XML text
}
\description{
Creates the XML of an MRCA prior,
  as used in the \code{state} section
}
\examples{
created <- mrca_prior_to_xml_state(
 mrca_prior = create_mrca_prior(
    alignment_id = "test_output_0",
    mrca_distr = create_normal_distr(id = 42)
  ),
  has_non_strict_clock_model = FALSE
)
expect_match(created, "<parameter id=\"clockRate.c:")
}
\author{
Richèl J.C. Bilderbeek
}
