% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrca_prior_to_xml_state.R
\name{mrca_prior_to_xml_state}
\alias{mrca_prior_to_xml_state}
\title{Creates the XML of an MRCA prior,
  as used in the \code{state} section}
\usage{
mrca_prior_to_xml_state(mrca_prior, has_non_strict_clock_model = FALSE,
  is_first = TRUE)
}
\arguments{
\item{mrca_prior}{a Most Recent Common Ancestor prior,
as returned by \code{\link{create_mrca_prior}}}

\item{has_non_strict_clock_model}{boolean to indicate that the is
already at least one non-strict (i.e. relaxed log-normal) clock model}

\item{is_first}{is this the first MRCA prior?}
}
\value{
the tree prior as XML text
}
\description{
Creates the XML of an MRCA prior,
  as used in the \code{state} section
}
\author{
Richel J.C. Bilderbeek
}
