% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrca_prior_to_xml_tracelog.R
\name{mrca_prior_to_xml_tracelog}
\alias{mrca_prior_to_xml_tracelog}
\title{Creates the mrca prior's XML for the tracelog section}
\usage{
mrca_prior_to_xml_tracelog(mrca_prior, has_non_strict_clock_model = FALSE,
  is_first = TRUE)
}
\arguments{
\item{mrca_prior}{a Most Recent Common Ancestor prior,
as returned by \code{\link{create_mrca_prior}}}

\item{has_non_strict_clock_model}{boolean to indicate that the is
already at least one non-strict (i.e. relaxed log-normal) clock model}

\item{is_first}{is this the first MRCA prior?}
}
\description{
Creates the mrca prior's XML for the tracelog section
}
\examples{
# <logger id="tracelog" ...>
#'   # Here
# </logger>
}
\seealso{
all mrca priors' tracelog section is created
  by \code{\link{mrca_priors_to_xml_tracelog}}
}
\author{
Richel J.C. Bilderbeek
}
