% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrca_prior_to_xml_tracelog.R
\name{mrca_prior_to_xml_tracelog}
\alias{mrca_prior_to_xml_tracelog}
\title{Creates the MRCA prior's XML for the tracelog section}
\usage{
mrca_prior_to_xml_tracelog(clock_models, mrca_prior, tipdates_filename = NA)
}
\arguments{
\item{clock_models}{a list of one or more clock models,
as returned by \code{\link{create_clock_model}}}

\item{mrca_prior}{a Most Recent Common Ancestor prior,
as returned by \code{\link{create_mrca_prior}}}

\item{tipdates_filename}{name of the file containing the tip dates.
This file is assumed to have two columns, separated by a tab.
The first column contains the taxa names, the second column contains
the date.}
}
\value{
lines of XML text
}
\description{
Creates the MRCA prior's XML for the tracelog section
}
\examples{
# <logger id="tracelog" ...>
#'   # Here
# </logger>
}
\seealso{
all MRCA priors' tracelog section is created
  by \code{\link{mrca_priors_to_xml_tracelog}}
}
\author{
Richèl J.C. Bilderbeek
}
