% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrca_priors_to_xml_distr.R
\name{mrca_priors_to_xml_prior_distr}
\alias{mrca_priors_to_xml_prior_distr}
\title{Creates the distribution section in the prior section of the
distribution section of a BEAST2 XML parameter file.
These lines start with '<distribution id='}
\usage{
mrca_priors_to_xml_prior_distr(mrca_priors, has_non_strict_clock_model)
}
\arguments{
\item{mrca_priors}{a list of one or more Most Recent Common Ancestor priors,
as returned by \code{\link{create_mrca_prior}}}

\item{has_non_strict_clock_model}{boolean to indicate that the is
already at least one non-strict (i.e. relaxed log-normal) clock model}
}
\description{
Creates the distribution section in the prior section of the
distribution section of a BEAST2 XML parameter file.
These lines start with '<distribution id='
}
\examples{
 # <distribution id="posterior" spec="util.CompoundDistribution">
 #     <distribution id="prior" spec="util.CompoundDistribution">
 #       HERE, where the ID of the distribution is 'prior'
 #     </distribution>
 #     <distribution id="likelihood" ...>
 #     </distribution>
 # </distribution>
}
\author{
Richel J.C. Bilderbeek
}
