% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrca_priors_to_xml_state.R
\name{mrca_priors_to_xml_state}
\alias{mrca_priors_to_xml_state}
\title{Converts one or more MRCA priors to the \code{state} section of the
XML as text}
\usage{
mrca_priors_to_xml_state(mrca_priors, has_non_strict_clock_model = FALSE)
}
\arguments{
\item{mrca_priors}{a list of one or more Most Recent Common Ancestor priors,
as returned by \code{\link{create_mrca_prior}}}

\item{has_non_strict_clock_model}{boolean to indicate that the is
already at least one non-strict (i.e. relaxed log-normal) clock model}
}
\value{
lines of XML text, without indentation nor \code{state}
  tags
}
\description{
Converts one or more MRCA priors to the \code{state} section of the
XML as text
}
\author{
Richèl J.C. Bilderbeek
}
