% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrca_priors_to_xml_tracelog.R
\name{mrca_priors_to_xml_tracelog}
\alias{mrca_priors_to_xml_tracelog}
\title{Creates the mrca priors' XML for the tracelog section}
\usage{
mrca_priors_to_xml_tracelog(mrca_priors, has_non_strict_clock_model = FALSE)
}
\arguments{
\item{mrca_priors}{a list of one or more Most Recent Common Ancestor priors,
as returned by \code{\link{create_mrca_prior}}}

\item{has_non_strict_clock_model}{boolean to indicate that the is
already at least one non-strict (i.e. relaxed log-normal) clock model}
}
\description{
Creates the mrca priors' XML for the tracelog section
}
\examples{
# <logger id="tracelog" ...>
#'   # Here
# </logger>
}
\seealso{
the complete tracelog section is created
  by \code{\link{create_beast2_input_tracelog}}
}
\author{
Richel J.C. Bilderbeek
}
