% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needs_trait_set_str.R
\name{needs_trait_set_str}
\alias{needs_trait_set_str}
\title{Is it needed to add the tip dates as a string to the BEAST2 XML input file?}
\usage{
needs_trait_set_str(inference_model)
}
\arguments{
\item{inference_model}{a Bayesian phylogenetic inference model.
An inference model is the complete model setup in which a site model,
clock model, tree prior and more are specified.
Use \link{create_inference_model} to create an inference model.
Use \link{check_inference_model} to check if  an inference model is valid.
Use \link{rename_inference_model_filenames} to rename the files in an
inference model.}
}
\value{
TRUE if this is needed
}
\description{
The hypothesis is that this is needed when
all times are zero.
}
\author{
Richèl J.C. Bilderbeek
}
