% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_to_xml_tree.R
\name{no_taxa_to_xml_tree}
\alias{no_taxa_to_xml_tree}
\title{Creates the \code{tree} section
(part of the \code{state} section)
when there is no tip-dating}
\usage{
no_taxa_to_xml_tree(id)
}
\arguments{
\item{id}{an alignment's IDs.
An ID can be extracted from its FASTA filename
with \code{\link{get_alignment_ids_from_fasta_filenames}})}
}
\value{
the random phylogeny as XML text
}
\description{
Creates the \code{tree} section
(part of the \code{state} section)
when there is no tip-dating
}
\author{
Richèl J.C. Bilderbeek
}
