% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml}
\alias{parameter_to_xml}
\title{Internal function}
\usage{
parameter_to_xml(parameter, beauti_options)
}
\arguments{
\item{parameter}{a parameter,
as created by \code{\link{create_param}})}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the parameter as XML text
}
\description{
Converts a parameter to XML
}
\examples{
check_empty_beautier_folder()

parameter_to_xml(
  create_alpha_param(id = 1),
  beauti_options = create_beauti_options()
)

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
