% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml}
\alias{parameter_to_xml}
\title{Converts a parameter to XML}
\usage{
parameter_to_xml(parameter)
}
\arguments{
\item{parameter}{a parameter,
as created by \code{\link{create_param}})}
}
\value{
the parameter as XML text
}
\description{
Converts a parameter to XML
}
\examples{
parameter_to_xml(create_alpha_param(id = 1))
}
\author{
Richèl J.C. Bilderbeek
}
