% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml_clock_rate}
\alias{parameter_to_xml_clock_rate}
\title{Converts a \code{clockRate} parameter to XML}
\usage{
parameter_to_xml_clock_rate(
  parameter,
  beauti_options = create_beauti_options_v2_4()
)
}
\arguments{
\item{parameter}{a \code{clockRate} parameter,
a numeric value.
For advanced usage, use the structure
as created by \code{\link{create_clock_rate_param}})}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the parameter as XML text
}
\description{
Converts a \code{clockRate} parameter to XML
}
\author{
Richèl J.C. Bilderbeek
}
