% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml_lambda}
\alias{parameter_to_xml_lambda}
\title{Converts a lambda parameter to XML}
\usage{
parameter_to_xml_lambda(parameter)
}
\arguments{
\item{parameter}{a lambda parameter,
as created by \code{\link{create_lambda_param}})}
}
\value{
the parameter as XML text
}
\description{
Converts a lambda parameter to XML
}
\author{
Richel J.C. Bilderbeek
}
