% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml_m}
\alias{parameter_to_xml_m}
\title{Converts a m parameter to XML}
\usage{
parameter_to_xml_m(parameter)
}
\arguments{
\item{parameter}{a m parameter,
as created by \code{\link{create_m_param}})}
}
\value{
the parameter as XML text
}
\description{
Converts a m parameter to XML
}
\author{
Richel J.C. Bilderbeek
}
