% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml_mu}
\alias{parameter_to_xml_mu}
\title{Converts a mu parameter to XML}
\usage{
parameter_to_xml_mu(parameter)
}
\arguments{
\item{parameter}{a mu parameter,
a numeric value.
For advanced usage, use the structure
as created by \code{\link{create_mu_param}})}
}
\value{
the parameter as XML text
}
\description{
Converts a mu parameter to XML
}
\author{
Richèl J.C. Bilderbeek
}
