% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml_scale}
\alias{parameter_to_xml_scale}
\title{Converts a scale parameter to XML}
\usage{
parameter_to_xml_scale(parameter)
}
\arguments{
\item{parameter}{a scale parameter,
a numeric value.
For advanced usage, use the structure
as created by \code{\link{create_scale_param}})}
}
\value{
the parameter as XML text
}
\description{
Converts a scale parameter to XML
}
\author{
Richèl J.C. Bilderbeek
}
