% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_to_xml.R
\name{parameter_to_xml_sigma}
\alias{parameter_to_xml_sigma}
\title{Converts a sigma parameter to XML}
\usage{
parameter_to_xml_sigma(parameter)
}
\arguments{
\item{parameter}{a sigma parameter,
a numeric value.
For advanced usage, use the structure
as created by \code{\link{create_sigma_param}})}
}
\value{
the parameter as XML text
}
\description{
Converts a sigma parameter to XML
}
\author{
Richèl J.C. Bilderbeek
}
