% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_to_xml_state.R
\name{phylo_to_xml_state}
\alias{phylo_to_xml_state}
\title{Creates the XML of a phylogeny,
  as used in the \code{state} section}
\usage{
phylo_to_xml_state(id, tipdates_filename = NA)
}
\arguments{
\item{id}{the ID of the alignment}

\item{tipdates_filename}{name of the file containing the tip dates.
This file is assumed to have two columns, separated by a tab.
The first column contains the taxa names, the second column contains
the date.}
}
\value{
the random phylogeny as XML text
}
\description{
Creates the XML of a phylogeny,
  as used in the \code{state} section
}
\author{
Richèl J.C. Bilderbeek
}
