% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_to_xml_state.R
\name{phylo_to_xml_state}
\alias{phylo_to_xml_state}
\title{Creates the XML of a phylogeny,
  as used in the \code{state} section}
\usage{
phylo_to_xml_state(phylo, id)
}
\arguments{
\item{phylo}{the phylogeny. If NA, a random phylogeny is used}

\item{id}{the ID of the alignment}
}
\value{
the random phylogeny as XML text
}
\description{
Creates the XML of a phylogeny,
  as used in the \code{state} section
}
\author{
Richel J.C. Bilderbeek
}
