% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_parameter_to_xml.R
\name{s_parameter_to_xml}
\alias{s_parameter_to_xml}
\title{Internal function}
\usage{
s_parameter_to_xml(parameter, beauti_options)
}
\arguments{
\item{parameter}{a s parameter,
a numeric value.
For advanced usage, use the structure
as created by \code{\link{create_s_param}})}

\item{beauti_options}{one BEAUti options object,
as returned by \code{\link{create_beauti_options}}}
}
\value{
the parameter as XML text
}
\description{
Converts an `s_param` to XML
}
\examples{
s_parameter_to_xml(
  create_s_param(id = 4, value = 1.25),
  beauti_options = create_beauti_options_v2_4()
)
s_parameter_to_xml(
  create_s_param(id = 4, value = 1.25),
  beauti_options = create_beauti_options_v2_6()
)
}
\author{
Richèl J.C. Bilderbeek
}
