% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_model_to_xml_lh_distr.R
\name{site_model_to_xml_lh_distr}
\alias{site_model_to_xml_lh_distr}
\title{Converts a site model to XML,
  used in the \code{siteModel} section}
\usage{
site_model_to_xml_lh_distr(site_model)
}
\arguments{
\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}
}
\value{
the site model as XML text
}
\description{
Converts a site model to XML,
  used in the \code{siteModel} section
}
\examples{
 # <distribution id="posterior" spec="util.CompoundDistribution">
 #     <distribution id="prior" spec="util.CompoundDistribution">
 #     </distribution>
 #     <distribution id="likelihood" ...>
 #       HERE, where the ID of the distribution is 'likelihood'
 #     </distribution>
 # </distribution>
}
\author{
Richèl J.C. Bilderbeek
}
