% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_model_to_xml_prior.R
\name{site_model_to_xml_prior_distr}
\alias{site_model_to_xml_prior_distr}
\title{Converts a site model to XML,
  used in the \code{prior} section}
\usage{
site_model_to_xml_prior_distr(site_model)
}
\arguments{
\item{site_model}{a site model,
as returned by \code{\link{create_site_model}}}
}
\value{
the site model as XML text
}
\description{
Converts a site model to XML,
  used in the \code{prior} section
}
\author{
Richel J.C. Bilderbeek
}
