% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_models_to_xml_tracelog.R
\name{site_models_to_xml_tracelog}
\alias{site_models_to_xml_tracelog}
\title{Creates the site models' XML for the tracelog section}
\usage{
site_models_to_xml_tracelog(site_models)
}
\arguments{
\item{site_models}{one or more site models,
as returned by \code{\link{create_site_model}}}
}
\value{
lines of XML text
}
\description{
Creates the site models' XML for the tracelog section
}
\examples{
check_empty_beautier_folder()

# <logger id="tracelog" ...>
#'   # Here
# </logger>

check_empty_beautier_folder()
}
\seealso{
the complete tracelog section is created
  by \code{\link{create_tracelog_xml}}
}
\author{
Richèl J.C. Bilderbeek
}
