% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_to_xml_tree.R
\name{tipdate_taxa_to_xml_tree}
\alias{tipdate_taxa_to_xml_tree}
\title{Creates the \code{tree} section
(part of the \code{state} section)
when there is tip-dating}
\usage{
tipdate_taxa_to_xml_tree(id, tipdates_filename)
}
\arguments{
\item{id}{an alignment's IDs.
An ID can be extracted from its FASTA filename
with \code{\link{get_alignment_ids_from_fasta_filenames}})}

\item{tipdates_filename}{name of the file containing the tip dates.
This file is assumed to have two columns, separated by a tab.
The first column contains the taxa names, the second column contains
the date.}
}
\value{
the random phylogeny as XML text
}
\description{
Creates the \code{tree} section
(part of the \code{state} section)
when there is tip-dating
}
\author{
Richèl J.C. Bilderbeek
}
