% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_prior_to_xml_operators.R
\name{tree_prior_to_xml_operators}
\alias{tree_prior_to_xml_operators}
\title{Creates the XML of a tree prior,
  as used in the \code{operators} section}
\usage{
tree_prior_to_xml_operators(tree_prior, fixed_crown_age = FALSE)
}
\arguments{
\item{tree_prior}{a tree prior,
as returned by \code{\link{create_tree_prior}})}
}
\value{
the tree prior as XML text
}
\description{
Creates the XML of a tree prior,
  as used in the \code{operators} section
}
\author{
Richel J.C. Bilderbeek
}
