% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_prior_to_xml_operators.R
\name{tree_prior_to_xml_operators}
\alias{tree_prior_to_xml_operators}
\title{Creates the XML of a tree prior,
  as used in the \code{operators} section}
\usage{
tree_prior_to_xml_operators(tree_prior, fixed_crown_age = FALSE)
}
\arguments{
\item{tree_prior}{a tree priors,
as returned by \code{\link{create_tree_prior}}}

\item{fixed_crown_age}{determines if the phylogeny's crown age is
fixed. If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}
}
\value{
the tree prior as XML text
}
\description{
Creates the XML of a tree prior,
  as used in the \code{operators} section
}
\author{
Richel J.C. Bilderbeek
}
