% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_prior_to_xml_state.R
\name{tree_prior_to_xml_state}
\alias{tree_prior_to_xml_state}
\title{Creates the XML of a tree prior,
  as used in the \code{state} section}
\usage{
tree_prior_to_xml_state(tree_prior)
}
\arguments{
\item{tree_prior}{a tree priors,
as returned by \code{\link{create_tree_prior}}}
}
\value{
the tree prior as XML text
}
\description{
Creates the XML of a tree prior,
  as used in the \code{state} section
}
\author{
Richèl J.C. Bilderbeek
}
