% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_prior_to_xml_tracelog.R
\name{tree_prior_to_xml_tracelog}
\alias{tree_prior_to_xml_tracelog}
\title{Creates the tree prior's XML for the tracelog section}
\usage{
tree_prior_to_xml_tracelog(tree_prior)
}
\arguments{
\item{tree_prior}{a tree priors,
as returned by \code{\link{create_tree_prior}}}
}
\description{
Creates the tree prior's XML for the tracelog section
}
\examples{
# <logger id="tracelog" ...>
#'   # Here
# </logger>
}
\seealso{
all tree priors' tracelog section is created
  by \code{\link{tree_priors_to_xml_tracelog}}
}
\author{
Richel J.C. Bilderbeek
}
