% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_priors_to_xml_tracelog.R
\name{tree_priors_to_xml_tracelog}
\alias{tree_priors_to_xml_tracelog}
\title{Creates the tree priors' XML for the tracelog section}
\usage{
tree_priors_to_xml_tracelog(tree_priors)
}
\arguments{
\item{tree_priors}{one or more tree priors,
as returned by \code{\link{create_tree_prior}}}
}
\value{
lines of XML text
}
\description{
Creates the tree priors' XML for the tracelog section
}
\examples{
# <logger id="tracelog" ...>
#'   # Here
# </logger>
}
\seealso{
the complete tracelog section is created
  by \code{\link{create_tracelog_xml}}
}
\author{
Richèl J.C. Bilderbeek
}
