% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_input_distr.R
\name{yule_tree_prior_to_xml_prior}
\alias{yule_tree_prior_to_xml_prior}
\title{Creates the \code{prior} section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file for a Yule tree prior}
\usage{
yule_tree_prior_to_xml_prior(yule_tree_prior)
}
\arguments{
\item{yule_tree_prior}{a Yule tree_prior,
as created by \code{\link{create_yule_tree_prior}}}
}
\description{
Creates the \code{prior} section in the prior section of
the prior section of the distribution section
of a BEAST2 XML parameter file for a Yule tree prior
}
\author{
Richel J.C. Bilderbeek
}
